### 2023-10-15 #5952 @editorialbot
[PRE REVIEW]: SODA: Software to Support the Curation and Sharing of FAIR Autonomic Nervous System Data
**Submitting author:** @bvhpatel (<a href="http://orcid.org/0000-0002-0307-262X">Bhavesh Patel</a>)
**Repository:** https://github.com/fairdataihub/SODA-for-SPARC
**Branch with paper.md** (empty if default branch): paper
**Version:** v12.3.1
