+++ scripts/others/renderer.js
@@ -3239,7 +3239,6 @@ const postCurationListChange = () => {
 
 // upload banner image //
 const Cropper = require("cropperjs");
-const { default: Swal } = require("sweetalert2");
 const { waitForDebugger } = require("inspector");
 const { resolve } = require("path");
 const { background } = require("jimp");
@@ -4904,127 +4903,6 @@ const deleteProblematicFilesFromDatasetStructure = (datasetStructure) => {
   const currentFoldersAtPath = Object.keys(datasetStructure.folders);
   for (const folderKey of currentFoldersAtPath) {
     deleteProblematicFilesFromDatasetStructure(datasetStructure["folders"][folderKey]);
-  }
-};
-
-const swalShowError = async (title, errorText) => {
-  await Swal.fire({
-    title: title,
-    html: errorText,
-    icon: "error",
-    width: 600,
-    heightAuto: false,
-    backdrop: "rgba(0,0,0, 0.4)",
-    showConfirmButton: true,
-    confirmButtonText: "OK",
-  });
-};
-
-const swalFileListSingleAction = async (fileList, title, helpText, postActionText) => {
-  await Swal.fire({
-    title: title,
-    html: `
-      ${helpText ? `<p>${helpText}</p>` : ""}
-      <div class="swal-file-list">
-        ${fileList
-          .map(
-            (file) => `<div class="swal-file-row"><span class="swal-file-text">${file}</span></div>`
-          )
-          .join("")}
-      </div>
-      ${postActionText ? `<b>${postActionText}</b>` : ""}
-    `,
-    width: 800,
-    heightAuto: false,
-    width: 800,
-    heightAuto: false,
-    backdrop: "rgba(0,0,0, 0.4)",
-    allowOutsideClick: false,
-    allowEscapeKey: false,
-    showCancelButton: false,
-    showCloseButton: false,
-  });
-};
-
-const swalFileListDoubleAction = async (
-  fileList,
-  title,
-  helpText,
-  confirmButtonText,
-  cancelButtonText,
-  confirmationText
-) => {
-  const { value: action } = await Swal.fire({
-    title: title,
-    html: `
-      ${helpText}
-      <div class="swal-file-list">
-        ${fileList
-          .map(
-            (file) => `<div class="swal-file-row"><span class="swal-file-text">${file}</span></div>`
-          )
-          .join("")}
-      </div>
-      <b>${confirmationText}</b>
-    `,
-    width: 800,
-    heightAuto: false,
-    backdrop: "rgba(0,0,0, 0.4)",
-    allowOutsideClick: false,
-    allowEscapeKey: false,
-    showCancelButton: true,
-    showCloseButton: false,
-    confirmButtonText: confirmButtonText,
-    cancelButtonText: cancelButtonText,
-  });
-  return action;
-};
-
-const swalFileListTripleAction = async (
-  fileList,
-  title,
-  helpText,
-  confirmButtonText,
-  denyButtonText,
-  cancelButtonText,
-  confirmationText
-) => {
-  const { value: action } = await Swal.fire({
-    title: title,
-    html: `
-      ${helpText}
-      <div class="swal-file-list">
-        ${fileList
-          .map(
-            (file) => `<div class="swal-file-row"><span class="swal-file-text">${file}</span></div>`
-          )
-          .join("")}
-      </div>
-      <b>${confirmationText}</b>
-    `,
-    width: 800,
-    heightAuto: false,
-    backdrop: "rgba(0,0,0, 0.4)",
-    allowOutsideClick: false,
-    allowEscapeKey: false,
-    showCloseButton: false,
-    showCancelButton: true,
-    showDenyButton: true,
-    confirmButtonText: confirmButtonText,
[…]
