+++ .github/workflows/Build-and-deploy-linux.yml
@@ -4,7 +4,9 @@ run-name: ${{ github.actor }} is building a linux release for SODA - should work
 on:
   push:
     branches:
-      - path-agent
+      - main
+      - staging
+      - get-datasets-pagination
 
 jobs:
   deploy-on-linux:
