### 2023-10-16 9ae4fae
feat: ✨ detect, start, stop agent on Ventura Macs
