### 2023-10-16 181f316
fix: 🐛 Handle case where agent is already running during agent start
