+++ .github/workflows/Build-and-deploy-mac.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - main
       - staging
-      - get-datasets-pagination
 jobs:
   deploy-on-mac:
     runs-on: macos-latest
