+++ .github/workflows/Build-and-deploy-win.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - main
       - staging
-      - get-datasets-pagination
 
 jobs:
   deploy-on-windows:
