### 2023-10-16 48339f0
build: 📦️ spawn testable build
