### 2023-10-16 f7b7864
fix: 🐛 spawn agent after installation
