### 2023-10-16 6f15694
fix: 🐛 revert PATH Pennsieve startup logic
