### 2023-10-17 6ed62a9
build: 📦️ test windows shell inheritance
