+++ pyflask/curate/curate.py
@@ -1314,13 +1314,15 @@ def create_high_lvl_manifest_files_existing_ps(
             ):
 
                 relative_path = ""
-                dict_folder_manifest = {}
+
                 # Initialize dict where manifest info will be stored
-                dict_folder_manifest["filename"] = []
-                dict_folder_manifest["timestamp"] = []
-                dict_folder_manifest["description"] = []
-                dict_folder_manifest["file type"] = []
-                dict_folder_manifest["Additional Metadata"] = []
+                dict_folder_manifest = {
+                    "filename": [],
+                    "timestamp": [],
+                    "description": [],
+                    "file type": [],
+                    "Additional Metadata": [],
+                }
 
                 # pull manifest file into if exists 
                 manifest_df = pd.DataFrame()
@@ -1363,6 +1365,10 @@ def create_high_lvl_manifest_files_existing_ps(
         existing_file_option = soda_json_structure["generate-dataset"][
             "if-existing-files"
         ]
+
+        namespace_logger.info(f"Existing folder option: {existing_folder_option}")
+        namespace_logger.info(f"Existing file option: {existing_file_option}")
+
         for folder_key, folder in dataset_structure["folders"].items():
             relative_path = ""
 
@@ -1817,8 +1823,8 @@ def ps_update_existing_dataset(soda_json_structure, ds, ps):
 
     soda_json_structure["generate-dataset"] = {
         "destination": "bf",
-        "if-existing": "merge",
-        "if-existing-files": "replace",
+        "if-existing": soda_json_structure["generate-dataset"]["if-existing"] or "merge",
+        "if-existing-files": soda_json_structure["generate-dataset"]["if-existing-files"] or "replace",
         "generate-option": "existing-bf"
     }
 
@@ -2037,6 +2043,7 @@ def ps_upload_to_dataset(soda_json_structure, ps, ds):
                 my_tracking_folder: Tracks what folders have been created on Pennsieve thus far. Starts as an empty dictionary.
                 existing_folder_option: Dictates whether to merge, duplicate, replace, or skip existing folders.
             """
+            namespace_logger.info("recursive_create_folder_for_ps existing folder option: " + existing_folder_option)
 
             # Check if the current folder has any subfolders that already exist on Pennsieve. Important step to appropriately handle replacing and merging folders.
             if len(my_tracking_folder["children"]["folders"]) == 0 and my_tracking_folder["content"]["id"].find("N:dataset") == -1:
@@ -2109,6 +2116,7 @@ def ps_upload_to_dataset(soda_json_structure, ps, ds):
             """
                 Delete files that are marked to be replaced in the dataset. Create a list of files to upload to Pennsieve.
             """
+            namespace_logger.info("recursive_dataset_scan_for_ps existing file option: " + existing_file_option)
 
             global main_total_generate_dataset_size
 
