+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated urllib3 and request versions to fix issues with the back-end processing responses from Pennsieve API.
 - Modified the request to retrieve a list of datasets by utilizing Pennsieve's dataset pagination endpoint. This change was implemented to address an issue that was affecting users with a large number of datasets.
 - Migrated the notarization tool from alttool to notarytool.
+- SODA for SPARC launches the Pennsieve Agent for OS X Ventura.
+- Duplicate files were being created when uploading a dataset through `Organize Datasets` and selecting the 'skip' existing files option. The 'skip' option now works as intended. Local files that exist on a Pennsieve dataset are no longer duplicated.
 
 ## v.12.3.1 - 2023-09-21
 
