### 2023-10-18 4ec6961
fix: 🐛 remove test logs
