### 2023-10-20 8dfe55d
feat: ✨  find agent with path instead of install location
