+++ scripts/others/renderer.js
@@ -587,21 +587,13 @@ const getPennsieveAgentPath = () => {
   }
 };
 
-const stopPennsieveAgent = async (pathToPennsieveAgent) => {
+const stopPennsieveAgent = async () => {
   return new Promise((resolve, reject) => {
     try {
-      let agentStopSpawn;
-      if (
-        process.platform === "win32" ||
-        process.platform === "cygwin" ||
-        process.platform === "linux"
-      ) {
-        agentStopSpawn = spawn("pennsieve", ["agent", "stop"], {
-          shell: true,
-        });
-      } else {
-        agentStopSpawn = spawn(pathToPennsieveAgent, ["agent", "stop"]);
-      }
+      let agentStopSpawn = spawn("pennsieve", ["agent", "stop"], {
+        shell: true,
+      });
+
       agentStopSpawn.stdout.on("data", (data) => {
         log.info(data.toString());
         resolve();
@@ -616,7 +608,7 @@ const stopPennsieveAgent = async (pathToPennsieveAgent) => {
     }
   });
 };
-const startPennsieveAgent = async (pathToPennsieveAgent) => {
+const startPennsieveAgent = async () => {
   return new Promise((resolve, reject) => {
     // Keep track of the output from the agent
     // (output is added as strings to the array)
@@ -632,20 +624,11 @@ const startPennsieveAgent = async (pathToPennsieveAgent) => {
         )
       );
     }, agentStartTimeout);
-    // Start the agent by running the command "agent start" at the path of the agent
-    let agentStartSpawn;
-    // check if linux
-    if (
-      process.platform === "win32" ||
-      process.platform === "cygwin" ||
-      process.platform === "linux"
-    ) {
-      agentStartSpawn = spawn("pennsieve", ["agent", "start"], {
-        shell: true,
-      });
-    } else {
-      agentStartSpawn = spawn(pathToPennsieveAgent, ["agent", "start"]);
-    }
+
+    let agentStartSpawn = spawn("pennsieve", ["agent", "start"], {
+      shell: true,
+    });
+
     // Listen to the output from the agent and resolve the promise if the agent outputs
     // "Running Agent NOT as daemon" or "Pennsieve Agent started"
     agentStartSpawn.stdout.on("data", (data) => {
@@ -669,7 +652,7 @@ const startPennsieveAgent = async (pathToPennsieveAgent) => {
         agentMessage.includes("Pennsieve Agent started")
       ) {
         setTimeout(() => {
-          const secondAgentStartSpawn = spawn(pathToPennsieveAgent, ["agent", "start"]);
+          const secondAgentStartSpawn = spawn("pennsieve", ["agent", "start"]);
           secondAgentStartSpawn.stdout.on("data", (data) => {
             const secondAgentMessage = `[Pennsieve Agent Output] ${data.toString()}`;
             if (secondAgentMessage.includes("Pennsieve Agent is already running")) {
@@ -696,7 +679,7 @@ const startPennsieveAgent = async (pathToPennsieveAgent) => {
   });
 };
 
-const getPennsieveAgentVersion = (pathToPennsieveAgent) => {
+const getPennsieveAgentVersion = () => {
   log.info("Getting Pennsieve agent version");
 
   return new Promise((resolve, reject) => {
@@ -715,18 +698,7 @@ const getPennsieveAgentVersion = (pathToPennsieveAgent) => {
       );
     }, timeout);
 
-    let agentVersionSpawn;
-
-    // check if linux
-    if (
-      process.platform === "win32" ||
-      process.platform === "cygwin" ||
-      process.platform === "linux"
-    ) {
-      agentVersionSpawn = execFile("pennsieve", ["version"]);
-    } else {
[…]
