+++ main-process/restart/life-cycle.js
@@ -1,0 +1,6 @@
+const { ipcMain, app } = require("electron");
+
+ipcMain.handle("restart-app", (event) => {
+  app.relaunch();
+  app.quit();
+});
