+++ .github/workflows/Build-and-deploy-mac.yml
@@ -4,8 +4,7 @@ run-name: ${{ github.actor }} is building a MAC release for SODA
 on:
   push:
     branches:
-      - main
-      - staging
+      - final-agent-test
 jobs:
   deploy-on-mac:
     runs-on: macos-latest
