### 2023-10-20 d84df54
feat: ✨ check for agent install miss on Windows
