### 2023-10-23 80afaf7
feat: ✨ handle linux ps not found message
