### 2023-10-23 8e99b11
fix: 🐛 show all elements of the path
