### 2023-10-23 fc32e5c
fix: 🐛 add extra path bases to MAC OS built GUIs in runtime
