### 2023-10-23 b2abb93
fix: 🐛 pass updated env to the spawn
