### 2023-10-23 fca3261
refactor: ♻️ fix path script moved into separate module
