### 2023-10-23 c0c8eb3
feat: ✨ stop and get agent version use path
