### 2023-10-23 93c9932
fix: 🐛 restarting soda programmatically dores not update env variables
