### 2023-10-23 1b1d9b1
feat: ✨ second agent restart attempt uses PATH
