### 2023-10-23 168b87f
refactor: ♻️ agent installed check and fail UI flow moved to function
