### 2023-10-24 27d1c1d
feat: ✨ check for agent after user adds account
