### 2023-10-24 8d7af14
chore: 🔨 remove unused function
