### 2023-10-24 6f8bd31
build: 📦️ update deploy scripts
