+++ .github/workflows/Build-and-deploy-linux.yml
@@ -4,7 +4,7 @@ run-name: ${{ github.actor }} is building a linux release for SODA - should work
 on:
   push:
     branches:
-      - final-agent-test
+      - staging
 
 jobs:
   deploy-on-linux:
