+++ .github/workflows/Build-and-deploy-mac.yml
@@ -4,7 +4,7 @@ run-name: ${{ github.actor }} is building a MAC release for SODA
 on:
   push:
     branches:
-      - final-agent-test
+      - staging
 jobs:
   deploy-on-mac:
     runs-on: macos-latest
