### 2023-10-25 520b073
fix: 🐛 close notyf after failed pre_flight_checks
