### 2023-10-25 6c86dae
feat: ✨ hide subjects table when no subjects
