### 2023-10-26 81b347c
feat: ✨ Added templates for all pages
