### 2023-10-26 6850cbe
feat: ✨ Added function to get subject names from sodaJSONObj
