### 2023-10-26 50f63cb
feat: ✨ Added subject focusing on swal open
