### 2023-10-27 4c51362
feat: bundled version of application
