+++ src/pyflask/startup/minimumApiVersion.py
@@ -1,0 +1,7 @@
+import os
+
+def get_api_version():
+    """
+    Returns the version of the API
+    """
+    return {'version': os.getenv('API_VERSION', "12.3.0-beta")}
