+++ src/pyflask/disseminate/disseminate.py
@@ -1,0 +1,245 @@
+# -*- coding: utf-8 -*-
+
+### Import required python modules
+from venv import create
+from flask import abort 
+import requests
+from permissions import pennsieve_get_current_user_permissions, has_edit_permissions
+from utils import connect_pennsieve_client, get_dataset_id, authenticate_user_with_client, create_request_headers
+from errorHandlers import handle_http_error
+from authentication import get_access_token
+
+
+
+PENNSIEVE_URL = "https://api.pennsieve.io"
+
+
+def bf_get_doi(selected_bfaccount, selected_bfdataset):
+    """
+    Function to get current doi for a selected dataset
+
+    Args:
+        selected_bfaccount: name of selected Pennsieve acccount (string)
+        selected_bfdataset: name of selected Pennsieve dataset (string)
+    Return:
+        Current doi or "None"
+    """
+    token = get_access_token()
+
+    selected_dataset_id = get_dataset_id(token, selected_bfdataset)
+
+    if not has_edit_permissions(token, selected_dataset_id):
+        abort(403, "You do not have permission to edit this dataset.")
+
+    try:
+        r = requests.get(f"{PENNSIEVE_URL}/datasets/{str(selected_dataset_id)}/doi", headers=create_request_headers(token))
+        r.raise_for_status()
+        result = r.json()
+
+        return {"doi": result["doi"]}
+    except Exception as e:
+        if "404" in str(e):
+            return {"doi": "None"}
+        handle_http_error(e)
+
+
+def bf_reserve_doi(selected_bfaccount, selected_bfdataset):
+    """
+    Function to reserve doi for a selected dataset
+
+    Args:
+        selected_bfaccount: name of selected Pennsieve acccount (string)
+        selected_bfdataset: name of selected Pennsieve dataset (string)
+    Return:
+        Success or error message
+    """
+
+    token = get_access_token()
+
+    selected_dataset_id = get_dataset_id(token, selected_bfdataset)
+
+    if not has_edit_permissions(token, selected_dataset_id):
+        abort(403, "You do not have permission to edit this dataset.")
+
+    try:
+        res = bf_get_doi(selected_bfaccount, selected_bfdataset)
+        if res["doi"] != "None":
+            abort(400, "A DOI has already been reserved for this dataset")
+    except Exception as e:
+        raise e
+
+    try:
+        r = requests.get(f"{PENNSIEVE_URL}/datasets/{str(selected_dataset_id)}/contributors", headers=create_request_headers(token))
+        r.raise_for_status()
+        contributors = r.json()
+        creators_list = [
+            item["firstName"] + " " + item["lastName"]
+            for item in contributors
+        ]
+    except Exception as e:
+        handle_http_error(e)
+    
+    try:
+        jsonfile = {
+            "title": selected_bfdataset,
+            "creators": creators_list,
+        }
+        
+        r = requests.post(f"{PENNSIEVE_URL}/datasets/{str(selected_dataset_id)}/doi", headers=create_request_headers(token), json=jsonfile)
+        r.raise_for_status()
+
+        return {"message": "Done!"}
+    except Exception as e:
+        handle_http_error(e)
+
+
+
+def bf_get_publishing_status(selected_bfaccount, selected_bfdataset):
+    """
+    Function to get the review request status and publishing status of a dataset
+
+    Args:
+        selected_bfaccount: name of selected Pennsieve account (string)
+        selected_bfdataset: name of selected Pennsieve dataset (string)
+    Return:
+        Current req publishing status
+    """
+
+    token = get_access_token()
+
+    selected_dataset_id = get_dataset_id(token, selected_bfdataset)
+
[…]
