### 2023-10-27 4c32717
feat: fix windows server launch
