+++ out/main/index.js
@@ -1,158 +1,0 @@
-"use strict";
-const electron = require("electron");
-const path = require("path");
-const utils = require("@electron-toolkit/utils");
-const ElectronLog = require("electron-log");
-require("axios");
-const fp = require("find-free-port");
-const node_child_process = require("node:child_process");
-const fs = require("fs");
-const icon = path.join(__dirname, "../../resources/icon.png");
-electron.ipcMain.handle("get-app-path", async (event, arg) => {
-  return electron.app.getPath(arg);
-});
-function createWindow() {
-  const mainWindow = new electron.BrowserWindow({
-    width: 900,
-    height: 670,
-    show: false,
-    nodeIntegration: true,
-    autoHideMenuBar: true,
-    ...process.platform === "linux" ? { icon } : {},
-    webPreferences: {
-      preload: path.join(__dirname, "../preload/index.js"),
-      sandbox: false,
-      contextIsolation: true
-    }
-  });
-  mainWindow.on("ready-to-show", () => {
-    mainWindow.show();
-  });
-  mainWindow.webContents.setWindowOpenHandler((details) => {
-    electron.shell.openExternal(details.url);
-    return { action: "deny" };
-  });
-  if (utils.is.dev && process.env["ELECTRON_RENDERER_URL"]) {
-    mainWindow.loadURL(process.env["ELECTRON_RENDERER_URL"]);
-  } else {
-    mainWindow.loadFile(path.join(__dirname, "../renderer/index.html"));
-  }
-}
-electron.app.whenReady().then(async () => {
-  utils.electronApp.setAppUserModelId("com.electron");
-  electron.app.on("browser-window-created", (_, window) => {
-    utils.optimizer.watchWindowShortcuts(window);
-  });
-  await createPyProc();
-  createWindow();
-  electron.app.on("activate", function() {
-    if (electron.BrowserWindow.getAllWindows().length === 0)
-      createWindow();
-  });
-});
-electron.app.on("window-all-closed", () => {
-  if (process.platform !== "darwin") {
-    electron.app.quit();
-  }
-});
-const PY_FLASK_DIST_FOLDER = "pyflaskdist";
-const PY_FLASK_FOLDER = "../src/pyflask";
-const PY_FLASK_MODULE = "app";
-let PORT = 4242;
-const portRange = 100;
-let pyflaskProcess = null;
-const guessPackaged = () => {
-  const windowsPath = path.join(process.resourcesPath, PY_FLASK_DIST_FOLDER);
-  ElectronLog.info("Windows path: " + windowsPath);
-  const unixPath = path.join(process.resourcesPath, PY_FLASK_MODULE);
-  if (process.platform === "darwin" || process.platform === "linux") {
-    if (fs.existsSync(unixPath)) {
-      return true;
-    } else {
-      return false;
-    }
-  }
-  if (process.platform === "win32") {
-    if (fs.existsSync(windowsPath)) {
-      ElectronLog.info("App is packaged returning true [ Windows ]");
-      return true;
-    } else {
-      return false;
-    }
-  }
-};
-const getScriptPath = () => {
-  if (!guessPackaged()) {
-    ElectronLog.info("App is not packaged returning path: ");
-    ElectronLog.info(path.join(__dirname, "..", PY_FLASK_FOLDER, PY_FLASK_MODULE + ".py"));
-    return path.join(__dirname, "..", PY_FLASK_FOLDER, PY_FLASK_MODULE + ".py");
-  }
-  if (process.platform === "win32") {
-    const winPath = path.join(process.resourcesPath, PY_FLASK_DIST_FOLDER, PY_FLASK_MODULE + ".exe");
[…]
