+++ .github/workflows/Build-and-deploy-mac.yml
@@ -4,7 +4,7 @@ run-name: ${{ github.actor }} is building a MAC release for SODA
 on:
   push:
     branches:
-      - main
+      - bundled-soda
 jobs:
   deploy-on-mac:
     runs-on: macos-latest
@@ -23,18 +23,18 @@ jobs:
       - uses: conda-incubator/setup-miniconda@v2
         with:
           miniconda-version: "latest"
-          activate-environment: env-electron-python
+          activate-environment: bundle-up-too
           environment-file: tools/anaconda-env/environment-MAC.yml
           auto-update-conda: true
           auto-activate-base: false
       - run: |
-          conda activate env-electron-python
+          conda activate bundle-up-too
           conda info
           conda list
 
       - uses: actions/setup-node@v3
         with:
-          node-version: "18.14.0"
+          node-version: "18.16.0"
 
       - name: Install package.json modules and their dependencies
         run: npm install --ignore-scripts
@@ -43,10 +43,6 @@ jobs:
         with:
           p12-file-base64: ${{ secrets.MACOS_CERTIFICATE }}
           p12-password: ${{ secrets.MACOS_CERTIFICATE_PWD }}
-
-      - name: Build the python server executable
-        run: |
-          npm run python-onefile-build-unix
 
       - name: Deploy the app to GitHub
         env:
