### 2023-10-27 8b53a7f
feat: ✨ Added ability to delete subjects
