+++ src/renderer/src/scripts/others/renderer.js
@@ -6,24 +6,18 @@ while (!window.htmlPagesAdded) {
   await new Promise((resolve) => setTimeout(resolve, 100))
 }
 
-// import fs from "fs-extra";
 import * as os from "os";
-console.log(os)
-
-// console.log(fs)
-// const os = require("os");
-// const path = require("path");
-// const { ipcRenderer, BrowserWindow } = require("electron");
-// const Editor = require("@toast-ui/editor");
+import * as  path from "path";
+import Editor from "@toast-ui/editor"
 // const remote = require("@electron/remote");
-// const { Notyf } = require("notyf");
-// const imageDataURI = require("image-data-uri");
-// const log = require("electron-log");
-// require("v8-compile-cache");
-// const Tagify = require("@yaireo/tagify");
+import { Notyf } from "notyf"
+import Tagify from "@yaireo/tagify";
 // const https = require("https");
 // const electron = require("electron");
-// const bootbox = require("bootbox");
+import jQuery from 'jquery'
+import bootstrap from 'bootstrap'
+import * as popper from '@popperjs/core'
+import * as bootbox from "bootbox";
 // const DragSelect = require("dragselect");
 // const excelToJson = require("convert-excel-to-json");
 // const csvToJson = require("convert-csv-to-json");
