### 2023-10-31 6b64bb4
wip: install all dependencies in renderer
