+++ src/renderer/src/scripts/others/api/api.js
@@ -1,0 +1,501 @@
+/*
+Purpose: An abstraction layer between the client and making HTTP requests via Axios. This layer handles the error parsing and logging.
+*/
+
+const getUserInformation = async () => {
+  let userResponse = await client.get(`/user`, {
+    params: {
+      pennsieve_account: defaultBfAccount,
+    },
+  });
+
+  return userResponse.data;
+};
+
+/**
+ *
+ * @param {string} datasetNameOrID - the current dataset name or id
+ * @returns {datasetObject} dataset - the dataset object
+ */
+const getDataset = async (datasetNameOrID) => {
+  let datasetResponse = await client.get(`/datasets/${datasetNameOrID}`);
+  return datasetResponse.data;
+};
+
+const getDatasetBannerImageURL = async (selected_account, selected_dataset) => {
+  let bannerResponse = await client.get(`/manage_datasets/bf_banner_image`, {
+    params: {
+      selected_account,
+      selected_dataset,
+    },
+  });
+
+  let { banner_image } = bannerResponse.data;
+
+  return banner_image;
+};
+
+const isDatasetLocked = async (account, datasetNameOrId) => {
+  try {
+    // get the logged in user's information which will be used to check if the user is a member of the "Publishers" team
+    const currentUserInformation = await getUserInformation();
+    const currentUserID = currentUserInformation.id;
+    const teamsReq = await client.get(
+      `manage_datasets/ps_get_teams?selected_account=${defaultBfAccount}`
+    );
+    const teamsInCurrentUsersOrganization = teamsReq.data.teams;
+
+    // Get the team with the name "Publishers" (if it exists)
+    const publishersTeam = teamsInCurrentUsersOrganization.find(
+      (teamElement) => teamElement.team.name === "Publishers"
+    );
+
+    // If a "Publishers" team exists, get the IDs of the team's administrators
+    if (publishersTeam) {
+      const publishersTeamIDs = publishersTeam.administrators.map(
+        (administrator) => administrator.id
+      );
+      // Check too see if the current user is a member of the "Publishers" team
+      if (publishersTeamIDs.includes(currentUserID)) {
+        // If the user is a member of the "Publishers" team, return false since the dataset should not be locked for them
+        return false;
+      }
+    }
+
+    // If the user is not a member of the "Publishers" team, check to see if the dataset is locked
+    const datasetRoleResponse = await client.get(`/datasets/${datasetNameOrId}`);
+    // Return the dataset's lock status (true or false)
+    return datasetRoleResponse.data.locked;
+  } catch (err) {
+    clientError(err);
+    // If the dataset is locked, the server will return a 423 status code, so return true if that is the case
+    if (err.response.status == 423) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+};
+
+const getDatasetRole = async (datasetNameOrId) => {
+  if (datasetNameOrId != undefined || datasetNameOrId != "") {
+    defaultBfDataset = datasetNameOrId;
+  }
+
+  let datasetRoleResponse = await client.get(`/datasets/${defaultBfDataset}/role`);
+
+  let { role } = datasetRoleResponse.data;
+
+  return role;
+};
+
+const getDatasetInformation = async (account, datasetNameOrId) => {
+  const datasetInformationResponse = await client.get(`/datasets/${datasetNameOrId}`, {
+    params: {
+      pennsieve_account: account,
+    },
+  });
+  // Returns information about the dataset (locked, published, etc.)
+  return datasetInformationResponse.data;
+};
+
+/**
+ * Withdraw any dataset from a pre-publishing review submission
+ * @param {string} datasetIdOrName
[…]
