### 2023-11-01 af5a038
refactor: fix path uses context bridge
