### 2023-11-01 8d2b546
refactor: renderer lotties and logger refactpred
