### 2023-11-01 a1b7eaa
refactor: announcements in renderer via ipc
