### 2023-11-01 #5952 @AJQuinn
Hi @bvhpatel - apologies for the delay getting started on this. Will start finding reviewers over the next few days.

In the meantime, could you double check the DOIs in your paper draft. You have valid bibtex on your branch https://github.com/fairdataihub/SODA-for-SPARC/blob/paper/paper/paper.bib but it looks like the entries don't have a doi key or a url. For example - from a quick glance, the Wilkinson et al 2016 ref should include the following lines

```
  doi = {10.1038/sdata.2016.18},
  url = {https://doi.org/10.1038/sdata.2016.18},
```

Could you review the bibtex and let me know when you have an update?
