### 2023-11-02 cf9d58f
wip: startup code refactor
