+++ src/renderer/src/scripts/others/renderer.js
@@ -32,6 +32,7 @@ import select2 from "select2"; // TODO: select2()
 select2()
 import DragSort from "@yaireo/dragsort";
 import axios from "axios";
+import Swal from "sweetalert2";
 import DatePicker from "tui-date-picker"; /* CommonJS */
 import  datasetUploadSession from "../analytics/upload-session-tracker";
 import kombuchaEnums from"../analytics/analytics-enums"
@@ -134,288 +135,293 @@ if (autoUpdateLaunch == false || autoUpdateLaunch == null || autoUpdateLaunch ==
 //   timeout: 300000,
 // });
 
-// const notyf = new Notyf({
-//   position: { x: "right", y: "bottom" },
-//   dismissible: true,
-//   ripple: false,
-//   types: [
-//     {
-//       type: "checking_server_is_live",
-//       background: "grey",
-//       icon: {
-//         className: "fas fa-wifi",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 1000,
-//     },
-//     {
-//       type: "checking_server_api_version",
-//       background: "grey",
-//       icon: {
-//         className: "fas fa-wifi",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 1000,
-//     },
-//     {
-//       type: "loading_internet",
-//       background: "grey",
-//       icon: {
-//         className: "fas fa-wifi",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 10000,
-//     },
-//     {
-//       type: "ps_agent",
-//       background: "grey",
-//       icon: {
-//         className: "fas fa-cogs",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 5000,
-//     },
-//     {
-//       type: "app_update",
-//       background: "grey",
-//       icon: {
-//         className: "fas fa-sync-alt",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 0,
-//     },
-//     {
-//       type: "api_key_search",
-//       background: "grey",
-//       icon: {
-//         className: "fas fa-users-cog",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 0,
-//     },
-//     {
-//       type: "success",
-//       background: "#13716D",
-//       icon: {
-//         className: "fas fa-check-circle",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 800,
-//     },
-//     {
-//       type: "final",
-//       background: "#13716D",
-//       icon: {
-//         className: "fas fa-check-circle",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 3000,
-//     },
-//     {
-//       type: "warning",
-//       background: "#fa8c16",
-//       icon: {
-//         className: "fas fa-exclamation-triangle",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 3000,
-//     },
-//     {
-//       type: "info",
-//       background: "#13716D",
-//       icon: {
-//         className: "fas fa-info-circle",
-//         tagName: "i",
-//         color: "white",
-//       },
-//       duration: 3000,
-//     },
+const notyf = new Notyf({
+  position: { x: "right", y: "bottom" },
+  dismissible: true,
+  ripple: false,
+  types: [
+    {
+      type: "checking_server_is_live",
+      background: "grey",
+      icon: {
+        className: "fas fa-wifi",
+        tagName: "i",
+        color: "white",
+      },
+      duration: 1000,
+    },
+    {
+      type: "checking_server_api_version",
+      background: "grey",
+      icon: {
+        className: "fas fa-wifi",
+        tagName: "i",
+        color: "white",
+      },
[…]
