### 2023-11-02 db1d201
wip: 🚧 Added example folder image
