### 2023-11-02 8e75a04
feat: ✨ Added function to add a list of subjects for gm
