### 2023-11-02 b9c36c2
refactor: api version and server live check in bundle
