+++ src/pyflask/manageDatasets/manage_datasets.py
@@ -246,7 +246,9 @@ def bf_default_account_load():
     try:
         accountlist = []
         if exists(configpath):
+            namespace_logger.info("Config file exists")
             valid_account = bf_get_accounts()
+            namespace_logger.info("Valid account: " + valid_account)
             if valid_account != "":
                 accountlist.append(valid_account)
         return {"defaultAccounts": accountlist}
@@ -272,15 +274,17 @@ def bf_get_accounts():
             get_access_token()
             return SODA_SPARC_API_KEY.lower()
     elif "global" in sections:
+        namespace_logger.info("Global section exists")
         if "default_profile" in config["global"]:
             default_profile = config["global"]["default_profile"]
+            namespace_logger.info("Default profile: " + default_profile)
             if default_profile in sections:
                 lowercase_account_names(config, default_profile, configpath)
                 try:
                     get_access_token()
                     return default_profile.lower()
                 except Exception as e:
-                    print(e)
+                    namespace_logger.error(e)
     else:
         for account in sections:
             if account != 'agent':

[…]
