+++ src/renderer/src/main.js
@@ -34,6 +34,7 @@ import './assets/css/spreadSheetTools.css'
 import './assets/css/spur.css'
 import './assets/css/tablepath.css'
 import './assets/css/variables.css'
+import 'notyf/notyf.min.css'; // for React, Vue and Svelte
 
 
 
