### 2023-11-03 0195167
wip: run pre flight checks
