+++ src/main/main.js
@@ -1,0 +1,572 @@
+const { app, BrowserWindow, dialog, shell } = require("electron");
+require("@electron/remote/main").initialize();
+app.showExitPrompt = true;
+const path = require("path");
+const glob = require("glob");
+const fp = require("find-free-port");
+const os = require("os");
+const contextMenu = require("electron-context-menu");
+const log = require("electron-log");
+require("v8-compile-cache");
+const { ipcMain } = require("electron");
+const { autoUpdater } = require("electron-updater");
+const { JSONStorage } = require("node-localstorage");
+const { trackEvent, trackKombuchaEvent } = require("./scripts/others/analytics/analytics");
+const { fstat } = require("fs");
+const { resolve } = require("path");
+const axios = require("axios");
+const { info } = require("console");
+const { node } = require("prop-types");
+const { kombuchaEnums } = require("./scripts/others/analytics/analytics-enums");
+const uuid = require("uuid").v4;
+
+const sodaVersion = app.getVersion();
+// If the version includes "beta", the app will not check for updates
+const buildIsBeta = sodaVersion.includes("beta");
+if (buildIsBeta) {
+  log.info("This is a beta build. Updates will not be checked.");
+}
+autoUpdater.channel = buildIsBeta ? "beta" : "latest";
+
+log.transports.console.level = false;
+log.transports.file.level = "debug";
+autoUpdater.logger = log;
+global.trackEvent = trackEvent;
+global.trackKombuchaEvent = trackKombuchaEvent;
+const nodeStorage = new JSONStorage(app.getPath("userData"));
+// nodeStorage.setItem("announcements", true);
+
+/*************************************************************
+ * Python Process
+ *************************************************************/
+// flask setup environment variables
+const PY_FLASK_DIST_FOLDER = "pyflaskdist";
+const PY_FLASK_FOLDER = "pyflask";
+const PY_FLASK_MODULE = "app";
+let pyflaskProcess = null;
+let PORT = 4242;
+let selectedPort = null;
+const portRange = 100;
+const kombuchaURL = "https://analytics-nine-ashen.vercel.app/api";
+const localKombuchaURL = "http://localhost:3000/api";
+const kombuchaServer = axios.create({
+  baseURL: kombuchaURL,
+  timeout: 0,
+});
+/**
+ * Determine if the application is running from a packaged version or from a dev version.
+ * The resources path is used for Linux and Mac builds and the app.getAppPath() is used for Windows builds.
+ * @returns {boolean} True if the app is packaged, false if it is running from a dev version.
+ */
+const guessPackaged = () => {
+  log.info("Guessing if packaged");
+  const windowsPath = path.join(__dirname, PY_FLASK_DIST_FOLDER);
+  const unixPath = path.join(process.resourcesPath, PY_FLASK_MODULE);
+  log.info(unixPath);
+  if (process.platform === "darwin" || process.platform === "linux") {
+    if (require("fs").existsSync(unixPath)) {
+      log.info("Unix path exists");
+      return true;
+    } else {
+      log.info("Unix path does not exist");
+      return false;
+    }
+  }
+  if (process.platform === "win32") {
+    if (require("fs").existsSync(windowsPath)) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+};
+/**
+ * Get the system path to the api server script.
+ * The script is located in the resources folder for packaged Linux and Mac builds and in the app.getAppPath() for Windows builds.
+ * It is relative to the main.js file directory when in dev mode.
+ * @returns {string} The path to the api server script that needs to be executed to start the Python server
+ */
+const getScriptPath = () => {
[…]
