+++ src/renderer/src/scripts/others/renderer.js
@@ -382,7 +382,6 @@ const startupServerAndApiCheck = async () => {
   try {
     await apiVersionsMatch();
   } catch (e) {
-    console.log("SODA will exit")
     // api versions do not match
     await window.electron.ipcRenderer.invoke("exit-soda")
   }
@@ -420,221 +419,228 @@ const startupServerAndApiCheck = async () => {
 
 startupServerAndApiCheck();
 
-// // Check if we are connected to the Pysoda server
-// // Check app version on current app and display in the side bar
-// // Also check the core systems to make sure they are all operational
-// ipcRenderer.on("start_pre_flight_checks", async (event, arg) => {
-//   // run pre flight checks once the server connection is confirmed
-//   // wait until soda is connected to the backend server
-//   while (!sodaIsConnected || !apiVersionChecked) {
-//     await wait(1000);
-//   }
+// Check if we are connected to the Pysoda server
+// Check app version on current app and display in the side bar
+// Also check the core systems to make sure they are all operational
+window.electron.ipcRenderer.on("start_pre_flight_checks", async (event, arg) => {
+  console.log("Starting pre flight checks************************")
 
-//   log.info("Done with startup");
+  // run pre flight checks once the server connection is confirmed
+  // wait until soda is connected to the backend server
+  while (!sodaIsConnected || !apiVersionChecked) {
+    await wait(1000);
+  }
 
-//   // check integrity of all the core systems
-//   await run_pre_flight_checks();
+  log.info("Done with startup");
 
-//   log.info("Running pre flight checks finished");
-// });
+  // check integrity of all the core systems
+  await run_pre_flight_checks();
 
-// const stopPennsieveAgent = async () => {
-//   return new Promise((resolve, reject) => {
-//     try {
-//       let agentStopSpawn = spawn("pennsieve", ["agent", "stop"], {
-//         shell: true,
-//         env: process.env,
-//       });
+  log.info("Running pre flight checks finished");
+});
 
-//       agentStopSpawn.stdout.on("data", (data) => {
-//         log.info(data.toString());
-//         resolve();
-//       });
-//       agentStopSpawn.stderr.on("data", (data) => {
-//         log.info(data.toString());
-//         reject(new Error(data.toString()));
-//       });
-//     } catch (error) {
-//       log.info(error);
-//       reject(error);
-//     }
-//   });
-// };
-// const startPennsieveAgent = async () => {
-//   return new Promise((resolve, reject) => {
-//     // Keep track of the output from the agent
-//     // (output is added as strings to the array)
-//     const pennsieveAgentOutputLog = [];
-//     // Throw an error if the agent doesn't start within 15 seconds
-//     const agentStartTimeout = 15000; // 15 seconds
-//     const versionCheckTimeout = setTimeout(() => {
-//       reject(
-//         new Error(
-//           `Pennsieve Agent output while trying to start the agent:<br />${pennsieveAgentOutputLog.join(
-//             "<br />"
-//           )}`
-//         )
-//       );
-//     }, agentStartTimeout);
+const stopPennsieveAgent = async () => {
+  return new Promise((resolve, reject) => {
+    try {
+      let agentStopSpawn = spawn("pennsieve", ["agent", "stop"], {
+        shell: true,
+        env: process.env,
+      });
 
-//     let agentStartSpawn = spawn("pennsieve", ["agent", "start"], {
-//       shell: true,
-//       env: process.env,
-//     });
+      agentStopSpawn.stdout.on("data", (data) => {
+        log.info(data.toString());
+        resolve();
+      });
+      agentStopSpawn.stderr.on("data", (data) => {
+        log.info(data.toString());
[…]
