### 2023-11-03 7b6d2b9
refactor: renderer startup checks do not rely on main process signal
