+++ src/renderer/src/scripts/others/renderer.js
@@ -34,8 +34,8 @@ import DragSort from "@yaireo/dragsort";
 import axios from "axios";
 import Swal from "sweetalert2";
 import DatePicker from "tui-date-picker"; /* CommonJS */
-import  datasetUploadSession from "../analytics/upload-session-tracker";
-import kombuchaEnums from"../analytics/analytics-enums"
+import datasetUploadSession from "../analytics/upload-session-tracker";
+import kombuchaEnums from "../analytics/analytics-enums"
 
 import {
   createEventData,
@@ -47,10 +47,15 @@ import {
   clientError,
   userErrorMessage,
 } from "./http-error-handler/error-handler"
-import  hasConnectedAccountWithPennsieve from "./authentication/auth";
+import hasConnectedAccountWithPennsieve from "./authentication/auth";
 import fixPath from "./update-path-darwin"
 import api from "./api/api"
-import {confirm_click_account_function, showHideDropdownButtons, updateDatasetList} from '../globals'
+import {
+  confirm_click_account_function, showHideDropdownButtons,
+  updateDatasetList, openDropdownPrompt, bfAccountOptionsStatus,
+  bfAccountOptions, defaultBfAccount, defaultBfDataset, defaultBfDatasetId,
+  reverseSwalButtons
+} from '../globals'
 import checkForAnnouncements from './announcements'
 
 fixPath();
@@ -63,7 +68,6 @@ fixPath();
 // let nodeStorage = new JSONStorage(app.getPath("userData")); NOTE: Done in main now. Need to expose to main world via contextBridge.
 
 var nextBtnDisabledVariable = true;
-var reverseSwalButtons = false;
 let organizeDSglobalPath = "";
 
 let datasetStructureJSONObj = {
@@ -85,7 +89,7 @@ let introStatus = {
 
 // // Log file settings //
 window.log.setupRendererLogOptions()
-const homeDirectory = await window.electron.ipcRenderer.invoke('get-app-path', 'home') 
+const homeDirectory = await window.electron.ipcRenderer.invoke('get-app-path', 'home')
 const SODA_SPARC_API_KEY = "SODA-Pennsieve";
 
 // // get port number from the main process
@@ -276,8 +280,8 @@ const wait = async (delay) => {
 };
 
 let client = axios.create({
-    baseURL: `http://127.0.0.1:${port}`,
-    timeout: 300000,
+  baseURL: `http://127.0.0.1:${port}`,
+  timeout: 300000,
 });
 
 // check that the client connected to the server using exponential backoff
@@ -386,11 +390,11 @@ const startupServerAndApiCheck = async () => {
   if (launchAnnouncement) {
     await checkForAnnouncements("announcements");
     launchAnnouncement = false;
-    window.electron.ipcRenderer.invoke("set-nodestorage-key", "announcements", false) 
+    window.electron.ipcRenderer.invoke("set-nodestorage-key", "announcements", false)
   }
 
   // get apps base path
-  const basepath = await window.electron.ipcRenderer.invoke("get-app-path", "home") 
+  const basepath = await window.electron.ipcRenderer.invoke("get-app-path", "home")
   const resourcesPath = window.process.resourcesPath();
 
   // set the templates path
@@ -430,26 +434,14 @@ const initializeSODARenderer = async () => {
 initializeSODARenderer()
 
 const stopPennsieveAgent = async () => {
-  return new Promise((resolve, reject) => {
-    try {
-      let agentStopSpawn = spawn("pennsieve", ["agent", "stop"], {
-        shell: true,
-        env: process.env,
-      });
+  try {
+    let agentStopSpawn = await window.spawn.stopPennsieveAgent()
+    console.log("Agent stop spawn: ", agentStopSpawn)
[…]
