+++ src/renderer/src/scripts/globals.js
@@ -1,3 +1,5 @@
+import Swal from 'sweetalert2'
+import api from './others/api/api'
 // // Purpose: Will become preload.js in the future. For now it is a place to put global variables/functions that are defined in javascript files
 // //          needed by the renderer process in order to run.
 
@@ -146,13 +148,17 @@ const updateDatasetList = (bfaccount) => {
   }, 100)
 }
 
-export {
-  currentConTable,
-  showHideDropdownButtons,
-  confirm_click_account_function,
-  initializeBootstrapSelect,
-  updateDatasetList
-}
+
+
+// global variables to be modularized
+let bfAccountOptionsStatus;
+let bfAccountOptions;
+let defaultBfAccount;
+let defaultBfDataset = "Select dataset";
+let defaultBfDatasetId = undefined;
+var reverseSwalButtons = false;
+
+
 
 // const updateOrganizationList = async (bfaccount) => {
 // let organizations = []
@@ -854,1185 +860,1202 @@ export {
 //   $('#para-review-dataset-info-disseminate').text('None')
 // }
 
-// const addBfAccount = async (ev, verifyingOrganization = False) => {
-//   var resolveMessage = ''
-//   if (bfAccountOptionsStatus === '') {
-//     if (Object.keys(bfAccountOptions).length === 1) {
-//       footerMessage = 'No existing accounts to load. Please add an account.'
-//     } else {
-//       footerMessage = ''
-//     }
-//   } else {
-//     footerMessage = bfAccountOptionsStatus
-//   }
-//   var bfacct
-//   let bfAccountSwal = false
-//   if (bfAccountSwal === null) {
-//     if (bfacct !== 'Select') {
-//       Swal.fire({
-//         allowEscapeKey: false,
-//         backdrop: 'rgba(0,0,0, 0.4)',
-//         heightAuto: false,
-//         showConfirmButton: false,
-//         timer: 2000,
-//         timerProgressBar: true,
-//         title: 'Loading your account details...',
-//         didOpen: () => {
-//           Swal.showLoading()
-//         }
-//       })
-//       $('#Question-getting-started-BF-account')
-//         .nextAll()
-//         .removeClass('show')
-//         .removeClass('prev')
-//         .removeClass('test2')
-//       $('#Question-generate-dataset-BF-account')
-//         .nextAll()
-//         .removeClass('show')
-//         .removeClass('prev')
-//         .removeClass('test2')
-//       $('#current-bf-account').text('')
-//       $('#current-bf-account-generate').text('')
-//       $('#create_empty_dataset_BF_account_span').text('')
-//       $('.bf-account-span').text('')
-//       $('#current-bf-dataset').text('None')
-//       $('#current-bf-dataset-generate').text('None')
-//       $('.bf-dataset-span').html('None')
-//       defaultBfDataset = 'Select dataset'
-//       document.getElementById('ds-description').innerHTML = ''
-//       refreshDatasetList()
-//       $($('#button-confirm-bf-dataset-getting-started').parents()[0]).css('display', 'none')
-//       $('#button-confirm-bf-dataset-getting-started').hide()
-
-//       $('#para-account-detail-curate').html('')
-//       $('#current-bf-dataset').text('None')
-//       $('.bf-dataset-span').html('None')
-//       showHideDropdownButtons('dataset', 'hide')
-
-//       try {
-//         let bf_account_details_req = await client.get(`/manage_datasets/bf_account_details`, {
-//           params: {
-//             selected_account: bfacct
-//           }
-//         })
-//         let user_email = bf_account_details_req.data.email
-//         $('#current-bf-account').text(user_email)
-//         $('#current-bf-account-generate').text(user_email)
[…]
