+++ src/renderer/src/main.js
@@ -1,6 +1,7 @@
 import './assets/imports'
 import './assets/demo-btns'
 import './assets/nav'
+import './scripts/client'
 import './scripts/guided-mode/guided-curate-dataset'
 import './scripts/globals'
 import './scripts/others/renderer'
