+++ scripts/organize-dataset/curate-functions.js
@@ -788,25 +788,7 @@ $(document).ready(async function () {
   $("#guided_bf_list_users_and_teams").selectpicker("refresh");
 });
 
-const get_api_key = (login, password, key_name) => {
-  return new Promise(async (resolve) => {
-    try {
-      let bf_get_pennsieve_secret_key = await client.post(
-        `/manage_datasets/pennsieve_api_key_secret`,
-        {
-          username: login,
-          password: password,
-          api_key: key_name,
-        }
-      );
-      let res = bf_get_pennsieve_secret_key.data;
-      resolve(res);
-    } catch (error) {
-      clientError(error);
-      resolve(["failed", userErrorMessage(error)]);
-    }
-  });
-};
+
 
 $("#select-permission-list-2").change((e) => {
   // updateDatasetList(defaultBfAccount);
