+++ src/renderer/src/scripts/globals.js
@@ -1,7 +1,13 @@
 import Swal from 'sweetalert2'
 import api from './others/api/api'
+import {clientError, userErrorMessage} from './others/http-error-handler/error-handler'
+import client from './client'
 // // Purpose: Will become preload.js in the future. For now it is a place to put global variables/functions that are defined in javascript files
 // //          needed by the renderer process in order to run.
+
+
+
+
 
 // // Contributors table for the dataset description editing page
 const currentConTable = document.getElementById('table-current-contributors')
@@ -151,12 +157,18 @@ const updateDatasetList = (bfaccount) => {
 
 
 // global variables to be modularized
-let bfAccountOptionsStatus;
-let bfAccountOptions;
-let defaultBfAccount;
-let defaultBfDataset = "Select dataset";
-let defaultBfDatasetId = undefined;
-var reverseSwalButtons = false;
+window.bfAccountOptionsStatus = ''
+window.myitem;
+
+let bfAccountOptions = []
+window.defaultBfAccount;
+window.defaultBfDataset = "Select dataset";
+window.defaultBfDatasetId = undefined;
+window.reverseSwalButtons = false;
+window.SODA_SPARC_API_KEY = "SODA-Pennsieve";
+let datasetList = [];
+let organizationList = [];
+
 
 
 
@@ -861,132 +873,132 @@ var reverseSwalButtons = false;
 // }
 
 const addBfAccount = async (ev, verifyingOrganization = False) => {
-  var resolveMessage = ''
-  let footerMessage = ''
-  if (bfAccountOptionsStatus === '') {
+  var resolveMessage = "";
+  let footerMessage = "No existing accounts to load. Please add an account.";
+  if (bfAccountOptionsStatus === "") {
     if (Object.keys(bfAccountOptions).length === 1) {
-      footerMessage = 'No existing accounts to load. Please add an account.'
+      footerMessage = "No existing accounts to load. Please add an account.";
     } else {
-      footerMessage = ''
+      footerMessage = "";
     }
   } else {
-    footerMessage = bfAccountOptionsStatus
+    footerMessage = bfAccountOptionsStatus;
   }
-  var bfacct
-  let bfAccountSwal = false
+  var bfacct;
+  let bfAccountSwal = false;
   if (bfAccountSwal === null) {
-    if (bfacct !== 'Select') {
+    if (bfacct !== "Select") {
       Swal.fire({
         allowEscapeKey: false,
-        backdrop: 'rgba(0,0,0, 0.4)',
+        backdrop: "rgba(0,0,0, 0.4)",
         heightAuto: false,
         showConfirmButton: false,
         timer: 2000,
         timerProgressBar: true,
-        title: 'Loading your account details...',
+        title: "Loading your account details...",
         didOpen: () => {
-          Swal.showLoading()
-        }
-      })
-      $('#Question-getting-started-BF-account')
+          Swal.showLoading();
+        },
+      });
+      $("#Question-getting-started-BF-account")
         .nextAll()
-        .removeClass('show')
-        .removeClass('prev')
-        .removeClass('test2')
-      $('#Question-generate-dataset-BF-account')
+        .removeClass("show")
+        .removeClass("prev")
+        .removeClass("test2");
+      $("#Question-generate-dataset-BF-account")
         .nextAll()
-        .removeClass('show')
-        .removeClass('prev')
-        .removeClass('test2')
-      $('#current-bf-account').text('')
-      $('#current-bf-account-generate').text('')
-      $('#create_empty_dataset_BF_account_span').text('')
-      $('.bf-account-span').text('')
-      $('#current-bf-dataset').text('None')
-      $('#current-bf-dataset-generate').text('None')
[…]
