### 2023-11-03 7e5119a
refactor: make immutable via export vars windowed
