+++ scripts/others/tab-effects.js
@@ -553,7 +553,7 @@ const nextPrev = (pageIndex) => {
       confirmButtonText: "Continue",
       heightAuto: false,
       backdrop: "rgba(0,0,0, 0.4)",
-      reverseButtons: reverseSwalButtons,
+      reverseButtons: window.reverseSwalButtons,
       showClass: {
         popup: "animate__animated animate__zoomIn animate__faster",
       },
@@ -628,7 +628,7 @@ const nextPrev = (pageIndex) => {
         confirmButtonText: "Continue",
         showCancelButton: "No",
         focusCancel: true,
-        reverseButtons: reverseSwalButtons,
+        reverseButtons: window.reverseSwalButtons,
         heightAuto: false,
         customClass: "swal-wide",
         backdrop: "rgba(0,0,0, 0.4)",
@@ -911,7 +911,7 @@ const raiseWarningGettingStarted = (ev) => {
         showCancelButton: "Cancel",
         focusCancel: true,
         confirmButtonText: "Continue",
-        reverseButtons: reverseSwalButtons,
+        reverseButtons: window.reverseSwalButtons,
         heightAuto: false,
         backdrop: "rgba(0,0,0, 0.4)",
         showClass: {
@@ -1776,7 +1776,7 @@ const transitionSubQuestionsButton = async (ev, currentDiv, parentDiv, button, c
     };
 
     // Set the default Pennsieve account and dataset
-    sodaJSONObj["bf-account-selected"]["account-name"] = defaultBfAccount;
+    sodaJSONObj["bf-account-selected"]["account-name"] = window.defaultBfDataset;
     sodaJSONObj["bf-dataset-selected"]["dataset-name"] = selectedDataset;
 
     $("#para-continue-bf-dataset-getting-started").text("");
@@ -1787,7 +1787,7 @@ const transitionSubQuestionsButton = async (ev, currentDiv, parentDiv, button, c
     $("#bf-dataset-spinner").css("visibility", "visible");
 
     // Check if dataset is locked before trying to import
-    const isDatasetLocked = await api.isDatasetLocked(defaultBfAccount, selectedDataset);
+    const isDatasetLocked = await api.isDatasetLocked(window.defaultBfDataset, selectedDataset);
     if (isDatasetLocked) {
       Swal.fire({
         icon: "info",
@@ -1892,7 +1892,7 @@ const transitionSubQuestionsButton = async (ev, currentDiv, parentDiv, button, c
         confirmButtonText: "Continue",
         heightAuto: false,
         backdrop: "rgba(0,0,0, 0.4)",
-        reverseButtons: reverseSwalButtons,
+        reverseButtons: window.reverseSwalButtons,
         showClass: {
           popup: "animate__animated animate__zoomIn animate__faster",
         },
@@ -2425,7 +2425,7 @@ const switchMetadataRCImportBFQuestions = async (metadataRCFileType) => {
       backdrop: "rgba(0,0,0, 0.4)",
       confirmButtonText: "Yes",
       cancelButtonText: "No",
-      reverseButtons: reverseSwalButtons,
+      reverseButtons: window.reverseSwalButtons,
     });
     if (!continueProgress) {
       return;
@@ -2447,7 +2447,7 @@ async function switchMetadataRCQuestion(metadataRCFileType) {
       backdrop: "rgba(0,0,0, 0.4)",
       confirmButtonText: "Yes",
       cancelButtonText: "No",
-      reverseButtons: reverseSwalButtons,
+      reverseButtons: window.reverseSwalButtons,
     });
     if (continueProgress) {
       $(`#existing-${metadataRCFileType}-file-destination`).attr("placeholder", "Browse here");
[…]
