+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={1--9},
   year={2016},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1038/sdata.2016.18}
 }
 
 @article{osanlouy2021sparc,
@@ -16,7 +17,8 @@
   volume={12},
   pages={693735},
   year={2021},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi = {10.3389/fphys.2021.693735}
 }
 
 @article{patel2023making,
@@ -27,7 +29,8 @@
   number={1},
   pages={557},
   year={2023},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/s41597-023-02463-x}
 }
 
 @article{quey2021knowmore,
@@ -36,7 +39,8 @@
   journal={bioRxiv},
   pages={2021--08},
   year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi = {10.12688/f1000research.73492.1}
 }
 
 @article{bandrowski2021sparc,
@@ -45,7 +49,8 @@
   journal={bioRxiv},
   pages={2021--02},
   year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi = {10.1101/2021.02.10.430563}
 }
 
 @article{patel2020sparc,
@@ -56,7 +61,8 @@
   number={S1},
   pages={1--1},
   year={2020},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1096/fasebj.2020.34.s1.02483}
 }
 
 @article{soundarajan2022sparclink,
@@ -65,7 +71,8 @@
   journal={F1000Research},
   volume={11},
   year={2022},
-  publisher={Faculty of 1000 Ltd}
+  publisher={Faculty of 1000 Ltd},
+  doi = {10.12688/f1000research.75071.1)}
 }
 
 @misc{Patel_SODA_Software_to_2023,
