+++ src/renderer/src/scripts/others/renderer.js
@@ -435,164 +435,47 @@ const stopPennsieveAgent = async () => {
     let agentStopSpawn = await window.spawn.stopPennsieveAgent()
     console.log("Agent stop spawn: ", agentStopSpawn)
   } catch (error) {
-    log.info(error);
+    window.log.info(error);
     throw error;
   }
 
 };
 const startPennsieveAgent = async () => {
-  return new Promise((resolve, reject) => {
-    // Keep track of the output from the agent
-    // (output is added as strings to the array)
-    const pennsieveAgentOutputLog = [];
-    // Throw an error if the agent doesn't start within 15 seconds
-    const agentStartTimeout = 15000; // 15 seconds
-    const versionCheckTimeout = setTimeout(() => {
-      reject(
-        new Error(
-          `Pennsieve Agent output while trying to start the agent:<br />${pennsieveAgentOutputLog.join(
-            "<br />"
-          )}`
-        )
-      );
-    }, agentStartTimeout);
-
-    let agentStartSpawn = spawn("pennsieve", ["agent", "start"], {
-      shell: true,
-      env: process.env,
-    });
-
-    // Listen to the output from the agent and resolve the promise if the agent outputs
-    // "Running Agent NOT as daemon" or "Pennsieve Agent started"
-    agentStartSpawn.stdout.on("data", (data) => {
-      const agentMessage = `[Pennsieve Agent Output] ${data.toString()}`;
-      log.info(agentMessage);
-      // Add to message to the output log which will be used to display the output to the user if the agent fails to start
-      pennsieveAgentOutputLog.push(agentMessage);
-
-      // Resolve the promise if the agent is already running
-      if (agentMessage.includes("Pennsieve Agent is already running")) {
-        log.info(`Pennsieve Agent is confirmed to be running: ${agentMessage}`);
-        clearTimeout(versionCheckTimeout);
-        resolve();
-      }
-
-      // If the error message contains "Running Agent NOT as daemon" or "Pennsieve Agent started", then the agent was able to
-      // start successfully with the caveat that the agent might throw an error if the agent has issues while starting up.
-      // To alleviate this, we will try to start the agent again after 5 seconds and make sure the agent was started successfully.
-      if (
-        agentMessage.includes("Running Agent NOT as daemon") ||
-        agentMessage.includes("Pennsieve Agent started")
-      ) {
-        setTimeout(() => {
-          const secondAgentStartSpawn = spawn("pennsieve", ["agent", "start"], {
-            shell: true,
-            env: process.env,
-          });
-          secondAgentStartSpawn.stdout.on("data", (data) => {
-            const secondAgentMessage = `[Pennsieve Agent Output] ${data.toString()}`;
-            if (secondAgentMessage.includes("Pennsieve Agent is already running")) {
-              log.info(`Pennsieve Agent is confirmed to be running: ${secondAgentMessage}`);
-              clearTimeout(versionCheckTimeout);
-              resolve();
-            }
-          });
-        }, 5000);
-      }
-    });
-    // Capture standard error output and add it to the output log
-    agentStartSpawn.stderr.on("data", (data) => {
-      const agentStdErr = `[Pennsieve Agent Error] ${data.toString()}`;
-      log.info(agentStdErr);
-      pennsieveAgentOutputLog.push(agentStdErr);
-    });
-    // Capture error output and add it to the output log
-    agentStartSpawn.on("error", (error) => {
-      const agentSpawnError = `[Pennsieve Agent Error] ${error.toString()}`;
-      log.info(agentSpawnError);
-      pennsieveAgentOutputLog.push(agentSpawnError);
-    });
[…]
