### 2023-11-05 582ba20
feat: load account when one is present
