+++ src/renderer/src/scripts/globals.js
@@ -89,8 +89,8 @@ const updateDatasetList = (bfaccount) => {
 
   $('#div-filter-datasets-progress-2').css('display', 'none')
 
-  removeOptions(curateDatasetDropdown)
-  addOption(curateDatasetDropdown, 'Search here...', 'Select dataset')
+  window.removeOptions(window.curateDatasetDropdown)
+  window.addOption(window.curateDatasetDropdown, 'Search here...', 'Select dataset')
 
   initializeBootstrapSelect('#curatebfdatasetlist', 'disabled')
 
@@ -103,20 +103,20 @@ const updateDatasetList = (bfaccount) => {
 
   // waiting for dataset list to load first before initiating BF dataset dropdown list
   setTimeout(() => {
-    var myPermission = $(datasetPermissionDiv).find('#select-permission-list-2').val()
+    var myPermission = $(window.datasetPermissionDiv).find('#select-permission-list-2').val()
 
     if (!myPermission) {
       myPermission = 'All'
     }
 
     if (myPermission.toLowerCase() === 'all') {
-      for (var i = 0; i < datasetList.length; i++) {
-        filteredDatasets.push(datasetList[i].name)
+      for (var i = 0; i < window.datasetList.length; i++) {
+        filteredDatasets.push(window.datasetList[i].name)
       }
     } else {
-      for (var i = 0; i < datasetList.length; i++) {
-        if (datasetList[i].role === myPermission.toLowerCase()) {
-          filteredDatasets.push(datasetList[i].name)
+      for (var i = 0; i < window.datasetList.length; i++) {
+        if (window.datasetList[i].role === myPermission.toLowerCase()) {
+          filteredDatasets.push(window.datasetList[i].name)
         }
       }
     }
@@ -125,7 +125,7 @@ const updateDatasetList = (bfaccount) => {
       return a.toLowerCase().localeCompare(b.toLowerCase())
     })
 
-    // The removeoptions() wasn't working in some instances (creating a double dataset list) so second removal for everything but the first element.
+    // The window.removeOptions() wasn't working in some instances (creating a double dataset list) so second removal for everything but the first element.
     $('#curatebfdatasetlist').find('option:not(:first)').remove()
 
     for (myitem in filteredDatasets) {
@@ -133,7 +133,7 @@ const updateDatasetList = (bfaccount) => {
       var option = document.createElement('option')
       option.textContent = myitemselect
       option.value = myitemselect
-      curateDatasetDropdown.appendChild(option)
+      window.curateDatasetDropdown.appendChild(option)
     }
 
     initializeBootstrapSelect('#curatebfdatasetlist', 'show')
@@ -163,75 +163,76 @@ window.bfAccountOptionsStatus = ''
 window.myitem;
 
 let bfAccountOptions = []
-window.defaultBfDataset;
+window.defaultBfAccount = undefined;
 window.defaultBfDataset = "Select dataset";
 window.defaultBfDatasetId = undefined;
 window.reverseSwalButtons = false;
 window.SODA_SPARC_API_KEY = "SODA-Pennsieve";
-let datasetList = [];
+window.datasetList = [];
 let organizationList = [];
 
 
 
 
-// const updateOrganizationList = async (bfaccount) => {
-// let organizations = []
 
-// $('#div-filter-datasets-progress-2').css('display', 'none')
+window.updateOrganizationList = async (bfaccount) => {
+let organizations = []
 
-// removeOptions(curateOrganizationDropdown)
-// addOption(curateOrganizationDropdown, 'Search here...', 'Select organization')
+$('#div-filter-datasets-progress-2').css('display', 'none')
 
-// initializeBootstrapSelect('#curatebforganizationlist', 'disabled')
[…]
