+++ src/renderer/src/scripts/others/renderer.js
@@ -273,7 +273,7 @@ let update_available_notification = "";
 let update_downloaded_notification = "";
 
 // utility function for async style set timeout
-const wait = async (delay) => {
+window.wait = async (delay) => {
   return new Promise((resolve) => setTimeout(resolve, delay));
 };
 
@@ -283,7 +283,7 @@ const wait = async (delay) => {
 // verify the api versions match
 const startupServerAndApiCheck = async () => {
   // wait for SWAL to be loaded in
-  await wait(2000);
+  await window.wait(2000);
 
   // notify the user that the application is starting connecting to the server
   Swal.fire({
@@ -324,7 +324,7 @@ const startupServerAndApiCheck = async () => {
     if (time_pass > 300000) {
       break;
     } //break after five minutes
-    await wait(2000);
+    await window.wait(2000);
   }
 
   if (!status) {
@@ -789,7 +789,7 @@ const run_pre_flight_checks = async (check_update = true) => {
 
     // make an api request to change to the organization members. If it fails with a 401 then ask them to go through the workspace change flow as SODA does not have access to the workspace.
     try {
-      await client.get(`/manage_datasets/ps_get_users?selected_account=${window.defaultBfDataset}`);
+      await client.get(`/manage_datasets/ps_get_users?selected_account=${window.defaultBfAccount}`);
     } catch (err) {
       clientError(err);
       if (err.response.status) {
@@ -990,7 +990,7 @@ const check_api_key = async () => {
     type: "api_key_search",
     message: "Checking for Pennsieve account...",
   });
-  await wait(800);
+  await window.wait(800);
   // If no accounts are found, return false.
   let responseObject;
 
@@ -1263,7 +1263,7 @@ window.electron.ipcRenderer.on("app_version", (event, arg) => {
 // const pathSubmitDataset = document.querySelector("#selected-local-dataset-submit");
 // const progressUploadBf = document.getElementById("div-progress-submit");
 // const progressBarUploadBf = document.getElementById("progress-bar-upload-bf");
-// const datasetPermissionDiv = document.getElementById("div-permission-list-2");
+window.datasetPermissionDiv = document.getElementById("div-permission-list-2");
 // const bfDatasetSubtitle = document.querySelector("#bf-dataset-subtitle");
 // const bfDatasetSubtitleCharCount = document.querySelector("#para-char-count-metadata");
 
@@ -2494,7 +2494,7 @@ const bfAddPermissionTeamBtn = document.getElementById("button-add-permission-te
 // };
 
 // Function to add options to dropdown list
-function addOption(selectbox, text, value) {
+window.addOption = (selectbox, text, value)  => {
   var opt = document.createElement("OPTION");
   opt.text = text;
   opt.value = value;
@@ -2507,19 +2507,19 @@ function addOption(selectbox, text, value) {
 // //// get datasets and append that to option list for parent datasets
 // function getParentDatasets() {
 //   var parentDatasets = [];
-//   for (var i = 0; i < datasetList.length; i++) {
-//     parentDatasets.push(datasetList[i].name);
+//   for (var i = 0; i < window.datasetList.length; i++) {
+//     parentDatasets.push(window.datasetList[i].name);
 //   }
 //   return parentDatasets;
 // }
 
 // function changeAwardInputDsDescription() {
 //   if (dsContributorArrayLast1) {
-//     removeOptions(dsContributorArrayLast1);
+//     window.removeOptions(dsContributorArrayLast1);
 //   }
 //   if (dsContributorArrayFirst1) {
-//     removeOptions(dsContributorArrayFirst1);
-//     addOption(dsContributorArrayFirst1, "Select an option", "Select an option");
[…]
