### 2023-11-06 fairdataihub/SODA-for-SPARC#273 @AJQuinn
DOI Tweaks for JOSS submission
Hello JOSS editor here - thanks for fixing up the DOIs, there is one final tweak needed on this line:

https://github.com/fairdataihub/SODA-for-SPARC/blob/f51ab22d0174b0e0c3b65b0af00864be78b6b3b1/paper/paper.bib#L75C40-L75C41

Need to drop the close parenthesis at the end of the DOI.
```
  doi = {10.12688/f1000research.75071.1)}
```

Our paper compiling tests should pass once that is through

https://github.com/openjournals/joss-reviews/issues/5952
