### 2023-11-06 d1886b2
feat: GM landing page navigable
