+++ src/renderer/src/scripts/organize-dataset/validation-functions.js
@@ -1,0 +1,405 @@
+// Purpose: Logic for Organize Dataset Step 7: Validate Dataset
+const { v4: uuid } = require("uuid");
+
+// Validate the dataset that has just been organized in Organize Dataset Step 6: Validate Dataset
+// TODO: Pennsieve vs local considerations for result parsing and error handling
+const validateOrganizedDataset = async () => {
+  let validationErrorsTable = document.querySelector("#organize--table-validation-errors tbody");
+  let datasetOrigin = "";
+  let datasetDestination = "";
+
+  if ("bf" in sodaJSONObj["starting-point"]["type"]) {
+    datasetOrigin = "Pennsieve";
+  }
+
+  if ("local" in sodaJSONObj["starting-point"]["type"]) {
+    datasetOrigin = "Local";
+  }
+
+  if ("new" in sodaJSONObj["starting-point"]["type"]) {
+    datasetOrigin = "New";
+  }
+
+  if ("save-progress" in sodaJSONObj) {
+    datasetOrigin = "Saved";
+  }
+
+  if (validationErrorsTable.childElementCount > 0) {
+    // ask the user to confirm they want to reset their validation progress
+    let resetValidationResult = await Swal.fire({
+      icon: "warning",
+      text: "This will reset your current validation results. Do you wish to continue?",
+      heightAuto: false,
+      showCancelButton: true,
+      cancelButtonText: "No",
+      focusCancel: true,
+      confirmButtonText: "Yes",
+      backdrop: "rgba(0,0,0, 0.4)",
+      reverseButtons: window.reverseSwalButtons,
+      showClass: {
+        popup: "animate__animated animate__zoomIn animate__faster",
+      },
+      hideClass: {
+        popup: "animate__animated animate__zoomOut animate__faster",
+      },
+    });
+
+    // user does not want to reset
+    if (!resetValidationResult.isConfirmed) {
+      return;
+    }
+
+    // get validation table body
+    clearValidationResults(validationErrorsTable);
+  }
+
+  Swal.fire({
+    title: "Validating Dataset",
+    text: "Please wait while your dataset is validated.",
+    allowEscapeKey: false,
+    allowOutsideClick: false,
+    heightAuto: false,
+    backdrop: "rgba(0,0,0, 0.4)",
+    didOpen: () => {
+      Swal.showLoading();
+    },
+  });
+
+  let sodaJSONObjCopy = JSON.parse(JSON.stringify(sodaJSONObj));
+  formatForDatasetGeneration(sodaJSONObjCopy);
+
+  // if the user performed move, rename, delete on files in an imported dataset we need to perform those actions before creating the validation report;
+  // rationale for this can be found in the function definition
+  if (sodaJSONObjCopy["starting-point"]["type"] === "bf") {
+    await api.performUserActions(sodaJSONObjCopy);
+  }
+
+  // check size before doing this
+  // situations:
+  // 1. Local dataset -> count sodaJSONObj
+  // 2. Pennsieve dataset -> count sodaJSONObj
+  // 3. Saved dataset -> count sodaJSONObj
+  // 4. Merge -> count sodaJSONObj + packageTypeCounts to get the total number of files
+  // NOTE: I do have to consider deleted files since they will not be included, but can ignore moved/renamed files since they will be included in the count
+  // for now lets count then handle option 4 later
+
+  // get the number of files and folders in the dataset that have been added in the virutal organizer
+  file_counter = 0;
+  get_num_files_and_folders(sodaJSONObjCopy["dataset-structure"]);
+
+  // check if the virutal files will be merged with a Pennsieve dataset
+  if (
+    $('input[name="generate-4"]:checked')[0] &&
+    $('input[name="generate-4"]:checked')[0].id === "generate-BF-dataset-options-existing"
+  ) {
[…]
