+++ src/renderer/src/scripts/others/renderer.js
@@ -69,7 +69,6 @@ fixPath();
 // let nodeStorage = new JSONStorage(app.getPath("userData")); NOTE: Done in main now. Need to expose to main world via contextBridge.
 
 var nextBtnDisabledVariable = true;
-let organizeDSglobalPath = "";
 
 let datasetStructureJSONObj = {
   folders: {},
@@ -4092,14 +4091,14 @@ const loadDefaultAccount = async () => {
 
 // /// back button Curate
 // organizeDSbackButton.addEventListener("click", function () {
-//   let slashCount = organizeDSglobalPath.value.trim().split("/").length - 1;
+//   let slashCount = window.organizeDSglobalPath.value.trim().split("/").length - 1;
 //   if (slashCount !== 1) {
-//     let filtered = getGlobalPath(organizeDSglobalPath);
+//     let filtered = getGlobalPath(window.organizeDSglobalPath);
 
 //     if (filtered.length === 1) {
-//       organizeDSglobalPath.value = filtered[0] + "/";
+//       window.organizeDSglobalPath.value = filtered[0] + "/";
 //     } else {
-//       organizeDSglobalPath.value = filtered.slice(0, filtered.length - 1).join("/") + "/";
+//       window.organizeDSglobalPath.value = filtered.slice(0, filtered.length - 1).join("/") + "/";
 //     }
 //     let myPath = datasetStructureJSONObj;
 //     for (var item of filtered.slice(1, filtered.length - 1)) {
@@ -4113,14 +4112,14 @@ const loadDefaultAccount = async () => {
 //     listItems(myPath, "#items", 500, (reset = true));
 //     organizeLandingUIEffect();
 //     // reconstruct div with new elements
-//     getInFolder(".single-item", "#items", organizeDSglobalPath, datasetStructureJSONObj);
+//     getInFolder(".single-item", "#items", window.organizeDSglobalPath, datasetStructureJSONObj);
 //   }
 // });
 
 // // Add folder button
 // organizeDSaddNewFolder.addEventListener("click", function (event) {
 //   event.preventDefault();
-//   let slashCount = organizeDSglobalPath.value.trim().split("/").length - 1;
+//   let slashCount = window.organizeDSglobalPath.value.trim().split("/").length - 1;
 //   if (slashCount !== 1) {
 //     let newFolderName = "New Folder";
 //     Swal.fire({
@@ -4194,7 +4193,7 @@ const loadDefaultAccount = async () => {
 //           );
 //         } else {
 //           /// update datasetStructureJSONObj
-//           let currentPath = organizeDSglobalPath.value;
+//           let currentPath = window.organizeDSglobalPath.value;
 //           let jsonPathArray = currentPath.split("/");
 //           let filtered = jsonPathArray.slice(1).filter(function (el) {
 //             return el != "";
@@ -4211,7 +4210,7 @@ const loadDefaultAccount = async () => {
 //           };
 
 //           listItems(myPath, "#items", 500, (reset = true));
-//           getInFolder(".single-item", "#items", organizeDSglobalPath, datasetStructureJSONObj);
+//           getInFolder(".single-item", "#items", window.organizeDSglobalPath, datasetStructureJSONObj);
 
 //           // log that the folder was successfully added
 //           logCurationForAnalytics(
@@ -5118,7 +5117,7 @@ const showDefaultBFAccount = async () => {
 //   try {
 //     // STEP 1: Build the JSON object from the imported data
 //     // (This function handles bad folders/files, inaccessible folders/files, etc and returns a clean dataset structure)
-//     const currentFileExplorerPath = organizeDSglobalPath.value.trim();
+//     const currentFileExplorerPath = window.organizeDSglobalPath.value.trim();
 
[…]
