### 2023-11-06 52e7b8c
fix: DOI error in paper.bib
