+++ paper/paper.bib
@@ -72,7 +72,7 @@
   volume={11},
   year={2022},
   publisher={Faculty of 1000 Ltd},
-  doi = {10.12688/f1000research.75071.1)}
+  doi = {10.12688/f1000research.75071.1}
 }
 
 @misc{Patel_SODA_Software_to_2023,
