### 2023-11-06 f7bd8ac
feat: GM continue btn after new radio selected
