+++ src/renderer/src/scripts/metadata-files/readme-changes.js
@@ -1,0 +1,731 @@
+// event listeners for changes open dropdown prompts
+document.querySelectorAll(".changes-change-current-account").forEach((element) => {
+  element.addEventListener("click", function () {
+    openDropdownPrompt(null, "bf");
+  });
+});
+
+document.querySelectorAll(".changes-change-current-ds").forEach((element) => {
+  element.addEventListener("click", function () {
+    openDropdownPrompt(null, "dataset");
+  });
+});
+
+// event listeners for readme open dropdown prompts
+document.querySelectorAll(".readme-change-current-account").forEach((element) => {
+  element.addEventListener("click", function () {
+    openDropdownPrompt(null, "bf");
+  });
+});
+
+document.querySelectorAll(".readme-change-current-ds").forEach((element) => {
+  element.addEventListener("click", function () {
+    openDropdownPrompt(null, "dataset");
+  });
+});
+
+// function to raise a warning for empty fields before generating changes or readme
+const generateRCFilesHelper = (type) => {
+  let textValue = $(`#textarea-create-${type}`).val().trim();
+  if (textValue === "") {
+    Swal.fire({
+      title: "Incomplete information",
+      text: "Plase fill in the textarea.",
+      heightAuto: false,
+      backdrop: "rgba(0,0,0, 0.4)",
+      icon: "error",
+      showCancelButton: false,
+      showClass: {
+        popup: "animate__animated animate__zoomIn animate__faster",
+      },
+      hideClass: {
+        popup: "animate__animated animate__zoomOut animate__faster",
+      },
+    });
+    return "empty";
+  }
+};
+
+// generate changes or readme either locally (uploadBFBoolean=false) or onto Pennsieve (uploadBFBoolean=true)
+const generateRCFiles = async (uploadBFBoolean, fileType) => {
+  let result = generateRCFilesHelper(fileType);
+  let bfDataset = document.getElementById(`bf_dataset_load_${fileType}`).innerText.trim();
+  let upperCaseLetters = fileType.toUpperCase() + ".txt";
+
+  if (result === "empty") {
+    return;
+  }
+
+  if (uploadBFBoolean) {
+    // Run pre-flight checks before uploading the changes or readme file to Pennsieve
+    const supplementary_checks = await run_pre_flight_checks(false);
+    if (!supplementary_checks) {
+      return;
+    }
+
+    // Check if dataset is locked after running pre-flight checks
+    const isLocked = await api.isDatasetLocked(window.defaultBfDataset, bfDataset);
+    if (isLocked) {
+      await Swal.fire({
+        icon: "info",
+        title: `${bfDataset} is locked from editing`,
+        html: `
+              This dataset is currently being reviewed by the SPARC curation team, therefore, has been set to read-only mode. No changes can be made to this dataset until the review is complete.
+              <br />
+              <br />
+              If you would like to make changes to this dataset, please reach out to the SPARC curation team at <a href="mailto:curation@sparc.science" target="_blank">curation@sparc.science.</a>
+            `,
+        width: 600,
+        heightAuto: false,
+        backdrop: "rgba(0,0,0, 0.4)",
+        confirmButtonText: "Ok",
+        focusConfirm: true,
+        allowOutsideClick: false,
+      });
+
+      return;
+    }
+
+    let { value: continueProgress } = await Swal.fire({
+      title: `Any existing ${upperCaseLetters} file in the high-level folder of the selected dataset will be replaced.`,
+      text: "Are you sure you want to continue?",
+      allowEscapeKey: false,
[…]
