+++ src/renderer/src/scripts/metadata-files/manifest.js
@@ -1,0 +1,2545 @@
+const { platform } = require("os");
+const { copyFile, readdir } = require("fs").promises;
+const { join } = require("path");
+let openedEdit = false;
+
+// opendropdown event listeners
+document.querySelectorAll(".manifest-change-current-account").forEach((element) => {
+  element.addEventListener("click", () => {
+    openDropdownPrompt(null, "bf");
+  });
+});
+
+document.querySelectorAll(".manifest-change-current-ds").forEach((element) => {
+  element.addEventListener("click", () => {
+    openDropdownPrompt(null, "dataset");
+  });
+});
+
+const jstreePreviewManifest = document.getElementById("div-dataset-tree-preview-manifest");
+
+const guidedJsTreePreviewManifest = document.getElementById(
+  "guided-div-dataset-tree-preview-manifest"
+);
+
+const showLocalDatasetManifest = () => {
+  ipcRenderer.send("open-file-dialog-local-dataset-manifest-purpose");
+};
+
+const selectManifestGenerationLocation = () => {
+  ipcRenderer.send("open-file-dialog-local-dataset-manifest-generate-purpose");
+};
+
+const openDirectoryAtManifestGenerationLocation = (generationLocation) => {
+  openFolder(generationLocation);
+  return;
+};
+
+const openFolder = (generationLocation) => {
+  // create the folder path
+  try {
+    client.get("/datasets/open", {
+      params: {
+        dataset_path: generationLocation,
+      },
+    });
+  } catch (error) {
+    console.error(error);
+    clientError(error);
+    Swal.fire({
+      title: "Error",
+      text: `Could not open the new manifest folder for you. Please view your manifest files by navigating to ${generationLocation}.`,
+      icon: "error",
+      confirmButtonText: "Ok",
+    });
+  }
+};
+
+$(document).ready(function () {
+  let localDataSetImport = false;
+  let fileOpenedOnce = {};
+  ipcRenderer.on("selected-local-dataset-manifest-purpose", (event, folderPath) => {
+    if (folderPath.length > 0) {
+      if (folderPath !== null) {
+        localDataSetImport = true;
+        fileOpenedOnce = {};
+        document.getElementById("input-manifest-local-folder-dataset").placeholder = folderPath[0];
+        localDatasetFolderPath = folderPath[0];
+        $("#div-confirm-manifest-local-folder-dataset").css("display", "flex");
+        $($("#div-confirm-manifest-local-folder-dataset button")[0]).show();
+      } else {
+        document.getElementById("input-manifest-local-folder-dataset").placeholder = "Browse here";
+        localDatasetFolderPath = "";
+        $("#div-confirm-manifest-local-folder-dataset").hide();
+        $("#Question-prepare-manifest-2").nextAll().removeClass("show").removeClass("prev");
+      }
+    } else {
+      document.getElementById("input-manifest-local-folder-dataset").placeholder = "Browse here";
+      localDatasetFolderPath = "";
+      $("#div-confirm-manifest-local-folder-dataset").hide();
+      $("#Question-prepare-manifest-2").nextAll().removeClass("show").removeClass("prev");
+    }
+  });
+
+  ipcRenderer.on("selected-local-dataset-manifest-generate-purpose", (event, folderPath) => {
+    if (folderPath.length <= 0 || folderPath === null) {
+      document.getElementById("input-manifest-local-gen-location").placeholder = "Browse here";
+      return;
+    }
+
+    document.getElementById("input-manifest-local-gen-location").placeholder = folderPath[0];
+  });
+
+  $("#bf_dataset_create_manifest").on("DOMSubtreeModified", () => {
[…]
