+++ src/renderer/src/main.js
@@ -5,6 +5,7 @@ import './scripts/client'
 import './scripts/guided-mode/guided-curate-dataset'
 import './scripts/globals'
 import './scripts/others/renderer'
+import './scripts/metadata-files/submission'
 
 
 // Application Lotties

[…]
