+++ src/renderer/src/scripts/globals.js
@@ -180,6 +180,9 @@ window.organizeDSglobalPath = "";
 window.dataset_path = document.getElementById("input-global-path");
 
 
+
+
+
 // used in renderer and guided Mode
 window.createDragSort = (tagify) => {
   const onDragEnd = () => {
