+++ src/renderer/src/scripts/manage-dataset/manage-dataset.js
@@ -1,0 +1,3552 @@
+// event listeners for opening dataset or account selection dropdown
+import Accordion from 'accordion-js';
+// TODO: Follow up that this is the way to import it
+import 'accordion-js/dist/accordion.min.css';
+
+
+while (!window.htmlPagesAdded) {
+  await new Promise((resolve) => setTimeout(resolve, 100))
+}
+
+// TODO: Add logic so this doesnt apply to the organization fields
+document.querySelectorAll(".ds-dd:not(.organization)").forEach((dropdownElement) => {
+  dropdownElement.addEventListener("click", function () {
+    openDropdownPrompt(this, "dataset");
+  });
+});
+
+document.querySelectorAll(".md-change-current-account").forEach((dropdownElement) => {
+  dropdownElement.addEventListener("click", function () {
+    openDropdownPrompt(this, "bf");
+  });
+});
+
+document.querySelectorAll(".ds-dd.organization").forEach((dropdownElement) => {
+  dropdownElement.addEventListener("click", function () {
+    openDropdownPrompt(this, "organization");
+  });
+});
+
+var forbidden_characters_bf = '/:*?"<>';
+
+const check_forbidden_characters_ps = (my_string) => {
+  // Args:
+  // my_string: string with characters (string)
+  // Returns:
+  // False: no forbidden character
+  // True: presence of forbidden character(s)
+  let check = false;
+
+  for (let i = 0; i < forbidden_characters_bf.length; i++) {
+    if (my_string.indexOf(forbidden_characters_bf[i]) > -1) {
+      return true;
+    }
+  }
+
+  return check;
+};
+
+const determineSwalLoadingMessage = (addEditButton) => {
+  let loadingMessage = "";
+  switch (addEditButton.text()) {
+    case "Add subtitle":
+      loadingMessage = "Adding subtitle to dataset";
+      break;
+    case "Edit subtitle":
+      loadingMessage = "Editing your dataset's subtitle";
+      break;
+    case "Add description":
+      loadingMessage = "Adding description to dataset";
+      break;
+    case "Edit description":
+      loadingMessage = "Editing your dataset's description";
+      break;
+    case "Add tags":
+      loadingMessage = "Adding tags to dataset";
+      break;
+    case "Edit tags":
+      loadingMessage = "Editing your dataset's tags";
+      break;
+  }
+  return loadingMessage;
+};
+
+const determineSwalSuccessMessage = (addEditButton) => {
+  let successMessage = "";
+  switch (addEditButton.text()) {
+    case "Add subtitle":
+      successMessage = "Successfully added subtitle to dataset";
+      break;
+    case "Edit subtitle":
+      successMessage = "Successfully edited dataset's subtitle";
+      break;
+    case "Add description":
+      successMessage = "Successfully added description to dataset";
+      break;
+    case "Edit description":
+      successMessage = "Successfully edited dataset's description";
+      break;
+    case "Add tags":
+      successMessage = "Successfully added tags to dataset";
+      break;
+    case "Edit tags":
+      successMessage = "Successfully edited dataset's tags";
+      break;
+  }
+  return successMessage;
+};
+
+// illegal character name warning for new dataset names
+$("#bf-new-dataset-name").on("keyup", () => {
+  // Clear success lottie
+  $("#dataset-created-success-lottie").empty();
+  $("#dataset-success-container").addClass("hidden");
+  let newName = $("#bf-new-dataset-name").val().trim();
+
+  if (newName !== "") {
+    if (check_forbidden_characters_ps(newName)) {
+      Swal.fire({
[…]
