+++ src/renderer/src/scripts/others/renderer.js
@@ -6525,7 +6525,7 @@ const showDefaultBFAccount = async () => {
 //   $("#para-continue-name-dataset-generate").text("");
 
 //   if (newName !== "") {
-//     if (check_forbidden_characters_ps(newName)) {
+//     if (window.check_forbidden_characters_ps(newName)) {
 //       document.getElementById("div-confirm-inputNewNameDataset").style.display = "none";
 //       $("#btn-confirm-new-dataset-name").hide();
 //       document.getElementById("para-new-name-dataset-message").innerHTML =
