+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -491,7 +491,7 @@ const objectsHaveSameKeys = (...objects) => {
 };
 
 const getGuidedProgressFileNames = () => {
-  return fs
+  return window.fs
     .readdirSync(guidedProgressFilePath)
     .map((progressFileName) => progressFileName.replace(".json", ""));
 };
@@ -502,13 +502,13 @@ const updateGuidedDatasetName = (newDatasetName) => {
   //update old progress file with new dataset name
   const oldProgressFilePath = `${guidedProgressFilePath}/${previousDatasetName}.json`;
   const newProgressFilePath = `${guidedProgressFilePath}/${newDatasetName}.json`;
-  fs.renameSync(oldProgressFilePath, newProgressFilePath);
+  window.fs.renameSync(oldProgressFilePath, newProgressFilePath);
 
   const bannerImagePathToUpdate = sodaJSONObj["digital-metadata"]["banner-image-path"];
   if (bannerImagePathToUpdate) {
     const newBannerImagePath = bannerImagePathToUpdate.replace(previousDatasetName, newDatasetName);
     //Rename the old banner image folder to the new dataset name
-    fs.renameSync(bannerImagePathToUpdate, newBannerImagePath);
+    window.fs.renameSync(bannerImagePathToUpdate, newBannerImagePath);
     //change the banner image path in the JSON obj
     sodaJSONObj["digital-metadata"]["banner-image-path"] = newBannerImagePath;
   }
@@ -1016,7 +1016,7 @@ const savePageChanges = async (pageBeingLeftID) => {
           message: "Please enter a dataset name.",
         });
       }
-      if (check_forbidden_characters_ps(datasetNameInput)) {
+      if (window.check_forbidden_characters_ps(datasetNameInput)) {
         errorArray.push({
           type: "notyf",
           message:
@@ -1424,7 +1424,7 @@ const savePageChanges = async (pageBeingLeftID) => {
 
           const codeDescriptionPath = codeDescriptionPathElement.innerHTML;
           //Check if the code description file is valid
-          if (!fs.existsSync(codeDescriptionPath)) {
+          if (!window.fs.existsSync(codeDescriptionPath)) {
             errorArray.push({
               type: "notyf",
               message: "The imported code_description file does not exist at the selected path",
@@ -1475,7 +1475,7 @@ const savePageChanges = async (pageBeingLeftID) => {
 
           const codeDescriptionPath = codeDescriptionPathElement.innerHTML;
           //Check if the code description file is valid
-          if (!fs.existsSync(codeDescriptionPath)) {
+          if (!window.fs.existsSync(codeDescriptionPath)) {
             errorArray.push({
               type: "notyf",
               message: "The imported code_description file does not exist at the selected path",
@@ -2855,7 +2855,7 @@ const saveGuidedProgress = async (guidedProgressFileName) => {
 
 const readDirAsync = async (path) => {
   return new Promise((resolve, reject) => {
-    fs.readdir(path, (error, result) => {
+    window.fs.readdir(path, (error, result) => {
       if (error) {
         throw new Error(error);
       } else {
@@ -2867,7 +2867,7 @@ const readDirAsync = async (path) => {
 
 const readFileAsync = async (path) => {
   return new Promise((resolve, reject) => {
-    fs.readFile(path, "utf-8", (error, result) => {
+    window.fs.readFile(path, "utf-8", (error, result) => {
       if (error) {
         throw new Error(error);
       } else {
@@ -2895,7 +2895,7 @@ const deleteProgresFile = async (progressFileName) => {
[…]
