+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -4625,7 +4625,7 @@ let removeEventListener = false;
 const copyLink = (link) => {
   Clipboard.writeText(link);
 
-  notyf.open({
+  window.notyf.open({
     duration: "2000",
     type: "success",
     message: "Link copied!",
@@ -10329,7 +10329,7 @@ const openCopySampleMetadataPopup = async () => {
     });
 };
 
-const specifySubject = (event, subjectNameInput) => {
+window.specifySubject = (event, subjectNameInput) => {
   if (event.which == 13) {
     try {
       const subjectName = `sub-${subjectNameInput.val().trim()}`;
@@ -10339,7 +10339,7 @@ const specifySubject = (event, subjectNameInput) => {
           <i
             class="far fa-edit"
             style="cursor: pointer; margin-top: .2rem;"
-            onclick="openSubjectRenameInput($(this))"
+            onclick="window.openSubjectRenameInput($(this))"
           >
           </i>
         </div>
@@ -10350,7 +10350,7 @@ const specifySubject = (event, subjectNameInput) => {
       trashCanElement.style.display = "block";
 
       if (subjectName.length > 0) {
-        const subjectNameIsValid = evaluateStringAgainstSdsRequirements(
+        const subjectNameIsValid = window.evaluateStringAgainstSdsRequirements(
           subjectName,
           "string-adheres-to-identifier-conventions"
         );
@@ -10369,10 +10369,10 @@ const specifySubject = (event, subjectNameInput) => {
           sodaJSONObj.addSubject(subjectName);
         }
         subjectIdCellToAddNameTo.html(subjectNameElement);
-        addSubjectSpecificationTableRow();
+        window.addSubjectSpecificationTableRow();
       }
     } catch (error) {
-      notyf.open({
+      window.notyf.open({
         duration: "3000",
         type: "error",
         message: error,
@@ -10412,7 +10412,7 @@ const specifySample = (event, sampleNameInput) => {
       const subjectToAddSampleTo = subjectSampleAdditionTable.find(".samples-subject-name").text();
 
       if (sampleName.length > 0) {
-        const sampleNameIsValid = evaluateStringAgainstSdsRequirements(
+        const sampleNameIsValid = window.evaluateStringAgainstSdsRequirements(
           sampleName,
           "string-adheres-to-identifier-conventions"
         );
@@ -10448,7 +10448,7 @@ const specifySample = (event, sampleNameInput) => {
       }
     } catch (error) {
       console.log(error);
-      notyf.open({
+      window.notyf.open({
         duration: "3000",
         type: "error",
         message: error,
@@ -10485,12 +10485,12 @@ const specifyPool = (event, poolNameInput) => {
       const poolIdCellToAddNameTo = poolNameInput.parent();
       let poolsTable = $("#pools-table");
       if (poolName !== "pool-") {
-        const poolNameIsValid = evaluateStringAgainstSdsRequirements(
+        const poolNameIsValid = window.evaluateStringAgainstSdsRequirements(
           poolName,
           "string-adheres-to-identifier-conventions"
         );
         if (!poolNameIsValid) {
-          notyf.open({
+          window.notyf.open({
             duration: "3000",
             type: "error",
             message: "Pool IDs may not contain spaces or special characters",
@@ -10544,7 +10544,7 @@ const specifyPool = (event, poolNameInput) => {
         poolIdCellToAddNameTo.html(poolNameElement);
       }
     } catch (error) {
-      notyf.open({
+      window.notyf.open({
         duration: "3000",
         type: "error",
         message: error,
@@ -10571,10 +10571,10 @@ const updatePoolDropdown = (poolDropDown, poolName) => {
 };
 
 //On edit button click, creates a new subject ID rename input box
[…]
