+++ src/renderer/src/scripts/others/renderer.js
@@ -59,6 +59,10 @@ import {
 } from '../globals'
 import checkForAnnouncements from './announcements'
 
+// add jquery to the window object
+window.$ = jQuery;
+window.jQuery = jQuery;
+
 fixPath();
 
 // // const prevent_sleep_id = "";
@@ -128,7 +132,7 @@ if (autoUpdateLaunch == false || autoUpdateLaunch == null || autoUpdateLaunch ==
 // // Connect to Python back-end
 // //////////////////////////////////
 
-const notyf = new Notyf({
+window.notyf = new Notyf({
   position: { x: "right", y: "bottom" },
   dismissible: true,
   ripple: false,
@@ -484,7 +488,7 @@ const run_pre_flight_checks = async (check_update = true) => {
     console.log("Initiating pre flight checks")
 
     if (!preFlightCheckNotyf) {
-      preFlightCheckNotyf = notyf.open({
+      preFlightCheckNotyf = window.notyf.open({
         duration: 25000,
         type: "info",
         message: "Checking SODA's connection to Pennsieve...",
@@ -509,7 +513,7 @@ const run_pre_flight_checks = async (check_update = true) => {
     if (!account_present) {
       // Dismiss the preflight check notification if it is still open
       if (preFlightCheckNotyf) {
-        notyf.dismiss(preFlightCheckNotyf);
+        window.notyf.dismiss(preFlightCheckNotyf);
         preFlightCheckNotyf = null;
       }
 
@@ -590,7 +594,7 @@ const run_pre_flight_checks = async (check_update = true) => {
 
         // Dismiss the preflight check notification if it is still open
         if (preFlightCheckNotyf) {
-          notyf.dismiss(preFlightCheckNotyf);
+          window.notyf.dismiss(preFlightCheckNotyf);
           preFlightCheckNotyf = null;
         }
 
@@ -630,7 +634,7 @@ const run_pre_flight_checks = async (check_update = true) => {
 
       // Dismiss the preflight check notification if it is still open
       if (preFlightCheckNotyf) {
-        notyf.dismiss(preFlightCheckNotyf);
+        window.notyf.dismiss(preFlightCheckNotyf);
         preFlightCheckNotyf = null;
       }
 
@@ -682,7 +686,7 @@ const run_pre_flight_checks = async (check_update = true) => {
 
       // Dismiss the preflight check notification if it is still open
       if (preFlightCheckNotyf) {
-        notyf.dismiss(preFlightCheckNotyf);
+        window.notyf.dismiss(preFlightCheckNotyf);
         preFlightCheckNotyf = null;
       }
       // If the user clicks the skip button, return false which will cause the pre flight checks to fail
@@ -725,7 +729,7 @@ const run_pre_flight_checks = async (check_update = true) => {
 
       // Dismiss the preflight check notification if it is still open
       if (preFlightCheckNotyf) {
-        notyf.dismiss(preFlightCheckNotyf);
+        window.notyf.dismiss(preFlightCheckNotyf);
         preFlightCheckNotyf = null;
       }
       // If the user clicks the skip button, return false which will cause the pre flight checks to fail
@@ -774,7 +778,7 @@ const run_pre_flight_checks = async (check_update = true) => {
       }
       // Dismiss the preflight check notification if it is still open
       if (preFlightCheckNotyf) {
-        notyf.dismiss(preFlightCheckNotyf);
+        window.notyf.dismiss(preFlightCheckNotyf);
         preFlightCheckNotyf = null;
       }
 
@@ -803,11 +807,11 @@ const run_pre_flight_checks = async (check_update = true) => {
 
     // Dismiss the preflight check notification if it is still open
     if (preFlightCheckNotyf) {
-      notyf.dismiss(preFlightCheckNotyf);
[…]
