### 2023-11-07 1a797a4
feat: ✨ Implemented dynamic template generation
