+++ src/renderer/src/scripts/others/renderer.js
@@ -4081,7 +4081,7 @@ const loadDefaultAccount = async () => {
 //     "<b>protocol</b>: This folder contains supplementary files to accompany the experimental protocols submitted to Protocols.io. Please note that this is not a substitution for the experimental protocol which must be submitted to <b><a target='_blank' href='https://www.protocols.io/groups/sparc'> Protocols.io/sparc </a></b>.",
 // };
 
-// var sodaJSONObj = {};
+window.sodaJSONObj = {};
 
 // /// back button Curate
 // organizeDSbackButton.addEventListener("click", function () {
