### 2023-11-08 c7245f3
wip: 🚧 Properly set buton config from existing progress datasets
