### 2023-11-08 bf2bfdd
fix: 🐛 fixed workbook styling
