### 2023-11-08 0c36a26
refactor: sample selectable from file viewer GM
