+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -5,6 +5,7 @@ import { resetLazyLoading, guidedUnLockSideBar } from "../../assets/nav";
 import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
 import api from '../others/api/api'
 import lottie from "lottie-web"
+import {dragDrop} from '../../assets/lotties/lotties'
 import Swal from "sweetalert2"
 import Tagify from "@yaireo/tagify";
 import DragSort from '@yaireo/dragsort'
@@ -314,7 +315,7 @@ const guidedCheckHighLevelFoldersForImproperFiles = (datasetStructure) => {
 };
 
 document.getElementById("guided-button-dataset-contains-code").addEventListener("click", () => {
-  const codeFolder = datasetStructureJSONObj["folders"]["code"];
+  const codeFolder = window.datasetStructureJSONObj["folders"]["code"];
   if (codeFolder) {
     if (folderImportedFromPennsieve(codeFolder)) {
       // If the code folder is imported from Pennsieve, unmark it as deleted
@@ -325,7 +326,7 @@ document.getElementById("guided-button-dataset-contains-code").addEventListener(
 });
 
 document.getElementById("guided-button-has-protocol-data").addEventListener("click", () => {
-  const protocolFolder = datasetStructureJSONObj["folders"]["protocol"];
+  const protocolFolder = window.datasetStructureJSONObj["folders"]["protocol"];
   if (protocolFolder) {
     if (folderImportedFromPennsieve(protocolFolder)) {
       // If the protocol folder is imported from Pennsieve, unmark it as deleted and update the UI
@@ -336,7 +337,7 @@ document.getElementById("guided-button-has-protocol-data").addEventListener("cli
 });
 
 document.getElementById("guided-button-has-docs-data").addEventListener("click", () => {
-  const docsFolder = datasetStructureJSONObj["folders"]["docs"];
+  const docsFolder = window.datasetStructureJSONObj["folders"]["docs"];
   if (docsFolder) {
     if (folderImportedFromPennsieve(docsFolder)) {
       // If the protocol folder is imported from Pennsieve, unmark it as deleted and update the UI
@@ -644,7 +645,7 @@ const savePageChanges = async (pageBeingLeftID) => {
           window.sodaJSONObj["initially-pulled-dataset-structure"] = JSON.parse(
             JSON.stringify(data["soda_object"]["dataset-structure"])
           );
-          datasetStructureJSONObj = data["soda_object"]["dataset-structure"];
+          window.datasetStructureJSONObj = data["soda_object"]["dataset-structure"];
         } catch (error) {
           console.log(error);
           errorArray.push({
@@ -657,7 +658,7 @@ const savePageChanges = async (pageBeingLeftID) => {
 
         // Reject if anyh non-sparc folders are in the root of the dataset
         let invalidBaseFolders = [];
-        for (const baseFolder of Object.keys(datasetStructureJSONObj["folders"])) {
+        for (const baseFolder of Object.keys(window.datasetStructureJSONObj["folders"])) {
           if (
             !guidedHighLevelFolders.includes(baseFolder) &&
             !nonGuidedHighLevelFolders.includes(baseFolder)
@@ -679,7 +680,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         // Datasets pulled into Guided Mode should only have pool-folders or sub-folders inside of the primary, source,
         // and derivative high level folders. If this is not the case with the pulled dataset, reject it.
         const [invalidFolders, invalidFiles] =
-          guidedCheckHighLevelFoldersForImproperFiles(datasetStructureJSONObj);
[…]
