+++ src/renderer/src/scripts/organize-dataset/organizeDS.js
@@ -1,3 +1,8 @@
+while (!window.htmlPagesAdded) {
+  await new Promise((resolve) => setTimeout(resolve, 100))
+}
+
+
 //// option to show tool-tips for high-level folders
 const showTooltips = (ev) => {
   let folderName = ev.parentElement.innerText;
@@ -128,7 +133,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
     }).then((result) => {
       if (result.isConfirmed) {
         let itemToRestore = itemToDelete;
-        let filtered = getGlobalPath(organizeCurrentLocation);
+        let filtered = window.getGlobalPath(organizeCurrentLocation);
         let myPath = getRecursivePath(filtered.slice(1), inputGlobal);
 
         if (filtered.length == 1) {
@@ -187,7 +192,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
         delete myPath[type][itemToRestore];
 
         // update UI with updated jsonobj
-        listItems(myPath, uiItem, 500, (reset = true));
+        window.listItems(myPath, uiItem, 500, true);
         getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
         beginScrollListen();
       }
@@ -213,7 +218,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
         },
       }).then((result) => {
         if (result.isConfirmed) {
-          let filtered = getGlobalPath(organizeCurrentLocation);
+          let filtered = window.getGlobalPath(organizeCurrentLocation);
           let myPath = getRecursivePath(filtered.slice(1), inputGlobal);
 
           $("div.single-item.selected-item > .folder_desc").each(function (index, current_element) {
@@ -250,7 +255,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
           });
 
           // update UI with updated jsonobj
-          listItems(myPath, uiItem, 500, (reset = true));
+          window.listItems(myPath, uiItem, 500, true);
           getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
           beginScrollListen();
         }
@@ -276,7 +281,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
       }).then((result) => {
         if (result.isConfirmed) {
           /// get current location of folders or files
-          var filtered = getGlobalPath(organizeCurrentLocation);
+          var filtered = window.getGlobalPath(organizeCurrentLocation);
           var myPath = getRecursivePath(filtered.slice(1), inputGlobal);
           // update Json object with new folder created
           if (
@@ -300,7 +305,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
             delete myPath[type][itemToDelete];
           }
           // update UI with updated jsonobj
-          listItems(myPath, uiItem, 500, (reset = true));
+          window.listItems(myPath, uiItem, 500, true);
           getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
           beginScrollListen();
         }
@@ -398,7 +403,7 @@ const renameFolder = (
   event1, //this
   organizeCurrentLocation, //current section of My_folder
   itemElement, //the elements in the container with items
-  inputGlobal, //datasetStructureJSONObj
+  inputGlobal, //window.datasetStructureJSONObj
   uiItem, //container with the folders
   singleUIItem //class name
 ) => {
@@ -538,7 +543,7 @@ const renameFolder = (
           });
 
           // get location of current item in SODA JSON
-          let filtered = getGlobalPath(organizeCurrentLocation);
[…]
