### 2023-11-08 66c47fe
feat: ✨ Added the ability to read the json structure excel file
