### 2023-11-08 595514c
feat: ✨ Added ability to mute thrown error on duplicates
