+++ src/main/main-process/help/application-menu.js
@@ -1,0 +1,227 @@
+const { BrowserWindow, Menu, app, shell, dialog } = require("electron");
+
+let template = [
+  {
+    label: "Edit",
+    submenu: [
+      {
+        label: "Undo",
+        accelerator: "CmdOrCtrl+Z",
+        role: "undo",
+      },
+      {
+        label: "Redo",
+        accelerator: "Shift+CmdOrCtrl+Z",
+        role: "redo",
+      },
+      {
+        type: "separator",
+      },
+      {
+        label: "Cut",
+        accelerator: "CmdOrCtrl+X",
+        role: "cut",
+      },
+      {
+        label: "Copy",
+        accelerator: "CmdOrCtrl+C",
+        role: "copy",
+      },
+      {
+        label: "Paste",
+        accelerator: "CmdOrCtrl+V",
+        role: "paste",
+      },
+      {
+        label: "Select All",
+        accelerator: "CmdOrCtrl+A",
+        role: "selectall",
+      },
+    ],
+  },
+  {
+    label: "View",
+    submenu: [
+      {
+        label: "Reload",
+        accelerator: "CmdOrCtrl+R",
+        click: (item, focusedWindow) => {
+          if (focusedWindow) {
+            // on reload, start fresh and close any old
+            // open secondary windows
+            if (focusedWindow.id === 1) {
+              BrowserWindow.getAllWindows().forEach((win) => {
+                if (win.id > 1) win.close();
+              });
+            }
+            focusedWindow.reload();
+          }
+        },
+      },
+      {
+        label: "Toggle Full Screen",
+        accelerator: (() => {
+          if (process.platform === "darwin") {
+            return "Ctrl+Command+F";
+          } else {
+            return "F11";
+          }
+        })(),
+        click: (item, focusedWindow) => {
+          if (focusedWindow) {
+            focusedWindow.setFullScreen(!focusedWindow.isFullScreen());
+          }
+        },
+      },
+      {
+        label: "Toggle Developer Tools",
+        accelerator: (() => {
+          if (process.platform === "darwin") {
+            return "Alt+Command+I";
+          } else {
+            return "Ctrl+Shift+I";
+          }
+        })(),
+        click: (item, focusedWindow) => {
+          if (focusedWindow) {
+            focusedWindow.toggleDevTools();
+          }
+        },
+      },
+      {
+        type: "separator",
+      },
+    ],
+  },
+  {
+    label: "Window",
+    role: "window",
+    submenu: [
+      {
+        label: "Minimize",
+        accelerator: "CmdOrCtrl+M",
+        role: "minimize",
+      },
+      {
+        label: "Close",
+        accelerator: "CmdOrCtrl+W",
+        role: "close",
+      },
+      {
+        type: "separator",
+      },
+    ],
+  },
+];
+
+function addUpdateMenuItems(items, position) {
+  if (process.mas) return;
+
+  const version = app.getVersion();
+  let updateItems = [
+    {
+      label: `Version ${version}`,
+      enabled: false,
+    },
+  ];
+
+  items.splice.apply(items, [position, 0].concat(updateItems));
+}
+
+function findReopenMenuItem() {
+  const menu = Menu.getApplicationMenu();
+  if (!menu) return;
+
+  let reopenMenuItem;
+  menu.items.forEach((item) => {
+    if (item.submenu) {
+      item.submenu.items.forEach((item) => {
+        if (item.key === "reopenMenuItem") {
+          reopenMenuItem = item;
+        }
+      });
+    }
+  });
+  return reopenMenuItem;
+}
+
+if (process.platform === "darwin") {
+  const name = app.getName();
+  template.unshift({
[…]

[…]
