### 2023-11-08 ff98426
feat: ✨ convert spreadsheet json to strings
