### 2023-11-09 8a79d28
refactor: import folder file viewer GM
