+++ src/renderer/src/scripts/others/renderer.js
@@ -1223,13 +1223,13 @@ window.electron.ipcRenderer.on("app_version", (event, arg) => {
 // const organizeDSbackButton = document.getElementById("button-back");
 // const organizeDSaddFiles = document.getElementById("add-files");
 // const organizeDSaddNewFolder = document.getElementById("new-folder");
-// const organizeDSaddFolders = document.getElementById("add-folders");
+const organizeDSaddFolders = document.getElementById("add-folders");
 // const contextMenu = document.getElementById("mycontext");
-// const fullNameValue = document.querySelector(".hoverFullName");
+const fullNameValue = document.querySelector(".hoverFullName");
 // const homePathButton = document.getElementById("home-path");
-// const menuFolder = document.querySelector(".menu.reg-folder");
-// const menuFile = document.querySelector(".menu.file");
-// const menuHighLevelFolders = document.querySelector(".menu.high-level-folder");
+window.menuFolder = document.querySelector(".menu.reg-folder");
+window.menuFile = document.querySelector(".menu.file");
+window.menuHighLevelFolders = document.querySelector(".menu.high-level-folder");
 // const organizeNextStepBtn = document.getElementById("button-organize-confirm-create");
 // const organizePrevStepBtn = document.getElementById("button-organize-prev");
 // const manifestFileCheck = document.getElementById("generate-manifest-curate");
@@ -4069,19 +4069,19 @@ const loadDefaultAccount = async () => {
 // var backFolder = [];
 // var forwardFolder = [];
 
-// var highLevelFolders = ["code", "derivative", "docs", "source", "primary", "protocol"];
-// var highLevelFolderToolTip = {
-//   code: "<b>code</b>: This folder contains all the source code used in the study (e.g., Python, MATLAB, etc.)",
-//   derivative:
-//     "<b>derivative</b>: This folder contains data files derived from raw data (e.g., processed image stacks that are annotated via the MBF tools, segmentation files, smoothed overlays of current and voltage that demonstrate a particular effect, etc.)",
-//   docs: "<b>docs</b>: This folder contains all other supporting files that don't belong to any of the other folders (e.g., a representative image for the dataset, figures, etc.)",
-//   source:
-//     "<b>source</b>: This folder contains very raw data i.e. raw or untouched files from an experiment. For example, this folder may include the “truly” raw k-space data for an MR image that has not yet been reconstructed (the reconstructed DICOM or NIFTI files, for example, would be found within the primary folder). Another example is the unreconstructed images for a microscopy dataset.",
-//   primary:
-//     "<b>primary</b>: This folder contains all folders and files for experimental subjects and/or samples. All subjects will have a unique folder with a standardized name the same as the names or IDs as referenced in the subjects metadata file. Within each subject folder, the experimenter may choose to include an optional “session” folder if the subject took part in multiple experiments/ trials/ sessions. The resulting data is contained within data type-specific (Datatype) folders within the subject (or session) folders. The SPARC program’s Data Sharing Committee defines 'raw' (primary) data as one of the types of data that should be shared. This covers minimally processed raw data, e.g. time-series data, tabular data, clinical imaging data, genomic, metabolomic, microscopy data, which can also be included within their own folders.",
-//   protocol:
[…]
