+++ src/renderer/src/scripts/organize-dataset/organizeDS.js
@@ -193,7 +193,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
 
         // update UI with updated jsonobj
         window.listItems(myPath, uiItem, 500, true);
-        getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
+        window.getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
         beginScrollListen();
       }
     });
@@ -256,7 +256,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
 
           // update UI with updated jsonobj
           window.listItems(myPath, uiItem, 500, true);
-          getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
+          window.getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
           beginScrollListen();
         }
       });
@@ -306,7 +306,7 @@ const delFolder = (ev, organizeCurrentLocation, uiItem, singleUIItem, inputGloba
           }
           // update UI with updated jsonobj
           window.listItems(myPath, uiItem, 500, true);
-          getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
+          window.getInFolder(singleUIItem, uiItem, organizeCurrentLocation, inputGlobal);
           beginScrollListen();
         }
       });
@@ -603,7 +603,7 @@ const loadFileFolder = (myPath) => {
     }
     appendString =
       appendString +
-      '<div class="single-item" onmouseover="hoverForFullName(this)" onmouseleave="hideFullName()"><h1 oncontextmenu="folderContextMenu(this)" class="myFol' +
+      '<div class="single-item" onmouseover="window.hoverForFullName(this)" onmouseleave="window.hideFullName()"><h1 oncontextmenu="folderContextMenu(this)" class="myFol' +
       emptyFolder +
       '"></h1><div class="folder_desc">' +
       item +
@@ -656,7 +656,7 @@ const loadFileFolder = (myPath) => {
     }
     appendString =
       appendString +
-      '<div class="single-item" onmouseover="hoverForFullName(this)" onmouseleave="hideFullName()"><h1 class="myFile ' +
+      '<div class="single-item" onmouseover="window.hoverForFullName(this)" onmouseleave="window.hideFullName()"><h1 class="myFile ' +
       extension +
       '" oncontextmenu="fileContextMenu(this)" style="margin-bottom: 10px""></h1><div class="folder_desc">' +
       item +
@@ -1262,9 +1262,9 @@ const handleDuplicateImports = async (btnId, duplicateArray, curationMode) => {
                 action: ["new", "renamed"],
               };
               window.listItems(myPath, "#items");
-              getInFolder("#items", "#items", organizeDSglobalPath, window.datasetStructureJSONObj);
-              hideMenu("folder", menuFolder, menuHighLevelFolders, menuFile);
-              hideMenu("high-level-folder", menuFolder, menuHighLevelFolders, menuFile);
+              window.getInFolder("#items", "#items", organizeDSglobalPath, window.datasetStructureJSONObj);
+              window.hideMenu("folder", window.menuFolder, window.menuHighLevelFolders, window.menuFile);
+              window.hideMenu("high-level-folder", window.menuFolder, window.menuHighLevelFolders, window.menuFile);
             }
           }
         } else {
@@ -1287,13 +1287,13 @@ const handleDuplicateImports = async (btnId, duplicateArray, curationMode) => {
                 action: ["new", "renamed"],
               };
               var appendString =
[…]
