### 2023-11-09 907ebb6
refactor: navigate file viewer GM
