+++ scripts/others/update-path-darwin.js
@@ -6,7 +6,6 @@
  */
 const fixPath = () => {
   if (process.platform == "darwin") {
-
     process.env.PATH = [
       "./node_modules/.bin",
       "/.nodebrew/current/bin",
