### 2023-11-09 d70f477
wip: 🚧 Added generalized entity analysis function
