+++ src/renderer/src/scripts/others/renderer.js
@@ -1121,7 +1121,7 @@ const get_latest_agent_version = async () => {
     window.reverseSwalButtons = false;
     targetRelease.assets.forEach((asset, index) => {
       let file_name = asset.name;
-      if (path.extname(file_name) == ".deb") {
+      if (window.path.extname(file_name) == ".deb") {
         browser_download_url = asset.browser_download_url;
       }
     });
@@ -1371,28 +1371,28 @@ dragselect_area.subscribe("dragstart", ({ items, event, isDragging }) => {
 // ///// Global variables for this section
 
 // /////// Save and load award and milestone info
-// let metadataPath = path.join(homeDirectory, "SODA", "METADATA");
+// let metadataPath = window.path.join(homeDirectory, "SODA", "METADATA");
 // let awardFileName = "awards.json";
 // let affiliationFileName = "affiliations.json";
 // let milestoneFileName = "milestones.json";
 // let protocolConfigFileName = "protocol-config.json";
-// let affiliationConfigPath = path.join(metadataPath, affiliationFileName);
-// let milestonePath = path.join(metadataPath, milestoneFileName);
-// let progressFilePath = path.join(homeDirectory, "SODA", "Progress");
-// let guidedProgressFilePath = path.join(homeDirectory, "SODA", "Guided-Progress");
-// const guidedManifestFilePath = path.join(homeDirectory, "SODA", "guided_manifest_files");
-// let protocolConfigPath = path.join(metadataPath, protocolConfigFileName);
+// let affiliationConfigPath = window.path.join(metadataPath, affiliationFileName);
+// let milestonePath = window.path.join(metadataPath, milestoneFileName);
+window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
+// let guidedProgressFilePath = window.path.join(homeDirectory, "SODA", "Guided-Progress");
+// const guidedManifestFilePath = window.path.join(homeDirectory, "SODA", "guided_manifest_files");
+// let protocolConfigPath = window.path.join(metadataPath, protocolConfigFileName);
 // let allCollectionTags = {};
 // let currentTags = {};
 // let currentCollectionTags = [];
 
 // if (process.platform === "linux") {
 //   //check if data exists inside of the Soda folder, and if it does, move it into the capitalized SODA folder
-//   if (fs.existsSync(path.join(homeDirectory, "Soda"))) {
+//   if (fs.existsSync(window.path.join(homeDirectory, "Soda"))) {
 //     //copy the folder contents of home/Soda to home/SODA
-//     fs.copySync(path.join(homeDirectory, "Soda"), path.join(homeDirectory, "SODA"));
+//     fs.copySync(window.path.join(homeDirectory, "Soda"), window.path.join(homeDirectory, "SODA"));
 //     //delete the old folder
-//     fs.removeSync(path.join(homeDirectory, "Soda"));
+//     fs.removeSync(window.path.join(homeDirectory, "Soda"));
 //   }
 // }
 
@@ -1559,7 +1559,7 @@ dragselect_area.subscribe("dragstart", ({ items, event, isDragging }) => {
 // // generate subjects file
 // ipcRenderer.on("selected-generate-metadata-subjects", (event, dirpath, filename) => {
 //   if (dirpath.length > 0) {
-//     var destinationPath = path.join(dirpath[0], filename);
+//     var destinationPath = window.path.join(dirpath[0], filename);
 //     if (fs.existsSync(destinationPath)) {
 //       var emessage =
 //         "File '" + filename + "' already exists in " + dirpath[0] + ". Do you want to replace it?";
@@ -1768,7 +1768,7 @@ dragselect_area.subscribe("dragstart", ({ items, event, isDragging }) => {
 // // generate samples file
[…]
