+++ scripts/others/tab-effects.js
@@ -587,9 +587,9 @@ const nextPrev = (pageIndex) => {
 
     if (Object.keys(sodaJSONObj["metadata-files"]).length > 0) {
       Object.keys(sodaJSONObj["metadata-files"]).forEach((element) => {
-        let file_name = path.parse(element).name;
+        let file_name = window.path.parse(element).name;
         if (!element.includes("-DELETED")) {
-          withoutExtMetadataArray.push(path.parse(element).name);
+          withoutExtMetadataArray.push(window.path.parse(element).name);
         }
         if (requiredFiles.includes(file_name)) {
           let element_index = requiredFiles.indexOf(file_name);
@@ -1241,7 +1241,7 @@ const create_json_object = (action, sodaJSONObj, root_folder_path) => {
   let stats = "";
   // Get high level folders and metadata files first
   fs.readdirSync(root_folder_path).forEach((file) => {
-    full_current_path = path.join(root_folder_path, file);
+    full_current_path = window.path.join(root_folder_path, file);
     stats = fs.statSync(full_current_path);
     if (stats.isDirectory()) {
       if (highLevelFolders.includes(file) && !/(^|\/)\.[^\/\.]/g.test(file)) {
@@ -1271,8 +1271,8 @@ const create_json_object = (action, sodaJSONObj, root_folder_path) => {
   for (folder in sodaJSONObj["dataset-structure"]["folders"]) {
     sodaJSONObj["starting-point"][folder] = {};
     sodaJSONObj["starting-point"][folder]["path"] = "";
-    temp_file_path_xlsx = path.join(root_folder_path, folder, "manifest.xlsx");
-    temp_file_path_csv = path.join(root_folder_path, folder, "manifest.csv");
+    temp_file_path_xlsx = window.path.join(root_folder_path, folder, "manifest.xlsx");
+    temp_file_path_csv = window.path.join(root_folder_path, folder, "manifest.csv");
     if (fs.existsSync(temp_file_path_xlsx)) {
       sodaJSONObj["starting-point"][folder]["path"] = temp_file_path_xlsx;
       sodaJSONObj["starting-point"][folder]["manifest"] = excelToJson({
@@ -1289,7 +1289,7 @@ const create_json_object = (action, sodaJSONObj, root_folder_path) => {
       action,
       sodaJSONObj["dataset-structure"]["folders"][folder],
       folder,
-      path.join(root_folder_path, folder)
+      window.path.join(root_folder_path, folder)
     );
   }
 };
@@ -1319,7 +1319,7 @@ const create_json_object_include_manifest = (action, sodaJSONObj, root_folder_pa
   let stats = "";
   // Get high level folders and metadata files first
   fs.readdirSync(root_folder_path).forEach((file) => {
-    full_current_path = path.join(root_folder_path, file);
+    full_current_path = window.path.join(root_folder_path, file);
     stats = fs.statSync(full_current_path);
     if (stats.isDirectory()) {
       if (highLevelFolders.includes(file) && !/(^|\/)\.[^\/\.]/g.test(file)) {
@@ -1348,8 +1348,8 @@ const create_json_object_include_manifest = (action, sodaJSONObj, root_folder_pa
   for (folder in sodaJSONObj["dataset-structure"]["folders"]) {
     sodaJSONObj["starting-point"][folder] = {};
     sodaJSONObj["starting-point"][folder]["path"] = "";
-    // temp_file_path_xlsx = path.join(root_folder_path, folder, "manifest.xlsx");
-    // temp_file_path_csv = path.join(root_folder_path, folder, "manifest.csv");
+    // temp_file_path_xlsx = window.path.join(root_folder_path, folder, "manifest.xlsx");
+    // temp_file_path_csv = window.path.join(root_folder_path, folder, "manifest.csv");
[…]
