### 2023-11-09 fed1a69
fix: 🐛 fixed pool table styling
