### 2023-11-09 01cb1ff
refactor: jstree css + img assets inlcuded
