### 2023-11-10 1734242
fix: 🐛 syntax fix
