### 2023-11-10 c20d301
style: fomantic ui styles added back
