+++ src/renderer/src/scripts/analytics/analytics-utils.js
@@ -1,3 +1,11 @@
+export const Destinations = {
+  LOCAL: "Local",
+  PENNSIEVE: "Pennsieve",
+  SAVED: "Saved",
+  NEW: "New",
+};
+
+
 // determine if we are working with a Local, Saved, or Pennsieve dataset in the current Curation flow
 const determineDatasetLocation = () => {
   let location = "";
