+++ src/renderer/src/scripts/others/renderer.js
@@ -45,7 +45,7 @@ import {
   logSelectedUpdateExistingDatasetOptions,
 } from "../analytics/curation-analytics"
 import createEventDataPrepareMetadata from "../analytics/prepare-metadata-analytics";
-import determineDatasetLocation from "../analytics/analytics-utils"
+import determineDatasetLocation, {Destinations} from "../analytics/analytics-utils"
 import {
   clientError,
   userErrorMessage,
@@ -2046,7 +2046,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //         logMetadataForAnalytics(
 //           "Error",
 //           MetadataAnalyticsPrefix.SUBJECTS,
-//           AnalyticsGranularity.ALL_LEVELS,
+//           window.AnalyticsGranularity.ALL_LEVELS,
 //           "Existing",
 //           Destinations.LOCAL
 //         );
@@ -2055,7 +2055,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //       logMetadataForAnalytics(
 //         "Success",
 //         MetadataAnalyticsPrefix.SUBJECTS,
-//         AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
+//         window.AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
 //         "Existing",
 //         Destinations.LOCAL
 //       );
@@ -2064,7 +2064,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //       logMetadataForAnalytics(
 //         "Error",
 //         MetadataAnalyticsPrefix.SUBJECTS,
-//         AnalyticsGranularity.ALL_LEVELS,
+//         window.AnalyticsGranularity.ALL_LEVELS,
 //         "Existing",
 //         Destinations.LOCAL
 //       );
@@ -2089,7 +2089,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //     logMetadataForAnalytics(
 //       "Error",
 //       MetadataAnalyticsPrefix.SUBJECTS,
-//       AnalyticsGranularity.ALL_LEVELS,
+//       window.AnalyticsGranularity.ALL_LEVELS,
 //       "Existing",
 //       Destinations.LOCAL
 //     );
@@ -2129,7 +2129,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //         logMetadataForAnalytics(
 //           "Error",
 //           MetadataAnalyticsPrefix.SAMPLES,
-//           AnalyticsGranularity.ALL_LEVELS,
+//           window.AnalyticsGranularity.ALL_LEVELS,
 //           "Existing",
 //           Destinations.LOCAL
 //         );
@@ -2139,7 +2139,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //       logMetadataForAnalytics(
 //         "Success",
 //         MetadataAnalyticsPrefix.SAMPLES,
-//         AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
+//         window.AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
 //         "Existing",
 //         Destinations.LOCAL
 //       );
@@ -2149,7 +2149,7 @@ window.progressFilePath = window.path.join(homeDirectory, "SODA", "Progress");
 //       logMetadataForAnalytics(
 //         "Error",
 //         MetadataAnalyticsPrefix.SAMPLES,
-//         AnalyticsGranularity.ALL_LEVELS,
[…]
