### 2023-11-10 fairdataihub/SODA-for-SPARC#273 @bvhpatel
Addressed that a few days ago but forgot to close the issue: https://github.com/fairdataihub/SODA-for-SPARC/commit/52e7b8c520d90fdf71bb1ec5e59944fdcbbb0946. Thanks!
