### 2023-11-10 84ac02e
refactor: manifest editing in GM
