+++ src/main/main.js
@@ -472,6 +472,7 @@ ipcMain.on("restart_app", async () => {
 // passing in the spreadsheet data to pass to a modal
 // that will have a jspreadsheet for user edits
 ipcMain.handle("spreadsheet", (event, spreadsheet) => {
+  console.log("Spreadsheet invoked")
   const windowOptions = {
     minHeight: 450,
     width: 1120,
