+++ src/renderer/src/scripts/metadata-files/manifest.js
@@ -1,6 +1,10 @@
-const { platform } = require("os");
-const { copyFile, readdir } = require("fs").promises;
-const { join } = require("path");
+import client from '../client'
+import excelToJson from 'convert-excel-to-json';
+
+while (!window.htmlPagesAdded) {
+  await new Promise((resolve) => setTimeout(resolve, 100))
+}
+
 let openedEdit = false;
 
 // opendropdown event listeners
@@ -23,11 +27,11 @@ const guidedJsTreePreviewManifest = document.getElementById(
 );
 
 const showLocalDatasetManifest = () => {
-  ipcRenderer.send("open-file-dialog-local-dataset-manifest-purpose");
+  window.electron.ipcRenderer.send("open-file-dialog-local-dataset-manifest-purpose");
 };
 
 const selectManifestGenerationLocation = () => {
-  ipcRenderer.send("open-file-dialog-local-dataset-manifest-generate-purpose");
+  window.electron.ipcRenderer.send("open-file-dialog-local-dataset-manifest-generate-purpose");
 };
 
 const openDirectoryAtManifestGenerationLocation = (generationLocation) => {
@@ -58,7 +62,7 @@ const openFolder = (generationLocation) => {
 $(document).ready(function () {
   let localDataSetImport = false;
   let fileOpenedOnce = {};
-  ipcRenderer.on("selected-local-dataset-manifest-purpose", (event, folderPath) => {
+  window.electron.ipcRenderer.on("selected-local-dataset-manifest-purpose", (event, folderPath) => {
     if (folderPath.length > 0) {
       if (folderPath !== null) {
         localDataSetImport = true;
@@ -81,7 +85,7 @@ $(document).ready(function () {
     }
   });
 
-  ipcRenderer.on("selected-local-dataset-manifest-generate-purpose", (event, folderPath) => {
+  window.electron.ipcRenderer.on("selected-local-dataset-manifest-generate-purpose", (event, folderPath) => {
     if (folderPath.length <= 0 || folderPath === null) {
       document.getElementById("input-manifest-local-gen-location").placeholder = "Browse here";
       return;
@@ -247,6 +251,7 @@ $(document).ready(function () {
 
   //Event listener for the jstree to open manifest files
   $(jstreePreviewManifest).on("select_node.jstree", async function (evt, data) {
+    console.log("Doing this?")
     // Check if pennsieve option was selected to reset localDataSetImport
     if (document.getElementById("pennsieve-option-create-manifest").classList.contains("checked")) {
       localDataSetImport = false;
@@ -286,9 +291,9 @@ $(document).ready(function () {
       }
 
       // TODO: Check if this is needed everytime
-      var localFolderPath = path.join(homeDirectory, "SODA", "manifest_files", parentFolderName);
-      var selectedManifestFilePath = path.join(localFolderPath, "manifest.xlsx");
-      if (fs.existsSync(selectedManifestFilePath)) {
+      var localFolderPath = window.path.join(homeDirectory, "SODA", "manifest_files", parentFolderName);
+      var selectedManifestFilePath = window.path.join(localFolderPath, "manifest.xlsx");
+      if (window.fs.existsSync(selectedManifestFilePath)) {
         jsonManifest = excelToJson({
           sourceFile: selectedManifestFilePath,
           columnToKey: {
@@ -353,18 +358,18 @@ $(document).ready(function () {
             };
             // Will create an excel sheet of the manifest files in case they receive no edits
             let jsonManifest = {};
-            let manifestFolder = path.join(homeDirectory, "SODA", "manifest_files");
-            let localFolderPath = path.join(manifestFolder, highLevelFolderName);
-            let selectedManifestFilePath = path.join(localFolderPath, "manifest.xlsx");
[…]

[…]
