### 2023-11-10 9987be0
feat: ✨ Added spreadsheet header validation
