+++ src/renderer/src/main.js
@@ -44,6 +44,7 @@ import './assets/css/variables.css'
 import 'notyf/notyf.min.css'; // for React, Vue and Svelte
 import 'jstree/dist/themes/default/style.css';
 import 'fomantic-ui/dist/semantic.min.css';
+import '@yaireo/tagify/dist/tagify.css'
 
 
 
