+++ src/renderer/src/scripts/others/contributor-storage.js
@@ -26,7 +26,7 @@ window.loadStoredContributors = () => {
 
 // Add a new contributor to the JSON file
 // If a contributor with the same ORCiD already exists, update the existing contributor
-const addOrUpdateStoredContributor = (
+window.addOrUpdateStoredContributor = (
   firstName,
   lastName,
   ORCiD,
