+++ tools/anaconda-env/environment-MAC.yml
@@ -1,4 +1,4 @@
-name: bundle-up-too
+name: bundle-up-tooo
 channels:
   - defaults
   - conda-forge
