### 2023-11-12 091cd3a
feat: delete contributors GM
