### 2023-11-12 c9b7a4d
wip: subjects GM
