+++ src/renderer/src/scripts/metadata-files/subjects-samples.js
@@ -1,3 +1,7 @@
+while (!window.htmlPagesAdded) {
+  await new Promise((resolve) => setTimeout(resolve, 100))
+}
+
 // event listeners for open dropdown prompt
 document.querySelectorAll(".subjects-change-current-account").forEach((element) => {
   element.addEventListener("click", function () {
@@ -24,7 +28,7 @@ document.querySelectorAll(".samples-change-current-ds").forEach((element) => {
 });
 
 var subjectsFormDiv = document.getElementById("form-add-a-subject");
-var guidedSubjectsFormDiv = document.getElementById("guided-form-add-a-subject");
+window.guidedSubjectsFormDiv = document.getElementById("guided-form-add-a-subject");
 var samplesFormDiv = document.getElementById("form-add-a-sample");
 var guidedSamplesFormDiv = document.getElementById("guided-form-add-a-sample");
 var subjectsTableData = [];
@@ -37,7 +41,7 @@ let guidedSamplesTableData = [];
 
 const showForm = (type, editBoolean) => {
   if (type !== "edit") {
-    clearAllSubjectFormFields(subjectsFormDiv);
+    window.clearAllSubjectFormFields(subjectsFormDiv);
   }
   subjectsFormDiv.style.display = "flex";
   $("#create_subjects-tab").removeClass("show");
@@ -49,7 +53,7 @@ const showForm = (type, editBoolean) => {
 
 const showFormSamples = (type, editBoolean) => {
   if (type !== "edit") {
-    clearAllSubjectFormFields(samplesFormDiv);
+    window.clearAllSubjectFormFields(samplesFormDiv);
   }
   samplesFormDiv.style.display = "flex";
   $("#create_samples-tab").removeClass("show");
@@ -172,7 +176,7 @@ const promptImportPrevInfoSubject = (arr1) => {
     if (result.isConfirmed) {
       if ($("#previous-subject-single").val() !== "Select") {
         prevSubIDSingle = $("#previous-subject-single").val();
-        populateForms(prevSubIDSingle, "import", "free-form");
+        window.populateForms(prevSubIDSingle, "import", "free-form");
       }
     } else {
       hideForm("subject");
@@ -211,7 +215,7 @@ const confirmSample = () => {
 };
 
 // for "Done adding" button - subjects
-const addSubject = (curationMode) => {
+window.addSubject = (curationMode) => {
   let subjectID = "";
   if (curationMode === "free-form") {
     subjectID = $("#bootbox-subject-id").val().trim();
@@ -504,7 +508,7 @@ const addSampleIDtoDataBase = (samID, subID) => {
   }
 };
 
-const clearAllSubjectFormFields = (form) => {
+window.clearAllSubjectFormFields = (form) => {
   for (var field of $(form).children().find("input")) {
     $(field).val("");
   }
@@ -515,12 +519,12 @@ const clearAllSubjectFormFields = (form) => {
   $(form).find(".content").removeClass("active");
 
   // hide Strains and Species
-  if (form === subjectsFormDiv || form === guidedSubjectsFormDiv) {
+  if (form === subjectsFormDiv || form === window.guidedSubjectsFormDiv) {
     let curationModeSelectorPrefix = "";
     if (form === subjectsFormDiv) {
       curationModeSelectorPrefix = "";
     }
-    if (form === guidedSubjectsFormDiv) {
+    if (form === window.guidedSubjectsFormDiv) {
       curationModeSelectorPrefix = "guided-";
     }
 
@@ -528,7 +532,7 @@ const clearAllSubjectFormFields = (form) => {
     $(`#${curationModeSelectorPrefix}bootbox-${keyword}-species`).css("display", "none");
     $(`#${curationModeSelectorPrefix}bootbox-${keyword}-strain`).css("display", "none");
 
[…]
