+++ scripts/guided-mode/guided-curate-dataset.js
@@ -1313,7 +1313,7 @@ const savePageChanges = async (pageBeingLeftID) => {
 
     if (pageBeingLeftID === "guided-create-subjects-metadata-tab") {
       //Save the subject metadata from the subject currently being modified
-      addSubject("guided");
+      window.addSubject("guided");
 
       const subjectsAsideItemsCount = document.querySelectorAll(
         ".subjects-metadata-aside-item"
@@ -2232,7 +2232,7 @@ const extractPoolSubSamStructureFromDataset = (datasetStructure) => {
     for (const subjectFolder of subjectFoldersInBase) {
       if (!addedSubjects.includes(subjectFolder)) {
         try {
-          sodaJSONObj.addSubject(subjectFolder);
+          sodaJSONObj.window.addSubject(subjectFolder);
           addedSubjects.push(subjectFolder);
         } catch (error) {
           console.log(error);
@@ -2290,7 +2290,7 @@ const extractPoolSubSamStructureFromDataset = (datasetStructure) => {
       for (const subjectFolder of subjectFoldersInPool) {
         if (!addedSubjects.includes(subjectFolder)) {
           try {
-            sodaJSONObj.addSubject(subjectFolder);
+            sodaJSONObj.window.addSubject(subjectFolder);
             addedSubjects.push(subjectFolder);
           } catch (error) {
             console.log(error);
@@ -7730,7 +7730,7 @@ const attachGuidedMethodsToSodaJSONObj = () => {
 
     this["dataset-metadata"]["pool-subject-sample-structure"]["pools"][poolName] = {};
   };
-  sodaJSONObj.addSubject = function (subjectName) {
+  sodaJSONObj.window.addSubject = function (subjectName) {
     //check if subject with the same name already exists
     if (
       this["dataset-metadata"]["pool-subject-sample-structure"]["pools"][subjectName] ||
@@ -10060,7 +10060,7 @@ const guidedLoadSubjectMetadataIfExists = (subjectMetadataId) => {
     //check through elements of tableData to find a subject ID match
     if (subjectsTableData[i][0] === subjectMetadataId) {
       //if the id matches, load the metadata into the form
-      populateForms(subjectMetadataId, "", "guided");
+      window.populateForms(subjectMetadataId, "", "guided");
       return;
     }
   }
@@ -10131,7 +10131,7 @@ const openModifySampleMetadataPage = (sampleMetadataID, samplesSubjectID) => {
 
 const openCopySubjectMetadataPopup = async () => {
   //save current subject metadata entered in the form
-  addSubject("guided");
+  window.addSubject("guided");
 
   let copyFromMetadata = ``;
   let copyToMetadata = ``;
@@ -10214,7 +10214,7 @@ const openCopySubjectMetadataPopup = async () => {
         const currentSubjectOpenInView = document.getElementById("guided-bootbox-subject-id").value;
         if (currentSubjectOpenInView) {
           //If a subject was open in the UI, update it with the new metadata
-          populateForms(currentSubjectOpenInView, "", "guided");
+          window.populateForms(currentSubjectOpenInView, "", "guided");
         }
 
         await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
@@ -10362,7 +10362,7 @@ const specifySubject = (event, subjectNameInput) => {
         sodaJSONObj.renameSubject(subjectToRename, subjectName);
       } else {
         //case where subject name is valid and not being renamed:
-        sodaJSONObj.addSubject(subjectName);
+        sodaJSONObj.window.addSubject(subjectName);
       }
       subjectIdCellToAddNameTo.html(subjectNameElement);
       addSubjectSpecificationTableRow();
[…]
