+++ src/renderer/src/main.js
@@ -12,6 +12,7 @@ import './scripts/organize-dataset/curate-functions'
 import './scripts/metadata-files/manifest'
 import './scripts/spreadSheetTools/spreadSheetTools'
 import './scripts/others/contributor-storage'
+import './scripts/metadata-files/subjects-samples'
 
 // Application Lotties
 import './assets/lotties/activate-lotties'

[…]
