+++ scripts/metadata-files/subjects-samples.js
@@ -27,7 +27,7 @@ var subjectsFormDiv = document.getElementById("form-add-a-subject");
 var guidedSubjectsFormDiv = document.getElementById("guided-form-add-a-subject");
 var samplesFormDiv = document.getElementById("form-add-a-sample");
 var guidedSamplesFormDiv = document.getElementById("guided-form-add-a-sample");
-var subjectsTableData = [];
+var window.subjectsTableData = [];
 var subjectsFileData = [];
 var samplesTableData = [];
 var samplesFileData = [];
@@ -239,10 +239,10 @@ const window.addSubject = (curationMode) => {
     }
 
     addSubjectIDtoDataBase(subjectID);
-    if (subjectsTableData.length !== 0) {
+    if (window.subjectsTableData.length !== 0) {
       $("#div-import-primary-folder-subjects").hide();
     }
-    if (subjectsTableData.length === 2) {
+    if (window.subjectsTableData.length === 2) {
       onboardingMetadata("subject");
     }
   }
@@ -549,7 +549,7 @@ const addSubjectIDToJSON = (subjectID) => {
 };
 
 /// function to add Species - subjects + samples
-const addSpecies = async (ev, type, curationMode) => {
+window.addSpecies = async (ev, type, curationMode) => {
   let curationModeSelectorPrefix = "";
   if (curationMode == "guided") {
     curationModeSelectorPrefix = "guided-";
@@ -763,7 +763,7 @@ const populateRRID = (strain, type, curationMode) => {
 
 const addSubjectMetadataEntriesIntoJSON = (curationMode) => {
   let curationModeSelectorPrefix;
-  let dataLength = subjectsTableData.length;
+  let dataLength = window.subjectsTableData.length;
 
   if (curationMode === "free-form") {
     curationModeSelectorPrefix = "";
@@ -802,22 +802,22 @@ const addSubjectMetadataEntriesIntoJSON = (curationMode) => {
     }
     valuesArr.push(field.value);
   }
-  subjectsTableData[0] = headersArrSubjects;
+  window.subjectsTableData[0] = headersArrSubjects;
 
   if (valuesArr !== undefined && valuesArr.length !== 0) {
     if (curationMode === "free-form") {
-      if (subjectsTableData[dataLength] !== undefined) {
-        subjectsTableData[dataLength + 1] = valuesArr;
+      if (window.subjectsTableData[dataLength] !== undefined) {
+        window.subjectsTableData[dataLength + 1] = valuesArr;
       } else {
-        subjectsTableData[dataLength] = valuesArr;
+        window.subjectsTableData[dataLength] = valuesArr;
       }
     }
     if (curationMode === "guided") {
       let subjectID = document.getElementById("guided-bootbox-subject-id").value;
       //Overwrite existing subject data with new subject data
-      for (let i = 1; i < subjectsTableData.length; i++) {
-        if (subjectsTableData[i][0] === subjectID) {
-          subjectsTableData[i] = valuesArr;
+      for (let i = 1; i < window.subjectsTableData.length; i++) {
+        if (window.subjectsTableData[i][0] === subjectID) {
+          window.subjectsTableData[i] = valuesArr;
         }
       }
     }
@@ -1092,10 +1092,10 @@ const window.populateForms = (subjectID, type, curationMode) => {
     fieldArr = $(guidedSubjectsFormDiv).children().find(".subjects-form-entry");
   }
 
-  if (subjectsTableData.length > 1) {
-    for (var i = 1; i < subjectsTableData.length; i++) {
-      if (subjectsTableData[i][0] === subjectID) {
-        infoJson = subjectsTableData[i];
+  if (window.subjectsTableData.length > 1) {
[…]
