### 2023-11-12 d66acc2
refactor: main subjects functionality GM
