+++ scripts/metadata-files/subjects-samples.js
@@ -26,7 +26,7 @@ document.querySelectorAll(".samples-change-current-ds").forEach((element) => {
 var subjectsFormDiv = document.getElementById("form-add-a-subject");
 var guidedSubjectsFormDiv = document.getElementById("guided-form-add-a-subject");
 var samplesFormDiv = document.getElementById("form-add-a-sample");
-var guidedSamplesFormDiv = document.getElementById("guided-form-add-a-sample");
+var window.guidedSamplesFormDiv = document.getElementById("guided-form-add-a-sample");
 var window.subjectsTableData = [];
 var subjectsFileData = [];
 var samplesTableData = [];
@@ -582,14 +582,14 @@ window.addSpecies = async (ev, type, curationMode) => {
   if (value) {
     if (value !== "") {
       $(`#${curationModeSelectorPrefix}bootbox-${type}-species`).val(value);
-      switchSpeciesStrainInput("species", "edit", curationMode);
+      window.switchSpeciesStrainInput("species", "edit", curationMode);
     }
   } else {
-    switchSpeciesStrainInput("species", "add", curationMode);
+    window.switchSpeciesStrainInput("species", "add", curationMode);
   }
 };
 
-const switchSpeciesStrainInput = (type, mode, curationMode) => {
+const window.switchSpeciesStrainInput = (type, mode, curationMode) => {
   let curationModeSelectorPrefix = "";
   if (curationMode == "guided") {
     curationModeSelectorPrefix = "guided-";
@@ -626,7 +626,7 @@ const guidedSetStrainRRID = (RRID) => {
   rridInput.value = RRID;
 };
 
-const addStrain = async (ev, type, curationMode) => {
+const window.addStrain = async (ev, type, curationMode) => {
   let curationModeSelectorPrefix = "";
   if (curationMode == "guided") {
     curationModeSelectorPrefix = "guided-";
@@ -649,7 +649,7 @@ const addStrain = async (ev, type, curationMode) => {
     },
     didOpen: () => {
       $(".swal2-confirm").attr("id", "btn-confirm-strain");
-      createStrain("sweetalert-" + type + "-strain", type, curationMode);
+      window.createStrain("sweetalert-" + type + "-strain", type, curationMode);
     },
     preConfirm: () => {
       if (document.getElementById("sweetalert-" + type + "-strain").value === "") {
@@ -661,15 +661,15 @@ const addStrain = async (ev, type, curationMode) => {
   if (value) {
     if (value !== "") {
       $(`#${curationModeSelectorPrefix}bootbox-${type}-strain`).val(value);
-      switchSpeciesStrainInput("strain", "edit", curationMode);
+      window.switchSpeciesStrainInput("strain", "edit", curationMode);
     }
   } else {
-    switchSpeciesStrainInput("strain", "add", curationMode);
+    window.switchSpeciesStrainInput("strain", "add", curationMode);
   }
 };
 
 // populate RRID
-const populateRRID = (strain, type, curationMode) => {
+const window.populateRRID = (strain, type, curationMode) => {
   let curationModeSelectorPrefix = "";
[…]
