+++ src/renderer/src/scripts/metadata-files/subjects-samples.js
@@ -1,3 +1,4 @@
+import https from "https";
 import Swal from "sweetalert2";
 
 while (!window.htmlPagesAdded) {
@@ -32,7 +33,7 @@ document.querySelectorAll(".samples-change-current-ds").forEach((element) => {
 var subjectsFormDiv = document.getElementById("form-add-a-subject");
 window.guidedSubjectsFormDiv = document.getElementById("guided-form-add-a-subject");
 var samplesFormDiv = document.getElementById("form-add-a-sample");
-var guidedSamplesFormDiv = document.getElementById("guided-form-add-a-sample");
+window.guidedSamplesFormDiv = document.getElementById("guided-form-add-a-sample");
 window.subjectsTableData = [];
 var subjectsFileData = [];
 var samplesTableData = [];
@@ -588,14 +589,14 @@ window.addSpecies = async (ev, type, curationMode) => {
   if (value) {
     if (value !== "") {
       $(`#${curationModeSelectorPrefix}bootbox-${type}-species`).val(value);
-      switchSpeciesStrainInput("species", "edit", curationMode);
+      window.switchSpeciesStrainInput("species", "edit", curationMode);
     }
   } else {
-    switchSpeciesStrainInput("species", "add", curationMode);
+    window.switchSpeciesStrainInput("species", "add", curationMode);
   }
 };
 
-const switchSpeciesStrainInput = (type, mode, curationMode) => {
+window.switchSpeciesStrainInput = (type, mode, curationMode) => {
   let curationModeSelectorPrefix = "";
   if (curationMode == "guided") {
     curationModeSelectorPrefix = "guided-";
@@ -632,7 +633,7 @@ const guidedSetStrainRRID = (RRID) => {
   rridInput.value = RRID;
 };
 
-const addStrain = async (ev, type, curationMode) => {
+window.addStrain = async (ev, type, curationMode) => {
   let curationModeSelectorPrefix = "";
   if (curationMode == "guided") {
     curationModeSelectorPrefix = "guided-";
@@ -655,7 +656,7 @@ const addStrain = async (ev, type, curationMode) => {
     },
     didOpen: () => {
       $(".swal2-confirm").attr("id", "btn-confirm-strain");
-      createStrain("sweetalert-" + type + "-strain", type, curationMode);
+      window.createStrain("sweetalert-" + type + "-strain", type, curationMode);
     },
     preConfirm: () => {
       if (document.getElementById("sweetalert-" + type + "-strain").value === "") {
@@ -667,15 +668,15 @@ const addStrain = async (ev, type, curationMode) => {
   if (value) {
     if (value !== "") {
       $(`#${curationModeSelectorPrefix}bootbox-${type}-strain`).val(value);
-      switchSpeciesStrainInput("strain", "edit", curationMode);
+      window.switchSpeciesStrainInput("strain", "edit", curationMode);
     }
   } else {
-    switchSpeciesStrainInput("strain", "add", curationMode);
+    window.switchSpeciesStrainInput("strain", "add", curationMode);
   }
 };
 
 // populate RRID
-const populateRRID = (strain, type, curationMode) => {
[…]
