+++ src/renderer/src/scripts/others/renderer.js
@@ -2340,154 +2340,154 @@ window.createSpeciesAutocomplete = (id, curationMode) => {
   });
 };
 
-// function createStrain(id, type, curationMode) {
-//   var autoCompleteJS4 = new autoComplete({
-//     selector: "#" + id,
-//     data: {
-//       src: ["Wistar", "Yucatan", "C57/B6J", "C57 BL/6J", "mixed background", "Sprague-Dawley"],
-//     },
-//     events: {
-//       input: {
-//         focus: () => {
-//           autoCompleteJS4.start();
-//         },
-//       },
-//     },
-//     resultItem: {
-//       element: (item, data) => {
-//         // Modify Results Item Style
-//         item.style = "display: flex; justify-content: space-between;";
-//         // Modify Results Item Content
-//         item.innerHTML = `
-//         <span style="text-overflow: ellipsis; white-space: nowrap; overflow: hidden;">
-//           ${data.match}
-//         </span>`;
-//       },
-//       highlight: true,
-//     },
-//     threshold: 0,
-//     resultsList: {
-//       element: (list, data) => {
-//         const info = document.createElement("div");
+window.createStrain = (id, type, curationMode) => {
+  var autoCompleteJS4 = new autoComplete({
+    selector: "#" + id,
+    data: {
+      src: ["Wistar", "Yucatan", "C57/B6J", "C57 BL/6J", "mixed background", "Sprague-Dawley"],
+    },
+    events: {
+      input: {
+        focus: () => {
+          autoCompleteJS4.start();
+        },
+      },
+    },
+    resultItem: {
+      element: (item, data) => {
+        // Modify Results Item Style
+        item.style = "display: flex; justify-content: space-between;";
+        // Modify Results Item Content
+        item.innerHTML = `
+        <span style="text-overflow: ellipsis; white-space: nowrap; overflow: hidden;">
+          ${data.match}
+        </span>`;
+      },
+      highlight: true,
+    },
+    threshold: 0,
+    resultsList: {
+      element: (list, data) => {
+        const info = document.createElement("div");
 
-//         if (data.results.length === 0) {
-//           info.setAttribute("class", "no_results_species");
-//           info.setAttribute(
-//             "onclick",
-//             "populateRRID('" + data.query + "', '" + type + "', '" + curationMode + "')"
-//           );
-//           info.innerHTML = `Click here to check <strong>"${data.query}"</strong>`;
-//         }
-//         list.prepend(info);
-//       },
-//       noResults: true,
-//       maxResults: 5,
-//       tabSelect: true,
-//     },
-//   });
+        if (data.results.length === 0) {
+          info.setAttribute("class", "no_results_species");
+          info.setAttribute(
+            "onclick",
+            "window.populateRRID('" + data.query + "', '" + type + "', '" + curationMode + "')"
+          );
+          info.innerHTML = `Click here to check <strong>"${data.query}"</strong>`;
+        }
+        list.prepend(info);
+      },
+      noResults: true,
+      maxResults: 5,
+      tabSelect: true,
+    },
+  });
 
-//   autoCompleteJS4.input.addEventListener("selection", function (event) {
-//     var feedback = event.detail;
-//     var selection = feedback.selection.value;
-//     document.querySelector("#" + id).value = selection;
-//     var strain = $("#sweetalert-" + type + "-strain").val();
-//     if (strain !== "") {
[…]

[…]
