### 2023-11-12 e4daf11
refactor: samples basics GM
