+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -774,7 +774,7 @@ const savePageChanges = async (pageBeingLeftID) => {
           const [samplesInPools, samplesOutsidePools] = window.sodaJSONObj.getAllSamplesFromSubjects();
           const samples = [...samplesInPools, ...samplesOutsidePools];
 
-          samplesTableData = [];
+          window.samplesTableData = [];
 
           if (samples.length > 0) {
             try {
@@ -793,8 +793,8 @@ const savePageChanges = async (pageBeingLeftID) => {
                   },
                 }
               );
-              // Set the samplesTableData as the samples metadata response
-              samplesTableData = samplesMetadataResponse.data.sample_file_rows;
+              // Set the window.samplesTableData as the samples metadata response
+              window.samplesTableData = samplesMetadataResponse.data.sample_file_rows;
             } catch (error) {
               const emessage = userErrorMessage(error);
               errorArray.push({
@@ -812,7 +812,7 @@ const savePageChanges = async (pageBeingLeftID) => {
           // dataset structure. If subject and sample metadata were not found, reset it and we'll add the metadata later
           const metadataSubSamStructure = createGuidedStructureFromSubSamMetadata(
             window.subjectsTableData.slice(1),
-            samplesTableData.slice(1)
+            window.samplesTableData.slice(1)
           );
 
           if (!objectsHaveSameKeys(metadataSubSamStructure, datasetSubSamStructure)) {
@@ -1364,12 +1364,12 @@ const savePageChanges = async (pageBeingLeftID) => {
 
     if (pageBeingLeftID === "guided-create-samples-metadata-tab") {
       //Save the sample metadata from the sample currently being modified
-      addSample("guided");
+      window.addSample("guided");
 
       const samplesAsideItemsCount = document.querySelectorAll(
         ".samples-metadata-aside-item"
       ).length;
-      const samplesInTableDataCount = samplesTableData.length - 1;
+      const samplesInTableDataCount = window.samplesTableData.length - 1;
       if (samplesAsideItemsCount !== samplesInTableDataCount) {
         let result = await Swal.fire({
           heightAuto: false,
@@ -2855,7 +2855,7 @@ const saveGuidedProgress = async (guidedProgressFileName) => {
   //window.datasetStructureJSONObj when progress resumed
   window.sodaJSONObj["saved-datset-structure-json-obj"] = window.datasetStructureJSONObj;
   window.sodaJSONObj["subjects-table-data"] = window.subjectsTableData;
-  window.sodaJSONObj["samples-table-data"] = samplesTableData;
+  window.sodaJSONObj["samples-table-data"] = window.samplesTableData;
 
   window.fs.writeFileSync(guidedFilePath, JSON.stringify(window.sodaJSONObj, null, 2));
 };
@@ -3785,7 +3785,7 @@ const guidedResetProgressVariables = () => {
   window.window.sodaJSONObj = {};
   window.datasetStructureJSONObj = {};
   window.window.subjectsTableData = [];
-  window.samplesTableData = [];
+  window.window.samplesTableData = [];
 };
 
 const guidedPrepareHomeScreen = async () => {
@@ -3859,7 +3859,7 @@ const guidedShowTreePreview = (new_dataset_name, targetElement) => {
   }
 
   //Add the Samples metadata file to the preview if at least one sample has been added
-  if (samplesTableData.length > 0) {
+  if (window.samplesTableData.length > 0) {
     dsJsonObjCopy["files"]["samples.xlsx"] = {
       action: ["new"],
       path: "",
@@ -7567,7 +7567,7 @@ window.guidedResumeProgress = async (datasetNameToResume) => {
 
[…]
