+++ src/renderer/src/scripts/others/renderer.js
@@ -29,7 +29,7 @@ import diskCheck from "check-disk-space";
 import validator from "validator";
 import doiRegex from "doi-regex";
 import lottie from "lottie-web";
-import {dragDrop} from '../../assets/lotties/lotties'
+import {dragDrop, successCheck} from '../../assets/lotties/lotties'
 import select2 from "select2"; // TODO: select2()
 select2()
 import autoComplete from "@tarekraafat/autocomplete.js/dist/autoComplete.min.js"
@@ -7932,121 +7932,121 @@ const manageDesc = (ev) => {
 //   return final_message;
 // }
 
-// var metadataIndividualFile = "";
-// var metadataAllowedExtensions = [];
-// var metadataParaElement = "";
-// var metadataCurationMode = "";
+var metadataIndividualFile = "";
+var metadataAllowedExtensions = [];
+var metadataParaElement = "";
+var metadataCurationMode = "";
 
-// function importMetadataFiles(ev, metadataFile, extensionList, paraEle, curationMode) {
-//   document.getElementById(paraEle).innerHTML = "";
-//   metadataIndividualFile = metadataFile;
-//   metadataAllowedExtensions = extensionList;
-//   metadataParaElement = paraEle;
-//   metadataCurationMode = curationMode;
-//   ipcRenderer.send("open-file-dialog-metadata-curate");
-// }
+window.importMetadataFiles = (ev, metadataFile, extensionList, paraEle, curationMode) => {
+  document.getElementById(paraEle).innerHTML = "";
+  metadataIndividualFile = metadataFile;
+  metadataAllowedExtensions = extensionList;
+  metadataParaElement = paraEle;
+  metadataCurationMode = curationMode;
+  window.electron.ipcRenderer.send("open-file-dialog-metadata-curate");
+}
 
-// function importPennsieveMetadataFiles(ev, metadataFile, extensionList, paraEle) {
-//   extensionList.forEach((file_type) => {
-//     file_name = metadataFile + file_type;
-//     if (
-//       file_name in sodaJSONObj["metadata-files"] &&
-//       sodaJSONObj["metadata-files"][file_name]["type"] != "bf"
-//     ) {
-//       delete sodaJSONObj["metadata-files"][file_name];
-//     }
-//     deleted_file_name = file_name + "-DELETED";
-//     if (
-//       deleted_file_name in sodaJSONObj["metadata-files"] &&
-//       sodaJSONObj["metadata-files"][deleted_file_name]["type"] === "bf"
-//     ) {
-//       // update Json object with the restored object
-//       let index = sodaJSONObj["metadata-files"][deleted_file_name]["action"].indexOf("deleted");
-//       sodaJSONObj["metadata-files"][deleted_file_name]["action"].splice(index, 1);
-//       let deleted_file_name_new_key = deleted_file_name.substring(
-//         0,
-//         deleted_file_name.lastIndexOf("-")
-//       );
-//       sodaJSONObj["metadata-files"][deleted_file_name_new_key] =
-//         sodaJSONObj["metadata-files"][deleted_file_name];
-//       delete sodaJSONObj["metadata-files"][deleted_file_name];
-//     }
-//   });
-//   populate_existing_metadata(sodaJSONObj);
-// }
+function importPennsieveMetadataFiles(ev, metadataFile, extensionList, paraEle) {
+  extensionList.forEach((file_type) => {
+    file_name = metadataFile + file_type;
+    if (
+      file_name in window.sodaJSONObj["metadata-files"] &&
+      window.sodaJSONObj["metadata-files"][file_name]["type"] != "bf"
+    ) {
+      delete window.sodaJSONObj["metadata-files"][file_name];
+    }
[…]
