+++ scripts/guided-mode/guided-curate-dataset.js
@@ -5917,8 +5917,8 @@ const openPage = async (targetPageID) => {
             document.getElementById("guided-div-img-container").style.display = "block";
 
             //set new cropper for imported image
-            myCropper.destroy();
-            myCropper = new Cropper(guidedBfViewImportedImage, guidedCropOptions);
+            window.myCropper.destroy();
+            window.myCropper = new Cropper(guidedBfViewImportedImage, guidedCropOptions);
 
             $("#guided-save-banner-image").css("visibility", "visible");
           }
@@ -12530,8 +12530,8 @@ document
 // function for importing a banner image if one already exists
 $("#guided-button-add-banner-image").click(async () => {
   $("#guided-banner-image-modal").modal("show");
-  myCropper.destroy();
-  myCropper = new Cropper(guidedBfViewImportedImage, guidedCropOptions);
+  window.myCropper.destroy();
+  window.myCropper = new Cropper(guidedBfViewImportedImage, guidedCropOptions);
 });
 
 // Action when user click on "Import image" button for banner image
@@ -14503,7 +14503,7 @@ const guidedSaveBannerImage = async () => {
   }
   let datasetName = sodaJSONObj["digital-metadata"]["name"];
   let imagePath = path.join(imageFolder, `${datasetName}.` + imageExtension);
-  let croppedImageDataURI = myCropper.getCroppedCanvas().toDataURL(imageType);
+  let croppedImageDataURI = window.myCropper.getCroppedCanvas().toDataURL(imageType);
 
   imageDataURI.outputFile(croppedImageDataURI, imagePath).then(async () => {
     let image_file_size = fs.statSync(imagePath)["size"];
