### 2023-11-13 fbfdd04
wip: 🚧 Added basic sub, pool, sam validation
