+++ scripts/guided-mode/guided-curate-dataset.js
@@ -2115,9 +2115,8 @@ const renderSideBar = (activePage) => {
         const pageWithErrorName = CURRENT_PAGE.dataset.pageName;
         const { value: continueWithoutSavingCurrPageChanges } = await Swal.fire({
           title: "The current page was not able to be saved",
-          html: `The following error${
-            error.length > 1 ? "s" : ""
-          } occurred when attempting to save the ${pageWithErrorName} page:
+          html: `The following error${error.length > 1 ? "s" : ""
+            } occurred when attempting to save the ${pageWithErrorName} page:
             <br />
             <br />
             <ul>
@@ -2323,7 +2322,7 @@ const extractPoolSubSamStructureFromDataset = (datasetStructure) => {
 
         const potentialSampleFolderNames = Object.keys(
           datasetStructure["folders"][hlf]["folders"][poolFolder]["folders"][subjectFolder][
-            "folders"
+          "folders"
           ]
         );
         const sampleFoldersInSubject = potentialSampleFolderNames.filter((folder) =>
@@ -2409,8 +2408,8 @@ const guidedUnLockSideBar = () => {
   guidedNav.style.display = "none";
 };
 
-guidedHideSidebar = () => {};
-guidedUnHideSidebar = () => {};
+guidedHideSidebar = () => { };
+guidedUnHideSidebar = () => { };
 
 // This function reads the innerText of the textSharedWithCurationTeamStatus element
 // and hides or shows the share and unshare buttons accordingly
@@ -2633,9 +2632,8 @@ const guidedSaveAndExit = async () => {
 
       const { value: continueWithoutSavingCurrPageChanges } = await Swal.fire({
         title: "The current page was not able to be saved before exiting",
-        html: `The following error${
-          error.length > 1 ? "s" : ""
-        } occurred when attempting to save the ${pageWithErrorName} page:
+        html: `The following error${error.length > 1 ? "s" : ""
+          } occurred when attempting to save the ${pageWithErrorName} page:
             <br />
             <br />
             <ul>
@@ -3067,11 +3065,10 @@ const generateProgressCardElement = (progressFileJSONObj) => {
             data-trigger="hover"
             style="font-weight: 400;"
           >
-              ${
-                progressFileSubtitle.length > 70
-                  ? `${progressFileSubtitle.substring(0, 70)}...`
-                  : progressFileSubtitle
-              }
+              ${progressFileSubtitle.length > 70
+      ? `${progressFileSubtitle.substring(0, 70)}...`
+      : progressFileSubtitle
+    }
           </h1>
         </div>
         <div class="dataset-card-row">
@@ -3085,22 +3082,21 @@ const generateProgressCardElement = (progressFileJSONObj) => {
             ></i>
           </h2>
           <h1 class="dataset-card-date-text">${progressFileLastModified}</h1>
-          ${
-            savedUploadDataProgress
-              ? `
+          ${savedUploadDataProgress
+      ? `
                 <span class="badge badge-warning mx-2">Incomplete upload</span>
               `
-              : ``
-          }
+      : ``
+    }
         </div>
       </div>
       <div class="dataset-card-button-container align-right">
         ${generateProgressResumptionButton(
-          datasetStartingPoint,
-          alreadyUploadedToPennsieve,
-          progressFileName,
-          workspaceUserNeedsToSwitchTo
-        )}
+      datasetStartingPoint,
+      alreadyUploadedToPennsieve,
+      progressFileName,
+      workspaceUserNeedsToSwitchTo
+    )}
         <h2 class="dataset-card-button-delete" onclick="deleteProgressCard(this)">
[…]
