### 2023-11-13 085e9af
wip: banner image uploads GM
