+++ src/renderer/src/main.js
@@ -49,6 +49,7 @@ import '@yaireo/tagify/dist/tagify.css'
 import "@tarekraafat/autocomplete.js/dist/css/autoComplete.css";
 import "@tarekraafat/autocomplete.js/dist/css/autoComplete.01.css";
 import "@tarekraafat/autocomplete.js/dist/css/autoComplete.02.css";
+import "cropperjs/src/css/cropper.css";
 
 
 
