+++ src/renderer/src/scripts/others/renderer.js
@@ -33,6 +33,7 @@ import {dragDrop, successCheck} from '../../assets/lotties/lotties'
 import select2 from "select2"; // TODO: select2()
 select2()
 import autoComplete from "@tarekraafat/autocomplete.js/dist/autoComplete.min.js"
+import Cropper from 'cropperjs'
 import DragSort from "@yaireo/dragsort";
 import axios from "axios";
 import Swal from "sweetalert2";
@@ -969,7 +970,7 @@ const apiVersionsMatch = async () => {
   if (hasConnectedAccountWithPennsieve()) {
 
     try {
-      updateBfAccountList();
+      window.updateBfAccountList();
     } catch (error) {
       clientError(error);
     }
@@ -1275,13 +1276,13 @@ window.datasetPermissionDiv = document.getElementById("div-permission-list-2");
 
 // const bfCurrentBannerImg = document.getElementById("current-banner-img");
 
-// const bfViewImportedImage = document.querySelector("#image-banner");
-// const guidedBfViewImportedImage = document.querySelector("#guided-image-banner");
+const bfViewImportedImage = document.querySelector("#image-banner");
+window.guidedBfViewImportedImage = document.querySelector("#guided-image-banner");
 
 // const bfSaveBannerImageBtn = document.getElementById("save-banner-image");
 // const datasetBannerImageStatus = document.querySelector("#para-dataset-banner-image-status");
 // const formBannerHeight = document.getElementById("form-banner-height");
-// const guidedFormBannerHeight = document.getElementById("guided-form-banner-height");
+window.guidedFormBannerHeight = document.getElementById("guided-form-banner-height");
 // const currentDatasetLicense = document.querySelector("#para-dataset-license-current");
 // const bfListLicense = document.querySelector("#bf-license-list");
 // const bfAddLicenseBtn = document.getElementById("button-add-license");
@@ -3032,80 +3033,81 @@ function populateDatasetDropdownCurate(datasetDropdown, datasetList) {
 // const { rename } = require("fs");
 // const { resolveSoa } = require("dns");
 // const internal = require("stream");
-// var cropOptions = {
-//   aspectRatio: 1,
-//   movable: false,
-//   // Enable to rotate the image
-//   rotatable: false,
-//   // Enable to scale the image
-//   scalable: false,
-//   // Enable to zoom the image
-//   zoomable: false,
-//   // Enable to zoom the image by dragging touch
-//   zoomOnTouch: false,
-//   // Enable to zoom the image by wheeling mouse
-//   zoomOnWheel: false,
-//   // preview: '.preview',
-//   viewMode: 1,
-//   responsive: true,
-//   crop: function (event) {
-//     var data = event.detail;
-//     let image_height = Math.round(data.height);
+var cropOptions = {
+  aspectRatio: 1,
+  movable: false,
+  // Enable to rotate the image
+  rotatable: false,
+  // Enable to scale the image
+  scalable: false,
+  // Enable to zoom the image
+  zoomable: false,
+  // Enable to zoom the image by dragging touch
+  zoomOnTouch: false,
+  // Enable to zoom the image by wheeling mouse
+  zoomOnWheel: false,
+  // preview: '.preview',
+  viewMode: 1,
+  responsive: true,
+  crop: function (event) {
+    var data = event.detail;
+    let image_height = Math.round(data.height);
 
-//     formBannerHeight.value = image_height;
-//     //if image-height exceeds 2048 then prompt about scaling image down
-//     if (image_height < 512) {
-//       $("#save-banner-image").prop("disabled", true);
-//       $("#form-banner-height").css("color", "red");
-//       $("#form-banner-height").css("border", "1px solid red");
[…]
