### 2023-11-13 04a8c56
feat: ✨ finished spreadsheet validation
