### 2023-11-14 397041c
fix: 🐛 remove duplicate validation logic
