+++ src/renderer/src/scripts/others/renderer.js
@@ -430,7 +430,7 @@ const initializeSODARenderer = async () => {
   log.info("Server is live and API versions match");
 
   // check integrity of all the core systems
-  await run_pre_flight_checks();
+  await window.run_pre_flight_checks();
 
   log.info("Pre flight checks finished");
 }
@@ -486,9 +486,9 @@ const agent_installed = async () => {
 };
 
 // Run a set of functions that will check all the core systems to verify that a user can upload datasets with no issues.
-const run_pre_flight_checks = async (check_update = true) => {
+window.run_pre_flight_checks = async (check_update = true) => {
   try {
-    log.info("Running pre flight checks");
+    window.log.info("Running pre flight checks");
     console.log("Initiating pre flight checks")
 
     if (!preFlightCheckNotyf) {
@@ -685,7 +685,7 @@ const run_pre_flight_checks = async (check_update = true) => {
       });
       // If the user clicks the retry button, rerun the pre flight checks
       if (rerunPreFlightChecks) {
-        return await run_pre_flight_checks();
+        return await window.run_pre_flight_checks();
       }
 
       // Dismiss the preflight check notification if it is still open
@@ -728,7 +728,7 @@ const run_pre_flight_checks = async (check_update = true) => {
       });
       // If the user clicks the retry button, rerun the pre flight checks
       if (rerunPreFlightChecks) {
-        return await run_pre_flight_checks();
+        return await window.run_pre_flight_checks();
       }
 
       // Dismiss the preflight check notification if it is still open
@@ -778,7 +778,7 @@ const run_pre_flight_checks = async (check_update = true) => {
       });
       // If the user clicks the retry button, rerun the pre flight checks
       if (rerunPreFlightChecks) {
-        return await run_pre_flight_checks();
+        return await window.run_pre_flight_checks();
       }
       // Dismiss the preflight check notification if it is still open
       if (preFlightCheckNotyf) {
@@ -855,7 +855,7 @@ const run_pre_flight_checks = async (check_update = true) => {
     });
     // If the user clicks retry, then run the preflight checks again
     if (retryChecks) {
-      return await run_pre_flight_checks();
+      return await window.run_pre_flight_checks();
     }
     // If the user clicks skip for now, then return false
     return false;
@@ -1614,7 +1614,7 @@ window.createDragSort(window.guidedSubmissionTagsTagifyManual);
 //   let bfdataset = document.getElementById("bf_dataset_load_subjects").innerText.trim();
 //   if (uploadBFBoolean) {
 //     // Run pre-flight checks before uploading the subjects file to Pennsieve
-//     let supplementary_checks = await run_pre_flight_checks(false);
+//     let supplementary_checks = await window.run_pre_flight_checks(false);
 //     if (!supplementary_checks) {
 //       return;
 //     }
@@ -1823,7 +1823,7 @@ window.createDragSort(window.guidedSubmissionTagsTagifyManual);
 //   let bfDataset = $("#bf_dataset_load_samples").text().trim();
 //   if (uploadBFBoolean) {
 //     // Run pre-flight checks before uploading the samples file to Pennsieve
-//     const supplementary_checks = await run_pre_flight_checks(false);
+//     const supplementary_checks = await window.run_pre_flight_checks(false);
 //     if (!supplementary_checks) {
 //       return;
 //     }
@@ -2967,7 +2967,7 @@ async function updateDatasetCurate(datasetDropdown, bfaccountDropdown) {
     window.datasetList = [];
     window.datasetList = responseObject.data.datasets;
     populateDatasetDropdownCurate(datasetDropdown, window.datasetList);
-    refreshDatasetList();
[…]
