### 2023-11-14 f23fd6d
feat: ✨ generalized subject addition swal for pools and sams
