+++ scripts/guided-mode/guided-curate-dataset.js
@@ -11317,65 +11317,83 @@ ipcRenderer.on("selected-sample-names-from-dialog", async (event, folders) => {
   console.log(sampleNames);
 });
 
-const guidedOpenSubjectAdditionSwal = async () => {
-  // Get a list of the existing subject names so we can check for duplicates
+const guidedOpenEntityAdditionSwal = async (entityName) => {
+  // Get a list of the existing entities so we can check for duplicates
   // const subjects = getExistingSubjectNames();
-  const preExistingSubjects = ["sub-1", "sub-2", "sub-3"];
+  let preExistingEntities;
+  let entityNameSingular;
+  let entityPrefix;
 
-  const subjects = [...preExistingSubjects, "sub-4", "sub-5", "sub-6"];
+  if (entityName === "subjects") {
+    preExistingEntities = getExistingSubjectNames();
+    entityNameSingular = "subject";
+    entityPrefix = "sub-";
+  }
+  if (entityName === "samples") {
+    preExistingEntities = getExistingSampleNames();
+    entityNameSingular = "sample";
+    entityPrefix = "sam-";
+  }
+  if (entityName === "pools") {
+    preExistingEntities = getExistingPoolNames();
+    entityNameSingular = "pool";
+    entityPrefix = "pool-";
+  }
 
-  const addSwalSubject = (subjectName) => {
-    if (subjectName.length < 1) {
-      throw new Error("Subject name cannot be empty");
+  const newEntities = [];
+
+  const handleSwalEntityAddition = (entityName) => {
+    if (entityName.length < 1) {
+      throw new Error(`Please enter a ${entityNameSingular} name`);
     }
     // Check to see if the subjectName starts with sub- otherwise prepend sub- to it
-    if (!subjectName.startsWith("sub-")) {
-      subjectName = `sub-${subjectName}`;
+    if (!entityName.startsWith(entityPrefix)) {
+      entityName = `${entityPrefix}${entityName}`;
     }
     // Check to see if the subjectName already exists
-    if (subjects.includes(subjectName)) {
-      throw new Error("Subject name already exists");
+    if (preExistingEntities.includes(entityName)) {
+      throw new Error(`${entityNameSingular} name has already been added`);
     }
 
-    if (preExistingSubjects.includes(subjectName)) {
-      throw new Error("Subject name already exists");
+    if (newEntities.includes(entityName)) {
+      throw new Error(`${entityNameSingular} name has already been added`);
     }
 
     // Hide any validation messages that may exist in the sweet alert
     Swal.resetValidationMessage();
 
     // Add the subject to the beginning of the subjects array
-    subjects.unshift(subjectName);
+    newEntities.unshift(entityName);
     // Re-render the subjects in the Swal
-    renderSubjectsInSwal();
+    renderEntitiesInSwal();
   };
 
-  const deleteSwalSubject = (subjectName) => {
-    console.log("deleteSwalSubject", subjectName);
+  const deleteSwalEntity = (entityName) => {
+    console.log("deleteSwalEntity", entityName);
     // Remove subject from subjects array
-    const index = subjects.indexOf(subjectName);
+    const index = newEntities.indexOf(entityName);
     if (index > -1) {
-      subjects.splice(index, 1);
+      newEntities.splice(index, 1);
     }
-    resetValidationMessage();
+    Swal.resetValidationMessage();
   };
 
-  const renderSubjectsInSwal = () => {
-    const subjectsList = document.getElementById("subjects-list");
-    subjectsList.innerHTML = subjects
+  const renderEntitiesInSwal = () => {
+    const entitiesList = document.getElementById("entities-liest");
+    entitiesList.innerHTML = newEntities
       .map(
-        (subject) => `
[…]
