+++ src/renderer/src/scripts/others/renderer.js
@@ -1382,7 +1382,7 @@ dragselect_area.subscribe("dragstart", ({ items, event, isDragging }) => {
 // let milestonePath = window.path.join(metadataPath, milestoneFileName);
 window.progressFilePath = window.path.join(window.homeDirectory, "SODA", "Progress");
 // let guidedProgressFilePath = window.path.join(window.homeDirectory, "SODA", "Guided-Progress");
-// const guidedManifestFilePath = window.path.join(window.homeDirectory, "SODA", "guided_manifest_files");
+window.guidedManifestFilePath = window.path.join(window.homeDirectory, "SODA", "guided_manifest_files");
 // let protocolConfigPath = window.path.join(metadataPath, protocolConfigFileName);
 // let allCollectionTags = {};
 // let currentTags = {};
