### 2023-11-15 1de5ab1
fix: linux get agent version fix
