### 2023-11-15 4a33c79
fix: synchronicity issue with variable instantiation and availability
