### 2023-11-15 7fccfc1
fix: get path to the application for template setup
