+++ src/main/index.js
@@ -28,7 +28,8 @@ console.log("Test up[date")
 
 // TODO: move to a separate file that handles all the ipcMain handlers
 ipcMain.handle('get-app-path', async (event, arg) => {
-  return app.getPath(arg) 
+  if(arg) return app.getPath(arg) 
+  return app.getAppPath()
 })
 
 ipcMain.handle("get-port", () => {
