### 2023-11-15 1fbb5c6
fix: template paths for metadata
