### 2023-11-15 5dbb30a
refactor: move to disseminate page GM
