### 2023-11-15 936588e
feat: ✨ Safely lowercase spreadsheet entities
