+++ scripts/guided-mode/guided-curate-dataset.js
@@ -49,6 +49,7 @@ const handleGuidedModeOrgSwitch = async (buttonClicked) => {
 };
 
 const guidedGetCurrentUserWorkSpace = () => {
+  return null;
   // Returns null if the user has not selected a workspace
   const workSpaceFromUI = document.getElementById(
     "guided-pennsive-selected-organization"
@@ -60,6 +61,13 @@ const guidedGetCurrentUserWorkSpace = () => {
     return null;
   }
   return workSpaceFromUI;
+};
+
+const lowercaseFirstLetter = (string) => {
+  if (!string) {
+    return string;
+  }
+  return string.charAt(0).toLowerCase() + string.slice(1);
 };
 
 const lottieAnimationManager = {
@@ -1057,6 +1065,25 @@ const savePageChanges = async (pageBeingLeftID) => {
         }
         sodaJSONObj["digital-metadata"]["name"] = datasetNameInput;
         sodaJSONObj["digital-metadata"]["subtitle"] = datasetSubtitleInput;
+      }
+    }
+
+    if (pageBeingLeftID === "guided-subjects-specification-tab") {
+      const subjectsArePooled = document
+        .getElementById("guided-button-subjects-are-pooled")
+        .classList.contains("selected");
+      if (subjectsArePooled) {
+        guidedUnSkipPage("guided-subjects-pooling-tab");
+      } else {
+        guidedSkipPage("guided-subjects-pooling-tab");
+      }
+      const subjectsHaveSamples = document
+        .getElementById("guided-button-subjects-have-samples")
+        .classList.contains("selected");
+      if (subjectsHaveSamples) {
+        guidedUnSkipPage("guided-samples-addition-tab");
+      } else {
+        guidedSkipPage("guided-samples-addition-tab");
       }
     }
 
@@ -6741,17 +6768,6 @@ const renderSubjectsTable = () => {
   document.getElementById("subject-specification-table-body").innerHTML = subjectElementRows;
 };
 
-document.getElementById("guided-button-add-a-subject").addEventListener("click", async () => {
-  const subjectName = await swalGetUserTextInput(
-    "Enter a name for the subject",
-    "Subject name",
-    "Subject name"
-  );
-  if (subjectName) {
-    console.log("subjectName", subjectName);
-  }
-});
-
 const renderPoolsTable = () => {
   const pools = sodaJSONObj.getPools();
   const poolElementRows = Object.keys(pools)
@@ -10993,13 +11009,6 @@ ipcRenderer.on("selected-create-dataset-structure-spreadsheet-path", async (even
       worksheet.column(i + 1).setWidth(30);
     }
 
-    // write the subjects to the spreadsheet
-    const subjects = ["sub-1", "sub-2", "sub-3"];
-    for (i = 0; i < subjects.length; i++) {
-      worksheet.cell(i + 2, 1).string(subjects[i]);
-    }
-    console.log("subjects", subjects);
-
     // write the spreadsheet to the selected
     const filePath = path + "/dataset_structure.xlsx";
     const buffer = await workbook.writeToBuffer();
@@ -11009,8 +11018,6 @@ ipcRenderer.on("selected-create-dataset-structure-spreadsheet-path", async (even
       "Template successfully generated",
       `
         Would you like to open the template now?
-        <br />
-        Template Path: ${filePath}
       `,
       "Yes",
       "No"
@@ -11033,7 +11040,7 @@ const validateDatasetStructureSpreadsheet = async (sheetData) => {
   const pooledSubjects = new Map();
   // 1. Loop through the spreadsheet rows and find subs, pools, and sams that do not have valid names
   for (const row of sheetData) {
-    const subjectName = row["Subject ID"];
+    const subjectName = lowercaseFirstLetter(row["Subject ID"]);
     if (!subjectName) {
[…]
