### 2023-11-15 d038247
feat: reserve doi and publish thru GM
