+++ scripts/others/renderer.js
@@ -2311,7 +2311,7 @@ async function loadSubjectsFileToDataframe(filePath) {
         logMetadataForAnalytics(
           "Error",
           MetadataAnalyticsPrefix.SUBJECTS,
-          AnalyticsGranularity.ALL_LEVELS,
+          window.AnalyticsGranularity.ALL_LEVELS,
           "Existing",
           Destinations.LOCAL
         );
@@ -2320,7 +2320,7 @@ async function loadSubjectsFileToDataframe(filePath) {
       logMetadataForAnalytics(
         "Success",
         MetadataAnalyticsPrefix.SUBJECTS,
-        AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
+        window.AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
         "Existing",
         Destinations.LOCAL
       );
@@ -2329,7 +2329,7 @@ async function loadSubjectsFileToDataframe(filePath) {
       logMetadataForAnalytics(
         "Error",
         MetadataAnalyticsPrefix.SUBJECTS,
-        AnalyticsGranularity.ALL_LEVELS,
+        window.AnalyticsGranularity.ALL_LEVELS,
         "Existing",
         Destinations.LOCAL
       );
@@ -2354,7 +2354,7 @@ async function loadSubjectsFileToDataframe(filePath) {
     logMetadataForAnalytics(
       "Error",
       MetadataAnalyticsPrefix.SUBJECTS,
-      AnalyticsGranularity.ALL_LEVELS,
+      window.AnalyticsGranularity.ALL_LEVELS,
       "Existing",
       Destinations.LOCAL
     );
@@ -2394,7 +2394,7 @@ async function loadSamplesFileToDataframe(filePath) {
         logMetadataForAnalytics(
           "Error",
           MetadataAnalyticsPrefix.SAMPLES,
-          AnalyticsGranularity.ALL_LEVELS,
+          window.AnalyticsGranularity.ALL_LEVELS,
           "Existing",
           Destinations.LOCAL
         );
@@ -2404,7 +2404,7 @@ async function loadSamplesFileToDataframe(filePath) {
       logMetadataForAnalytics(
         "Success",
         MetadataAnalyticsPrefix.SAMPLES,
-        AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
+        window.AnalyticsGranularity.ACTION_AND_ACTION_WITH_DESTINATION,
         "Existing",
         Destinations.LOCAL
       );
@@ -2414,7 +2414,7 @@ async function loadSamplesFileToDataframe(filePath) {
       logMetadataForAnalytics(
         "Error",
         MetadataAnalyticsPrefix.SAMPLES,
-        AnalyticsGranularity.ALL_LEVELS,
+        window.AnalyticsGranularity.ALL_LEVELS,
         "Existing",
         Destinations.LOCAL
       );
@@ -2440,7 +2440,7 @@ async function loadSamplesFileToDataframe(filePath) {
     logMetadataForAnalytics(
       "Error",
       MetadataAnalyticsPrefix.SAMPLES,
-      AnalyticsGranularity.ALL_LEVELS,
+      window.AnalyticsGranularity.ALL_LEVELS,
       "Existing",
       Destinations.LOCAL
     );
@@ -3421,7 +3421,7 @@ const setupPublicationOptionsPopover = () => {
   });
 };
 
-const submitReviewDatasetCheck = async (res, curationMode) => {
+const window.submitReviewDatasetCheck = async (res, curationMode) => {
   let reviewstatus = res["review_request_status"];
   let publishingStatus = res["publishing_status"];
[…]
