### 2023-11-15 dd1cc08
refactor: delete progress files
