### 2023-11-16 7ad2c96
wip: 🚧 cleaned up entity addition swal
