+++ src/main/index.js
@@ -10,6 +10,7 @@ import { existsSync } from 'fs'
 import { JSONStorage } from "node-localstorage";
 import log from 'electron-log/main';
 import  "./manifest-workbook"
+import "./banner-image"
 
 import "./main-process/native-ui/dialogs/open-file"
 
