### 2023-11-16 c882a67
fix: import banner image from PS
