### 2023-11-16 077b9d3
fix: continue with no subs & samples GM
