+++ scripts/guided-mode/guided-curate-dataset.js
@@ -14070,16 +14070,16 @@ const guidedUploadDatasetToPennsieve = async () => {
         setGuidedProgressBarValue(percentOfDatasetUploaded);
 
         let totalSizePrint;
-        if (main_total_generate_dataset_size < displaySize) {
+        if (main_total_generate_dataset_size < window.displaySIze) {
           totalSizePrint = main_total_generate_dataset_size.toFixed(2) + " B";
-        } else if (main_total_generate_dataset_size < displaySize * displaySize) {
-          totalSizePrint = (main_total_generate_dataset_size / displaySize).toFixed(2) + " KB";
-        } else if (main_total_generate_dataset_size < displaySize * displaySize * displaySize) {
+        } else if (main_total_generate_dataset_size < window.displaySIze * window.displaySIze) {
+          totalSizePrint = (main_total_generate_dataset_size / window.displaySIze).toFixed(2) + " KB";
+        } else if (main_total_generate_dataset_size < window.displaySIze * window.displaySIze * window.displaySIze) {
           totalSizePrint =
-            (main_total_generate_dataset_size / displaySize / displaySize).toFixed(2) + " MB";
+            (main_total_generate_dataset_size / window.displaySIze / window.displaySIze).toFixed(2) + " MB";
         } else {
           totalSizePrint =
-            (main_total_generate_dataset_size / displaySize / displaySize / displaySize).toFixed(
+            (main_total_generate_dataset_size / window.displaySIze / window.displaySIze / window.displaySIze).toFixed(
               2
             ) + " GB";
         }
