### 2023-11-16 f5f2d58
fix: manifest ivokation no longer uses the event argument in a function
