### 2023-11-16 476c683
fix: banner image preview
