### 2023-11-16 8486fa2
fix: GM Pools not addable
