### 2023-11-17 309bb8f
feat: ✨ Map all structure texts to their spans
