### 2023-11-17 b8a5ab6
build: update environment files for Mac
