+++ tools/anaconda-env/environment-MAC.yml
@@ -11,7 +11,7 @@ dependencies:
 - pip:
   - chardet == 4.0.0
   - pandas == 1.4.2
-  - openpyxl == 3.0.9
+  - openpyxl == 3.0.10
   - configparser == 4.0.2
   - python-docx == 0.8.10
   - xlrd == 2.0.1
