+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -2986,7 +2986,7 @@ const generateProgressCardElement = (progressFileJSONObj) => {
           <button
             class="ui positive button guided--progress-button-login-to-pennsieve"
             data-reset-guided-mode-page="true"
-            onclick="openDropdownPrompt(this, 'bf')"
+            onclick="window.openDropdownPrompt(this, 'bf')"
           >
             Log in to Pennsieve to resume curation
           </button>
@@ -2997,7 +2997,7 @@ const generateProgressCardElement = (progressFileJSONObj) => {
       return `
         <button
           class="ui positive button guided-change-workspace guided--progress-button-switch-workspace"
-          onClick="openDropdownPrompt(this, 'organization')"
+          onClick="window.openDropdownPrompt(this, 'organization')"
         >
           Switch to ${workspaceUserNeedsToSwitchTo} workspace to resume curation
         </button>
@@ -10811,7 +10811,7 @@ window.confirmEnter = (button) => {
     //for adding a new sample row
     let clickSampleButton = true;
     for (let i = 0; i < sampleTableContainers.length; i++) {
-      sampleEntries = sampleTableContainers[i].children[1];
+      let sampleEntries = sampleTableContainers[i].children[1];
       if (sampleEntries.children.length > 0) {
         //entries have been create so look at the last one if an input is there
         let lastEntryCount = sampleEntries.children.length - 1;
