+++ src/renderer/src/scripts/metadata-files/readme-changes.js
@@ -1,26 +1,26 @@
 // event listeners for changes open dropdown prompts
 document.querySelectorAll(".changes-change-current-account").forEach((element) => {
   element.addEventListener("click", function () {
-    openDropdownPrompt(null, "bf");
+    window.openDropdownPrompt(null, "bf");
   });
 });
 
 document.querySelectorAll(".changes-change-current-ds").forEach((element) => {
   element.addEventListener("click", function () {
-    openDropdownPrompt(null, "dataset");
+    window.openDropdownPrompt(null, "dataset");
   });
 });
 
 // event listeners for readme open dropdown prompts
 document.querySelectorAll(".readme-change-current-account").forEach((element) => {
   element.addEventListener("click", function () {
-    openDropdownPrompt(null, "bf");
+    window.openDropdownPrompt(null, "bf");
   });
 });
 
 document.querySelectorAll(".readme-change-current-ds").forEach((element) => {
   element.addEventListener("click", function () {
-    openDropdownPrompt(null, "dataset");
+    window.openDropdownPrompt(null, "dataset");
   });
 });
 
