### 2023-11-17 4a14547
feat: ✨ Migrated sample addition to swal
