### 2023-11-20 fb10fcb
refactor: user details card + workpskace selection
