+++ scripts/guided-mode/guided-curate-dataset.js
@@ -38,7 +38,7 @@ const guidedGetDatasetOrigin = (sodaJSON) => {
   return "New";
 };
 
-const handleGuidedModeOrgSwitch = async (buttonClicked) => {
+const window.handleGuidedModeOrgSwitch = async (buttonClicked) => {
   const clickedButtonId = buttonClicked.id;
   if (clickedButtonId === "guided-button-change-workspace-dataset-import") {
     renderGuidedResumePennsieveDatasetSelectionDropdown();
