+++ src/renderer/src/scripts/globals.js
@@ -1197,7 +1197,7 @@ const addBfAccount = async (ev, verifyingOrganization = False) => {
             // Get the current page that the user is on in the guided mode
             const currentPage = window.CURRENT_PAGE.id;
             if (currentPage) {
-              await openPage(currentPage);
+              await window.window.openPage(currentPage);
             }
           }
         } catch (error) {
