+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -565,7 +565,7 @@ const savePageChanges = async (pageBeingLeftID) => {
 
         // Open the first page
         const firstPage = getNonSkippedGuidedModePages(document)[0];
-        await openPage(firstPage.id);
+        await window.openPage(firstPage.id);
       }
 
       if (resumingExistingProgress) {
@@ -2101,7 +2101,7 @@ const renderSideBar = (activePage) => {
             await checkIfPageIsValid(page);
           } catch (error) {
             const pageWithErrorName = document.getElementById(page).getAttribute("data-page-name");
-            await openPage(page);
+            await window.openPage(page);
             await Swal.fire({
               title: `An error occured on an intermediate page: ${pageWithErrorName}`,
               html: `Please address the issues before continuing to ${pageToNaviatetoName}:
@@ -2123,7 +2123,7 @@ const renderSideBar = (activePage) => {
         }
 
         //All pages have been validated. Open the target page.
-        await openPage(pageToNavigateTo);
+        await window.openPage(pageToNavigateTo);
       } catch (error) {
         const pageWithErrorName = window.CURRENT_PAGE.dataset.pageName;
         const { value: continueWithoutSavingCurrPageChanges } = await Swal.fire({
@@ -2150,7 +2150,7 @@ const renderSideBar = (activePage) => {
           width: 700,
         });
         if (continueWithoutSavingCurrPageChanges) {
-          await openPage(pageToNavigateTo);
+          await window.openPage(pageToNavigateTo);
         }
       }
     });
@@ -4642,7 +4642,7 @@ const validatePageArray = async (arrayOfPagesToCheck) => {
     try {
       await checkIfPageIsValid(page.id);
     } catch (error) {
-      await openPage(page.id);
+      await window.openPage(page.id);
       break;
     }
   }
@@ -4650,7 +4650,7 @@ const validatePageArray = async (arrayOfPagesToCheck) => {
 
 const checkIfPageIsValid = async (pageID) => {
   try {
-    await openPage(pageID);
+    await window.openPage(pageID);
     await savePageChanges(pageID);
   } catch (error) {
     throw error;
@@ -4701,7 +4701,7 @@ const guidedShowOptionalRetrySwal = async (errorMessage, pageIdToRetryOpening) =
   });
 
   if (!addDataManually) {
-    await openPage(pageIdToRetryOpening);
+    await window.openPage(pageIdToRetryOpening);
   }
 };
 
@@ -4781,7 +4781,7 @@ const handleNextButtonVisibility = (targetPageID) => {
 // //The general flow is to check if there is values for the keys relevant to the page
 // //If the keys exist, extract the data from the window.sodaJSONObj and populate the page
 // //If the keys do not exist, reset the page (inputs, tables etc.) to the default state
-const openPage = async (targetPageID) => {
+window.openPage = async (targetPageID) => {
   //NOTE: 2 Bottom back buttons (one handles sub pages, and the other handles main pages)
   //Back buttons should be disabled and the function setLoading should be (set as false?)
 
@@ -7277,7 +7277,7 @@ const guidedCheckIfUserNeedsToReconfirmAccountDetails = () => {
 };
 
 const guidedGetPageToReturnTo = (sodaJSONObj) => {
-  // Set by openPage function
+  // Set by window.openPage function
   const usersPageBeforeExit = sodaJSONObj["page-before-exit"];
 
   //If the dataset was successfully uploaded, send the user to the share with curation team
@@ -7598,13 +7598,13 @@ window.guidedResumeProgress = async (datasetNameToResume) => {
 
[…]
