+++ preload.js
@@ -1124,7 +1124,7 @@ const addBfAccount = async (ev, verifyingOrganization = False) => {
             // Get the current page that the user is on in the guided mode
             const currentPage = CURRENT_PAGE.id;
             if (currentPage) {
-              await openPage(currentPage);
+              await window.openPage(currentPage);
             }
           }
         } catch (error) {
