+++ src/renderer/src/scripts/validator/validator-UI.js
@@ -1,0 +1,345 @@
+/*
+*******************************************************************************************************************
+// Purpose: Presentation logic regarding transitioning from one question to another and/or resetting state upon user action in the validator
+*******************************************************************************************************************
+*/
+
+// Presentation logic for transitioning from question one to question two
+const transitionToValidateQuestionTwo = async () => {
+  // hide both local and pennsieve sections
+  let pennsieveSection = document.querySelector("#pennsieve-question-2-container");
+
+  pennsieveSection.style = "display: none !important;";
+
+  let localSection = document.querySelector("#validate_dataset-question-1-local-container");
+  localSection.style = "display: none !important";
+
+  // allow time for the check box to get checked
+  await window.wait(300);
+
+  // check if the local validation option has been checked
+  let localDatasetCard = document.querySelector("#validate-1-Local");
+  let validatingLocalDataset = localDatasetCard.checked;
+
+  // perform the transition for a local dataset
+  if (validatingLocalDataset) {
+    // show local section
+    localSection.style = "display: flex;";
+
+    // show the confirm button if it was hidden
+    let confirmDatasetBtn = document.querySelector("#validator-confirm-local-dataset-btn");
+
+    confirmDatasetBtn.parentElement.style.display = "none";
+
+    // hide the confirm dataset btn -- it will appear after the user selects a local folder
+    // confirmDatasetBtn.style.display = "none"
+
+    // confirm that the input holding the local dataset path's placeholder is reset
+    let input = document.querySelector("#validate-local-dataset-path");
+    input.setAttribute("placeholder", "Browse here");
+    input.value = "";
+  } else {
+    // hide the local dataset section
+    localSection.style = "display: none !important;";
+
+    // transition for pennsieve dataset
+    pennsieveSection.style = "display: flex;";
+
+    // show the pennsieve track's confirm button
+    document.querySelector("#confirm-dataset-selection--validator").style.visibility = "visible";
+  }
+
+  return true;
+};
+
+// Presentation logic for transitioning from question 2 to question 3
+const transitionToValidateQuestionThree = async () => {
+  let userWantsToReset = await userWantsToResetValidation();
+
+  if (userWantsToReset === false) {
+    return userWantsToReset;
+  }
+
+  // hide the confirm buttons
+  let confirmDatasetBtn = document.querySelector("#validator-confirm-local-dataset-btn");
+
+  // set the field display property to none to remove the margins
+  confirmDatasetBtn.parentElement.style.display = "none";
+
+  return true;
+};
+
+// check the local dataset input
+document.querySelector("#validate_dataset-1-local").addEventListener("click", async (e) => {
+  // if there is validation work done check if the user wants to reset progress
+  let userWantsToReset = await userWantsToResetValidation();
+  if (!userWantsToReset) {
+    // deselect local option card and reselect pennsieve option card
+    undoOptionCardSelection(this);
+    // user does not want to reset
+    return;
+  }
+
+  // reset validation table
+  clearValidationResults();
+
+  // transition to the next question - uses transitionToValidateQuestionTwo
+  transitionFreeFormMode(
+    document.querySelector("#validate_dataset-1-local"),
+    "validate_dataset-question-1",
+    "validate_dataset-tab",
+    "",
+    "individual-question "
+  );
+
+  // check the input
+  document.querySelector("#validate-1-Local").checked = true;
+
+  document.querySelector("#validate-1-Pennsieve").checked = false;
+});
+
+// check the pennsieve dataset input
+document
+  .querySelector("#validate_dataset-1-pennsieve")
+  .addEventListener("click", async function () {
[…]
