+++ src/renderer/src/scripts/others/tab-effects.js
@@ -2,6 +2,8 @@
 import Accordion from 'accordion-js';
 // TODO: Follow up that this is the way to import it
 import 'accordion-js/dist/accordion.min.css';
+import { showHideDropdownButtons } from '../globals';
+import {clearValidationResults} from '../validator/validate';
 
 while (!window.htmlPagesAdded) {
   await new Promise((resolve) => setTimeout(resolve, 100))
@@ -970,9 +972,9 @@ var divList = [];
 const transitionSubQuestions = async (ev, currentDiv, parentDiv, button, category) => {
   if (currentDiv === "Question-getting-started-1") {
     // log the start of a new curation process from scratch
-    // logCurationForAnalytics(
+    // window.logCurationForAnalytics(
     //   "Success",
-    //   PrepareDatasetsAnalyticsPrefix.CURATE,
+    //   window.PrepareDatasetsAnalyticsPrefix.CURATE,
     //   window.AnalyticsGranularity.ACTION,
     //   ["New"],
     //   "Local",
@@ -983,7 +985,7 @@ const transitionSubQuestions = async (ev, currentDiv, parentDiv, button, categor
       $("#progress-files-dropdown").val("Select");
       $("#para-progress-file-status").text("");
       $("#nextBtn").prop("disabled", true);
-      exitCurate(false);
+      window.exitCurate(false);
       globalGettingStarted1stQuestionBool = false;
     } else {
       globalGettingStarted1stQuestionBool = false;
@@ -1141,7 +1143,7 @@ const transitionSubQuestions = async (ev, currentDiv, parentDiv, button, categor
     $("#nextBtn").prop("disabled", true);
     $("#para-continue-prepare-new-getting-started").text("");
     if ($("#prepare-new").prop("checked")) {
-      exitCurate(false);
+      window.exitCurate(false);
       $("#prepare-new").prop("checked", true);
       $($("#prepare-new").parents()[2]).addClass("checked");
       $($($("#div-getting-started-prepare-new").parents()[0]).siblings().children()).addClass(
@@ -1159,7 +1161,7 @@ const transitionSubQuestions = async (ev, currentDiv, parentDiv, button, categor
       }, 600);
     } else if ($("#existing-bf").is(":checked")) {
       $("#nextBtn").prop("disabled", true);
-      // this exitCurate function gets called in the beginning here
+      // this window.exitCurate function gets called in the beginning here
       // in case users have existing, non-empty SODA object structure due to previous progress option was selected prior to this "existing-bf" option
       $("#Question-getting-started-existing-BF-account").show();
       $("#Question-getting-started-existing-BF-account").children().show();
@@ -1754,7 +1756,7 @@ const verify_sparc_folder = (root_folder_path, type) => {
 };
 
 // function similar to transitionSubQuestions, but for buttons
-const transitionSubQuestionsButton = async (ev, currentDiv, parentDiv, button, category) => {
+window.transitionSubQuestionsButton = async (ev, currentDiv, parentDiv, button, category) => {
   /*
     ev: the button being clicked
     currentDiv: current option-card (question)
@@ -1765,7 +1767,7 @@ const transitionSubQuestionsButton = async (ev, currentDiv, parentDiv, button, c
   if (currentDiv === "Question-getting-started-BF-dataset") {
     let selectedDataset = $("#current-bf-dataset").text();
     $("#nextBtn").prop("disabled", true);
-    sodaJSONObj = {
+    window.sodaJSONObj = {
       "bf-account-selected": {
         "account-name": {},
       },
[…]

[…]
