+++ scripts/others/renderer.js
@@ -1680,20 +1680,20 @@ createDragSort(guidedSubmissionTagsTagifyManual);
 
 // initiate Tagify input fields for Dataset description file
 var keywordInput = document.getElementById("ds-keywords"),
-  keywordTagify = new Tagify(keywordInput, {
+  window.keywordTagify = new Tagify(keywordInput, {
     duplicates: false,
   });
 
-createDragSort(keywordTagify);
+createDragSort(window.keywordTagify);
 
 var otherFundingInput = document.getElementById("ds-other-funding"),
-  otherFundingTagify = new Tagify(otherFundingInput, {
+  window.otherFundingTagify = new Tagify(otherFundingInput, {
     duplicates: false,
   });
-createDragSort(otherFundingTagify);
+createDragSort(window.otherFundingTagify);
 
 var collectionDatasetInput = document.getElementById("tagify-collection-tags"),
-  collectionDatasetTags = new Tagify(collectionDatasetInput, {
+  window.collectionDatasetTags = new Tagify(collectionDatasetInput, {
     whitelist: [],
     duplicates: false,
     dropdown: {
@@ -1707,10 +1707,10 @@ var collectionDatasetInput = document.getElementById("tagify-collection-tags"),
       rightKey: true,
     },
   });
-createDragSort(collectionDatasetTags);
+createDragSort(window.collectionDatasetTags);
 
 var studyOrganSystemsInput = document.getElementById("ds-study-organ-system"),
-  studyOrganSystemsTagify = new Tagify(studyOrganSystemsInput, {
+  window.studyOrganSystemsTagify = new Tagify(studyOrganSystemsInput, {
     whitelist: [
       "autonomic ganglion",
       "brain",
@@ -1739,25 +1739,25 @@ var studyOrganSystemsInput = document.getElementById("ds-study-organ-system"),
       closeOnSelect: true,
     },
   });
-createDragSort(studyOrganSystemsTagify);
+createDragSort(window.studyOrganSystemsTagify);
 
 var studyTechniquesInput = document.getElementById("ds-study-technique"),
-  studyTechniquesTagify = new Tagify(studyTechniquesInput, {
+  window.studyTechniquesTagify = new Tagify(studyTechniquesInput, {
     duplicates: false,
   });
-createDragSort(studyTechniquesTagify);
+createDragSort(window.studyTechniquesTagify);
 
 var studyApproachesInput = document.getElementById("ds-study-approach"),
-  studyApproachesTagify = new Tagify(studyApproachesInput, {
+  window.studyApproachesTagify = new Tagify(studyApproachesInput, {
     duplicates: false,
   });
-createDragSort(studyApproachesTagify);
+createDragSort(window.studyApproachesTagify);
 
 // tagify the input inside of the "Add/edit tags" manage dataset section
 var datasetTagsInput = document.getElementById("tagify-dataset-tags"),
   // initialize Tagify on the above input node reference
-  datasetTagsTagify = new Tagify(datasetTagsInput);
-createDragSort(datasetTagsTagify);
+  window.datasetTagsTagify = new Tagify(datasetTagsInput);
+createDragSort(window.datasetTagsTagify);
 
 /////////////////// Provide Grant Information section /////////////////////////
 //////////////// //////////////// //////////////// //////////////// ///////////
@@ -2782,9 +2782,9 @@ function changeAwardInputDsDescription() {
     addOption(dsContributorArrayFirst1, "Select an option", "Select an option");
   }
 
-  currentContributorsLastNames = [];
+  window.currentContributorsLastNames = [];
   currentContributorsFirstNames = [];
-  globalContributorNameObject = {};
+  window.globalContributorNameObject = {};
 
   /// delete old table
   $("#table-current-contributors").find("tr").slice(1, -1).remove();
[…]
