+++ scripts/metadata-files/datasetDescription.js
@@ -16,9 +16,9 @@ const dsAwardArray = document.getElementById("ds-description-award-list");
 const dsContributorArrayLast1 = document.getElementById("ds-description-contributor-list-last-1");
 const dsContributorArrayFirst1 = document.getElementById("ds-description-contributor-list-first-1");
 
-var currentContributorsLastNames = [];
+var window.currentContributorsLastNames = [];
 var currentContributorsFirstNames = [];
-var globalContributorNameObject = {};
+var window.globalContributorNameObject = {};
 
 // const affiliationInput = document.getElementById("input-con-affiliation-1");
 const addCurrentContributorsBtn = document.getElementById("button-ds-add-contributor");
@@ -146,8 +146,8 @@ const checkContributorNameDuplicates = (table, currentRow) => {
 const cloneConNamesSelect = (selectLast) => {
   removeOptions(document.getElementById(selectLast));
   addOption(document.getElementById(selectLast), "Select an option", "Select");
-  for (var i = 0; i < currentContributorsLastNames.length; i++) {
-    var opt = currentContributorsLastNames[i];
+  for (var i = 0; i < window.currentContributorsLastNames.length; i++) {
+    var opt = window.currentContributorsLastNames[i];
     if (document.getElementById(selectLast)) {
       addOption(document.getElementById(selectLast), opt, opt);
     }
@@ -585,7 +585,7 @@ const grabConInfoEntries = () => {
     fundingArray = [funding];
   }
   /// other funding sources
-  var otherFunding = otherFundingTagify.value;
+  var otherFunding = window.otherFundingTagify.value;
   for (var i = 0; i < otherFunding.length; i++) {
     fundingArray.push(otherFunding[i].value);
   }
@@ -594,7 +594,7 @@ const grabConInfoEntries = () => {
 
   contributorInfo["funding"] = fundingArray;
   contributorInfo["acknowledgment"] = acknowledgment;
-  contributorInfo["contributors"] = contributorArray;
+  contributorInfo["contributors"] = window.contributorArray;
   return contributorInfo;
 };
 
@@ -845,14 +845,14 @@ var contributorElement = `<div id="contributor-popup"><div style="display:flex">
 var contributorElementRaw =
[…]
