+++ scripts/validator/validate.js
@@ -770,7 +770,7 @@ document.querySelector("#validate_dataset-1-local").addEventListener("click", as
 
   // reset validation table
   let validationErrorsTable = document.querySelector("#validation-errors-container tbody");
-  clearValidationResults(validationErrorsTable);
+  window.clearValidationResults(validationErrorsTable);
 
   document.querySelector("#run_validator_btn").style.display = "none";
 
@@ -808,7 +808,7 @@ document
 
     // reset validation table
     let validationErrorsTable = document.querySelector("#validation-errors-container tbody");
-    clearValidationResults(validationErrorsTable);
+    window.clearValidationResults(validationErrorsTable);
 
     document.querySelector("#run_validator_btn").style.display = "none";
 
@@ -846,7 +846,7 @@ document.querySelector("#validate-local-dataset-path").addEventListener("click",
     let validationErrorsTable = document.querySelector("#validation-errors-container tbody");
 
     // reset validation table
-    clearValidationResults(validationErrorsTable);
+    window.clearValidationResults(validationErrorsTable);
 
     // clear the input value
     this.value = "";
@@ -919,7 +919,7 @@ document.querySelector("#run_validator_btn").addEventListener("click", async fun
 
     // get validation table body
     let validationErrorsTable = document.querySelector("#validation-errors-container tbody");
-    clearValidationResults(validationErrorsTable);
+    window.clearValidationResults(validationErrorsTable);
 
     // hide question 3
     // let questionThreeSection = document.querySelector("#validate_dataset-question-3");
@@ -993,7 +993,7 @@ document
 
       // get validation table body
       let validationErrorsTable = document.querySelector("#validation-errors-container tbody");
-      clearValidationResults(validationErrorsTable);
+      window.clearValidationResults(validationErrorsTable);
 
       // hide question 3
       let questionThreeSection = document.querySelector("#validate_dataset-question-3");
@@ -1062,7 +1062,7 @@ const undoOptionCardSelection = (activeOptionCard) => {
   activeOptionCard.classList.remove("checked");
 };
 
-const clearValidationResults = (validationTableElement) => {
+const window.clearValidationResults = (validationTableElement) => {
   // remove its children
   while (validationTableElement.firstChild) {
     validationTableElement.removeChild(validationTableElement.firstChild);
