+++ scripts/guided-mode/guided-curate-dataset.js
@@ -4729,7 +4729,7 @@ const handleValidationTableUi = (errors) => {
       "#guided-section-dataset-validation-table tbody"
     );
     // clear the table
-    clearValidationResults(validationErrorsTable);
+    window.clearValidationResults(validationErrorsTable);
     // display errors onto the page
     displayValidationErrors(
       errors,
@@ -5428,12 +5428,12 @@ const window.openPage = async (targetPageID) => {
 
           // Loop through the contributorData array besides the first row (which is the header)
           for (let i = 1; i < contributorData.length; i++) {
-            const contributorArray = contributorData[i];
+            const window.contributorArray = contributorData[i];
             // split the name into first and last name with the first name being the first element and last name being the rest of the elements
-            const contributorFullName = contributorArray[0];
-            const contributorID = contributorArray[1];
-            const contributorAffiliation = contributorArray[2].split(", ");
-            const contributorRoles = contributorArray[3].split(", ");
+            const contributorFullName = window.contributorArray[0];
+            const contributorID = window.contributorArray[1];
+            const contributorAffiliation = window.contributorArray[2].split(", ");
+            const contributorRoles = window.contributorArray[3].split(", ");
             try {
               addContributor(
                 contributorFullName,

[…]
