+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -2804,7 +2804,7 @@ const guidedTransitionFromHome = async () => {
 
 const guidedTransitionToHome = () => {
   guidedUnLockSideBar();
-  guidedPrepareHomeScreen();
+  window.guidedPrepareHomeScreen();
 
   document.getElementById("guided-home").classList.remove("hidden");
   // Hide all of the parent tabs
@@ -3793,7 +3793,7 @@ const guidedResetProgressVariables = () => {
   window.window.samplesTableData = [];
 };
 
-const guidedPrepareHomeScreen = async () => {
+window.guidedPrepareHomeScreen = async () => {
   //Wipe out existing progress if it exists
   guidedResetProgressVariables();
   //Check if Guided-Progress folder exists. If not, create it.
@@ -12085,15 +12085,10 @@ document.getElementById("button-homepage-guided-mode").addEventListener("click",
 });
 
 
-const directToFreeFormMode = () => {
-  const freeFormModeLinkButton = document.getElementById("main_tabs_view");
-  freeFormModeLinkButton.click();
-};
-
 
 document.querySelector("#button-homepage-freeform-mode").addEventListener("click", async () => {
   //Free form mode will open through here
-  guidedPrepareHomeScreen();
+  window.guidedPrepareHomeScreen();
 
   // guidedResetSkippedPages();
 
