### 2023-11-21 838e3ba
feat: ✨ Added spreadsheet instructions templates for all options
