+++ src/renderer/src/scripts/manage-dataset/manage-dataset.js
@@ -1,4 +1,13 @@
 import axios from "axios"
+import Swal from "sweetalert2"
+import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
+import client from '../client'
+import lottie from 'lottie-web'
+import {licenseLottie} from '../../assets/lotties/license-lottie'
+import kombuchaEnums from "../analytics/analytics-enums"
+import { updateDatasetList } from "../globals"
+import api from '../others/api/api'
+
 // event listeners for opening dataset or account selection dropdown
 import Accordion from 'accordion-js';
 // TODO: Follow up that this is the way to import it
@@ -146,10 +155,10 @@ $("#bf-rename-dataset-name").on("keyup", () => {
 // Add new dataset folder (empty) on bf //
 $("#button-create-bf-new-dataset").click(async () => {
   setTimeout(async () => {
-    let selectedbfaccount = window.defaultBfDataset;
+    let selectedbfaccount = window.defaultBfAccount;
     let bfNewDatasetName = $("#bf-new-dataset-name").val();
 
-    log.info(`Creating a new dataset with the name: ${bfNewDatasetName}`);
+    window.log.info(`Creating a new dataset with the name: ${bfNewDatasetName}`);
 
     $("#button-create-bf-new-dataset").prop("disabled", true);
 
@@ -205,27 +214,27 @@ $("#button-create-bf-new-dataset").click(async () => {
       }
       $("#dataset-success-container").removeClass("hidden");
 
-      log.info(`Created dataset successfully`);
+      window.log.info(`Created dataset successfully`);
 
       $("#button-create-bf-new-dataset").hide();
 
       window.defaultBfDataset = bfNewDatasetName;
-      window.window.defaultBfDatasetId = res;
+      window.defaultBfDatasetId = res;
       // log a map of datasetId to dataset name to analytics
       // this will be used to help us track private datasets which are not trackable using a datasetId alone
-      ipcRenderer.send(
+      window.electron.ipcRenderer.send(
         "track-event",
         "Dataset ID to Dataset Name Map",
-        window.window.defaultBfDatasetId,
+        window.defaultBfDatasetId,
         window.defaultBfDataset
       );
       window.refreshDatasetList();
-      currentDatasetPermission.innerHTML = "";
-      currentAddEditDatasetPermission.innerHTML = "";
+      window.currentDatasetPermission.innerHTML = "";
+      window.currentAddEditDatasetPermission.innerHTML = "";
       $("#button-create-bf-new-dataset").prop("disabled", false);
 
       window.addNewDatasetToList(bfNewDatasetName);
-      ipcRenderer.send(
+      window.electron.ipcRenderer.send(
         "track-kombucha",
         kombuchaEnums.Category.MANAGE_DATASETS,
         kombuchaEnums.Action.CREATE_NEW_DATASET,
@@ -233,22 +242,22 @@ $("#button-create-bf-new-dataset").click(async () => {
         kombuchaEnums.Status.SUCCESS,
         {
           value: 1,
-          dataset_id: window.window.defaultBfDatasetId,
+          dataset_id: window.defaultBfDatasetId,
         }
       );
 
-      ipcRenderer.send(
+      window.electron.ipcRenderer.send(
         "track-event",
         "Success",
-        ManageDatasetsAnalyticsPrefix.MANAGE_DATASETS_CREATE_DATASET,
+        window.ManageDatasetsAnalyticsPrefix.MANAGE_DATASETS_CREATE_DATASET,
         bfNewDatasetName
       );
 
[…]
