### 2023-11-22 2d7d421
refactor: nodestorage uses ipc calls
