### 2023-11-22 7ab4224
refactor: nodestorage in main script
