### 2023-11-22 330079c
fix: replace default dataset with default account when appropriately
