+++ src/renderer/src/scripts/manage-dataset/manage-dataset.js
@@ -310,8 +310,8 @@ window.addNewDatasetToList = (newDataset) => {
 // Rename dataset on pennsieve
 $("#button-rename-dataset").on("click", async () => {
   setTimeout(async function () {
-    var selectedbfaccount = window.defaultBfDataset;
-    var currentDatasetName = window.defaultBfAccount;
+    var selectedbfaccount = window.defaultBfAccount;
+    var currentDatasetName = window.defaultBfDataset;
     var renamedDatasetName = $("#bf-rename-dataset-name").val();
 
     Swal.fire({
@@ -440,7 +440,7 @@ $("#button-rename-dataset").on("click", async () => {
 
       try {
         window.datasetList = [];
-        window.datasetList = await api.getDatasetsForAccount(window.defaultBfDataset);
+        window.datasetList = await api.getDatasetsForAccount(window.defaultBfAccount);
         window.refreshDatasetList();
       } catch (error) {
         clientError(error);
@@ -491,7 +491,7 @@ $("#button-add-permission-pi").click(async () => {
         },
       });
 
-      let selectedBfAccount = window.defaultBfDataset;
+      let selectedBfAccount = window.defaultBfAccount;
       let selectedBfDataset = window.defaultBfDataset;
       let selectedUser = $("#bf_list_users_pi").val();
       let selectedRole = "owner";
@@ -591,7 +591,7 @@ const changeDatasetRolePI = (selectedDataset) => {
 };
 
 const showCurrentPermission = async () => {
-  let selectedBfAccount = window.defaultBfDataset;
+  let selectedBfAccount = window.defaultBfAccount;
   let selectedBfDataset = window.defaultBfDataset;
 
   if (selectedBfDataset === null) {
@@ -773,7 +773,7 @@ $("#button-add-permission-user").click(() => {
       },
     });
 
-    let selectedBfAccount = window.defaultBfDataset;
+    let selectedBfAccount = window.defaultBfAccount;
     let selectedBfDataset = window.defaultBfDataset;
     let selectedUser = $("#bf_list_users").val();
     let selectedRole = $("#bf_list_roles_user").val();
@@ -812,7 +812,7 @@ $("#button-add-permission-team").click(async () => {
         },
         {
           params: {
-            selected_account: window.defaultBfDataset,
+            selected_account: window.defaultBfAccount,
             selected_dataset: window.defaultBfDataset,
             scope: "team",
             name: selectedTeam,
@@ -918,7 +918,7 @@ $("#button-add-subtitle").click(async () => {
       },
     });
 
-    let selectedBfAccount = window.defaultBfDataset;
+    let selectedBfAccount = window.defaultBfAccount;
     let selectedBfDataset = window.defaultBfDataset;
     let inputSubtitle = $("#bf-dataset-subtitle").val().trim();
 
@@ -1015,7 +1015,7 @@ $("#button-add-subtitle").click(async () => {
 });
 
 const showCurrentSubtitle = async () => {
-  let selectedBfAccount = window.defaultBfDataset;
+  let selectedBfAccount = window.defaultBfAccount;
   let selectedBfDataset = window.defaultBfDataset;
 
   if (selectedBfDataset === null) {
@@ -1077,7 +1077,7 @@ dsAccordion.open( 0);
 
 // fires whenever a user selects a dataset, from any card
 const showCurrentDescription = async () => {
-  let selectedBfAccount = window.defaultBfDataset;
+  let selectedBfAccount = window.defaultBfAccount;
   let selectedBfDataset = window.defaultBfDataset;
 
   if (selectedBfDataset === "Select dataset" || selectedBfDataset === null) {
@@ -1195,7 +1195,7 @@ const showCurrentDescription = async () => {
 
[…]
