### 2023-11-22 53b34c8
refactor: ffm dataset selection
