+++ src/renderer/src/scripts/organize-dataset/curate-functions.js
@@ -1,5 +1,7 @@
 import Swal from "sweetalert2";
 import checkDiskSpace from "check-disk-space";
+import { updateDatasetList } from "../globals"
+
 
 
 while (!window.htmlPagesAdded) {
