+++ scripts/guided-mode/guided-curate-dataset.js
@@ -877,6 +877,176 @@ const savePageChanges = async (pageBeingLeftID) => {
       guidedSkipPage("guided-select-starting-point-tab");
     }
 
+    if (pageBeingLeftID === "guided-ask-if-submission-is-sparc-funded-tab") {
+      // NOTE: We use the button config generated by this page to determine if the dataset is SPARC funded
+      // See the function: datasetIsSparcFunded()
+      const userSelectedDatasetIsSparcFunded = document
+        .getElementById("guided-button-dataset-is-sparc-funded")
+        .classList.contains("selected");
+      const userSelectedDatasetIsReJoinFunded = document
+        .getElementById("guided-button-dataset-is-re-join-funded")
+        .classList.contains("selected");
+      const userSelectedDatasetIsOtherFunded = document
+        .getElementById("guided-button-dataset-is-not-sparc-funded")
+        .classList.contains("selected");
+
+      if (
+        !userSelectedDatasetIsSparcFunded &&
+        !userSelectedDatasetIsReJoinFunded &&
+        !userSelectedDatasetIsOtherFunded
+      ) {
+        errorArray.push({
+          type: "notyf",
+          message: "Please indicate the funding source for this dataset.",
+        });
+        throw errorArray;
+      }
+
+      // If the user selected that the dataset is SPARC funded, unskip the submission metadata page
+      if (userSelectedDatasetIsSparcFunded) {
+        // Set the consortium data standard value in the JSON
+        sodaJSONObj["dataset-metadata"]["submission-metadata"]["consortium-data-standard"] =
+          "SPARC";
+        // Get the funding source value from the dropdown
+        const selectedFuncingSourceFromDropdown = $(
+          "#guided-select-sparc-funding-consortium"
+        ).val();
+
+        // Throw an error if the user did not select a funding source from the dropdown
+        if (!selectedFuncingSourceFromDropdown) {
+          errorArray.push({
+            type: "notyf",
+            message: "Please select a funding source from the dropdown.",
+          });
+          throw errorArray;
+        } else {
+          // Set the funding consortium value in the JSON
+          sodaJSONObj["dataset-metadata"]["submission-metadata"]["funding-consortium"] =
+            selectedFuncingSourceFromDropdown;
+        }
+      }
+
+      if (userSelectedDatasetIsReJoinFunded) {
+        sodaJSONObj["dataset-metadata"]["submission-metadata"]["consortium-data-standard"] = "HEAL";
+        sodaJSONObj["dataset-metadata"]["submission-metadata"]["funding-consortium"] =
+          "REJOIN-HEAL";
+      }
+
+      if (userSelectedDatasetIsOtherFunded) {
+        const userSelectedTheyHaveReachedOutToCurationTeam = document
+          .getElementById("guided-button-non-sparc-user-has-contacted-sparc")
+          .classList.contains("selected");
+        const userSelectedTheyHaveNotReachedOutToCurationTeam = document
+          .getElementById("guided-button-non-sparc-user-has-not-contacted-sparc")
+          .classList.contains("selected");
+
+        if (
+          !userSelectedTheyHaveReachedOutToCurationTeam &&
+          !userSelectedTheyHaveNotReachedOutToCurationTeam
+        ) {
+          errorArray.push({
+            type: "notyf",
+            message: "Please indicate if you have reached out to the curation team",
+          });
+          throw errorArray;
+        }
+        if (userSelectedTheyHaveNotReachedOutToCurationTeam) {
+          errorArray.push({
+            type: "notyf",
+            message: "Please reach out to the curation team before continuing the curation process",
+          });
+          throw errorArray;
+        }
+
[…]
