### 2023-11-23 4d7c086
feat: ✨ Updated samples table headers
