+++ src/renderer/src/scripts/metadata-files/manifest.js
@@ -1,5 +1,7 @@
 import client from '../client'
 import excelToJson from 'convert-excel-to-json';
+import determineDatasetLocation, { Destinations } from "../analytics/analytics-utils"
+
 
 while (!window.htmlPagesAdded) {
   await new Promise((resolve) => setTimeout(resolve, 100))
@@ -845,17 +847,17 @@ const generateManifest = async (action, type, manifestEditBoolean, ev) => {
           icon: "warning",
           showConfirmButton: "OK",
         });
-        logMetadataForAnalytics(
+        window.logMetadataForAnalytics(
           "Error",
-          MetadataAnalyticsPrefix.MANIFEST,
+          window.MetadataAnalyticsPrefix.MANIFEST,
           window.AnalyticsGranularity.ALL_LEVELS,
           "Generate",
           Destinations.LOCAL
         );
 
-        logMetadataForAnalytics(
+        window.logMetadataForAnalytics(
           "Error",
-          MetadataAnalyticsPrefix.MANIFEST,
+          window.MetadataAnalyticsPrefix.MANIFEST,
           window.AnalyticsGranularity.ACTION,
           "Generate - Check Storage Space",
           Destinations.LOCAL
@@ -865,9 +867,9 @@ const generateManifest = async (action, type, manifestEditBoolean, ev) => {
           generateAfterEdits();
           return;
         }
-        logMetadataForAnalytics(
+        window.logMetadataForAnalytics(
           "Success",
-          MetadataAnalyticsPrefix.MANIFEST,
+          window.MetadataAnalyticsPrefix.MANIFEST,
           window.AnalyticsGranularity.ACTION,
           "Generate - Check Storage Space",
           Destinations.LOCAL
@@ -904,9 +906,9 @@ const generateManifest = async (action, type, manifestEditBoolean, ev) => {
           }).then((result) => {});
 
           // log the error to analytics
-          logMetadataForAnalytics(
+          window.logMetadataForAnalytics(
             "Error",
-            MetadataAnalyticsPrefix.MANIFEST,
+            window.MetadataAnalyticsPrefix.MANIFEST,
             window.AnalyticsGranularity.ALL_LEVELS,
             "Generate",
             Destinations.LOCAL
@@ -932,9 +934,9 @@ const generateManifest = async (action, type, manifestEditBoolean, ev) => {
             },
           }).then((result) => {});
           // log the error to analytics
-          logMetadataForAnalytics(
+          window.logMetadataForAnalytics(
             "Error",
-            MetadataAnalyticsPrefix.MANIFEST,
+            window.MetadataAnalyticsPrefix.MANIFEST,
             window.AnalyticsGranularity.ALL_LEVELS,
             "Generate",
             Destinations.LOCAL
@@ -960,9 +962,9 @@ const generateManifest = async (action, type, manifestEditBoolean, ev) => {
             },
           }).then((result) => {});
           // log the error to analytics
-          logMetadataForAnalytics(
+          window.logMetadataForAnalytics(
             "Error",
-            MetadataAnalyticsPrefix.MANIFEST,
+            window.MetadataAnalyticsPrefix.MANIFEST,
             window.AnalyticsGranularity.ALL_LEVELS,
             "Generate",
             Destinations.LOCAL
@@ -1123,9 +1125,9 @@ const initiate_generate_manifest_local = async (manifestEditBoolean, originalDat
       });
       //////////// Tracking analytics /////////////
[…]
