+++ scripts/others/analytics/analytics-utils.js
@@ -1,3 +1,6 @@
+import { Destinations } from "./analytics"
+
+
 // determine if we are working with a Local, Saved, or Pennsieve dataset in the current Curation flow
 const determineDatasetLocation = () => {
   let location = "";

[…]
