+++ src/renderer/src/scripts/others/tab-effects.js
@@ -3662,7 +3662,7 @@ const saveOrganizeProgressPrompt = () => {
         if (result.value !== null && result.value !== "") {
           window.sodaJSONObj["save-progress"] = result.value.trim();
           saveSODAJSONProgress(result.value.trim());
-          addOption(progressFileDropdown, result.value.trim(), result.value.trim() + ".json");
+          window.addOption(progressFileDropdown, result.value.trim(), result.value.trim() + ".json");
         }
       }
     });
