+++ src/renderer/src/scripts/metadata-files/readme-changes.js
@@ -1,5 +1,15 @@
 import determineDatasetLocation, { Destinations } from "../analytics/analytics-utils"
+import api from "../others/api/api"
+import Swal from "sweetalert2";
+import client from '../client'
+import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
+import kombuchaEnums from "../analytics/analytics-enums";
+import createEventDataPrepareMetadata from "../analytics/prepare-metadata-analytics";
 
+
+while (!window.htmlPagesAdded) {
+  await new Promise((resolve) => setTimeout(resolve, 100))
+}
 
 // event listeners for changes open dropdown prompts
 document.querySelectorAll(".changes-change-current-account").forEach((element) => {
@@ -50,7 +60,7 @@ window.generateRCFilesHelper = (type) => {
 };
 
 // generate changes or readme either locally (uploadBFBoolean=false) or onto Pennsieve (uploadBFBoolean=true)
-const generateRCFiles = async (uploadBFBoolean, fileType) => {
+window.generateRCFiles = async (uploadBFBoolean, fileType) => {
   let result = window.generateRCFilesHelper(fileType);
   let bfDataset = document.getElementById(`bf_dataset_load_${fileType}`).innerText.trim();
   let upperCaseLetters = fileType.toUpperCase() + ".txt";
@@ -67,7 +77,7 @@ const generateRCFiles = async (uploadBFBoolean, fileType) => {
     }
 
     // Check if dataset is locked after running pre-flight checks
-    const isLocked = await api.isDatasetLocked(window.defaultBfDataset, bfDataset);
+    const isLocked = await api.isDatasetLocked(window.defaultBfAccount, bfDataset);
     if (isLocked) {
       await Swal.fire({
         icon: "info",
@@ -130,7 +140,7 @@ const generateRCFiles = async (uploadBFBoolean, fileType) => {
         {
           params: {
             file_type: upperCaseLetters,
-            selected_account: window.defaultBfDataset,
+            selected_account: window.defaultBfAccount,
             selected_dataset: bfDataset,
           },
         }
@@ -145,7 +155,7 @@ const generateRCFiles = async (uploadBFBoolean, fileType) => {
         backdrop: "rgba(0,0,0, 0.4)",
       });
 
-      ipcRenderer.send(
+      window.electron.ipcRenderer.send(
         "track-kombucha",
         kombuchaEnums.Category.PREPARE_METADATA,
         kombuchaEnums.Action.GENERATE_METADATA,
@@ -156,7 +166,7 @@ const generateRCFiles = async (uploadBFBoolean, fileType) => {
         createEventDataPrepareMetadata("Pennsieve", 1)
       );
 
-      ipcRenderer.send(
+      window.electron.ipcRenderer.send(
         "track-kombucha",
         kombuchaEnums.Category.PREPARE_METADATA,
         kombuchaEnums.Action.GENERATE_METADATA,
@@ -178,7 +188,7 @@ const generateRCFiles = async (uploadBFBoolean, fileType) => {
         backdrop: "rgba(0,0,0, 0.4)",
       });
 
-      ipcRenderer.send(
+      window.electron.ipcRenderer.send(
         "track-kombucha",
         kombuchaEnums.Category.PREPARE_METADATA,
         kombuchaEnums.Action.GENERATE_METADATA,
@@ -190,7 +200,7 @@ const generateRCFiles = async (uploadBFBoolean, fileType) => {
       );
     }
   } else {
-    ipcRenderer.send(`open-destination-generate-${fileType}-locally`);
+    window.electron.ipcRenderer.send(`open-destination-generate-${fileType}-locally`);
   }
 };
 
[…]
