+++ src/renderer/src/scripts/metadata-files/readme-changes.js
@@ -471,7 +471,7 @@ window.saveRCFile = async (type) =>  {
 }
 
 // show filebrowser for existing local Changes/README file
-function showExistingRCFile(type) {
+window.showExistingRCFile = (type) => {
   if (
     $(`#existing-${type}-file-destination`).prop("placeholder") !== "Browse here" &&
     $(`#Question-prepare-${type}-2`).hasClass("show")
@@ -632,7 +632,7 @@ window.getRC = async (type) => {
 };
 
 // helper function to import a local readme/changes file
-function importExistingRCFile(type) {
+window.importExistingRCFile = (type) => {
   var filePath = $(`#existing-${type}-file-destination`).prop("placeholder");
   if (type === "changes") {
     var upperCaseLetter = "CHANGES";
