+++ preload.js
@@ -407,7 +407,7 @@ function resetDDFields() {
   $("#div-confirm-destination-dd-locally").css("display", "none");
 }
 
-function resetSubjects(askToReset = true) {
+function window.resetSubjects(askToReset = true) {
   if (!askToReset) {
     // 1. remove Prev and Show from all individual-question except for the first one
     // 2. empty all input, textarea, select, para-elements
@@ -801,7 +801,7 @@ window.resetFFMUI = (ev) => {
   }
 
   if (resetSubjectsTab) {
-    resetSubjects(false);
+    window.resetSubjects(false);
   }
 
   if (resetSamplesTab) {
