+++ src/renderer/src/scripts/others/renderer.js
@@ -1802,7 +1802,7 @@ window.electron.ipcRenderer.on("selected-generate-metadata-samples", (event, dir
               Swal.showLoading();
             },
           }).then((result) => {});
-          generateSamplesFileHelper(uploadBFBoolean);
+          window.generateSamplesFileHelper(uploadBFBoolean);
         }
       });
     } else {
@@ -1818,12 +1818,12 @@ window.electron.ipcRenderer.on("selected-generate-metadata-samples", (event, dir
           Swal.showLoading();
         },
       }).then((result) => {});
-      generateSamplesFileHelper(uploadBFBoolean);
+      window.generateSamplesFileHelper(uploadBFBoolean);
     }
   }
 });
 
-const generateSamplesFileHelper = async (uploadBFBoolean) => {
+window.generateSamplesFileHelper = async (uploadBFBoolean) => {
   let bfDataset = $("#bf_dataset_load_samples").text().trim();
   if (uploadBFBoolean) {
     // Run pre-flight checks before uploading the samples file to Pennsieve
@@ -1905,7 +1905,7 @@ const generateSamplesFileHelper = async (uploadBFBoolean) => {
     let samplesFileResponse = await client.post(
       "prepare_metadata/samples_file",
       {
-        filepath: samplesDestinationPath,
+        filepath: window.samplesDestinationPath,
         selected_account: window.defaultBfAccount,
         selected_dataset: $("#bf_dataset_load_samples").text().trim(),
         samples_str: window.samplesTableData,
