+++ src/renderer/src/scripts/globals.js
@@ -598,7 +598,7 @@ window.resetSubjects = (askToReset = true) => {
   })
 }
 
-function resetSamples(askToReset = true) {
+window.resetSamples = (askToReset = true) => {
   if (!askToReset) {
     // 1. remove Prev and Show from all individual-question except for the first one
     // 2. empty all input, textarea, select, para-elements
@@ -878,7 +878,7 @@ window.resetFFMUI = (ev) => {
   }
 
   if (resetSamplesTab) {
-    resetSamples(false)
+    window.resetSamples(false)
   }
 
   if (resetManifestTab) {
