### 2023-11-27 bc89249
fix: 🐛 Added identifier convention validation check to entities
