### 2023-11-27 5c314f4
refactor: ffm create local manifest
