+++ src/renderer/src/scripts/metadata-files/manifest.js
@@ -1,6 +1,13 @@
 import client from '../client'
 import excelToJson from 'convert-excel-to-json';
 import determineDatasetLocation, { Destinations } from "../analytics/analytics-utils"
+import https from "https";
+import Swal from "sweetalert2";
+import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
+import kombuchaEnums from "../analytics/analytics-enums";
+import createEventDataPrepareMetadata from "../analytics/prepare-metadata-analytics";
+import api from '../others/api/api'
+
 
 
 while (!window.htmlPagesAdded) {
@@ -28,11 +35,11 @@ const guidedJsTreePreviewManifest = document.getElementById(
   "guided-div-dataset-tree-preview-manifest"
 );
 
-const showLocalDatasetManifest = () => {
+window.showLocalDatasetManifest = () => {
   window.electron.ipcRenderer.send("open-file-dialog-local-dataset-manifest-purpose");
 };
 
-const selectManifestGenerationLocation = () => {
+window.selectManifestGenerationLocation = () => {
   window.electron.ipcRenderer.send("open-file-dialog-local-dataset-manifest-generate-purpose");
 };
 
@@ -61,7 +68,7 @@ const openFolder = (generationLocation) => {
   }
 };
 
-$(document).ready(function () {
+$(document).ready(async function () {
   let localDataSetImport = false;
   let fileOpenedOnce = {};
   window.electron.ipcRenderer.on("selected-local-dataset-manifest-purpose", (event, folderPath) => {
@@ -296,16 +303,16 @@ $(document).ready(function () {
       var localFolderPath = window.path.join(homeDirectory, "SODA", "manifest_files", parentFolderName);
       var selectedManifestFilePath = window.path.join(localFolderPath, "manifest.xlsx");
       if (window.fs.existsSync(selectedManifestFilePath)) {
-        jsonManifest = excelToJson({
+        jsonManifest = await window.electron.ipcRenderer.invoke("excelToJsonSheet1Options", {
           sourceFile: selectedManifestFilePath,
           columnToKey: {
             "*": "{{columnHeader}}",
           },
-        })["Sheet1"];
+        })
       }
 
       window.sodaCopy = window.sodaJSONObj;
-      datasetStructCopy = window.sodaCopy["dataset-structure"];
+      let datasetStructCopy = window.sodaCopy["dataset-structure"];
 
       //Save content of manifest file to a variable to add to soda json at the end
       let originalManifestFilesValue = window.sodaCopy["manifest-files"];
@@ -323,7 +330,7 @@ $(document).ready(function () {
         let response = cleanJson.data.soda_json_structure;
 
         window.sodaCopy = response;
-        datasetStructCopy = window.sodaCopy["dataset-structure"];
+        let datasetStructCopy = window.sodaCopy["dataset-structure"];
       } catch (e) {
         clientError(e);
         userErrorMessage(e);
@@ -373,12 +380,12 @@ $(document).ready(function () {
             }
             if (!window.fs.existsSync(selectedManifestFilePath)) {
             } else {
-              jsonManifest = excelToJson({
+              jsonManifest =  await window.electron.ipcRenderer.invoke("excelToJsonSheet1Options", {
                 sourceFile: selectedManifestFilePath,
                 columnToKey: {
                   "*": "{{columnHeader}}",
                 },
-              })["Sheet1"];
+              })
             }
             // If file doesn't exist then that means it didn't get imported properly
             let sortedJSON = window.processManifestInfo(manifestHeader, manifestFileData);
@@ -418,7 +425,7 @@ $(document).ready(function () {
                 window.sodaCopy["manifest-files"][highLevelFolderName] =
[…]
