+++ src/renderer/src/scripts/others/renderer.js
@@ -61,7 +61,7 @@ import {
   bfAccountOptions,
 } from '../globals'
 import checkForAnnouncements from './announcements'
-import { swalFileListSingleAction, swalFileListTripleAction, swalFileListDoubleAction } from "../utils/swal-utils"
+import { swalFileListSingleAction, swalFileListTripleAction, swalFileListDoubleAction, swalShowError } from "../utils/swal-utils"
 
 // add jquery to the window object
 window.$ = jQuery;
@@ -1377,7 +1377,7 @@ dragselect_area.subscribe("dragstart", ({ items, event, isDragging }) => {
 // ///// Global variables for this section
 
 // /////// Save and load award and milestone info
-// let metadataPath = window.path.join(window.homeDirectory, "SODA", "METADATA");
+let metadataPath = window.path.join(window.homeDirectory, "SODA", "METADATA");
 // let awardFileName = "awards.json";
 // let affiliationFileName = "affiliations.json";
 // let milestoneFileName = "milestones.json";
@@ -4248,13 +4248,13 @@ window.sodaJSONObj = {};
 //     let addedElement = window.path.join(folderPath, itemWithinFolder);
 
 //     if (statsObj.isDirectory() && !/(^|\/)\[^\/\.]/g.test(itemWithinFolder)) {
-//       if (irregularFolderArray.includes(addedElement)) {
+//       if (window.irregularFolderArray.includes(addedElement)) {
 //         let renamedFolderName = "";
 //         if (action !== "ignore" && action !== "") {
 //           if (action === "remove") {
-//             renamedFolderName = removeIrregularFolders(itemWithinFolder);
+//             renamedFolderName = window.removeIrregularFolders(itemWithinFolder);
 //           } else if (action === "replace") {
-//             renamedFolderName = replaceIrregularFolders(itemWithinFolder);
+//             renamedFolderName = window.replaceIrregularFolders(itemWithinFolder);
 //           }
 //           jsonObject["folders"][renamedFolderName] = {
 //             type: "local",
@@ -5076,28 +5076,28 @@ const mergeLocalAndRemoteDatasetStructure = async (
   }
 };
 
-// const checkForDuplicateFolderAndFileNames = async (importedFolders, itemsAtPath) => {
-//   const duplicateFolderNames = [];
-//   const duplicateFileNames = [];
+const checkForDuplicateFolderAndFileNames = async (importedFolders, itemsAtPath) => {
+  const duplicateFolderNames = [];
+  const duplicateFileNames = [];
 
-//   const checkForDuplicateNames = async (importedFolders, itemsAtPath) => {
-//     const currentFoldersAtPath = Object.keys(itemsAtPath.folders);
-//     const currentFilesAtPath = Object.keys(itemsAtPath.files);
-//     fg;
-//     for (const folder of importedFolders) {
-//       folderName = window.path.basename(folder);
-//       if (currentFoldersAtPath.includes(folderName)) {
-//         duplicateFolderNames.push(folderName);
-//       }
-//       const folderContents = await fs.readdir(folder);
-//     }
-//     for (const fileName of importedFiles) {
-//       if (currentFilesAtPath.includes(fileName)) {
-//         duplicateFileNames.push(fileName);
-//       }
-//     }
-//   };
-// };
+  const checkForDuplicateNames = async (importedFolders, itemsAtPath) => {
+    const currentFoldersAtPath = Object.keys(itemsAtPath.folders);
+    const currentFilesAtPath = Object.keys(itemsAtPath.files);
+    fg;
+    for (const folder of importedFolders) {
+      folderName = window.path.basename(folder);
+      if (currentFoldersAtPath.includes(folderName)) {
[…]
