+++ scripts/others/renderer.js
@@ -4943,7 +4943,7 @@ const swalShowError = async (title, errorText) => {
     title: title,
     html: errorText,
     icon: "error",
-    width: 600,
+    width: 800,
     heightAuto: false,
     backdrop: "rgba(0,0,0, 0.4)",
     showConfirmButton: true,
@@ -4956,7 +4956,7 @@ const swalShowInfo = async (title, infoText) => {
     title: title,
     html: infoText,
     icon: "info",
-    width: 600,
+    width: 800,
     heightAuto: false,
     backdrop: "rgba(0,0,0, 0.4)",
     showConfirmButton: true,
