### 2023-11-27 ceb15e6
refactor: ffm upload manifest to Pennsieve standalone
