### 2023-11-28 8d75e00
feat: ✨ Make user regenerate spreadsheet on option switch
