### 2023-11-28 b29fbfa
wip: 🚧 Show multiple swals when multiple spreadsheet errors exist
