### 2023-11-28 6c86422
feat: ✨ Skip/unskip pool and sample data organization pages
