### 2023-11-28 ed8ebbe
fix: 🐛 throw notyf when empty pools were added
