### 2023-11-28 0eef15d
fix: 🐛 Require at least one sample on sample addition page
