+++ src/renderer/src/scripts/others/renderer.js
@@ -1226,50 +1226,50 @@ window.electron.ipcRenderer.on("app_version", (event, arg) => {
 
 // /////// New Organize Datasets /////////////////////
 
-// const organizeDSbackButton = document.getElementById("button-back");
-// const organizeDSaddFiles = document.getElementById("add-files");
-// const organizeDSaddNewFolder = document.getElementById("new-folder");
+const organizeDSbackButton = document.getElementById("button-back");
+const organizeDSaddFiles = document.getElementById("add-files");
+const organizeDSaddNewFolder = document.getElementById("new-folder");
 const organizeDSaddFolders = document.getElementById("add-folders");
-// const contextMenu = document.getElementById("mycontext");
+const contextMenu = document.getElementById("mycontext");
 const fullNameValue = document.querySelector(".hoverFullName");
-// const homePathButton = document.getElementById("home-path");
+const homePathButton = document.getElementById("home-path");
 window.menuFolder = document.querySelector(".menu.reg-folder");
 window.menuFile = document.querySelector(".menu.file");
 window.menuHighLevelFolders = document.querySelector(".menu.high-level-folder");
-// const organizeNextStepBtn = document.getElementById("button-organize-confirm-create");
-// const organizePrevStepBtn = document.getElementById("button-organize-prev");
-// const manifestFileCheck = document.getElementById("generate-manifest-curate");
+const organizeNextStepBtn = document.getElementById("button-organize-confirm-create");
+const organizePrevStepBtn = document.getElementById("button-organize-prev");
+window.manifestFileCheck = document.getElementById("generate-manifest-curate");
 
 
-// // Organize dataset //
-// const selectImportFileOrganizationBtn = document.getElementById(
-//   "button-select-upload-file-organization"
-// );
-// const tableMetadata = document.getElementById("metadata-table");
-// let tableMetadataCount = 0;
+// Organize dataset //
+const selectImportFileOrganizationBtn = document.getElementById(
+  "button-select-upload-file-organization"
+);
+const tableMetadata = document.getElementById("metadata-table");
+let tableMetadataCount = 0;
 
-// // Validate dataset //
-// const validateCurrentDSBtn = document.getElementById("button-validate-current-ds");
-// const validateCurrentDatasetReport = document.querySelector("#textarea-validate-current-dataset");
-// const currentDatasetReportBtn = document.getElementById("button-generate-report-current-ds");
-// const validateLocalDSBtn = document.getElementById("button-validate-local-ds");
-// const validateLocalDatasetReport = document.querySelector("#textarea-validate-local-dataset");
-// const localDatasetReportBtn = document.getElementById("button-generate-report-local-ds");
-// const validateLocalProgressBar = document.getElementById("div-indetermiate-bar-validate-local");
-// const validateSODAProgressBar = document.getElementById("div-indetermiate-bar-validate-soda");
+// Validate dataset //
+const validateCurrentDSBtn = document.getElementById("button-validate-current-ds");
+const validateCurrentDatasetReport = document.querySelector("#textarea-validate-current-dataset");
+const currentDatasetReportBtn = document.getElementById("button-generate-report-current-ds");
+const validateLocalDSBtn = document.getElementById("button-validate-local-ds");
+const validateLocalDatasetReport = document.querySelector("#textarea-validate-local-dataset");
+const localDatasetReportBtn = document.getElementById("button-generate-report-local-ds");
+const validateLocalProgressBar = document.getElementById("div-indetermiate-bar-validate-local");
+const validateSODAProgressBar = document.getElementById("div-indetermiate-bar-validate-soda");
 
[…]
