+++ src/renderer/src/scripts/others/renderer.js
@@ -6609,8 +6609,8 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
       let monitorProgressResponse = await client.get(`/organize_datasets/datasets/import/progress`);
 
       let { data } = monitorProgressResponse;
-      percentage_amount = data["progress_percentage"].toFixed(2);
-      finished = data["create_soda_json_completed"];
+      let percentage_amount = data["progress_percentage"].toFixed(2);
+      let finished = data["create_soda_json_completed"];
 
       numb.innerText = percentage_amount + "%";
       if (percentage_amount <= 50) {
@@ -6628,7 +6628,7 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
         clearInterval(local_progress);
         progressBar_rightSide.classList.remove("notransition");
         window.populate_existing_folders(window.datasetStructureJSONObj);
-        window.populate_existing_metadata(sodaJSONObj);
+        window.populate_existing_metadata(window.sodaJSONObj);
         $("#para-continue-location-dataset-getting-started").text("Please continue below.");
         $("#nextBtn").prop("disabled", false);
         // log the success to analytics
@@ -6652,14 +6652,14 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
   // Function begins here
   if (filepath.length > 0) {
     if (filepath != null) {
-      sodaJSONObj["starting-point"]["local-path"] = "";
+      window.sodaJSONObj["starting-point"]["local-path"] = "";
       document.getElementById("input-destination-getting-started-locally").placeholder =
         filepath[0];
       if (
-        sodaJSONObj["starting-point"]["type"] === "local" &&
-        sodaJSONObj["starting-point"]["local-path"] == ""
+        window.sodaJSONObj["starting-point"]["type"] === "local" &&
+        window.sodaJSONObj["starting-point"]["local-path"] == ""
       ) {
-        valid_dataset = window.verify_sparc_folder(
+        let valid_dataset = window.verify_sparc_folder(
           document.getElementById("input-destination-getting-started-locally").placeholder,
           "local"
         );
@@ -6711,7 +6711,7 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
               } else {
                 document.getElementById("input-destination-getting-started-locally").placeholder =
                   "Browse here";
-                sodaJSONObj["starting-point"]["local-path"] = "";
+                window.sodaJSONObj["starting-point"]["local-path"] = "";
                 $("#para-continue-location-dataset-getting-started").text("");
                 return;
               }
@@ -6725,7 +6725,7 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
               document.getElementById("loading_local_dataset").style.display = "block";
 
               // Show file path to user in the input box
-              sodaJSONObj["starting-point"]["local-path"] = filepath[0];
+              window.sodaJSONObj["starting-point"]["local-path"] = filepath[0];
               let root_folder_path = $("#input-destination-getting-started-locally").attr(
                 "placeholder"
               );
@@ -6737,7 +6737,7 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
                 let importLocalDatasetResponse = await client.post(
                   `/organize_datasets/datasets/import`,
                   {
-                    sodajsonobject: sodaJSONObj,
+                    sodajsonobject: window.sodaJSONObj,
                     root_folder_path: root_folder_path,
                     irregular_folders: window.irregularFolderArray,
                     replaced: replaced,
@@ -6745,8 +6745,8 @@ window.electron.ipcRenderer.on("selected-local-destination-datasetCurate", async
                   { timeout: 0 }
                 );
[…]
