### 2023-11-28 d88f119
wip: organize ds manifest creation done
