+++ src/renderer/src/scripts/organize-dataset/curate-functions.js
@@ -4,6 +4,7 @@ import { updateDatasetList } from "../globals"
 import determineDatasetLocation, { Destinations } from "../analytics/analytics-utils"
 import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
 import { successCheck } from "../../assets/lotties/lotties";
+import client from "../client"
 
 
 while (!window.htmlPagesAdded) {
@@ -87,8 +88,8 @@ window.confirmMetadataFilePath = (ev) => {
       [
         "Step 4",
         "Import",
-        `${getMetadataFileNameFromStatus(metadataFileStatus)}`,
-        determineLocationFromStatus(metadataFileStatus)
+        `${window.getMetadataFileNameFromStatus(metadataFileStatus)}`,
+        window.determineLocationFromStatus(metadataFileStatus)
           ? Destinations.PENNSIEVE
           : Destinations.LOCAL,
       ],
@@ -105,8 +106,8 @@ window.confirmMetadataFilePath = (ev) => {
       [
         "Step 4",
         "Import",
-        `${getMetadataFileNameFromStatus(metadataFileStatus)}`,
-        determineLocationFromStatus(metadataFileStatus)
+        `${window.getMetadataFileNameFromStatus(metadataFileStatus)}`,
+        window.determineLocationFromStatus(metadataFileStatus)
           ? Destinations.PENNSIEVE
           : Destinations.LOCAL,
       ],
@@ -1693,7 +1694,7 @@ const generateFFManifestEditCard = (highLevelFolderName) => {
           height: 3rem;
           font-size: 13px;
         "
-        onClick="ffOpenManifestEditSwal('${highLevelFolderName}')"
+        onClick="window.ffOpenManifestEditSwal('${highLevelFolderName}')"
       >
         Preview/Edit ${highLevelFolderName} manifest file
       </button>
@@ -1703,7 +1704,7 @@ const generateFFManifestEditCard = (highLevelFolderName) => {
 };
 
 const renderFFManifestCards = () => {
-  const manifestData = sodaCopy["manifest-files"];
+  const manifestData = window.sodaCopy["manifest-files"];
   const highLevelFoldersWithManifestData = Object.keys(manifestData);
 
   const manifestCards = highLevelFoldersWithManifestData
@@ -1718,11 +1719,11 @@ const renderFFManifestCards = () => {
   window.smoothScrollToElement(manifestFilesCardsContainer);
 };
 
-const ffOpenManifestEditSwal = async (highlevelFolderName) => {
+window.ffOpenManifestEditSwal = async (highlevelFolderName) => {
   let saveManifestFiles = false;
   let guidedManifestTable = [];
   // Function for when user wants to edit the manifest cards
-  const existingManifestData = sodaCopy["manifest-files"]?.[highlevelFolderName];
+  const existingManifestData = window.sodaCopy["manifest-files"]?.[highlevelFolderName];
 
   let ffmManifestContainer = document.getElementById("ffm-container-manifest-file-cards").children;
   //Lock manifest buttons
@@ -1730,19 +1731,19 @@ const ffOpenManifestEditSwal = async (highlevelFolderName) => {
     ffmManifestContainer[i].children[1].children[0].disabled = true;
   }
 
-  ipcRenderer.invoke("spreadsheet", existingManifestData);
+  window.electron.ipcRenderer.invoke("spreadsheet", existingManifestData);
 
   //upon receiving a reply of the spreadsheet, handle accordingly
-  ipcRenderer.on("spreadsheet-reply", async (event, result) => {
+  window.electron.ipcRenderer.on("spreadsheet-reply", async (event, result) => {
     for (let i = 0; i < ffmManifestContainer.length; i++) {
[…]
