### 2023-11-28 a5c62d3
fix: 🐛 fixed pool addition variable name bug
