### 2023-11-28 f92c71e
wip: organize ds ffm finished ps upload
