### 2023-11-28 e0c9eca
fix: 🐛 Select manual structure button for datasets with subjects
