### 2023-11-29 64b33c9
feat: ✨ replace sub- with proper entity name in swal
