### 2023-11-29 5a8b8ee
refactor: organize datasets FFM
