+++ scripts/guided-mode/guided-curate-dataset.js
@@ -6910,6 +6910,103 @@ const openPage = async (targetPageID) => {
 
   guidedSetNavLoadingState(false);
 };
+const guidedOpenEntityEditSwal = async (entityName) => {
+  let preExistingEntities;
+  let entityNameSingular;
+  let entityPrefix;
+
+  if (entityName.startsWith("sub-")) {
+    preExistingEntities = getExistingSubjectNames();
+    entityNameSingular = "subject";
+    entityPrefix = "sub-";
+  }
+  if (entityName.startsWith("pool-")) {
+    preExistingEntities = getExistingPoolNames();
+    entityNameSingular = "pool";
+    entityPrefix = "pool-";
+  }
+  if (entityName.startsWith("sam-")) {
+    preExistingEntities = getExistingSampleNames();
+    entityNameSingular = "sample";
+    entityPrefix = "sam-";
+  }
+
+  let newEntityName;
+
+  const entityEditConfirmed = await Swal.fire({
+    title: `Editing ${entityNameSingular} ${entityName}`,
+    html: `
+      <p class="help-text text-center">
+        Enter a new ${entityNameSingular} name in the input below.
+        <br />
+      </p>
+      <div class="space-between w-100 align-flex-center">
+        <p class="help-text m-0 mr-1">${entityPrefix}</p>
+        <input value="${entityName.replace(
+          entityPrefix,
+          ""
+        )}" id='input-new-entity-name' class='guided--input' type='text' placeholder='Enter new subject name and press edit'/>
+      </div>
+    `,
+    width: 800,
+    heightAuto: false,
+    backdrop: "rgba(0,0,0,0.4)",
+    showConfirmButton: true,
+    showCancelButton: true,
+    showCloseButton: false,
+    confirmButtonText: `Edit`,
+    cancelButtonText: `Cancel`,
+    didOpen: () => {
+      // Add event listener to the input to enable the confirm button when the input is changed
+      document.getElementById("input-new-entity-name").addEventListener("keyup", () => {
+        Swal.resetValidationMessage();
+        Swal.enableButtons();
+      });
+    },
+    preConfirm: () => {
+      let newEntityInputValue = document.getElementById("input-new-entity-name").value;
+      if (newEntityInputValue.length === 0) {
+        Swal.showValidationMessage(`Please enter a new ${entityNameSingular} name`);
+        return;
+      }
+
+      newEntityName = `${entityPrefix}${newEntityInputValue}`;
+      if (newEntityName === entityName) {
+        console.log("No changes made");
+        Swal.close();
+      }
+      const entityNameIsValid = evaluateStringAgainstSdsRequirements(
+        newEntityName,
+        "string-adheres-to-identifier-conventions"
+      );
+      if (!entityNameIsValid) {
+        Swal.showValidationMessage(
+          `${entityNameSingular} names can not contain spaces or special characters`
+        );
+        return;
+      }
+      if (preExistingEntities.includes(newEntityName)) {
+        Swal.showValidationMessage(`A ${entityNameSingular} with that name already exists`);
+        return;
+      }
+    },
+  });
+
+  if (entityEditConfirmed.isConfirmed) {
+    if (entityName.startsWith("sub-")) {
+      sodaJSONObj.renameSubject(entityName, newEntityName);
+      renderSubjectsTable();
+    }
+    if (entityName.startsWith("pool-")) {
+      sodaJSONObj.renamePool(entityName, newEntityName);
+      renderPoolsTable();
+    }
+    if (entityName.startsWith("sam-")) {
+      sodaJSONObj.renameSample(entityName, newEntityName);
+      renderSamplesTable();
+    }
+  }
+};
 
 const renderSubjectsTable = () => {
   const [subjectsInPools, subjectsOutsidePools] = sodaJSONObj.getAllSubjects();
[…]
