### 2023-11-29 a8171ee
refactor: ffm standalone validator local dataset
