+++ scripts/validator/validate.js
@@ -194,7 +194,7 @@ const validateLocalDataset = async () => {
     clientError(error);
     file_counter = 0;
     folder_counter = 0;
-    get_num_files_and_folders(localSodaJsonObject["dataset-structure"]);
+    window.get_num_files_and_folders(localSodaJsonObject["dataset-structure"]);
     // log successful validation run to analytics
     ipcRenderer.send(
       "track-event",
@@ -267,7 +267,7 @@ const validateLocalDataset = async () => {
 
   file_counter = 0;
   folder_counter = 0;
-  get_num_files_and_folders(localSodaJsonObject["dataset-structure"]);
+  window.get_num_files_and_folders(localSodaJsonObject["dataset-structure"]);
   // log successful validation run to analytics
   ipcRenderer.send(
     "track-event",
@@ -448,7 +448,7 @@ const validatePennsieveDatasetStandAlone = async () => {
     clientError(error);
     file_counter = 0;
     folder_counter = 0;
-    get_num_files_and_folders(localSodaJSONObj["dataset-structure"]);
+    window.get_num_files_and_folders(localSodaJSONObj["dataset-structure"]);
     // hide the validation question
     $("#validate_dataset-question-4").removeClass("show");
     // log successful validation run to analytics
@@ -534,7 +534,7 @@ const validatePennsieveDatasetStandAlone = async () => {
 
   file_counter = 0;
   folder_counter = 0;
-  get_num_files_and_folders(localSodaJSONObj["dataset-structure"]);
+  window.get_num_files_and_folders(localSodaJSONObj["dataset-structure"]);
   // log successful validation run to analytics
   ipcRenderer.send(
     "track-event",
