+++ src/renderer/src/scripts/others/renderer.js
@@ -7897,7 +7897,7 @@ const show_curation_shortcut = async () => {
   });
 };
 
-const get_num_files_and_folders = (dataset_folders) => {
+window.get_num_files_and_folders = (dataset_folders) => {
   if ("files" in dataset_folders) {
     for (let file in dataset_folders["files"]) {
       file_counter += 1;
@@ -7906,7 +7906,7 @@ const get_num_files_and_folders = (dataset_folders) => {
   if ("folders" in dataset_folders) {
     for (let folder in dataset_folders["folders"]) {
       folder_counter += 1;
-      get_num_files_and_folders(dataset_folders["folders"][folder]);
+      window.get_num_files_and_folders(dataset_folders["folders"][folder]);
     }
   }
   return;
