### 2023-11-29 06e7897
feat: ✨ Updated entity edit text
