### 2023-11-29 20927a5
fix: 🐛 entity replacement input placeholder text fix
