### 2023-11-30 b5ef09d
feat: ✨ Updated spreadsheet query flow
