+++ scripts/guided-mode/guided-curate-dataset.js
@@ -1131,7 +1131,7 @@ const savePageChanges = async (pageBeingLeftID) => {
       const datasetHasCode = sodaJSONObj["dataset-contains-code"];
 
       if (datasetHasSubjects) {
-        guidedUnSkipPage("guided-subjects-dataset-structure-specification-tab");
+        guidedUnSkipPage("guided-subject-structure-spreadsheet-importation-tab");
         guidedUnSkipPage("guided-subjects-addition-tab");
 
         guidedUnSkipPage("guided-primary-data-organization-tab");
@@ -1140,7 +1140,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         guidedUnSkipPage("guided-create-subjects-metadata-tab");
         guidedUnSkipPage("guided-create-samples-metadata-tab");
       } else {
-        guidedSkipPage("guided-subjects-dataset-structure-specification-tab");
+        guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
         guidedSkipPage("guided-subjects-addition-tab");
         guidedSkipPage("guided-subjects-pooling-tab");
         guidedSkipPage("guided-samples-addition-tab");
@@ -1231,82 +1231,82 @@ const savePageChanges = async (pageBeingLeftID) => {
       }
     }
 
-    if (pageBeingLeftID === "guided-subjects-dataset-structure-specification-tab") {
-      const userSelectedDatasetHasPools = document
-        .getElementById("guided-button-subjects-are-pooled")
-        .classList.contains("selected");
-      const userSelectedDatasetDoesNotHavePools = document
-        .getElementById("guided-button-subjects-are-not-pooled")
-        .classList.contains("selected");
-      if (!userSelectedDatasetHasPools && !userSelectedDatasetDoesNotHavePools) {
-        errorArray.push({
-          type: "notyf",
-          message: "Please indicate whether or not the dataset contains pools",
-        });
-        throw errorArray;
-      }
-      if (userSelectedDatasetHasPools) {
-        guidedUnSkipPage("guided-subjects-pooling-tab");
-        //Unskip the pool data organization pages
-        guidedUnSkipPage(`guided-primary-pools-organization-page`);
-        guidedUnSkipPage(`guided-source-pools-organization-page`);
-        guidedUnSkipPage(`guided-derivative-pools-organization-page`);
-      } else {
-        guidedSkipPage("guided-subjects-pooling-tab");
-        //Skip the pool data organization pages
-        guidedSkipPage(`guided-primary-pools-organization-page`);
-        guidedSkipPage(`guided-source-pools-organization-page`);
-        guidedSkipPage(`guided-derivative-pools-organization-page`);
-      }
-
-      const userSelectedSubjectsHaveSamples = document
-        .getElementById("guided-button-subjects-have-samples")
-        .classList.contains("selected");
-      const userSelectedSubjectsDoNotHaveSamples = document
-        .getElementById("guided-button-subjects-do-not-have-samples")
-        .classList.contains("selected");
-      if (!userSelectedSubjectsHaveSamples && !userSelectedSubjectsDoNotHaveSamples) {
-        errorArray.push({
-          type: "notyf",
-          message: "Please indicate whether or not the dataset contains samples",
-        });
-        throw errorArray;
-      }
-
-      if (userSelectedSubjectsHaveSamples) {
-        guidedUnSkipPage("guided-samples-addition-tab");
-        //Unskip the sample data organization pages
-        guidedUnSkipPage(`guided-primary-samples-organization-page`);
-        guidedUnSkipPage(`guided-source-samples-organization-page`);
-        guidedUnSkipPage(`guided-derivative-samples-organization-page`);
-        //Unskip the samples metadata page
-        guidedUnSkipPage(`guided-create-samples-metadata-tab`);
-      } else {
[…]
