### 2023-11-30 6d8a93a
wip: splash screen and startup refinement
