+++ src/main/index.js
@@ -1,22 +1,39 @@
 import { app, shell, BrowserWindow, ipcMain } from 'electron'
+import os from 'os'
 import { join } from 'path'
 import { electronApp, optimizer, is } from '@electron-toolkit/utils'
+import { autoUpdater } from "electron-updater";
+import { trackEvent, trackKombuchaEvent } from "./analytics"
 import icon from '../../resources/icon.png?asset'
 import ElectronLog from "electron-log"
 import axios from "axios"
 import fp from "find-free-port"
-import {spawn, execFile} from "node:child_process"
+import { spawn, execFile, spawnSync } from "node:child_process"
 import { existsSync } from 'fs'
 import { JSONStorage } from "node-localstorage";
 import log from 'electron-log/main';
-import  "./manifest-workbook"
+import "./manifest-workbook"
 import "./banner-image"
 import './node-storage'
-
 import "./main-process/native-ui/dialogs/open-file"
+
+const sodaVersion = app.getVersion();
+// If the version includes "beta", the app will not check for updates
+const buildIsBeta = sodaVersion.includes("beta");
+if (buildIsBeta) {
+  log.info("This is a beta build. Updates will not be checked.");
+}
+autoUpdater.channel = buildIsBeta ? "beta" : "latest";
+autoUpdater.logger = log;
+
+// setup event tracking
+global.trackEvent = trackEvent;
+global.trackKombuchaEvent = trackKombuchaEvent;
 
 // Optional, initialize the logger for any renderer process
 log.initialize({ preload: true });
+log.transports.console.level = false;
+log.transports.file.level = "debug";
 
 
 let nodeStorage = new JSONStorage(app.getPath("userData"))
@@ -30,7 +47,7 @@ console.log("Test up[date")
 
 // TODO: move to a separate file that handles all the ipcMain handlers
 ipcMain.handle('get-app-path', async (event, arg) => {
-  if(arg) return app.getPath(arg) 
+  if (arg) return app.getPath(arg)
   return app.getAppPath()
 })
 
@@ -59,44 +76,6 @@ ipcMain.handle("relaunch-soda", () => {
 ipcMain.handle("exit-soda", () => {
   app.exit();
 })
-
-
-
-
-function createWindow() {
-  // Create the browser window.
-  const mainWindow = new BrowserWindow({
-    width: 900,
-    height: 670,
-    show: false,
-    nodeIntegration: true,
-    autoHideMenuBar: true,
-    ...(process.platform === 'linux' ? { icon } : {}),
-    webPreferences: {
-      preload: join(__dirname, '../preload/index.js'),
-      sandbox: false,
-      contextIsolation: true,
-      webSecurity: false // TODO: set to true and make the Python server a proxy to add CORS headers
-    }
-  })
-
-  mainWindow.on('ready-to-show', () => {
-    mainWindow.show()
-  })
-
-  mainWindow.webContents.setWindowOpenHandler((details) => {
-    shell.openExternal(details.url)
-    return { action: 'deny' }
-  })
-
-  // HMR for renderer base on electron-vite cli.
-  // Load the remote URL for development or the local html file for production.
-  if (is.dev && process.env['ELECTRON_RENDERER_URL']) {
-    mainWindow.loadURL(process.env['ELECTRON_RENDERER_URL'])
-  } else {
-    mainWindow.loadFile(join(__dirname, '../renderer/index.html'))
-  }
-
 
 // passing in the spreadsheet data to pass to a modal
 // that will have a jspreadsheet for user edits
@@ -146,46 +125,8 @@ ipcMain.handle("spreadsheet", (event, spreadsheet) => {
       console.log(e);
     }
   });
-});
-
-
-
-}
-
-// This method will be called when Electron has finished
[…]
