### 2023-11-30 05b5175
feat: ✨ Added yes/no questions on sample and pool pages
