### 2023-12-01 481c160
build: update version numbers for beta build
