+++ src/main/index.js
@@ -400,7 +400,13 @@ const initialize = () => {
     })
 
     // TODO: Add dev check for this path
-    splashScreen.loadURL(process.env['ELECTRON_RENDERER_URL'] +  "/splash/splash-screen.html")
+    // HMR for renderer base on electron-vite cli.
+    // Load the remote URL for development or the local html file for production.
+    if (is.dev && process.env['ELECTRON_RENDERER_URL']) {
+      splashScreen.loadURL(process.env['ELECTRON_RENDERER_URL'] + "/splash/splash-screen.html")
+    } else {
+      splashScreen.loadFile(join(__dirname, '../renderer/public/splash/splash-screen.html'))
+    }
 
 
     splashScreen.once("ready-to-show", () => {
