### 2023-12-01 35c9f99
refactor: log renderer file
