+++ electron.vite.config.js
@@ -3,6 +3,7 @@ import { defineConfig, externalizeDepsPlugin } from 'electron-vite'
 import react from '@vitejs/plugin-react'
 import copy from 'rollup-plugin-copy'
 import inject from "@rollup/plugin-inject";
+import commonjs from '@rollup/plugin-commonjs';
 
 const commonjsPackages = ['image-data-uri']
 
@@ -25,7 +26,7 @@ export default defineConfig({
     plugins: [react(), inject({
       $: 'jquery',
       jQuery: 'jquery'
-    })], 
+    }), commonjs()], 
     optimizeDeps: {
       exclude: ['bootbox', 'Jimp/es']
     },
