### 2023-12-01 aec5192
fix: synchronicity issue with dependencies
