+++ src/renderer/src/main.js
@@ -2,8 +2,8 @@ import './assets/imports'
 import './assets/demo-btns'
 import './assets/nav'
 import './scripts/client'
-import './scripts/guided-mode/guided-curate-dataset'
 import './scripts/globals'
+import './scripts/guided-mode/guided-curate-dataset'
 import './scripts/others/renderer'
 import './scripts/metadata-files/submission'
 import './scripts/manage-dataset/manage-dataset'
