### 2023-12-01 b58a07a
build: halt win and mac builds
