### 2023-12-01 01fc353
fix: select2 is not a function
