### 2023-12-04 f8cf0b4
fix: splash screen path for builds
