### 2023-12-04 dca9ab2
refactor: nav bar dynamic resizing
