### 2023-12-04 af779e8
style: 🎨 fix code style issues with Prettier
