+++ scripts/guided-mode/guided-curate-dataset.js
@@ -871,7 +871,8 @@ const savePageChanges = async (pageBeingLeftID) => {
         sodaJSONObj["digital-metadata"]["dataset-workspace"] = guidedGetCurrentUserWorkSpace();
         guidedSkipPage("guided-pennsieve-intro-tab");
         // Skip the subject structure import page since the user is starting from Pennsieve
-        guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");      }
+        guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
+      }
 
       //Skip this page becausae we should not come back to it
       guidedSkipPage("guided-select-starting-point-tab");
@@ -1373,8 +1374,6 @@ const savePageChanges = async (pageBeingLeftID) => {
         // Skip the page so the user can add subjects manually
         guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
       }
-
-      
     }
 
     if (pageBeingLeftID === "guided-primary-data-organization-tab") {
