### 2023-12-04 526278e
feat: ✨ match spreadsheet id format with sds format
