### 2023-12-04 ca76b2e
fix: swal not defined in check for announcements
