+++ pyflask/startup/minimumApiVersion.py
@@ -4,4 +4,4 @@ def get_api_version():
     """
     Returns the version of the API
     """
-    return {'version': os.getenv('API_VERSION', "12.4.5-beta")}
+    return {'version': os.getenv('API_VERSION', "12.4.0-beta")}
