+++ .github/workflows/Build-and-deploy-mac.yml
@@ -5,6 +5,8 @@ on:
   push:
     branches:
       - main
+      - staging
+      
 jobs:
   deploy-on-mac:
     runs-on: macos-latest
