### 2023-12-04 8cd9efb
fix: auto updater launching twice
