+++ src/main/index.js
@@ -533,11 +533,7 @@ const initialize = () => {
     // show the splash screen
 
 
-    app.on('activate', function () {
-      // On macOS it's common to re-create a window in the app when the
-      // dock icon is clicked and there are no other windows open.
-      if (BrowserWindow.getAllWindows().length === 0) createWindow()
-    })
+ 
 
     // track app launch at Kombucha analytics server
     // trackKombuchaEvent(
