### 2023-12-04 36f995f
fix: 🐛 updated spreadsheet skipping logic
