+++ scripts/guided-mode/guided-curate-dataset.js
@@ -870,8 +870,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         // Skip the page where they confirm their log in and workspace because we should already have it
         sodaJSONObj["digital-metadata"]["dataset-workspace"] = guidedGetCurrentUserWorkSpace();
         guidedSkipPage("guided-pennsieve-intro-tab");
-        // Skip the subject structure import page since the user is starting from Pennsieve
-        guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
+
       }
 
       //Skip this page becausae we should not come back to it
@@ -1133,7 +1132,6 @@ const savePageChanges = async (pageBeingLeftID) => {
       const datasetHasCode = sodaJSONObj["dataset-contains-code"];
 
       if (datasetHasSubjects) {
-        guidedUnSkipPage("guided-subject-structure-spreadsheet-importation-tab");
         guidedUnSkipPage("guided-subjects-addition-tab");
         guidedUnSkipPage("guided-subjects-pooling-tab");
         guidedUnSkipPage("guided-samples-addition-tab");
@@ -1144,7 +1142,6 @@ const savePageChanges = async (pageBeingLeftID) => {
         guidedUnSkipPage("guided-create-subjects-metadata-tab");
         guidedUnSkipPage("guided-create-samples-metadata-tab");
       } else {
-        guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
         guidedSkipPage("guided-subjects-addition-tab");
         guidedSkipPage("guided-subjects-pooling-tab");
         guidedSkipPage("guided-samples-addition-tab");
@@ -1370,10 +1367,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         }
       }
 
-      if (userChoseToEnterSubsSamsPoolsManually) {
-        // Skip the page so the user can add subjects manually
-        guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
-      }
+     
     }
 
     if (pageBeingLeftID === "guided-primary-data-organization-tab") {
@@ -5108,6 +5102,16 @@ const openPage = async (targetPageID) => {
       importProgressCircle.classList.add("hidden");
     }
 
+    if (targetPageID === "guided-dataset-structure-intro-capsule") {
+      if (getExistingSubjectNames().length === 0 &&
+       sodaJSONObj["starting-point"]["type"] != "bf" && 
+       sodaJSONObj["button-config"]["dataset-contains-subjects"] === "yes") {
+        guidedUnSkipPage("guided-subject-structure-spreadsheet-importation-tab")
+      } else {
+          guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab")
+        }
+    }
+
     if (targetPageID === "guided-subject-structure-spreadsheet-importation-tab") {
       const savedSpreadSheetPath = sodaJSONObj["dataset-structure-spreadsheet-path"];
       setUiBasedOnSavedDatasetStructurePath(savedSpreadSheetPath);
@@ -5286,6 +5290,7 @@ const openPage = async (targetPageID) => {
     }
 
     if (targetPageID === "guided-subjects-addition-tab") {
+      guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab")
       renderSubjectsTable();
     }
 
@@ -7826,11 +7831,6 @@ const patchPreviousGuidedModeVersions = async () => {
 
   if (!sodaJSONObj["curation-mode"]) {
     sodaJSONObj["cuartion-mode"] = "guided";
-  }
-
-  // Skip the subject spreadsheet importatin page if subjects have already been imported
-  if (getExistingSubjectNames().length > 0) {
-    guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab");
   }
 
[…]
