### 2023-12-04 1288c91
fix: checking for updates twice one launch
