+++ scripts/metadata-files/downloadTemplates.js
@@ -99,7 +99,6 @@ downloadManifest.addEventListener("click", (event) => {
 
 document.querySelectorAll(".guided-data-deliverables-download-button").forEach((button) => {
   button.addEventListener("click", (event) => {
-    console.log("clicked");
     ipcRenderer.send("open-folder-dialog-save-metadata", "code_description.xlsx");
   });
 });
