+++ scripts/guided-mode/guided-curate-dataset.js
@@ -870,7 +870,6 @@ const savePageChanges = async (pageBeingLeftID) => {
         // Skip the page where they confirm their log in and workspace because we should already have it
         sodaJSONObj["digital-metadata"]["dataset-workspace"] = guidedGetCurrentUserWorkSpace();
         guidedSkipPage("guided-pennsieve-intro-tab");
-
       }
 
       //Skip this page becausae we should not come back to it
@@ -1366,8 +1365,6 @@ const savePageChanges = async (pageBeingLeftID) => {
           throw errorArray;
         }
       }
-
-     
     }
 
     if (pageBeingLeftID === "guided-primary-data-organization-tab") {
@@ -4267,7 +4264,6 @@ const guidedSkipPage = (pageId) => {
   if (page.classList.contains("guided--page")) {
     // replace -tab with -capsule  in pageId string
     const pagesCapsule = pageId.replace("-tab", "-capsule");
-    console.log(pagesCapsule);
     document.getElementById(pagesCapsule).classList.add("hidden");
   }
   if (page.classList.contains("sub-page")) {
@@ -4287,8 +4283,6 @@ const guidedUnSkipPage = (pageId) => {
   // If the page no longer exists, return
   if (!page) {
     return;
-  } else {
-    console.log(pageId);
   }
 
   page.dataset.skipPage = "false";
@@ -4296,10 +4290,9 @@ const guidedUnSkipPage = (pageId) => {
   //Show the parent page or sub page capsule
   if (page.classList.contains("guided--page")) {
     const pagesCapsuleID = pageId.replace("-tab", "-capsule");
-    console.log(pagesCapsuleID);
     const domElement = document.getElementById(pagesCapsuleID);
     if (!domElement) {
-      console.log(`Could not find element with id ${pagesCapsuleID}`);
+      console.error(`Could not find element with id ${pagesCapsuleID}`);
     }
     document.getElementById(pagesCapsuleID).classList.remove("hidden");
   }
@@ -5102,20 +5095,6 @@ const openPage = async (targetPageID) => {
       importProgressCircle.classList.add("hidden");
     }
 
-    if (targetPageID === "guided-dataset-structure-intro-capsule") {
-      if (getExistingSubjectNames().length === 0 &&
-       sodaJSONObj["starting-point"]["type"] != "bf" && 
-       sodaJSONObj["button-config"]["dataset-contains-subjects"] === "yes") {
-        guidedUnSkipPage("guided-subject-structure-spreadsheet-importation-tab")
-      } else {
-          guidedSkipPage("guided-subject-structure-spreadsheet-importation-tab")
-        }
-    }
-
-    if (targetPageID === "guided-subject-structure-spreadsheet-importation-tab") {
-      const savedSpreadSheetPath = sodaJSONObj["dataset-structure-spreadsheet-path"];
-      setUiBasedOnSavedDatasetStructurePath(savedSpreadSheetPath);
-    }
     if (targetPageID === "guided-prepare-dataset-structure-tab") {
       // If the user has already added subjects, disallow them from selecting no (they have to go to the subject
       // page to delete subjects but this would be a very strange case anyways)
@@ -5289,8 +5268,26 @@ const openPage = async (targetPageID) => {
       $("#guided-select-sparc-funding-consortium").trigger("change");
     }
 
+    if (targetPageID === "guided-dataset-structure-intro-tab") {
+      if (
+        getExistingSubjectNames().length === 0 &&
+        sodaJSONObj["starting-point"]["type"] != "bf" &&
+        sodaJSONObj["button-config"]["dataset-contains-subjects"] === "yes"
+      ) {
+        guidedUnSkipPage("guided-subject-structure-spreadsheet-importation-tab");
+      } else {
[…]
