### 2023-12-04 dd79ee4
wip: 🚧 Prepare channges and announcements
