+++ scripts/guided-mode/guided-curate-dataset.js
@@ -5269,6 +5269,10 @@ const openPage = async (targetPageID) => {
     }
 
     if (targetPageID === "guided-dataset-structure-intro-tab") {
+      // Handle whether or not the spreadsheet importation page should be skipped
+      // Note: this is done here to centralize the logic for skipping the page
+      // The page is unskipped only if the user has not added any subjects,
+      // indicated that they will be adding subjects, and the user is not starting from Pennsieve
       if (
         getExistingSubjectNames().length === 0 &&
         sodaJSONObj["starting-point"]["type"] != "bf" &&
