+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Feature Additions:
 
-- Added a new feature to Guided Mode that allows you to import subjects, pools, and samples for datasets with a large amount of subjects/samples
+- Added a new feature to Guided Mode that allows you to import subjects, pools, and samples from an external spreadsheet.
 
 ## v.12.3.3 - 2023-10-24
 
