### 2023-12-05 8d99f0b
build: change mac plist path
