### 2023-12-05 f7e527c
build: spawn mac builds
