### 2023-12-05 faefb1f
feat: Converted gm dataset name and subtitle page to react/zustand
