+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -1,33 +1,45 @@
 // // sourcery skip: merge-nested-ifs
 
-import { existingDataset, modifyDataset, blackArrow, questionList, datasetMetadataIntroLottie, addScienceData, startNew, resumeExisting } from "../../assets/lotties/lotties";
+import {
+  existingDataset,
+  modifyDataset,
+  blackArrow,
+  questionList,
+  datasetMetadataIntroLottie,
+  addScienceData,
+  startNew,
+  resumeExisting,
+} from "../../assets/lotties/lotties";
 import { resetLazyLoading, guidedUnLockSideBar } from "../../assets/nav";
-import determineDatasetLocation from "../analytics/analytics-utils"
-import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
+import determineDatasetLocation from "../analytics/analytics-utils";
+import { clientError, userErrorMessage } from "../others/http-error-handler/error-handler";
 import datasetUploadSession from "../analytics/upload-session-tracker";
-import api from '../others/api/api'
-import lottie from "lottie-web"
-import {dragDrop, successCheck, errorMark} from '../../assets/lotties/lotties'
-import kombuchaEnums from '../analytics/analytics-enums'
-import Swal from "sweetalert2"
+import api from "../others/api/api";
+import lottie from "lottie-web";
+import { dragDrop, successCheck, errorMark } from "../../assets/lotties/lotties";
+import kombuchaEnums from "../analytics/analytics-enums";
+import Swal from "sweetalert2";
 import Tagify from "@yaireo/tagify/dist/tagify.esm";
 // const Tagify = require("@yaireo/tagify/dist/tagify.esm");
 import tippy from "tippy.js";
-import client from '../client'
-import jQuery from 'jquery'
-import bootstrap from 'bootstrap'
-import 'bootstrap-select'
-import DragSort from '@yaireo/dragsort'
-import Cropper from 'cropperjs'
-
-import 'jstree'
-
-
+import client from "../client";
+import jQuery from "jquery";
+import bootstrap from "bootstrap";
+import "bootstrap-select";
+import DragSort from "@yaireo/dragsort";
+import Cropper from "cropperjs";
+import "jstree";
+import useTestStore from "../../stores/testStore";
 
 while (!window.htmlPagesAdded) {
-  await new Promise((resolve) => setTimeout(resolve, 100))
+  await new Promise((resolve) => setTimeout(resolve, 100));
 }
 
+const { getState, setState, subscribe, destroy } = useTestStore;
+
+// for testing remove later
+window.getState = getState;
+window.setState = setState;
 
 const returnToGuided = () => {
   document.getElementById("guided_mode_view").click();
@@ -254,7 +266,9 @@ const guidedMovePennsieveFolder = (movedFolderName, folderJSONPath, newFolderJSO
 
 // Returns a boolean that indicates whether or not the user selected that the dataset is SPARC funded
 const datasetIsSparcFunded = () => {
-  return window.sodaJSONObj["dataset-metadata"]["submission-metadata"]["funding-consortium"] === "SPARC";
+  return (
+    window.sodaJSONObj["dataset-metadata"]["submission-metadata"]["funding-consortium"] === "SPARC"
+  );
 };
 
 const checkIfPoolsFoldersAreCorrect = (poolFolderPath) => {
@@ -372,7 +386,10 @@ const checkIfChangesMetadataPageShouldBeShown = async (pennsieveDatasetID) => {
   } catch (error) {
     const emessage = userErrorMessage(error);
 
-    const datasetInfo = await api.getDatasetInformation(window.defaultBfAccount, pennsieveDatasetID);
+    const datasetInfo = await api.getDatasetInformation(
+      window.defaultBfAccount,
[…]
