+++ scripts/guided-mode/guided-curate-dataset.js
@@ -986,7 +986,7 @@ const savePageChanges = async (pageBeingLeftID) => {
       }
     }
 
-    if (pageBeingLeftID === "guided-name-subtitle-tab") {
+    /*if (pageBeingLeftID === "guided-name-subtitle-tab") {
       let datasetNameInput = document.getElementById("guided-dataset-name-input").value.trim();
       let datasetSubtitleInput = document
         .getElementById("guided-dataset-subtitle-input")
@@ -1044,7 +1044,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         sodaJSONObj["digital-metadata"]["name"] = datasetNameInput;
         sodaJSONObj["digital-metadata"]["subtitle"] = datasetSubtitleInput;
       }
-    }
+    }*/
 
     if (pageBeingLeftID === "guided-ask-if-submission-is-sparc-funded-tab") {
       // NOTE: We use the button config generated by this page to determine if the dataset is SPARC funded
