+++ src/renderer/src/stores/testStore.js
@@ -1,0 +1,13 @@
+import { create } from "zustand";
+
+const useTestStore = create((set) => ({
+  name: "",
+  setName: (name) => set({ name }),
+
+  subtitle: "",
+  setSubtitle: (subtitle) => set({ subtitle }),
+
+  reset: () => set({ name: "", subtitle: "" }),
+}));
+
+export default useTestStore;
