+++ src/main/index.js
@@ -1,21 +1,21 @@
-import { app, shell, BrowserWindow, ipcMain } from 'electron'
-import os from 'os'
-import { join } from 'path'
-import { electronApp, optimizer, is } from '@electron-toolkit/utils'
+import { app, shell, BrowserWindow, ipcMain } from "electron";
+import os from "os";
+import { join } from "path";
+import { electronApp, optimizer, is } from "@electron-toolkit/utils";
 import { autoUpdater } from "electron-updater";
-import { trackEvent, trackKombuchaEvent } from "./analytics"
-import icon from '../../resources/icon.png?asset'
-import ElectronLog from "electron-log"
-import axios from "axios"
-import fp from "find-free-port"
-import { spawn, execFile, spawnSync } from "node:child_process"
-import { existsSync } from 'fs'
+import { trackEvent, trackKombuchaEvent } from "./analytics";
+import icon from "../../resources/icon.png?asset";
+import ElectronLog from "electron-log";
+import axios from "axios";
+import fp from "find-free-port";
+import { spawn, execFile, spawnSync } from "node:child_process";
+import { existsSync } from "fs";
 import { JSONStorage } from "node-localstorage";
-import log from 'electron-log/main';
-import "./manifest-workbook"
-import "./banner-image"
-import './node-storage'
-import "./main-process/native-ui/dialogs/open-file"
+import log from "electron-log/main";
+import "./manifest-workbook";
+import "./banner-image";
+import "./node-storage";
+import "./main-process/native-ui/dialogs/open-file";
 
 const sodaVersion = app.getVersion();
 // If the version includes "beta", the app will not check for updates
@@ -35,53 +35,50 @@ log.initialize({ preload: true });
 log.transports.console.level = false;
 log.transports.file.level = "debug";
 
-
-let nodeStorage = new JSONStorage(app.getPath("userData"))
+let nodeStorage = new JSONStorage(app.getPath("userData"));
 
 // TODO: Move to ipcMain handler so renderer processes can talk to the nodestorage
 var mainWindow = null;
 
-
 // import "./appUtils"
-console.log("Test up[date")
-
+console.log("Test up[date");
 
 // TODO: move to a separate file that handles all the ipcMain handlers
-ipcMain.handle('get-app-path', async (event, arg) => {
-  if (arg) return app.getPath(arg)
-  return app.getAppPath()
-})
+ipcMain.handle("get-app-path", async (event, arg) => {
+  if (arg) return app.getPath(arg);
+  return app.getAppPath();
+});
 
 ipcMain.handle("get-port", () => {
   log.info("Renderer requested port: " + selectedPort);
-  return selectedPort
+  return selectedPort;
 });
 
 ipcMain.handle("app-version", () => {
   return app.getVersion();
-})
+});
 
 ipcMain.handle("set-nodestorage-key", (key, value) => {
   return nodeStorage.setItem(key, value);
-})
+});
 
 ipcMain.handle("get-nodestorage-key", (key) => {
   return nodeStorage.getItem(key);
-})
+});
 
 ipcMain.handle("relaunch-soda", () => {
   app.relaunch();
   app.exit();
-})
+});
 
 ipcMain.handle("exit-soda", () => {
   app.exit();
-})
+});
 
 // passing in the spreadsheet data to pass to a modal
 // that will have a jspreadsheet for user edits
 ipcMain.handle("spreadsheet", (event, spreadsheet) => {
-  console.log("Spreadsheet invoked")
+  console.log("Spreadsheet invoked");
   const windowOptions = {
     minHeight: 450,
     width: 1120,
@@ -126,8 +123,7 @@ ipcMain.handle("spreadsheet", (event, spreadsheet) => {
       console.log(e);
     }
   });
-})
-
+});
[…]
