+++ src/renderer/src/scripts/client.js
@@ -5,9 +5,8 @@ const port = await window.electron.ipcRenderer.invoke("get-port");
 window.log.info("Port is: " + port);
 
 let client = axios.create({
-    baseURL: `http://127.0.0.1:${port}`,
-    timeout: 300000,
-  });
+  baseURL: `http://127.0.0.1:${port}`,
+  timeout: 300000,
+});
 
-
-  export default client;
+export default client;
