+++ src/renderer/src/scripts/others/update-path-darwin.js
@@ -6,7 +6,6 @@
  */
 const fixPath = () => {
   if (window.process.platform == "darwin") {
-
     window.process.env.PATH = [
       "./node_modules/.bin",
       "/.nodebrew/current/bin",
