### 2023-12-05 3abcd2d
chore: update CITATION.cff for Zenodo release
