+++ CITATION.cff
@@ -21,11 +21,10 @@ authors:
     family-names: Portillo
     given-names: Dorian
 cff-version: 1.2.0
-date-released: '2023-09-21'
+date-released: '2023-12-05'
 identifiers:
   - description: DOI for this software's record on Zenodo
     type: doi
-    value: https://doi.org/10.5281/zenodo.8368612
 keywords:
   - SPARC
   - FAIR
@@ -42,4 +41,4 @@ repository-code: https://github.com/fairdataihub/SODA-for-SPARC
 title: SODA (Software to Organize Data Automatically) for SPARC
 type: software
 url: https://fairdataihub.org/
-version: 12.3.1
+version: 12.4.0
