### 2023-12-06 d004f0d
style: all fontawesome icons reintroduced
