### 2023-12-06 927c3bb
style: set main window max and min size
