### 2023-12-06 db3f559
refactor: sidebar closes and opens when enetering GM
