### 2023-12-06 f834cac
refactor: GM get strain movved to node side
