+++ src/main/index.js
@@ -17,6 +17,7 @@ import "./manifest-workbook"
 import "./banner-image"
 import './node-storage'
 import "./main-process/native-ui/dialogs/open-file"
+import "./strain"
 
 const sodaVersion = app.getVersion();
 // If the version includes "beta", the app will not check for updates
