### 2023-12-07 896e007
refactor: Guided Mode get strain
