### 2023-12-07 e3d2bb0
fix: select2 pools for GM
