### 2023-12-07 145f92f
fix: select2 not working in GM
