### 2023-12-08 15e3730
style: 🎨 fix code style issues with Prettier
