### 2023-12-08 8f97006
feat: ✨ fixed eventData for subtitle analytics
