### 2023-12-08 172d8aa
refactor: contributors accordion btn ffm
