+++ src/renderer/src/scripts/metadata-files/datasetDescription.js
@@ -6,6 +6,7 @@ import kombuchaEnums from "../analytics/analytics-enums";
 import createEventDataPrepareMetadata from "../analytics/prepare-metadata-analytics";
 import api from "../others/api/api"
 import Tagify from "@yaireo/tagify/dist/tagify.esm";
+import tippy from "tippy.js";
 
 
 
@@ -823,7 +824,7 @@ const generateContributorRowElement = (contributorLastName, contributorFirstName
 };
 
 const addContributortoTableDD = (name, contributorObject) => {
-  const contributorIsValid = contributorDataIsValid(contributorObject);
+  const contributorIsValid = window.contributorDataIsValid(contributorObject);
   const conRole = contributorObject.conRole;
   let conTable = document.getElementById("contributor-table-dd");
   let rowcount = conTable.rows.length;
@@ -852,8 +853,8 @@ const addContributortoTableDD = (name, contributorObject) => {
         ? `<span class="badge badge-pill badge-success">Valid</span>`
         : `<span class="badge badge-pill badge-warning">Missing Fields</span>`
     }
-    </td> <td><div class='ui small basic icon buttons contributor-helper-buttons' style='display: flex'><button class='ui button' onclick='edit_current_con_id(this)'><i class='pen icon' style='color: var(--tagify-dd-color-primary)'></i></button></div></td>
-    <td><div class='ui small basic icon buttons contributor-helper-buttons' style='display: flex'><button class='ui button' onclick='delete_current_con_id(this)'><i class='trash alternate outline icon' style='color: red'></i></button></div></td></tr>`);
+    </td> <td><div class='ui small basic icon buttons contributor-helper-buttons' style='display: flex'><button class='ui button' onclick='window.edit_current_con_id(this)'><i class='pen icon' style='color: var(--tagify-dd-color-primary)'></i></button></div></td>
+    <td><div class='ui small basic icon buttons contributor-helper-buttons' style='display: flex'><button class='ui button' onclick='window.delete_current_con_id(this)'><i class='trash alternate outline icon' style='color: red'></i></button></div></td></tr>`);
 };
 
[…]
