+++ scripts/guided-mode/guided-curate-dataset.js
@@ -1775,7 +1775,7 @@ const savePageChanges = async (pageBeingLeftID) => {
 
       // Make sure that all contributors have a valid fields
       for (const contributor of contributors) {
-        if (!contributorDataIsValid(contributor)) {
+        if (!window.contributorDataIsValid(contributor)) {
           errorArray.push({
             type: "notyf",
             message: "Please make sure all contributors have valid metadata",
@@ -9236,7 +9236,7 @@ const window.openGuidedAddContributorSwal = async () => {
   });
 };
 
-const contributorDataIsValid = (contributorObj) => {
+const window.contributorDataIsValid = (contributorObj) => {
   if (
     contributorObj.conAffliation.length > 0 &&
     contributorObj.conID &&
@@ -9312,7 +9312,7 @@ const handleContributorDrop = (event) => {
 };
 
 const generateContributorTableRow = (contributorObj, contributorIndex) => {
-  const contributorObjIsValid = contributorDataIsValid(contributorObj);
+  const contributorObjIsValid = window.contributorDataIsValid(contributorObj);
   const contributorFullName = contributorObj["conName"];
   const contributorOrcid = contributorObj["conID"];
   const contributorRoleString = contributorObj["conRole"].join(", ");
