+++ src/renderer/src/scripts/metadata-files/subjects-samples.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import validator from "validator";
 import doiRegex from "doi-regex";
 import Swal from "sweetalert2";
 import determineDatasetLocation, { Destinations } from "../analytics/analytics-utils"
@@ -1011,7 +1012,7 @@ window.edit_current_protocol_id = async (ev) => {
   }
 };
 
-const edit_current_additional_link_id = async (ev) => {
+window.edit_current_additional_link_id = async (ev) => {
   var currentRow = $(ev).parents()[2];
   var link = $(currentRow)[0].cells[1].innerText;
   var linkType = $(currentRow)[0].cells[2].innerText;
@@ -2737,7 +2738,8 @@ const loadDataFrametoUISamples = (type) => {
   }
 };
 
-const addAdditionalLink = async () => {
+window.addAdditionalLink = async () => {
+  let protocolLink = ""
   const { value: values } = await Swal.fire({
     title: "Add additional link",
     html:
@@ -2808,7 +2810,7 @@ const addAdditionalLink = async () => {
     },
   });
   if (values) {
-    addAdditionalLinktoTableDD(values[0], values[1], values[2], values[3]);
+    window.addAdditionalLinktoTableDD(values[0], values[1], values[2], values[3]);
   }
 };
 
