### 2023-12-09 5af49fe
wip: 🚧 Standardize local and pennsieve metadata gen
