+++ scripts/guided-mode/guided-curate-dataset.js
@@ -13880,6 +13880,13 @@ const guidedAddDatasetSubtitle = async (bfAccount, datasetName, datasetSubtitle)
         },
       }
     );
+
+    datasetSubtitleUploadText.innerHTML = `Successfully added dataset subtitle: ${datasetSubtitle}`;
+    guidedUploadStatusIcon("guided-dataset-subtitle-upload-status", "success");
+    sodaJSONObj["previously-uploaded-data"]["subtitle"] = datasetSubtitle;
+    await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
+
+    // Track the successful upload of a dataset subtitle via Kombucha
     ipcRenderer.send(
       "track-kombucha",
       kombuchaEnums.Category.GUIDED_MODE,
@@ -13888,20 +13895,16 @@ const guidedAddDatasetSubtitle = async (bfAccount, datasetName, datasetSubtitle)
       kombuchaEnums.Status.SUCCESS,
       {
         value: 1,
-        dataset_name: datasetName,
+        dataset_name: guidedGetDatasetName(sodaJSONObj),
+        dataset_id: guidedGetDatasetId(sodaJSONObj),
       }
     );
-    // TODO: EDIT CODE BELOW!!!! AND CHECK dataset_id above
     ipcRenderer.send(
       "track-event",
       "Success",
       ManageDatasetsAnalyticsPrefix.MANAGE_DATASETS_ADD_EDIT_SUBTITLE,
-      defaultBfDatasetId
+      guidedGetDatasetId(sodaJSONObj)
     );
-    datasetSubtitleUploadText.innerHTML = `Successfully added dataset subtitle: ${datasetSubtitle}`;
-    guidedUploadStatusIcon("guided-dataset-subtitle-upload-status", "success");
-    sodaJSONObj["previously-uploaded-data"]["subtitle"] = datasetSubtitle;
-    await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
   } catch (error) {
     console.error(error);
     let emessage = userErrorMessage(error);
@@ -14299,11 +14302,9 @@ const guidedAddTeamPermissions = async (bfAccount, datasetName, teamPermissionsA
   const result = await Promise.allSettled(promises);
 };
 
-//const add_dataset_permission = async();
-
 //********************************************************************************************************
 
-const guidedUploadSubjectsMetadata = async (bfAccount, datasetName, subjectsTableData) => {
+const guidedGenerateSubjectsMetadata = async (destination) => {
   document.getElementById("guided-subjects-metadata-upload-tr").classList.remove("hidden");
   const subjectsMetadataUploadText = document.getElementById(
     "guided-subjects-metadata-upload-text"
@@ -14346,7 +14347,7 @@ const guidedUploadSubjectsMetadata = async (bfAccount, datasetName, subjectsTabl
     throw new Error(userErrorMessage(error));
   }
 };
-const guidedUploadSamplesMetadata = async (bfAccount, datasetName, samplesTableData) => {
+const guidedGenerateSamplesMetadata = async (bfAccount, datasetName, samplesTableData) => {
   document.getElementById("guided-samples-metadata-upload-tr").classList.remove("hidden");
   const samplesMetadataUploadText = document.getElementById("guided-samples-metadata-upload-text");
   samplesMetadataUploadText.innerHTML = "Uploading samples metadata...";
@@ -14389,7 +14390,7 @@ const guidedUploadSamplesMetadata = async (bfAccount, datasetName, samplesTableD
   }
 };
 
-const guidedUploadSubmissionMetadata = async (bfAccount, datasetName, submissionMetadataJSON) => {
[…]
