### 2023-12-09 b405222
refactor: validation in organize dataset ffm
