+++ src/renderer/src/scripts/others/renderer.js
@@ -7054,7 +7054,7 @@ const checkEmptyFilesAndFolders = async (sodaJSONObj) => {
   return errorMessage;
 };
 
-const setSodaJSONStartingPoint = (sodaJSONObj) => {
+window.setSodaJSONStartingPoint = (sodaJSONObj) => {
   if (sodaJSONObj["starting-point"]["type"] === "local") {
     sodaJSONObj["starting-point"]["type"] = "new";
   }
@@ -7116,7 +7116,7 @@ document.getElementById("button-generate").addEventListener("click", async funct
   // updateJSON structure after Generate dataset tab
   window.updateJSONStructureGenerate(false, sodaJSONObj);
 
-  setSodaJSONStartingPoint(sodaJSONObj);
+  window.setSodaJSONStartingPoint(sodaJSONObj);
 
   let [dataset_name, dataset_destination] = setDatasetNameAndDestination(sodaJSONObj);
 
