+++ src/renderer/src/scripts/organize-dataset/curate-functions.js
@@ -1,5 +1,4 @@
 import Swal from "sweetalert2";
-// import checkDiskSpace from "check-disk-space";
 import { updateDatasetList } from "../globals"
 import determineDatasetLocation, { Destinations } from "../analytics/analytics-utils"
 import {clientError, userErrorMessage} from '../others/http-error-handler/error-handler'
@@ -258,106 +257,106 @@ window.dropHandler = async (
   }
 };
 
-const checkAvailableSpace = () => {
+const checkAvailableSpace = async () => {
   const roundToHundredth = (value) => {
     return Number(parseFloat(value.toFixed(2)));
   };
+
+
+  let freeMemory = await window.electron.ipcRenderer.invoke("getDiskSpace", location)
+  let freeMemoryMB = roundToHundredth(freeMemory / 1024 ** 2);
+
   let location = document
     .getElementById("input-destination-generate-dataset-locally")
     .getAttribute("placeholder");
 
-    // TODO: Place into mainjs
-  // checkDiskSpace(location).then(async (diskSpace) => {
-  //   log.info(`Checking available disk space for ${location}`);
-  //   let freeMemory = diskSpace.free; //returns in bytes
-  //   let freeMemoryMB = roundToHundredth(freeMemory / 1024 ** 2);
+    let datasetSizeResponse;
+    try {
+      datasetSizeResponse = await client.post(
+        "/curate_datasets/dataset_size",
+        {
+          soda_json_structure: sodaJSONObj,
+        },
+        { timeout: 0 }
+      );
 
-  //   let datasetSizeResponse;
-  //   try {
-  //     datasetSizeResponse = await client.post(
-  //       "/curate_datasets/dataset_size",
-  //       {
-  //         soda_json_structure: sodaJSONObj,
-  //       },
-  //       { timeout: 0 }
-  //     );
+      let tempFolderSize = datasetSizeResponse.data.dataset_size;
+      let folderSizeMB = roundToHundredth(tempFolderSize / 1024 ** 2);
+      let warningText =
+        "Please free up " +
+        roundToHundredth(folderSizeMB) +
+        "MB " +
+        "or consider uploading directly to Pennsieve.";
 
-  //     let tempFolderSize = datasetSizeResponse.data.dataset_size;
-  //     let folderSizeMB = roundToHundredth(tempFolderSize / 1024 ** 2);
-  //     let warningText =
-  //       "Please free up " +
-  //       roundToHundredth(folderSizeMB) +
-  //       "MB " +
-  //       "or consider uploading directly to Pennsieve.";
+      //converted to MB/GB/TB for user readability
+      if (folderSizeMB > 1000) {
+        //if bigger than a gb then convert to that
+        folderSizeMB = roundToHundredth(folderSizeMB / 1024);
+        freeMemoryMB = roundToHundredth(freeMemoryMB / 1024);
+        warningText =
+          "Please free up " +
+          roundToHundredth(folderSizeMB) +
+          "GB " +
+          "or consider uploading directly to Pennsieve.";
+        //if bigger than a tb then convert to that
+        if (folderSizeMB > 1000) {
+          folderSizeMB = roundToHundredth(folderSizeMB / 1024);
+          freeMemoryMB = roundToHundredth(freeMemoryMB / 1024);
+          warningText =
+            "Please free up " +
+            roundToHundredth(folderSizeMB) +
+            "TB " +
+            "or consider uploading directly to Pennsieve.";
+        }
+      }
 
-  //     //converted to MB/GB/TB for user readability
-  //     if (folderSizeMB > 1000) {
-  //       //if bigger than a gb then convert to that
-  //       folderSizeMB = roundToHundredth(folderSizeMB / 1024);
-  //       freeMemoryMB = roundToHundredth(freeMemoryMB / 1024);
-  //       warningText =
[…]
