+++ src/main/index.js
@@ -18,6 +18,7 @@ import "./banner-image"
 import './node-storage'
 import "./main-process/native-ui/dialogs/open-file"
 import "./strain"
+import "./checkDiskSpace"
 
 const sodaVersion = app.getVersion();
 // If the version includes "beta", the app will not check for updates
