+++ src/renderer/src/scripts/metadata-files/manifest.js
@@ -847,157 +847,156 @@ const generateManifest = async (action, type, manifestEditBoolean, ev) => {
     }
     window.sodaJSONObj["starting-point"]["local-path"] = window.finalManifestGenerationPath;
 
-    //checking size of local folder path
-    // checkDiskSpace(window.finalManifestGenerationPath).then(async (diskSpace) => {
-      // let freeMem = diskSpace.free;
-      //if free memory is less than 3MB
-      // if (freeMem < 3145728) {
-      //   Swal.fire({
-      //     title: "Not enough space in local storage",
-      //     html: "Please select another storage device or free up 3MB",
-      //     allowEscapeKey: true,
-      //     allowOutsideClick: false,
-      //     heightAuto: false,
-      //     backdrop: "rgba(0,0,0, 0.4)",
-      //     icon: "warning",
-      //     showConfirmButton: "OK",
-      //   });
-      //   window.logMetadataForAnalytics(
-      //     "Error",
-      //     window.MetadataAnalyticsPrefix.MANIFEST,
-      //     window.AnalyticsGranularity.ALL_LEVELS,
-      //     "Generate",
-      //     Destinations.LOCAL
-      //   );
+    let freeMem = await window.electron.ipcRenderer.invoke("getDiskSpace", window.finalManifestGenerationPath);
 
-      //   window.logMetadataForAnalytics(
-      //     "Error",
-      //     window.MetadataAnalyticsPrefix.MANIFEST,
-      //     window.AnalyticsGranularity.ACTION,
-      //     "Generate - Check Storage Space",
-      //     Destinations.LOCAL
-      //   );
-      // } else {
-        if (pennsievePreview) {
-          generateAfterEdits();
-          return;
-        }
+    //if free memory is less than 3MB
+    if (freeMem < 3145728) {
+      Swal.fire({
+        title: "Not enough space in local storage",
+        html: "Please select another storage device or free up 3MB",
+        allowEscapeKey: true,
+        allowOutsideClick: false,
+        heightAuto: false,
+        backdrop: "rgba(0,0,0, 0.4)",
+        icon: "warning",
+        showConfirmButton: "OK",
+      });
+      window.logMetadataForAnalytics(
+        "Error",
+        window.MetadataAnalyticsPrefix.MANIFEST,
+        window.AnalyticsGranularity.ALL_LEVELS,
+        "Generate",
+        Destinations.LOCAL
+      );
+
+      window.logMetadataForAnalytics(
+        "Error",
+        window.MetadataAnalyticsPrefix.MANIFEST,
+        window.AnalyticsGranularity.ACTION,
+        "Generate - Check Storage Space",
+        Destinations.LOCAL
+      );
+    } else {
+      if (pennsievePreview) {
+        generateAfterEdits();
+        return;
+      }
+    
+      window.logMetadataForAnalytics(
+        "Success",
+        window.MetadataAnalyticsPrefix.MANIFEST,
+        window.AnalyticsGranularity.ACTION,
+        "Generate - Check Storage Space",
+        Destinations.LOCAL
+      );
+
+      window.sodaJSONObj["starting-point"]["type"] = "local";
+      // if the manifest is going to be recreated post edits
+      if (manifestEditBoolean) {
+        localDatasetFolderPath = $("#input-manifest-local-folder-dataset").attr("placeholder");
+      }
+      await window.create_json_object(action, window.sodaJSONObj, localDatasetFolderPath);
+      console.log("FInished create json object")
+      window.datasetStructureJSONObj = window.sodaJSONObj["dataset-structure"];
+      window.populate_existing_folders(window.datasetStructureJSONObj);
[…]
