### 2023-12-11 #6140 @editorialbot
[REVIEW]: SODA: Software to Support the Curation and Sharing 1 of FAIR Autonomic Nervous System Data
**Submitting author:** @bvhpatel (<a href="http://orcid.org/0000-0002-0307-262X">Bhavesh Patel</a>)
**Repository:** https://github.com/fairdataihub/SODA-for-SPARC
**Branch with paper.md** (empty if default branch): paper
**Version:** v15.1.0
