### 2023-12-11 #6140 @AJQuinn
👋🏼 @bvhpatel @mattsouth, @yarikoptic, @abhishektiwari - this is the review thread for the paper. All of our communications will happen here from now on.

For @mattsouth, @yarikoptic and @abhishektiwari - As a reviewer, the first step is to create a checklist for your review by entering

as the top of a new comment in this thread. Please can you do this soon as a confirmation that you've seen the review starting.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. Summary conversation is great on this thread but try to avoid substantial discussion about the repository here, this should take place in issues on the source repository.

When discussing the submission on an issue thread, please mention https://github.com/openjournals/joss-reviews/issues/6140 so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
