### 2023-12-11 e47af11
refactor: check disk space using es6
