+++ src/main/checkDiskSpace.js
@@ -1,4 +1,4 @@
-import checkDiskSpace  from 'check-disk-space';
+import checkDiskSpace from 'check-disk-space'
 import { ipcMain } from "electron";
 
 
@@ -6,8 +6,7 @@ import { ipcMain } from "electron";
 
 ipcMain.handle('getDiskSpace', async (event, location) => {
   let freeMemory = await new Promise( (resolve, reject) => {
-    checkDiskSpace(location).then(async (diskSpace) => {
-      window.log.info(`Checking available disk space for ${location}`);
+    checkDiskSpace.default(location).then(async (diskSpace) => {
       let freeMemory = diskSpace.free; //returns in bytes
       resolve(freeMemory)
      })
