### 2023-12-11 c2c48c2
build: icons update
