+++ electron-builder.yml
@@ -28,6 +28,7 @@ nsis:
   uninstallDisplayName: ${productName}
   createDesktopShortcut: always
 mac:
+  icon: 'assets/menu-icon/soda_icon.png'
   entitlementsInherit: ./entitlements.mac.inherit.plist
   notarize: true
   extraResources:
