### 2023-12-11 dbca886
build: add icon to linux
