+++ electron-builder.yml
@@ -51,6 +51,7 @@ linux:
   asar: true
   maintainer: electronjs.org
   category: Utility
+  icon: "app-icon/png/icon.png"
 appImage:
   artifactName: ${name}-${version}.${ext}
 npmRebuild: false
