### 2023-12-11 63a625a
build: change icon to build directory
