+++ electron-builder.yml
@@ -28,7 +28,7 @@ nsis:
   uninstallDisplayName: ${productName}
   createDesktopShortcut: always
 mac:
-  icon: 'assets/menu-icon/soda_icon.png'
+  icon: 'app-icon/png/icon.png'
   entitlementsInherit: ./entitlements.mac.inherit.plist
   notarize: true
   extraResources:
