### 2023-12-11 d324bf0
build: update icon path
