### 2023-12-11 9f67da8
wip: 🚧 re-format metadata generation elements
