+++ scripts/guided-mode/guided-curate-dataset.js
@@ -7696,7 +7696,7 @@ const patchPreviousGuidedModeVersions = async () => {
   //Add key to track status of Pennsieve uploads
   if (!sodaJSONObj["pennsieve-upload-status"]) {
     sodaJSONObj["pennsieve-upload-status"] = {
-      "dataset-metadata-upload-status": "not-started",
+      "dataset-metadata-pennsieve-genration-status": "not-started",
     };
   }
 
@@ -14305,59 +14305,69 @@ const guidedAddTeamPermissions = async (bfAccount, datasetName, teamPermissionsA
 //********************************************************************************************************
 
 const guidedGenerateSubjectsMetadata = async (destination) => {
-  document.getElementById("guided-subjects-metadata-upload-tr").classList.remove("hidden");
-  const subjectsMetadataUploadText = document.getElementById(
-    "guided-subjects-metadata-upload-text"
+  const generationDestination = destination === "pennsieve" ? "pennsieve" : "local";
+
+  //// Unhide the subject metadata generation table row and set the rows status to loading ////
+  document
+    .getElementById(`guided-subjects-metadata-${generationDestination}-genration-tr`)
+    .classList.remove("hidden");
+  const subjectsMetadataGenerationText = document.getElementById(
+    `guided-subjects-metadata-${generationDestination}-genration-text`
   );
-  subjectsMetadataUploadText.innerHTML = "Uploading subjects metadata...";
-  guidedUploadStatusIcon("guided-subjects-metadata-upload-status", "loading");
+  subjectsMetadataGenerationText.innerHTML = "Uploading subjects metadata...";
+  guidedUploadStatusIcon(
+    `guided-subjects-metadata-${generationDestination}-genration-status`,
+    "loading"
+  );
+  //////////////////////////////////////////////////////////////////////////////////////////
 
-  const previouslyUpdatedSubjectsMetadata =
-    sodaJSONObj["previously-uploaded-data"]["subjects-metadata"];
-
-  if (JSON.stringify(previouslyUpdatedSubjectsMetadata) === JSON.stringify(subjectsTableData)) {
-    guidedUploadStatusIcon("guided-subjects-metadata-upload-status", "success");
-    subjectsMetadataUploadText.innerHTML = "Subjects metadata added to Pennsieve";
-    return;
-  }
-
+  //// Generate the subjects metadata file ////
   try {
     await client.post(
       `/prepare_metadata/subjects_file`,
       {
-        filepath: "",
-        selected_account: bfAccount,
-        selected_dataset: datasetName,
+        filepath: generationDestination === "pennsieve" ? "" : destination,
+        selected_account: defaultBfAccount,
+        selected_dataset:
+          generationDestination === "pennsieve" ? guidedGetDatasetName(sodaJSONObj) : "",
         subjects_header_row: subjectsTableData,
       },
       {
         params: {
-          upload_boolean: true,
+          upload_boolean: generationDestination === "pennsieve",
         },
       }
     );
-    guidedUploadStatusIcon("guided-subjects-metadata-upload-status", "success");
-    subjectsMetadataUploadText.innerHTML = `Subjects metadata successfully uploaded`;
-    sodaJSONObj["previously-uploaded-data"]["subjects-metadata"] = subjectsTableData;
-    await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
+    guidedUploadStatusIcon(
+      `guided-subjects-metadata-${generationDestination}-genration-status`,
+      "success"
+    );
[…]
