### 2023-12-11 8a6efbe
wip: update icon path linux
