+++ electron-builder.yml
@@ -51,7 +51,7 @@ linux:
   asar: true
   maintainer: electronjs.org
   category: Utility
-  icon: "./build/app-icon/png/icon.png"
+  icon: "build/app-icon/png/icon.png"
 appImage:
   artifactName: ${name}-${version}.${ext}
 npmRebuild: false
