### 2023-12-11 48fb2d6
chore: update icon path for Mac
