### 2023-12-11 eba9658
wip: add app icons
