+++ electron-builder.yml
@@ -28,7 +28,6 @@ nsis:
   uninstallDisplayName: ${productName}
   createDesktopShortcut: always
 mac:
-  icon: 'app-icon/png/icon.png'
   entitlementsInherit: ./entitlements.mac.inherit.plist
   notarize: true
   extraResources:
@@ -51,7 +50,6 @@ linux:
   asar: true
   maintainer: electronjs.org
   category: Utility
-  icon: "app-icon/png/icon.png"
 appImage:
   artifactName: ${name}-${version}.${ext}
 npmRebuild: false
