### 2023-12-12 64a5b6e
build: mac and linux icons
