### 2023-12-12 d23357f
build: run windows builds
