+++ .github/workflows/Build-and-deploy-win.yml
@@ -4,7 +4,7 @@ run-name: ${{ github.actor }} is building a Windows release for SODA
 on:
   push:
     branches:
-      - staging
+      - bundled-soda
 
 jobs:
   deploy-on-windows:
