### 2023-12-12 17a4d37
fix: add agent locations to path for Mac
