+++ main-process/native-ui/dialogs/open-file.js
@@ -14,6 +14,20 @@ ipcMain.on("open-create-dataset-structure-spreadsheet-path-selection-dialog", (e
   }
 });
 
+ipcMain.on("guided-select-local-dataset-generation-path", (event) => {
+  const mainWindow = BrowserWindow.getFocusedWindow();
+
+  // Get the path to the directory where the user wants to save the spreadsheet
+  const spreadsheetPath = dialog.showOpenDialogSync(mainWindow, {
+    properties: ["openDirectory"],
+    title: "Select a folder to copy the dataset to",
+  });
+
+  if (spreadsheetPath) {
+    event.sender.send("selected-guided-local-dataset-generation-path", spreadsheetPath[0]);
+  }
+});
+
 ipcMain.on("open-subject-multi-folder-import-dialog", (event) => {
   const mainWindow = BrowserWindow.getFocusedWindow();
 
