### 2023-12-12 9abfc93
fix: cannot iterate through samples in GM
