+++ scripts/guided-mode/guided-curate-dataset.js
@@ -8320,7 +8320,7 @@ const attachGuidedMethodsToSodaJSONObj = () => {
       this["dataset-metadata"]["pool-subject-sample-structure"]["pools"]
     )) {
       for (const [subjectName, subjectData] of Object.entries(poolData)) {
-        for (sampleName of Object.keys(subjectData)) {
+        for (const sampleName of Object.keys(subjectData)) {
           samplesInPools.push({
             sampleName: sampleName,
             subjectName: subjectName,
