### 2023-12-12 bce8a90
fix: next page adding samples inside of pools GM
