### 2023-12-13 a52b088
wip: 🚧 Added local dataset generation to guided mode
