### 2023-12-13 faa37ae
fix: file_counter + folder_counter undefined in GM validation
