### 2023-12-13 4397923
wip: 🚧 update gm progress bar when generating locally
