+++ scripts/guided-mode/guided-curate-dataset.js
@@ -245,44 +245,6 @@ const datasetIsSparcFunded = () => {
   return sodaJSONObj["dataset-metadata"]["submission-metadata"]["funding-consortium"] === "SPARC";
 };
 
-const checkIfPoolsFoldersAreCorrect = (poolFolderPath) => {
-  const poolFolders = Object.keys(poolFolderPath["folders"]);
-  const invalidPoolFolders = poolFolders.filter((folder) => {
-    !folder.startsWith("sub-");
-  });
-  const poolSubjectFolders = poolFolders.filter((folder) => {
-    folder.startsWith("sub-");
-  });
-  for (const subjectFolder of poolSubjectFolders) {
-    checkIfSubjectsFoldersAreCorrect(poolFolderPath["folders"][subjectFolder]);
-  }
-
-  for (const invalidPoolFolder of invalidPoolFolders) {
-    invalidFolders.push(invalidPoolFolder);
-  }
-
-  //Push all files in the pools to invalid files since the user should not have files in the pool folders
-  const poolFiles = Object.keys(poolFolderPath["files"]);
-  for (const poolFile of poolFiles) {
-    invalidFiles.push(poolFile);
-  }
-};
-
-const checkIfSubjectsFoldersAreCorrect = (subjectFolderPath) => {
-  const subjectFolders = Object.keys(subjectFolderPath["folders"]);
-  const invalidSubjectFolders = subjectFolders.filter((folder) => {
-    !folder.startsWith("sam-");
-  });
-
-  for (const invalidSubjectFolder of invalidSubjectFolders) {
-    invalidFolders.push(invalidSubjectFolder);
-  }
-  const subjectFiles = Object.keys(subjectFolderPath["files"]);
-  for (const subjectFile of subjectFiles) {
-    invalidFiles.push(subjectFile);
-  }
-};
-
 const guidedCheckHighLevelFoldersForImproperFiles = (datasetStructure) => {
   const invalidFolders = [];
   const invalidFiles = [];
@@ -2350,7 +2312,7 @@ const renderSideBar = (activePage) => {
   }
 };
 
-const updateDatasetUploadProgressTable = (progressObject) => {
+const updateDatasetUploadProgressTable = (generationDestination, progressObject) => {
   const datasetUploadTableBody = document.getElementById("guided-tbody-dataset-upload");
   //delete datasetUPloadTableBody children with class "upload-status-tr"
   const uploadStatusTRs = datasetUploadTableBody.querySelectorAll(".upload-status-tr");
@@ -2717,7 +2679,6 @@ const guidedModifyCurationTeamAccess = async (action) => {
   }
   if (action === "unshare") {
     // Add your dataset has been shared, to withdraw please do so from Pennsieve
-    // TODO: Dorian -> refactor this (it will be removed essentially)
     guidedUnshareWithCurationTeamButton.disabled = true;
     guidedUnshareWithCurationTeamButton.classList.add("loading");
 
@@ -7552,19 +7513,8 @@ const setActiveSubPage = (pageIdToActivate) => {
   // renderSideBar(CURRENT_PAGE.id);
 };
 
-const guidedIncreaseCurateProgressBar = (percentToIncrease) => {
-  $("#guided-progress-bar-new-curate").attr(
-    "value",
-    parseInt($("#guided-progress-bar-new-curate").attr("value")) + percentToIncrease
-  );
-};
-
 const setGuidedProgressBarValue = (value) => {
   $("#guided-progress-bar-new-curate").attr("value", value);
-};
-
-const isNumberBetween = (number, minVal, maxVal) => {
-  return !isNaN(parseFloat(number)) && isFinite(number) && number >= minVal && number <= maxVal;
 };
 
[…]
