### 2023-12-14 a545a37
chore: log windows location
