### 2023-12-14 c26d626
fix: GM detects when a user connetcs account with Pennsieve
