+++ src/renderer/src/scripts/guided-mode/guided-curate-dataset.js
@@ -392,7 +392,7 @@ const skipOrUnSkipCodeDescriptionPage = async (pennsieveDatasetID) => {
   try {
     await client.get(`/prepare_metadata/import_metadata_file`, {
       params: {
-        selected_account: window.defaultBfDataset,
+        selected_account: window.defaultBfAccount,
         selected_dataset: pennsieveDatasetID,
         file_type: "code_description.xlsx",
       },
@@ -628,7 +628,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         }
 
         const datasetIsLocked = await api.isDatasetLocked(
-          window.defaultBfDataset,
+          window.defaultBfAccount,
           selectedPennsieveDataset
         );
         if (datasetIsLocked) {
@@ -648,7 +648,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         //Pull the dataset folders and files from Pennsieve\
         window.sodaJSONObj["bf-dataset-selected"] = {};
         window.sodaJSONObj["bf-dataset-selected"]["dataset-name"] = selectedPennsieveDataset;
-        window.sodaJSONObj["bf-account-selected"]["account-name"] = window.defaultBfDataset;
+        window.sodaJSONObj["bf-account-selected"]["account-name"] = window.defaultBfAccount;
         const importProgressCircle = document.querySelector(
           "#guided_loading_pennsieve_dataset-organize"
         );
@@ -760,7 +760,7 @@ const savePageChanges = async (pageBeingLeftID) => {
               `/prepare_metadata/import_metadata_file`,
               {
                 params: {
-                  selected_account: window.defaultBfDataset,
+                  selected_account: window.defaultBfAccount,
                   selected_dataset: selectedPennsieveDatasetID,
                   file_type: "subjects.xlsx",
                   ui_fields: fieldEntries.toString(),
@@ -796,7 +796,7 @@ const savePageChanges = async (pageBeingLeftID) => {
                 `/prepare_metadata/import_metadata_file`,
                 {
                   params: {
-                    selected_account: window.defaultBfDataset,
+                    selected_account: window.defaultBfAccount,
                     selected_dataset: selectedPennsieveDatasetID,
                     file_type: "samples.xlsx",
                     ui_fields: fieldEntries.toString(),
@@ -1515,7 +1515,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         "guided-confirm-pennsieve-account-button"
       );
       if (!confirmAccountbutton.classList.contains("selected")) {
-        if (!window.defaultBfDataset) {
+        if (!window.defaultBfAccount) {
           // If the user has not logged in, throw an error
           errorArray.push({
             type: "notyf",
@@ -1557,7 +1557,7 @@ const savePageChanges = async (pageBeingLeftID) => {
         throw errorArray;
       }
 
-      window.sodaJSONObj["last-confirmed-bf-account-details"] = window.defaultBfDataset;
+      window.sodaJSONObj["last-confirmed-bf-account-details"] = window.defaultBfAccount;
       window.sodaJSONObj["last-confirmed-pennsieve-workspace-details"] = userSelectedWorkSpace;
     }
 
@@ -2562,7 +2562,7 @@ window.guidedModifyCurationTeamAccess = async (action) => {
 
 const checkIfDatasetExistsOnPennsieve = async (datasetNameOrID) => {
   let datasetName = null;
-  const datasetList = await api.getDatasetsForAccount(window.defaultBfDataset);
+  const datasetList = await api.getDatasetsForAccount(window.defaultBfAccount);
   for (const dataset of datasetList) {
[…]
