### 2023-12-14 3f76965
fix: show add account first message
