### 2023-12-14 c4c4102
wip: 🚧 Seperate generation in progress bars and tables
