+++ src/renderer/src/scripts/globals.js
@@ -1701,10 +1701,13 @@ window.openDropdownPrompt = async (ev, dropdown, show_timer = true) => {
         },
       });
 
+      console.log(result)
+
       if (result && result.isConfirmed) {
         await window.openDropdownPrompt(this, "bf");
         $(".ui.active.green.inline.loader.small").css("display", "none");
         $(".svg-change-current-account.dataset").css("display", "block");
+        return
       } else {
         $(".ui.active.green.inline.loader.small").css("display", "none");
         $(".svg-change-current-account.dataset").css("display", "block");
