### 2023-12-14 6a7c84f
refactor: add public directory at root level
