+++ src/renderer/src/scripts/organize-dataset/curate-functions.js
@@ -6,6 +6,13 @@ import { successCheck } from "../../assets/lotties/lotties";
 import client from "../client"
 import 'jstree'
 import fileTxt from "/img/txt-file.png"
+import filePng from "/img/png-file.png"
+import filePdf from "/img/pdf-file.png"
+import fileCsv from "/img/csv-file.png"
+import fileDoc from "/img/doc-file.png"
+import fileXlsx from "/img/excel-file.png"
+import fileJpeg from "/img/jpeg-file.png"
+import fileOther from "/img/other-file.png"
 
 while (!window.htmlPagesAdded) {
   await new Promise((resolve) => setTimeout(resolve, 100))
@@ -1075,55 +1082,55 @@ $(document).ready(function () {
         icon: "fas fa-folder fa-fw",
       },
       "file xlsx": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file xls": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file png": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file PNG": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file pdf": {
-        icon: "src/assets/img/pdf-file.png",
+        icon: filePdf,
       },
       "file txt": {
         icon: fileTxt,
       },
       "file csv": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file CSV": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file DOC": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file DOCX": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file docx": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file doc": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file jpeg": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file JPEG": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file jpg": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file JPG": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file other": {
-        icon: "src/assets/img/other-file.png",
+        icon: fileOther,
       },
     },
   });
