### 2023-12-14 d4b107a
fix: file paths for prod and dev fixed
