+++ src/renderer/src/scripts/metadata-files/manifest.js
@@ -7,6 +7,14 @@ import {clientError, userErrorMessage} from '../others/http-error-handler/error-
 import kombuchaEnums from "../analytics/analytics-enums";
 import createEventDataPrepareMetadata from "../analytics/prepare-metadata-analytics";
 import api from '../others/api/api'
+import fileTxt from "/img/txt-file.png"
+import filePng from "/img/png-file.png"
+import filePdf from "/img/pdf-file.png"
+import fileCsv from "/img/csv-file.png"
+import fileDoc from "/img/doc-file.png"
+import fileXlsx from "/img/excel-file.png"
+import fileJpeg from "/img/jpeg-file.png"
+import fileOther from "/img/other-file.png"
 
 
 
@@ -145,49 +153,49 @@ $(document).ready(async function () {
         icon: "fas fa-folder fa-fw",
       },
       "file xlsx": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file xls": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file png": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file PNG": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file pdf": {
-        icon: "src/assets/img/pdf-file.png",
+        icon: filePdf,
       },
       "file txt": {
-        icon: "src/assets/img/txt-file.png",
+        icon: fileTxt,
       },
       "file csv": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file CSV": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file DOC": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file DOCX": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file docx": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file doc": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file jpeg": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file JPEG": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file other": {
-        icon: "src/assets/img/other-file.png",
+        icon: fileOther,
       },
     },
   });
@@ -209,49 +217,49 @@ $(document).ready(async function () {
         icon: "fas fa-folder fa-fw",
       },
       "file xlsx": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file xls": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file png": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file PNG": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file pdf": {
-        icon: "src/assets/img/pdf-file.png",
+        icon: filePdf,
       },
       "file txt": {
-        icon: "src/assets/img/txt-file.png",
+        icon: fileTxt,
       },
       "file csv": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file CSV": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file DOC": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
[…]
