+++ src/renderer/src/scripts/organize-dataset/curate-functions.js
@@ -1569,49 +1569,49 @@ $(document).ready(function () {
         icon: "fas fa-folder fa-fw",
       },
       "file xlsx": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file xls": {
-        icon: "src/assets/img/excel-file.png",
+        icon: fileXlsx,
       },
       "file png": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file PNG": {
-        icon: "src/assets/img/png-file.png",
+        icon: filePng,
       },
       "file pdf": {
-        icon: "src/assets/img/pdf-file.png",
+        icon: filePdf,
       },
       "file txt": {
         icon: fileTxt,
       },
       "file csv": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file CSV": {
-        icon: "src/assets/img/csv-file.png",
+        icon: fileCsv,
       },
       "file DOC": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file DOCX": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file docx": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file doc": {
-        icon: "src/assets/img/doc-file.png",
+        icon: fileDoc,
       },
       "file jpeg": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file JPEG": {
-        icon: "src/assets/img/jpeg-file.png",
+        icon: fileJpeg,
       },
       "file other": {
-        icon: "src/assets/img/other-file.png",
+        icon: fileOther,
       },
     },
   });
