+++ src/renderer/src/scripts/organize-dataset/organizeDS.js
@@ -1,6 +1,13 @@
 import Swal from "sweetalert2";
 import determineDatasetLocation from "../analytics/analytics-utils"
-
+import fileTxt from "/img/txt-file.png"
+import filePng from "/img/png-file.png"
+import filePdf from "/img/pdf-file.png"
+import fileCsv from "/img/csv-file.png"
+import fileDoc from "/img/doc-file.png"
+import fileXlsx from "/img/excel-file.png"
+import fileJpeg from "/img/jpeg-file.png"
+import fileOther from "/img/other-file.png"
 
 while (!window.htmlPagesAdded) {
   await new Promise((resolve) => setTimeout(resolve, 100))
