### 2023-12-14 1a0086d
feat: ✨ Only show first level in jsTre preview
