### 2023-12-14 9a97872
feat: ✨ Standardize gm jsTree rendering
