### 2023-12-14 4ecebe2
fix: smile and sad can paths
