### 2023-12-14 b634e94
fix: manifest generation on second navigation organize ds
