### 2023-12-14 aa34a9a
fix: import primary folder ffm subjects
