+++ scripts/others/renderer.js
@@ -2236,7 +2236,7 @@ const window.generateSamplesFileHelper = async (uploadBFBoolean) => {
 // import Primary folder
 ipcRenderer.on("selected-local-primary-folder", (event, primaryFolderPath) => {
   if (primaryFolderPath.length > 0) {
-    importPrimaryFolderSubjects(primaryFolderPath[0]);
+    window.importPrimaryFolderSubjects(primaryFolderPath[0]);
   }
 });
 ipcRenderer.on("selected-local-primary-folder-samples", (event, primaryFolderPath) => {
