+++ src/renderer/src/scripts/others/renderer.js
@@ -1991,7 +1991,7 @@ window.electron.ipcRenderer.on("selected-local-primary-folder", (event, primaryF
 });
 window.electron.ipcRenderer.on("selected-local-primary-folder-samples", (event, primaryFolderPath) => {
   if (primaryFolderPath.length > 0) {
-    importPrimaryFolderSamples(primaryFolderPath[0]);
+    window.importPrimaryFolderSamples(primaryFolderPath[0]);
   }
 });
 
