### 2023-12-15 0ec24cf
wip: 🚧 Added primitive local generation progress tracking
