+++ scripts/others/renderer.js
@@ -7935,6 +7935,7 @@ const initiate_generate = async () => {
     }
 
     let { data } = mainCurationProgressResponse;
+    console.log(data);
 
     main_curate_status = data["main_curate_status"];
     var start_generate = data["start_generate"];
