+++ scripts/guided-mode/guided-curate-dataset.js
@@ -1727,156 +1727,23 @@ const savePageChanges = async (pageBeingLeftID) => {
       const selectedLicense = selectedLicenseButton.dataset.value;
       sodaJSONObj["digital-metadata"]["license"] = selectedLicense;
     }
-    /*
-    if (pageBeingLeftID === "guided-dataset-generate-location-tab") {
-      const buttonGenerateLocally = document.getElementById(
-        "guided-button-generate-dataset-locally"
-      );
-      const buttonGenerateOnPennsieve = document.getElementById(
-        "guided-button-generate-dataset-on-pennsieve"
-      );
 
-      // If the user did not select if they would like to import a SPARC award,
-      // throw an error
-      if (
-        !buttonGenerateLocally.classList.contains("selected") &&
-        !buttonGenerateOnPennsieve.classList.contains("selected")
-      ) {
-        errorArray.push({
-          type: "notyf",
-          message: "Please indicate where you would like to genrate your dataset",
-        });
-        throw errorArray;
-      }
-
-      if (buttonGenerateOnPennsieve.classList.contains("selected")) {
-        const accountName = document.getElementById("guided-bf-account");
-        if (accountName.innerHTML.trim() === "None" || accountName.innerHTML.trim() === "") {
-          errorArray.push({
-            type: "notyf",
-            message: "Please select a Pennsieve account to generate your dataset on",
-          });
-          throw errorArray;
-        }
-        sodaJSONObj["generate-dataset"]["destination"] = "bf";
-      }
-    }
-    */
-    /*
-    if (pageBeingLeftID === "guided-dataset-generate-destination-tab") {
-      const buttonGenerateOnExistingPennsieveDataset = document.getElementById(
-        "guided-button-pennsieve-generate-existing"
-      );
-      const buttonGenerateOnNewPennsieveDataset = document.getElementById(
-        "guided-button-pennsieve-generate-new"
-      );
-
-      // If the user did not select if they would like to import a SPARC award,
-      // throw an error
-      if (
-        !buttonGenerateOnExistingPennsieveDataset.classList.contains("selected") &&
-        !buttonGenerateOnNewPennsieveDataset.classList.contains("selected")
-      ) {
-        errorArray.push({
-          type: "notyf",
-          message:
-            "Please indicate if you would like to generate on a new or existing Pennsieve dataset",
-        });
-        throw errorArray;
-      }
-
-      if (buttonGenerateOnExistingPennsieveDataset.classList.contains("selected")) {
-        sodaJSONObj["generate-dataset"]["destination"] = "local";
-      }
-
-      if (buttonGenerateOnNewPennsieveDataset.classList.contains("selected")) {
-        confirmDatasetGenerationNameinput = document.getElementById("guided-input-dataset-name");
-        if (confirmDatasetGenerationNameinput.value.trim() === "") {
-          errorArray.push({
-            type: "notyf",
-            message: "Please enter a name for your new Pennsieve dataset",
-          });
-          throw errorArray;
-        }
-        sodaJSONObj["digital-metadata"]["name"] = confirmDatasetGenerationNameinput.value.trim();
-        sodaJSONObj["generate-dataset"]["destination"] = "bf";
-      }
-    }
-    */
-
-    if (pageBeingLeftID === "guided-folder-structure-preview-tab") {
+    if (pageBeingLeftID === "guided-dataset-structure-review-tab") {
       //if folders and files in datasetStruture json obj are empty, warn the user
       if (
         Object.keys(datasetStructureJSONObj["folders"]).length === 0 &&
         Object.keys(datasetStructureJSONObj["files"]).length === 0
       ) {
-        const { value: continueProgress } = await Swal.fire({
[…]
