+++ src/renderer/src/scripts/metadata-files/submission.js
@@ -777,7 +777,7 @@ window.generateSubmissionHelper = async (uploadBFBoolean) => {
     }
 
     // Check if dataset is locked after running pre-flight checks
-    const isLocked = await api.isDatasetLocked(window.defaultBfDataset, datasetName);
+    const isLocked = await api.isDatasetLocked(window.defaultBfAccount, datasetName);
 
     if (isLocked) {
       await Swal.fire({
