### 2023-12-15 b2214d3
fix: swal queue not defined
