### 2023-12-15 5068611
fix: reset submission fields not defined
