+++ src/renderer/src/scripts/globals.js
@@ -292,7 +292,7 @@ window.resetSubmission = (askToReset = true) => {
       .removeClass('disabled')
       .removeClass('non-selected')
     $('#Question-prepare-submission-1 .option-card .folder-input-check').prop('checked', false)
-    resetSubmissionFields()
+    window.resetSubmissionFields()
     return
   }
 
@@ -323,12 +323,12 @@ window.resetSubmission = (askToReset = true) => {
         .removeClass('disabled')
         .removeClass('non-selected')
       $('#Question-prepare-submission-1 .option-card .folder-input-check').prop('checked', false)
-      resetSubmissionFields()
+      window.resetSubmissionFields()
     }
   })
 }
 
-function resetSubmissionFields() {
+window.resetSubmissionFields = () => {
   $('#existing-submission-file-destination').attr('placeholder', 'Browse here')
 
   $('#div-confirm-existing-submission-import').hide()
