### 2023-12-15 c497033
fix: find windows executable in packaged app
