### 2023-12-15 477ed61
fix: use correct logger in main process
