### 2023-12-15 39ae3bf
refactor: add logs to startup code
