### 2023-12-16 7ebb48f
fix: executable path not defined
