### 2023-12-16 #6140 @abhishektiwari
@bvhpatel Given the software has hard dependency on Pennsieve platform for most of the features including login, use of OSI approved does not seem applicable here. To me this is more of hosted web view just wrapped in a Desktop app. Can you please clarify data architecture for this app?

@AJQuinn This software is not standalone OSI software. It needs registration, login and other details at a third party website [Pennsieve](pennsieve.io) which requires its own [T&C](https://docs.pennsieve.io/page/pennsieve-terms-of-use) to be accepted. What is JOSS guideline here ([submission requirements](https://docs.pennsieve.io/page/pennsieve-terms-of-use) are not explicit about this kind of use cases)?
