+++ src/renderer/src/main.js
@@ -26,6 +26,7 @@ import "./scripts/organize-dataset/validation-functions"
 import './assets/lotties/activate-lotties'
 
 // Application CSS 
+import "cropperjs/dist/cropper.css";
 import 'fomantic-ui/dist/semantic.min.css';
 import '@fortawesome/fontawesome-free/css/all.min.css'
 import './assets/css/animations.css'
@@ -57,7 +58,7 @@ import '@yaireo/tagify/dist/tagify.css'
 import "@tarekraafat/autocomplete.js/dist/css/autoComplete.css";
 import "@tarekraafat/autocomplete.js/dist/css/autoComplete.01.css";
 import "@tarekraafat/autocomplete.js/dist/css/autoComplete.02.css";
-import "cropperjs/src/css/cropper.css";
+
 import "select2/dist/css/select2.min.css"
 import "intro.js/minified/introjs.min.css"
 
