### 2023-12-17 f41ef05
refactor: guided mode add protocols
