+++ scripts/guided-mode/guided-curate-dataset.js
@@ -9579,7 +9579,7 @@ const determineIfLinkIsDOIorURL = (link) => {
   }
 };
 
-const openProtocolSwal = async (protocolElement) => {
+const window.openProtocolSwal = async (protocolElement) => {
   // True if adding a new protocol, false if editing an existing protocol
   let protocolURL = "";
   let protocolDescription = "";
@@ -9638,7 +9638,7 @@ const openProtocolSwal = async (protocolElement) => {
   });
 };
 
-const guidedDeleteProtocol = (protocolElement) => {
+const window.guidedDeleteProtocol = (protocolElement) => {
   const linkToDelete = protocolElement.dataset.protocolUrl;
   sodaJSONObj["dataset-metadata"]["description-metadata"]["protocols"] = sodaJSONObj[
     "dataset-metadata"
@@ -9708,7 +9708,7 @@ const generateProtocolField = (protocolUrl, protocolType, protocolDescription, i
           style="color: white; background-color: var(--color-light-green); border-color: var(--color-light-green);"
           data-protocol-url="${protocolUrl}"
           data-protocol-description="${protocolDescription}"
-          onclick="openProtocolSwal(this)"
+          onclick="window.openProtocolSwal(this)"
         >
         View/Edit
         </button>
@@ -9719,7 +9719,7 @@ const generateProtocolField = (protocolUrl, protocolType, protocolDescription, i
           class="btn btn-danger btn-sm"
           data-protocol-url="${protocolUrl}"
           data-protocol-description="${protocolDescription}"
-          onclick="guidedDeleteProtocol(this)"
+          onclick="window.guidedDeleteProtocol(this)"
         >
         Delete
         </button>
