### 2023-12-18 #6140 @AJQuinn
Thanks for the input @abhishektiwari - I'll clarify this with the editor team and let you know.

In the meantime @bvhpatel  - could you clarify the point above? My reading is that the JOSS submission is for SODA which is mainly used for SPARC/Pennsieve at the moment but/will can be extended to other platforms. Is this correct? and is there a way to run stand-alone tests of the platform?

Thanks all
