### 2023-12-18 045b316
feat: ✨ Get local generation progress percentage
