+++ scripts/guided-mode/guided-curate-dataset.js
@@ -13997,42 +13997,64 @@ const guidedAddTeamPermissions = async (bfAccount, datasetName, teamPermissionsA
 
 //********************************************************************************************************
 
+// Add click event listener to the button triggering local dataset generation
 document
   .getElementById("guided-button-generate-local-dataset-copy")
   .addEventListener("click", () => {
+    // Send an IPC message to select the local dataset generation path
     ipcRenderer.send("guided-select-local-dataset-generation-path");
   });
 
+const convertBytesToMb = (bytes) => {
+  return roundToHundredth(bytes / 1024 ** 2);
+};
+
+const convertBytesToGb = (bytes) => {
+  return roundToHundredth(bytes / 1024 ** 3);
+};
+
+// Listen for the selected path for local dataset generation
 ipcRenderer.on("selected-guided-local-dataset-generation-path", async (event, filePath) => {
   try {
-    // Check the available disk space on the drive where the user wants to generate the dataset
-    console.log("Checking disk space... for file path: ", filePath);
-    const availableDiskSpace = await checkDiskSpace(filePath);
-    console.log("Available disk space: ", availableDiskSpace);
-    const freeMemoryInMb = roundToHundredth(availableDiskSpace.free / 1024 ** 2);
-    console.log("Available disk space in MB: ", freeMemoryInMb);
+    // Check available free memory on disk
+    const diskSpaceRes = await checkDiskSpace(filePath);
+    const freeMemoryInBytes = diskSpaceRes.free;
+
+    console.log("Free memory in bytes: ", diskSpaceRes.free);
+    const freeMemoryInMb = convertBytesToMb(diskSpaceRes.free);
+    console.log("Free memory in MB: ", freeMemoryInMb);
+    console.log("Free memory in GB: ", convertBytesToGb(diskSpaceRes.free));
     // Get the size of the dataset that will be generated
     const localDatasetSizeReq = await client.post(
       "/curate_datasets/dataset_size",
-      {
-        soda_json_structure: sodaJSONObj,
-      },
+      { soda_json_structure: sodaJSONObj },
       { timeout: 0 }
     );
-    const localDatasetSize = localDatasetSizeReq.data.dataset_size;
-    const datasetSizeInMb = roundToHundredth(localDatasetSize / 1024 ** 2);
-    console.log("Dataset size in MB: ", datasetSizeInMb);
+    const localDatasetSizeInBytes = localDatasetSizeReq.data.dataset_size;
+    const datasetSizeInMb = convertBytesToMb(localDatasetSizeInBytes);
+    console.log("Dataset size in GB: ", convertBytesToGb(localDatasetSizeInBytes));
 
-    // Attach the manifest files to the dataset structure before generating the dataset locally
+    // Check if there is enough free space on disk for the dataset
+    if (freeMemoryInMb < datasetSizeInMb) {
+      throw new Error(
+        `Not enough free space on disk. Free space: ${freeMemoryInMb} MB. Dataset size: ${datasetSizeInMb} MB`
+      );
+    } else {
+      console.log("Free space besides dataset size in MB: ", freeMemoryInMb - datasetSizeInMb);
+      console.log(
+        "Free space besided dataset size in GB: ",
+        (freeMemoryInMb - datasetSizeInMb) / 1024
+      );
+    }
+
+    // Attach manifest files to the dataset structure before local generation
     await guidedCreateManifestFilesAndAddToDatasetStructure();
 
+    // Get the dataset name based on the sodaJSONObj
     const guidedDatasetName = guidedGetDatasetName(sodaJSONObj);
 
-    // Create a temp copy of the sodaJSONObj to be used for generating the dataset locally.
-    // We do this because there are keys we want to modify in the sodaJSONObj but we don't want to
-    // modify in the original sodaJSONObj
[…]
