### 2023-12-18 ee62121
wip: 🚧 Allow user to continue without latest agent
