+++ scripts/others/renderer.js
@@ -745,6 +745,8 @@ const agent_installed = () => {
   });
 };
 
+let userHasSelectedTheyAreOkWithOutdatedAgent = false;
+
 // Run a set of functions that will check all the core systems to verify that a user can upload datasets with no issues.
 const run_pre_flight_checks = async (check_update = true) => {
   try {
@@ -807,15 +809,34 @@ const run_pre_flight_checks = async (check_update = true) => {
       }
     }
 
-    // First get the latest Pennsieve agent version on GitHub
-    // This is to ensure the user has the latest version of the agent
-    let browser_download_url;
-    let latest_agent_version;
-    try {
-      [browser_download_url, latest_agent_version] = await get_latest_agent_version();
-    } catch (error) {
-      const emessage = userErrorMessage(error);
-      throw new Error(`Error getting latest Pennsieve agent version:<br />${emessage}`);
+    let platformSpecificAgentDownloadURL;
+    let latestPennsieveAgentVersion;
+    if (!userHasSelectedTheyAreOkWithOutdatedAgent) {
+      // First get the latest Pennsieve agent version on GitHub
+      // This is to ensure the user has the latest version of the agent
+      try {
+        [platformSpecificAgentDownloadURL, latestPennsieveAgentVersion] =
+          await getLatestPennsieveAgentVersion();
+      } catch (error) {
+        const emessage = userErrorMessage(error);
+        const continueWithPotentiallyOutdatedAgent = await swalConfirmAction(
+          "SODA was not able to get the latest Pennsieve agent version",
+          `
+            Error message:
+            <br />
+            <b>${emessage}</b>
+            <br /><br />
+            Would you like to continue without ensuring you have the latest Pennsieve agent version?
+          `,
+          "Continue",
+          "Cancel"
+        );
+        if (!continueWithPotentiallyOutdatedAgent) {
+          throw new Error(`Error getting latest Pennsieve agent version:<br />${emessage}`);
+        } else {
+          userHasSelectedTheyAreOkWithOutdatedAgent = true;
+        }
+      }
     }
 
     // check if the Pennsieve agent is installed [ here ]
@@ -829,9 +850,12 @@ const run_pre_flight_checks = async (check_update = true) => {
                   It looks like the Pennsieve Agent is not installed on your computer. It is recommended that you install the Pennsieve Agent now if you want to upload datasets to Pennsieve through SODA.
                   <br />
                   To install the Pennsieve Agent, please visit the link below and follow the instructions.
+                  <br /> 
                   <br />
-                  <br />
-                  <a href="${browser_download_url}" target="_blank">Download the Pennsieve agent</a>
+                  <a href="${
+                    platformSpecificAgentDownloadURL ??
+                    "https://github.com/Pennsieve/pennsieve-agent/releases"
+                  }" target="_blank">Download the Pennsieve agent</a>
                   <br />
                   <br />
                   Once you have installed the Pennsieve Agent, you will need to close and restart SODA before you can upload datasets. Would you like to close SODA now?
@@ -996,7 +1020,10 @@ const run_pre_flight_checks = async (check_update = true) => {
       return false;
     }
 
-    if (usersPennsieveAgentVersion !== latest_agent_version) {
+    if (
+      usersPennsieveAgentVersion !== latestPennsieveAgentVersion &&
+      !userHasSelectedTheyAreOkWithOutdatedAgent
+    ) {
       // Stop the Pennsieve agent if it is running to prevent any issues when updating while the agent is running
       try {
         await stopPennsieveAgent();
[…]
