### 2023-12-18 fd6a54e
fix: 🐛 Move agent version check to after agent start attempt
