+++ scripts/others/renderer.js
@@ -809,40 +809,11 @@ const run_pre_flight_checks = async (check_update = true) => {
       }
     }
 
-    let platformSpecificAgentDownloadURL;
-    let latestPennsieveAgentVersion;
-    if (!userHasSelectedTheyAreOkWithOutdatedAgent) {
-      // First get the latest Pennsieve agent version on GitHub
-      // This is to ensure the user has the latest version of the agent
-      try {
-        [platformSpecificAgentDownloadURL, latestPennsieveAgentVersion] =
-          await getLatestPennsieveAgentVersion();
-      } catch (error) {
-        const emessage = userErrorMessage(error);
-        const continueWithPotentiallyOutdatedAgent = await swalConfirmAction(
-          "SODA was not able to get the latest Pennsieve agent version",
-          `
-            Error message:
-            <br />
-            <b>${emessage}</b>
-            <br /><br />
-            Would you like to continue without ensuring you have the latest Pennsieve agent version?
-          `,
-          "Continue",
-          "Cancel"
-        );
-        if (!continueWithPotentiallyOutdatedAgent) {
-          throw new Error(`Error getting latest Pennsieve agent version:<br />${emessage}`);
-        } else {
-          userHasSelectedTheyAreOkWithOutdatedAgent = true;
-        }
-      }
-    }
-
     // check if the Pennsieve agent is installed [ here ]
     try {
       let installed = await agent_installed();
       if (!installed) {
+        const downloadUrl = await getPlatformSpecificAgentDownloadURL();
         const { value: restartSoda } = await Swal.fire({
           icon: "info",
           title: "Pennsieve Agent Not Found",
@@ -852,10 +823,7 @@ const run_pre_flight_checks = async (check_update = true) => {
                   To install the Pennsieve Agent, please visit the link below and follow the instructions.
                   <br /> 
                   <br />
-                  <a href="${
-                    platformSpecificAgentDownloadURL ??
-                    "https://github.com/Pennsieve/pennsieve-agent/releases"
-                  }" target="_blank">Download the Pennsieve agent</a>
+                  <a href="${downloadUrl}" target="_blank">Download the Pennsieve agent</a>
                   <br />
                   <br />
                   Once you have installed the Pennsieve Agent, you will need to close and restart SODA before you can upload datasets. Would you like to close SODA now?
@@ -1020,9 +988,39 @@ const run_pre_flight_checks = async (check_update = true) => {
       return false;
     }
 
+    let platformSpecificAgentDownloadURL;
+    let latestPennsieveAgentVersion;
+    if (!userHasSelectedTheyAreOkWithOutdatedAgent) {
+      // First get the latest Pennsieve agent version on GitHub
+      // This is to ensure the user has the latest version of the agent
+      try {
+        [platformSpecificAgentDownloadURL, latestPennsieveAgentVersion] =
+          await getLatestPennsieveAgentVersion();
+      } catch (error) {
+        const emessage = userErrorMessage(error);
+        const continueWithPotentiallyOutdatedAgent = await swalConfirmAction(
+          "SODA was not able to get the latest Pennsieve agent version",
+          `
+            Error message:
+            <br />
+            <b>${emessage}</b>
+            <br /><br />
+            Would you like to continue without ensuring you have the latest Pennsieve agent version?
+          `,
+          "Continue",
+          "Cancel"
+        );
+        if (!continueWithPotentiallyOutdatedAgent) {
+          throw new Error(`Error getting latest Pennsieve agent version:<br />${emessage}`);
+        } else {
+          userHasSelectedTheyAreOkWithOutdatedAgent = true;
+        }
+      }
+    }
+
     if (
[…]
