### 2023-12-18 ac33dba
fix: 🐛 Fix annoucements display
