### 2023-12-18 72903ac
fix: remove white background to mac icon
