+++ scripts/manage-dataset/manage-dataset.js
@@ -880,7 +880,7 @@ const countCharacters = (textelement, pelement) => {
 
 $(document).ready(() => {
   bfDatasetSubtitle.addEventListener("keyup", function () {
-    countCharacters(bfDatasetSubtitle, bfDatasetSubtitleCharCount);
+    countCharacters(bfDatasetSubtitle, window.bfDatasetSubtitleCharCount);
   });
 });
 
@@ -959,7 +959,7 @@ $("#button-add-subtitle").click(async () => {
       );
 
       // run the pre-publishing checklist validation -- this is displayed in the pre-publishing section
-      showPrePublishingStatus();
+      window.showPrePublishingStatus();
     } catch (error) {
       clientError(error);
 
@@ -1025,7 +1025,7 @@ const showCurrentSubtitle = async () => {
     );
     $("#bf-dataset-subtitle").val(subtitle);
     $("#ds-description").val(subtitle);
-    let result = countCharacters(bfDatasetSubtitle, bfDatasetSubtitleCharCount);
+    let result = countCharacters(bfDatasetSubtitle, window.bfDatasetSubtitleCharCount);
     if (result === 0) {
       $("#button-add-subtitle > .btn_animated-inside").html("Add subtitle");
     } else {
@@ -1234,7 +1234,7 @@ $("#button-add-description").click(() => {
         $("#ds-isa-warning").css("display", "none");
         await addDescription(selectedBfDataset, requiredFields.join("\n"));
         // run the pre-publishing checklist validation -- this is displayed in the pre-publishing section
-        showPrePublishingStatus();
+        window.showPrePublishingStatus();
       });
     } else {
       // hide the warning message if it exists
@@ -1242,7 +1242,7 @@ $("#button-add-description").click(() => {
       // add the user's description to Pennsieve
       await addDescription(selectedBfDataset, requiredFields.join("\n"));
       // run the pre-publishing checklist validation -- this is displayed in the pre-publishing section
-      showPrePublishingStatus();
+      window.showPrePublishingStatus();
     }
   }, window.delayAnimation);
 });
@@ -1977,7 +1977,7 @@ const uploadBannerImage = async () => {
         });
 
         // run the pre-publishing checklist validation -- this is displayed in the pre-publishing section
-        showPrePublishingStatus();
+        window.showPrePublishingStatus();
       } catch (error) {
         clientError(error);
         let emessage = userErrorMessage(error);
@@ -2357,7 +2357,7 @@ $("#button-add-tags").click(async () => {
   );
 
   // run the pre-publishing checklist items to update the list found in the "Submit for pre-publishing review" section/card
-  showPrePublishingStatus();
+  window.showPrePublishingStatus();
 
   //check if tags array is empty and set Add/Edit tags appropriately
   tags === undefined || tags.length == 0
@@ -2506,7 +2506,7 @@ $("#button-add-license").click(async () => {
       );
 
       // run the pre-publishing checklist validation -- this is displayed in the pre-publishing section
-      showPrePublishingStatus();
+      window.showPrePublishingStatus();
     } catch (error) {
       clientError(error);
       let emessage = userErrorMessage(error);
