### 2023-12-19 7d22ba7
refactor: pre publishing review script general refactors
