### 2023-12-19 e7b95fc
fix: get text value without endlines banner img
