+++ .github/workflows/Build-and-deploy-win.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - main
       - staging
-      - optional-outdated-agent-bypass
 
 jobs:
   deploy-on-windows:
