### 2023-12-19 84c7b87
fix: editing banner img button action
