### 2023-12-19 a570261
chore: update CITATION.cff for Zenodo release
