### 2023-12-19 aa1b3d9
refactor: make disseminate and orcid funcs global
