### 2023-12-20 e427447
fix: show banner image ffm
