### 2023-12-20 2134094
fix: banner image shows
