### 2023-12-20 ce78629
fix: show banner img after importing ffm
