### 2023-12-20 ed12522
fix: make banner image uploadable
