+++ src/renderer/src/scripts/manage-dataset/manage-dataset.js
@@ -1831,9 +1831,9 @@ const displayBannerImage = async (path) => {
                       document.getElementById("div-img-container").style.display = "block";
 
                       $("#para-path-image").html(image_path);
-                      bfViewImportedImage.src = converted_image_file;
+                      window.bfViewImportedImage.src = converted_image_file;
                       window.myCropper.destroy();
-                      window.myCropper = new Cropper(bfViewImportedImage, cropOptions);
+                      window.myCropper = new Cropper(window.bfViewImportedImage, cropOptions);
                       $("#save-banner-image").css("visibility", "visible");
                       $("body").removeClass("waiting");
                     });
@@ -1856,9 +1856,9 @@ const displayBannerImage = async (path) => {
               image_path = converted_image_file;
               imageExtension = "jpg";
               $("#para-path-image").html(image_path);
-              bfViewImportedImage.src = image_path;
+              window.bfViewImportedImage.src = image_path;
               window.myCropper.destroy();
-              window.myCropper = new Cropper(bfViewImportedImage, cropOptions);
+              window.myCropper = new Cropper(window.bfViewImportedImage, cropOptions);
               $("#save-banner-image").css("visibility", "visible");
             }
           });
@@ -1885,9 +1885,9 @@ const displayBannerImage = async (path) => {
       document.getElementById("div-img-container").style.display = "block";
 
       $("#para-path-image").html(image_path);
-      bfViewImportedImage.src = image_path;
+      window.bfViewImportedImage.src = image_path;
       window.myCropper.destroy();
-      window.myCropper = new Cropper(bfViewImportedImage, cropOptions);
+      window.myCropper = new Cropper(window.bfViewImportedImage, cropOptions);
 
       $("#save-banner-image").css("visibility", "visible");
     }
@@ -2122,9 +2122,9 @@ const uploadBannerImage = async () => {
 
 $("#save-banner-image").click((event) => {
   //save button for banner image (need on the size of the cropped image)
-  //bfViewImportedImage holds the entire image
+  //window.bfViewImportedImage holds the entire image
   $("#para-dataset-banner-image-status").html("");
-  if (bfViewImportedImage.src.length > 0) {
+  if (window.bfViewImportedImage.src.length > 0) {
     if (window.formBannerHeight.value > 511) {
       Swal.fire({
         icon: "warning",
@@ -2219,7 +2219,7 @@ const showCurrentBannerImage = async () => {
 
     bfCurrentBannerImg.src = "";
     document.getElementById("para-current-banner-img").innerHTML = "None";
-    bfViewImportedImage.src = "";
+    window.bfViewImportedImage.src = "";
 
     $("#div-img-container-holder").css("display", "block");
     $("#div-img-container").css("display", "none");
@@ -2248,7 +2248,7 @@ const showCurrentBannerImage = async () => {
     if (res === "No banner image") {
       bfCurrentBannerImg.src = "";
       document.getElementById("para-current-banner-img").innerHTML = "None";
-      bfViewImportedImage.src = "";
+      window.bfViewImportedImage.src = "";
 
       $("#div-img-container-holder").css("display", "block");
       $("#div-img-container").css("display", "none");
@@ -2266,7 +2266,7 @@ const showCurrentBannerImage = async () => {
 
     bfCurrentBannerImg.src = "assets/img/no-banner-image.png";
     document.getElementById("para-current-banner-img").innerHTML = "None";
-    bfViewImportedImage.src = "";
+    window.bfViewImportedImage.src = "";
 
     $("#div-img-container-holder").css("display", "block");
     $("#div-img-container").css("display", "none");
