### 2023-12-20 9c07bc9
feat: ✨ Modified agent config default_profile to replace . with _
