+++ src/renderer/src/scripts/others/renderer.js
@@ -1287,18 +1287,18 @@ window.datasetPermissionDiv = document.getElementById("div-permission-list-2");
 
 window.bfDatasetSubtitleCharCount = document.querySelector("#para-char-count-metadata");
 
-// const bfCurrentBannerImg = document.getElementById("current-banner-img");
+window.bfCurrentBannerImg = document.getElementById("current-banner-img");
 
 window.bfViewImportedImage = document.querySelector("#image-banner");
 window.guidedBfViewImportedImage = document.querySelector("#guided-image-banner");
 
-// const bfSaveBannerImageBtn = document.getElementById("save-banner-image");
-// const datasetBannerImageStatus = document.querySelector("#para-dataset-banner-image-status");
+const bfSaveBannerImageBtn = document.getElementById("save-banner-image");
+const datasetBannerImageStatus = document.querySelector("#para-dataset-banner-image-status");
 window.formBannerHeight = document.getElementById("form-banner-height");
 window.guidedFormBannerHeight = document.getElementById("guided-form-banner-height");
 window.currentDatasetLicense = document.querySelector("#para-dataset-license-current");
-// const bfListLicense = document.querySelector("#bf-license-list");
-// const bfAddLicenseBtn = document.getElementById("button-add-license");
+const bfListLicense = document.querySelector("#bf-license-list");
+const bfAddLicenseBtn = document.getElementById("button-add-license");
 
 // // Pennsieve dataset permission //
 window.currentDatasetPermission = document.querySelector("#para-dataset-permission-current");
@@ -8855,178 +8855,178 @@ window.logGeneralOperationsForAnalytics = (category, analyticsPrefix, granularit
 //   $("#dataset_validator_spinner").hide();
 // });
 
-// //function used to scale banner images
-// const scaleBannerImage = async (imagePath) => {
-//   try {
-//     let imageScaled = await client.post(
-//       `/manage_datasets/scale_image`,
-//       {
-//         image_file_path: imagePath,
-//       },
-//       {
-//         params: {
-//           selected_account: window.defaultBfAccount,
-//           selected_dataset: window.defaultBfDataset,
-//         },
-//       }
-//     );
-//     return imageScaled.data.scaled_image_path;
-//   } catch (error) {
-//     clientError(error);
-//     return error.response;
-//   }
-// };
+//function used to scale banner images
+window.scaleBannerImage = async (imagePath) => {
+  try {
+    let imageScaled = await client.post(
+      `/manage_datasets/scale_image`,
+      {
+        image_file_path: imagePath,
+      },
+      {
+        params: {
+          selected_account: window.defaultBfAccount,
+          selected_dataset: window.defaultBfDataset,
+        },
+      }
+    );
+    return imageScaled.data.scaled_image_path;
+  } catch (error) {
+    clientError(error);
+    return error.response;
+  }
+};
 
-// const gatherLogs = () => {
-//   //function will be used to gather all logs on all OS's
-//   let homedir = os.homedir();
-//   let file_path = "";
-//   let clientLogsPath = "";
-//   let serverLogsPath = window.path.join(homedir, "SODA", "logs");
-//   let logFiles = ["main.log", "renderer.log", "agent.log", "api.log"];
+window.gatherLogs = () => {
+  //function will be used to gather all logs on all OS's
+  let homedir = window.os.homedir();
+  let file_path = "";
+  let clientLogsPath = "";
+  let serverLogsPath = window.path.join(homedir, "SODA", "logs");
+  let logFiles = ["main.log", "renderer.log", "agent.log", "api.log"];
 
-//   if (os.platform() === "darwin") {
[…]
