### 2023-12-20 6ef9052
fix: 🐛 format bf_get_accounts returns
