### 2023-12-20 18db64b
fix: banner image size
