### 2023-12-20 c1ac44e
fix: 🐛 format profile name during profile creation
