### 2023-12-20 761ad30
fix: banner image size calculation
