### 2023-12-21 74d91dd
refactor: show ds details in ffm after seletion
