### 2023-12-21 e0c5db1
fix: ffm edit existing banner image
