+++ src/renderer/src/scripts/manage-dataset/manage-dataset.js
@@ -1670,7 +1670,9 @@ $("#edit_banner_image_button").click(async () => {
     //Do nothing... regular import
   } else {
     let img_src = $("#current-banner-img").attr("src");
-    img_src = "file://" + img_src;
+    if(!img_src.includes("https://pennsieve")) {
+      img_src = "file://" + img_src;
+    }
     let img_base64 = await window.getBase64(img_src); // encode image to base64
 
     $("#image-banner").attr("src", "data:image/jpg;base64," + img_base64);
