+++ scripts/manage-dataset/manage-dataset.js
@@ -1741,7 +1741,7 @@ $("#edit_banner_image_button").click(async () => {
     );
 
     window.myCropper.destroy();
-    window.myCropper = new Cropper(document.getElementById("image-banner"), cropOptions);
+    window.myCropper = new Cropper(document.getElementById("image-banner"), window.cropOptions);
   }
 });
 
@@ -1810,7 +1810,7 @@ const displayBannerImage = async (path) => {
                       $("#para-path-image").html(image_path);
                       window.bfViewImportedImage.src = converted_image_file;
                       window.myCropper.destroy();
-                      window.myCropper = new Cropper(window.bfViewImportedImage, cropOptions);
+                      window.myCropper = new Cropper(window.bfViewImportedImage, window.cropOptions);
                       $("#save-banner-image").css("visibility", "visible");
                       $("body").removeClass("waiting");
                     });
@@ -1835,7 +1835,7 @@ const displayBannerImage = async (path) => {
               $("#para-path-image").html(image_path);
               window.bfViewImportedImage.src = image_path;
               window.myCropper.destroy();
-              window.myCropper = new Cropper(window.bfViewImportedImage, cropOptions);
+              window.myCropper = new Cropper(window.bfViewImportedImage, window.cropOptions);
               $("#save-banner-image").css("visibility", "visible");
             }
           });
@@ -1864,7 +1864,7 @@ const displayBannerImage = async (path) => {
       $("#para-path-image").html(image_path);
       window.bfViewImportedImage.src = image_path;
       window.myCropper.destroy();
-      window.myCropper = new Cropper(window.bfViewImportedImage, cropOptions);
+      window.myCropper = new Cropper(window.bfViewImportedImage, window.cropOptions);
 
       $("#save-banner-image").css("visibility", "visible");
     }
