### 2023-12-21 339b546
fix: showcurrentpermissions not defined ffm
