+++ scripts/others/renderer.js
@@ -3276,7 +3276,7 @@ const metadataDatasetlistChange = () => {
 
 // Manage dataset permission
 const permissionDatasetlistChange = () => {
-  showCurrentPermission();
+  window.showCurrentPermission();
 };
 
 const datasetStatusListChange = () => {
