+++ scripts/manage-dataset/manage-dataset.js
@@ -3364,14 +3364,14 @@ $("body").on("change", "input[type=radio][name=dataset_status_radio]", function 
 
 // Change dataset status option change
 $("#bf_list_dataset_status").on("change", async () => {
-  $(bfCurrentDatasetStatusProgress).css("visibility", "visible");
+  $(window.bfCurrentDatasetStatusProgress).css("visibility", "visible");
   $("#bf-dataset-status-spinner").css("display", "block");
 
-  selectOptionColor(bfListDatasetStatus);
+  selectOptionColor(window.bfListDatasetStatus);
 
   let selectedBfAccount = window.defaultBfDataset;
   let selectedBfDataset = window.defaultBfDataset;
-  let selectedStatusOption = bfListDatasetStatus.options[bfListDatasetStatus.selectedIndex].text;
+  let selectedStatusOption = window.bfListDatasetStatus.options[window.bfListDatasetStatus.selectedIndex].text;
 
   log.info(`Changing dataset status to ${selectedStatusOption}`);
 
@@ -3402,7 +3402,7 @@ $("#bf_list_dataset_status").on("change", async () => {
       }
     );
 
-    $(bfCurrentDatasetStatusProgress).css("visibility", "hidden");
+    $(window.bfCurrentDatasetStatusProgress).css("visibility", "hidden");
     $("#bf-dataset-status-spinner").css("display", "none");
 
     Swal.fire({
@@ -3445,15 +3445,15 @@ $("#bf_list_dataset_status").on("change", async () => {
         backdrop: "rgba(0,0,0, 0.4)",
       });
 
-      $(bfCurrentDatasetStatusProgress).css("visibility", "hidden");
+      $(window.bfCurrentDatasetStatusProgress).css("visibility", "hidden");
       $("#bf-dataset-status-spinner").css("display", "none");
     }
 
-    showCurrentDatasetStatus(showErrorDatasetStatus);
+    window.showCurrentDatasetStatus(showErrorDatasetStatus);
   }
 });
 
-const showCurrentDatasetStatus = async (callback) => {
+const window.showCurrentDatasetStatus = async (callback) => {
   let selectedBfAccount = window.defaultBfDataset;
   let selectedBfDataset = window.defaultBfDataset;
 
@@ -3462,13 +3462,13 @@ const showCurrentDatasetStatus = async (callback) => {
   }
 
   if (selectedBfDataset === "Select dataset") {
-    $(bfCurrentDatasetStatusProgress).css("visibility", "hidden");
+    $(window.bfCurrentDatasetStatusProgress).css("visibility", "hidden");
     $("#bf-dataset-status-spinner").css("display", "none");
 
-    window.removeOptions(bfListDatasetStatus);
+    window.removeOptions(window.bfListDatasetStatus);
     removeRadioOptions("dataset_status_ul");
 
-    bfListDatasetStatus.style.color = "black";
+    window.bfListDatasetStatus.style.color = "black";
 
     return;
   }
@@ -3490,7 +3490,7 @@ const showCurrentDatasetStatus = async (callback) => {
       window.defaultBfDatasetId
     );
 
-    window.removeOptions(bfListDatasetStatus);
+    window.removeOptions(window.bfListDatasetStatus);
     removeRadioOptions("dataset_status_ul");
 
     for (let item in res["status_options"]) {
@@ -3500,7 +3500,7 @@ const showCurrentDatasetStatus = async (callback) => {
       option.value = res["status_options"][item]["name"];
       option.style.color = res["status_options"][item]["color"];
 
-      bfListDatasetStatus.appendChild(option);
+      window.bfListDatasetStatus.appendChild(option);
 
       addRadioOption(
[…]
