### 2023-12-21 1dfb6f0
wip: 🚧 add logging to trace authorization issues
