### 2023-12-21 bbba8d1
feat: ✨ Format agent account name when setting organization
