+++ .github/workflows/Build-and-deploy-mac.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - main
       - staging
-      - handle-ps-accounts-with-periods
 
 jobs:
   deploy-on-mac:
