### 2023-12-21 f69c8f9
fix: imageExtension not defined ffm banner image
