+++ scripts/disseminate/prePublishingReview.js
@@ -361,7 +361,7 @@ const window.showPrePublishingStatus = async (inPrePublishing = false, curationM
   if (curationMode === "guided") {
     // Check the status of checklist items for the dataset and alert
     // the user if any of the items are not completed through a sweet alert
-    let checklistItemCheck = allPrepublishingChecklistItemsCompleted("guided");
+    let checklistItemCheck = window.allPrepublishingChecklistItemsCompleted("guided");
     let allItemsChecked = checklistItemCheck[0];
     let checklistItems = checklistItemCheck[1];
     if (!allItemsChecked) {
@@ -476,7 +476,7 @@ const setPrepublishingChecklistItemIconByStatus = (iconElementId, status) => {
 // reads the pre-publishing checklist items from the UI and returns true if all are completed and false otherwise
 // This function checks elements with icon wrapper i class
 // If curationMode is guided the return will include an array of the checklist items that are not completed to alert user
-const allPrepublishingChecklistItemsCompleted = (curationMode) => {
+const window.allPrepublishingChecklistItemsCompleted = (curationMode) => {
   let curationModeID = "";
   let prePublishingChecklistItemNames = [];
   if (curationMode === "guided") {
@@ -517,7 +517,7 @@ const createPrepublishingChecklist = async (curationMode) => {
   }
 
   // check that the user completed all pre-publishing checklist items for the given dataset
-  if (!allPrepublishingChecklistItemsCompleted(curationMode)) {
+  if (!window.allPrepublishingChecklistItemsCompleted(curationMode)) {
     // alert the user they must complete all checklist items before beginning the prepublishing process
     Swal.fire({
       backdrop: "rgba(0,0,0, 0.4)",
