### 2023-12-21 75401a0
chore: update CITATION.cff for Zenodo release
