+++ CITATION.cff
@@ -21,7 +21,7 @@ authors:
     family-names: Portillo
     given-names: Dorian
 cff-version: 1.2.0
-date-released: '2023-12-19'
+date-released: '2023-12-21'
 identifiers:
   - description: DOI for this software's record on Zenodo
     type: doi
@@ -41,4 +41,4 @@ repository-code: https://github.com/fairdataihub/SODA-for-SPARC
 title: SODA (Software to Organize Data Automatically) for SPARC
 type: software
 url: https://fairdataihub.org/
-version: 12.4.1
+version: 12.4.2
