+++ scripts/disseminate/disseminate.js
@@ -48,7 +48,7 @@ const disseminatePublish = async (curationMode) => {
       // submit the dataset for review with the given embargoReleaseDate
       await window.submitReviewDataset(embargoReleaseDate, curationMode);
       $("#prepublishing-submit-btn-container").hide();
-      resetffmPrepublishingUI();
+      window.resetffmPrepublishingUI();
     }
   } catch (error) {
     log.error(error);
