+++ src/renderer/src/scripts/others/renderer.js
@@ -3482,7 +3482,7 @@ window.submitReviewDataset = async (embargoReleaseDate, curationMode) => {
   });
 
   if (curationMode != "guided") {
-    await resetffmPrepublishingUI();
+    await window.resetffmPrepublishingUI();
   } else {
     // Update the UI again and hide the flow
     $("#guided--prepublishing-checklist-container").addClass("hidden");
@@ -3639,7 +3639,7 @@ const withdrawReviewDataset = async (curationMode) => {
     if (curationMode != "guided") {
       // reveal the current section (question-3) again using the new publishing status value
       // TODO : INCLUDE for bundling
-      await resetffmPrepublishingUI();
+      await window.resetffmPrepublishingUI();
 
       bfRefreshPublishingDatasetStatusBtn.disabled = false;
       bfWithdrawReviewDatasetBtn.disabled = false;
