### 2023-12-21 b2b41e9
chore: add logs to disseminate
