+++ src/main/index.js
@@ -516,10 +516,10 @@ const initialize = () => {
         }
 
         // start_pre_flight_checks();
-        if (!buildIsBeta) {
+        // if (!buildIsBeta) {
           log.info("Checking for updates in initialize");
           autoUpdater.checkForUpdatesAndNotify();
-        }
+        // }
         updatechecked = true;
       }, 6000);
     })
