### 2023-12-21 907bf7a
build: 📦️ update build scripts
