### 2023-12-21 41a8e7f
style: 🎨 fix code style issues with Prettier
