### 2023-12-22 #6140 @bvhpatel
@abhishektiwari @AJQuinn I would not say SODA has a hard dependency on Pennsieve. The main purpose of SODA is to allow researchers to format their research data according to the SPARC Dataset Structure (SDS). If the data is funded by the NIH SPARC Program, they need researchers to upload the data to the Pennsieve platform after it is formatted according to the SDS. Since the development of SODA is funded by SPARC, many of our workflows end with telling users to upload their data to Pennsieve and this is why we make sure to remind them about connecting their Pennsieve account on start. We use the Pennsieve Python client (https://github.com/Pennsieve/pennsieve-python) in the backend of SODA to upload the SDS formatted data on Pennsieve if the user choose to do so. Now that the SDS is progressively being adopted outside of the SPARC Program, we anticipate SODA being used by users not necessarily having a Pennsieve account. There are many features that already allow that (for examples, all features under Free Form Mode --> Prepare Metadata) and all of the remaining workflows that require to upload to Pennsieve (e.g., End to End Curation) will also have the option to just safe the data locally on the user's computer. I hope that clarifies your point.
