### 2023-12-27 2c30258
feat: ✨ gm allow empty datasets for starting from Pennsieve flow
