+++ scripts/guided-mode/guided-curate-dataset.js
@@ -655,180 +655,189 @@ const savePageChanges = async (pageBeingLeftID) => {
           throw errorArray;
         }
 
-        // Reject if anyh non-sparc folders are in the root of the dataset
-        let invalidBaseFolders = [];
-        for (const baseFolder of Object.keys(datasetStructureJSONObj["folders"])) {
-          if (
-            !guidedHighLevelFolders.includes(baseFolder) &&
-            !nonGuidedHighLevelFolders.includes(baseFolder)
-          ) {
-            invalidBaseFolders.push(baseFolder);
+        // If there are folders and/or files in the dataset on Pennsieve, validate the structure and
+        // extract various metadata from the dataset structure to prepare the guided workflow.
+        if (
+          Object.keys(datasetStructureJSONObj["folders"]).length > 0 ||
+          Object.keys(datasetStructureJSONObj["files"]).length > 0
+        ) {
+          // Reject if any non-sparc folders are in the root of the dataset
+          let invalidBaseFolders = [];
+          for (const baseFolder of Object.keys(datasetStructureJSONObj["folders"])) {
+            if (
+              !guidedHighLevelFolders.includes(baseFolder) &&
+              !nonGuidedHighLevelFolders.includes(baseFolder)
+            ) {
+              invalidBaseFolders.push(baseFolder);
+            }
           }
-        }
-        if (invalidBaseFolders.length > 0) {
-          errorArray.push({
-            type: "swal",
-            title: "This dataset is not eligible to be edited via Guided Mode",
-            message: `The following folders are not allowed in the root of your dataset: ${invalidBaseFolders.join(
-              ", "
-            )}`,
-          });
-          throw errorArray;
-        }
+          if (invalidBaseFolders.length > 0) {
+            errorArray.push({
+              type: "swal",
+              title: "This dataset is not eligible to be edited via Guided Mode",
+              message: `The following folders are not allowed in the root of your dataset: ${invalidBaseFolders.join(
+                ", "
+              )}`,
+            });
+            throw errorArray;
+          }
 
-        // Datasets pulled into Guided Mode should only have pool-folders or sub-folders inside of the primary, source,
-        // and derivative high level folders. If this is not the case with the pulled dataset, reject it.
-        const [invalidFolders, invalidFiles] =
-          guidedCheckHighLevelFoldersForImproperFiles(datasetStructureJSONObj);
-        if (invalidFolders.length > 0 || invalidFiles.length > 0) {
-          errorArray.push({
-            type: "swal",
-            title: "This dataset is not eligible to be edited via Guided Mode",
-            message: `
+          // Datasets pulled into Guided Mode should only have pool-folders or sub-folders inside of the primary, source,
+          // and derivative high level folders. If this is not the case with the pulled dataset, reject it.
+          const [invalidFolders, invalidFiles] =
+            guidedCheckHighLevelFoldersForImproperFiles(datasetStructureJSONObj);
+          if (invalidFolders.length > 0 || invalidFiles.length > 0) {
+            errorArray.push({
+              type: "swal",
+              title: "This dataset is not eligible to be edited via Guided Mode",
+              message: `
               Your primary, source, and derivative folders must only contain pool- folders or sub- folders when resuming a Pennsieve dataset via Guided Mode
               <br />
               <br />
               Please remove any folders or files that are not pool- folders or sub- folders from your primary, source, and derivative folders and try again.
             `,
-          });
-          throw errorArray;
-        }
+            });
+            throw errorArray;
+          }
 
-        // Extract the pool/subject/sample structure from the folders and files pulled from Pennsieve
[…]
