### 2023-12-28 9c36419
feat: ✨ Added analytics tracking for Pennsieve metadata modifications
