+++ scripts/guided-mode/guided-curate-dataset.js
@@ -13789,6 +13789,7 @@ const guidedCreateOrRenameDataset = async (bfAccount, datasetName) => {
     throw new Error(userErrorMessage(error));
   }
 };
+
 const guidedAddDatasetSubtitle = async (bfAccount, datasetName, datasetSubtitle) => {
   document.getElementById("guided-dataset-subtitle-upload-tr").classList.remove("hidden");
   const datasetSubtitleUploadText = document.getElementById("guided-dataset-subtitle-upload-text");
@@ -13822,7 +13823,7 @@ const guidedAddDatasetSubtitle = async (bfAccount, datasetName, datasetSubtitle)
     sodaJSONObj["previously-uploaded-data"]["subtitle"] = datasetSubtitle;
     await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
 
-    // Track the successful upload of a dataset subtitle via Kombucha
+    // Send successful dataset subtitle upload event to Kombucha
     ipcRenderer.send(
       "track-kombucha",
       kombuchaEnums.Category.GUIDED_MODE,
@@ -13842,11 +13843,30 @@ const guidedAddDatasetSubtitle = async (bfAccount, datasetName, datasetSubtitle)
       guidedGetDatasetId(sodaJSONObj)
     );
   } catch (error) {
+    // Send failed dataset subtitle upload event to Kombucha
+    ipcRenderer.send(
+      "track-kombucha",
+      kombuchaEnums.Category.GUIDED_MODE,
+      kombuchaEnums.Action.ADD_EDIT_DATASET_METADATA,
+      kombuchaEnums.Label.SUBTITLE,
+      kombuchaEnums.Status.FAIL,
+      {
+        value: 1,
+        dataset_name: guidedGetDatasetName(sodaJSONObj),
+        dataset_id: guidedGetDatasetId(sodaJSONObj),
+      }
+    );
+    ipcRenderer.send(
+      "track-event",
+      "Error",
+      ManageDatasetsAnalyticsPrefix.MANAGE_DATASETS_ADD_EDIT_SUBTITLE,
+      guidedGetDatasetId(sodaJSONObj)
+    );
     console.error(error);
     let emessage = userErrorMessage(error);
     datasetSubtitleUploadText.innerHTML = "Failed to add a dataset subtitle.";
     guidedUploadStatusIcon("guided-dataset-subtitle-upload-status", "error");
-    throw new Error(userErrorMessage(error));
+    throw new Error(emessage);
   }
 };
 
@@ -13891,10 +13911,49 @@ const guidedAddDatasetDescription = async (
     guidedUploadStatusIcon("guided-dataset-description-upload-status", "success");
     sodaJSONObj["previously-uploaded-data"]["description"] = description;
     await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
+
+    // Send successful dataset description upload event to Kombucha
+    ipcRenderer.send(
+      "track-kombucha",
+      kombuchaEnums.Category.GUIDED_MODE,
+      kombuchaEnums.Action.ADD_EDIT_DATASET_METADATA,
+      kombuchaEnums.Label.DESCRIPTION,
+      kombuchaEnums.Status.SUCCESS,
+      {
+        value: 1,
+        dataset_name: guidedGetDatasetName(sodaJSONObj),
+        dataset_id: guidedGetDatasetId(sodaJSONObj),
+      }
+    );
+    ipcRenderer.send(
+      "track-event",
+      "Success",
+      ManageDatasetsAnalyticsPrefix.MANAGE_DATASETS_ADD_EDIT_SUBTITLE,
[…]
