### 2023-12-28 ea3d3e0
fix: 🐛 standardize enums with ffm
