### 2024-01-02 bd858f8
feat: ✨ Added subjects metadata analytics tracking
