### 2024-01-02 d32f47b
feat: ✨ Added analytics tracking for dataset metadata
