+++ scripts/guided-mode/guided-curate-dataset.js
@@ -14647,20 +14647,25 @@ const guidedGenerateSubjectsMetadata = async (destination) => {
       kombuchaEnums.Status.SUCCESS,
       guidedCreateEventDataPrepareMetadata(generationDestination, 1)
     );
-    ipcRenderer.send(
-      "track-event",
-      "Success",
-      ManageDatasetsAnalyticsPrefix.MANAGE_DATASETS_ADD_EDIT_PERMISSIONS,
-      guidedGetDatasetId(sodaJSONObj)
-    );
   } catch (error) {
-    guidedUploadStatusIcon(
-      `guided-subjects-metadata-${generationDestination}-genration-status`,
-      "error"
+    const emessage = userErrorMessage(error);
+    if (generationDestination === "Pennsieve") {
+      guidedUploadStatusIcon(
+        `guided-subjects-metadata-${generationDestination}-genration-status`,
+        "error"
+      );
+      subjectsMetadataGenerationText.innerHTML = `Failed to generate subjects metadata`;
+    }
+    // Send failed subjects metadata generation event to Kombucha
+    ipcRenderer.send(
+      "track-kombucha",
+      kombuchaEnums.Category.GUIDED_MODE,
+      kombuchaEnums.Action.GENERATE_METADATA,
+      kombuchaEnums.Label.SUBJECTS_XLSX,
+      kombuchaEnums.Status.FAIL,
+      guidedCreateEventDataPrepareMetadata(generationDestination, 1)
     );
-    subjectsMetadataGenerationText.innerHTML = `Failed to generate subjects metadata`;
-    clientError(error);
-    throw new Error(userErrorMessage(error));
+    throw new Error(emessage);
   }
 };
 const guidedGenerateSamplesMetadata = async (bfAccount, datasetName, samplesTableData) => {
@@ -14707,12 +14712,30 @@ const guidedGenerateSamplesMetadata = async (bfAccount, datasetName, samplesTabl
     samplesMetadataUploadText.innerHTML = `Samples metadata successfully uploaded`;
     sodaJSONObj["previously-uploaded-data"]["samples-metadata"] = samplesTableData;
     await saveGuidedProgress(sodaJSONObj["digital-metadata"]["name"]);
+    // Send successful samples metadata generation event to Kombucha
+    ipcRenderer.send(
+      "track-kombucha",
+      kombuchaEnums.Category.GUIDED_MODE,
+      kombuchaEnums.Action.GENERATE_METADATA,
+      kombuchaEnums.Label.SAMPLES_XLSX,
+      kombuchaEnums.Status.SUCCESS,
+      guidedCreateEventDataPrepareMetadata("Pennsieve", 1)
+    );
   } catch (error) {
+    const emessage = userErrorMessage(error);
     guidedUploadStatusIcon("guided-samples-metadata-pennsieve-genration-status", "error");
     samplesMetadataUploadText.innerHTML = `Failed to upload samples metadata`;
-    clientError(error);
+    // Send failed samples metadata generation event to Kombucha
+    ipcRenderer.send(
+      "track-kombucha",
+      kombuchaEnums.Category.GUIDED_MODE,
+      kombuchaEnums.Action.GENERATE_METADATA,
+      kombuchaEnums.Label.SAMPLES_XLSX,
+      kombuchaEnums.Status.FAIL,
+      guidedCreateEventDataPrepareMetadata("Pennsieve", 1)
+    );
 
-    throw new Error(userErrorMessage(error));
+    throw new Error(emessage);
   }
 };
 
@@ -14756,12 +14779,32 @@ const guidedGenerateSubmissionMetadata = async (bfAccount, datasetName, submissi
[…]
