### 2024-01-03 3a033ce
feat: ✨ handle previous account key formats when confirming worksapce
