+++ scripts/others/api/api.js
@@ -490,6 +490,12 @@ const getUserPoolAccessToken = async (email, password) => {
   return response.data;
 };
 
+
+const getUserWorkspaces = async () => {
+  const response = await client.get("/user/organizations");
+  return response.data;
+}
+
 const api = {
   getUserInformation,
   getDataset,
@@ -520,6 +526,7 @@ const api = {
   getUserPoolAccessToken,
   setDefaultProfile,
   createProfileName,
+  getUserWorkspaces
 };
 
 module.exports = api;
