### 2024-01-03 1254640
feat: ✨ workspace switch uses new api key format
