### 2024-01-03 3e704d3
wip: 🚧 comment out local generation for metadata analytics release
