+++ .github/workflows/Build-and-deploy-win.yml
@@ -4,8 +4,6 @@ run-name: ${{ github.actor }} is building a Windows release for SODA
 on:
   push:
     branches:
-      - main
-      - staging
       - switch-fix-reuse-valid-keys
 
 jobs:
