### 2024-01-04 e9d1214
fix: 🐛 remove depcrecated track-event ipcRenderer calls
