### 2024-01-04 f9f4233
wip: 🚧 delete agent db files
