+++ scripts/others/renderer.js
@@ -910,13 +910,15 @@ const run_pre_flight_checks = async (check_update = true) => {
       clientError(error);
       const emessage = userErrorMessage(error);
 
-      // check if the Agent is failing to start due to Unique constraint violation 
-      // if so then we prompt the user to allow us to remove the pennsieve Agent DB files and try again 
-      if (emessage.includes("UNIQUE constraint failed: user_settings.user_id, user_settings.profile")) {
-
+      // check if the Agent is failing to start due to Unique constraint violation
+      // if so then we prompt the user to allow us to remove the pennsieve Agent DB files and try again
+      if (
+        emessage.includes("UNIQUE constraint failed: user_settings.user_id, user_settings.profile")
+      ) {
         const { value: deleteFilesRerunChecks } = await Swal.fire({
           icon: "error",
-          title: "The Pennsieve Agent Failed to Start Due to A Duplicate User Profile Being Detected",
+          title:
+            "The Pennsieve Agent Failed to Start Due to A Duplicate User Profile Being Detected",
           html: `This is a known issue with the Pennsieve Agent. 
                  To resolve it SODA will need to delete the Pennsieve Agent database files. 
                  This will not affect your Pennsieve account, any of your existing datasets, or upload manifests stored on Pennsieve.
@@ -931,16 +933,16 @@ const run_pre_flight_checks = async (check_update = true) => {
           reverseButtons: reverseSwalButtons,
           confirmButtonText: "Delete Database Files and Try Again",
           cancelButtonText: "Exit SODA",
-        })
+        });
 
         if (!deleteFilesRerunChecks) {
           await ipcRenderer.invoke("quit-app");
         }
 
         // wait for the Agent to stop using the db files so they may be deleted
-        await wait(1000)
-        const fs = require('fs').promises
-        const fsSync = require('fs')
+        await wait(1000);
+        const fs = require("fs").promises;
+        const fsSync = require("fs");
         if (fsSync.existsSync(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db`))
           await fs.unlink(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db`);
         if (fsSync.existsSync(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db-shm`))
@@ -948,7 +950,7 @@ const run_pre_flight_checks = async (check_update = true) => {
         if (fsSync.existsSync(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db-wal`))
           await fs.unlink(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db-wal`);
 
-        // rerun checks 
+        // rerun checks
         return await run_pre_flight_checks();
       }
 
@@ -1922,7 +1924,7 @@ ipcRenderer.on("selected-generate-metadata-subjects", (event, dirpath, filename)
             didOpen: () => {
               Swal.showLoading();
             },
-          }).then((result) => { });
+          }).then((result) => {});
           generateSubjectsFileHelper(false);
         }
       });
@@ -1938,7 +1940,7 @@ ipcRenderer.on("selected-generate-metadata-subjects", (event, dirpath, filename)
         didOpen: () => {
           Swal.showLoading();
         },
-      }).then((result) => { });
+      }).then((result) => {});
       generateSubjectsFileHelper(false);
     }
   }
@@ -2021,7 +2023,7 @@ const generateSubjectsFileHelper = async (uploadBFBoolean) => {
     didOpen: () => {
       Swal.showLoading();
     },
-  }).then((result) => { });
+  }).then((result) => {});
 
   try {
[…]
