### 2024-01-04 6e0f885
feat: ✨ delete db files if duplicate user exists
