+++ scripts/others/renderer.js
@@ -917,7 +917,8 @@ const run_pre_flight_checks = async (check_update = true) => {
       ) {
         const { value: deleteFilesRerunChecks } = await Swal.fire({
           icon: "error",
-          title: "The Pennsieve Agent Failed to Start Due to A Duplicate User Profile Being Detected",
+          title:
+            "The Pennsieve Agent Failed to Start Due to A Duplicate User Profile Being Detected",
           html: `<p style="text-align: left">This is a known issue with the Pennsieve Agent. 
                  To resolve it SODA will need to delete the Pennsieve Agent database files. 
                  This will not affect your Pennsieve account, any of your existing datasets, or any existing upload manifests stored on Pennsieve.
@@ -939,9 +940,9 @@ const run_pre_flight_checks = async (check_update = true) => {
         }
 
         // wait for the Agent to stop using the db files so they may be deleted
-        await wait(1000)
-        const fs = require('fs').promises
-        const fsSync = require('fs')
+        await wait(1000);
+        const fs = require("fs").promises;
+        const fsSync = require("fs");
         // delete any db files that exist
         if (fsSync.existsSync(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db`))
           await fs.unlink(`${app.getPath("home")}/.pennsieve/pennsieve_agent.db`);
