### 2024-01-14 #6140 @mattsouth
Ive had a look at the software and worked my way through the checklist.   There are few points that prevent me from checking off all items, all of which I assume can be resolved.

authorship

Two of the largest contributors to the associated github project (Aaron M and Tram Ngo) arent listed as an author on the paper and the first author doesnt seem to have contributed to the code.   What is the contribution of Christopher Marroquin and should the named developers be added as authors?

installation

There are 65 vulnerabilities (inc 2 marked as critical) highlighted by npm on install.  There is an existing issue in the github project (#206) regarding updating dependencies that might help resolve this, though its now over a year old.  I suggest the authors review and ideally address these vulns.  Furthermore it may be an idea to introduce a mechanism for reviewing these in the build process before acceptance.

testing

I cant find any manual or automated tests to run and, similarly, I cant find an example project that I can use to check the functionality of the application.  Could the authors describe their testing procedure and would it be possible for the authors to provide an example that can be used to test the functionality provided?

API documentation

There is good documentation for this project that is end-user focussed, but not API focussed.  It seems to me that the API consists mainly of endpoints and functions within a python flask app, and comment coverage seems pretty good within the codebase.  It may be possible to automatically generate some useful API documentation with a tool like https://github.com/acoomans/flask-autodoc (or similar).

A note on the paper

The paper makes a general statement that I suggest needs to be qualified somewhat:

>  To our knowledge, SODA is the first researcher-oriented tool for making data FAIR.

As a counter example, Gorgolewski, K., Auer, T., Calhoun, V. et al (2016) - https://doi.org/10.1038/sdata.2016.44 describe a similar workflow to that which SODA supports with their BIDS validator.  This paper was published a year before the SPARC project started and is referenced in the SDS paper and I would argue is a research-oriented tool for making data FAIR.
