### 2024-01-31 #6140 @bvhpatel
Thanks for your comments @mattsouth! Please see our responses below:

**Authorship**
Christopher Marroquin is actually Aaron M. (he goes by both Christopher and Aaron). Tram Ngo left the team about 2-3 years ago and we have no way of reaching out to her to get her agreement to be listed as a co-author (which I understand is required as per JOSS’s authorship guidelines).

**Installation**
The existing issue #206 is rather old (we have now closed it) and some results are no longer applicable however we have looked at the recent results from running npm audit. We have now addressed both 'Critical' level security vulnerabilities and will include these updates into the next release. The 'High' vulnerabilities stem from 'Electron', 'url-regex', and 'xlsx', 'diff', and 'glob-parent'. Here is further information regarding these vulnerabilities:
* Updating Electron will be addressed in an upcoming milestone that focuses on addressing security vulnerabilities.
* 'xlsx' is used by 'convert-excel-to-json' which we use heavily throughout SODA. The alternative packages we could find are also outdated. Given that the vulnerability stems from crafted .xslx files from arbitrary sources we are unlikely to have issues due to the fact that all of our data comes from Pennsieve, the NIH SPARC supported data platform, which is a trusted source.
* 'url-regex' and 'glob-parent' and 'diff' vulnerable versions are used in a package we use in Free Form Mode for UIs called 'fomantic-ui'. We will be replacing 'fomantic-ui' with a more secure UI library as part of our next milestone.

Overall, most of the remaining vulnerabilities will be addressed around mid-2024 as part of a milestone we have in our current funding from SPARC specifically to do that. As for mechanisms for reviewing vulnerabilities we currently use dependabot and renovate for getting pull requests with dependency updates. However going forward we will likely add an action to run npm audit on PRs and add Snyk integration to create additional PRs for any security issues our other methods may have missed. In addition we also have several code level vulnerability detection tools: Sourcery and Sonarcloud.

**Testing**
Our testing procedure consists of manually checking the functionalities that have been touched in a given new release. We have a milestone as part of our current funding from SPARC for implementing automated testing approaches as manual testing has now become very time consuming with the expansion of the app. Our plan under that milestone is to use the Jest testing framework for testing app-wide web components. Additionally, Postman will be used for end-to-end testing of our Flask backend. We will also leverage Chromatic for UI regression testing of pages. For now, we have added a new page [here](https://docs.sodaforsparc.io/docs/how-to/how-to-test-soda-without-a-pennsieve-account) in our user documentation to provide instructions for testing some features of SODA that work without necessitating access to the Pennsieve platform. You may want to download and use the latest version of SODA (available [here](https://docs.sodaforsparc.io/docs/getting-started/download-soda)) for that.

**API documentation**
To view the API documentation run the application and then enter http://localhost:4242 in your browser to see the Swagger documentation. This has been added to the SODA for SPARC Docs [here](https://docs.sodaforsparc.io/docs/developer-documentation/viewing-api-documentation) so this is more clear and available going forward.

**Note on the paper**
By researcher-oriented tool, we mean a tool that guides through all the steps to make data FAIR via intuitive user interfaces and doesn’t require any coding knowledge. To our knowledge SODA is the first to do that as it guides users through all the steps for converting their raw data into SDS-compliant format with all required metadata, validating the dataset, and uploading it on the SPARC mandated repository. At the time we established SODA, creating a BIDS compliant dataset required manual effort (e.g., Step 2 described in the paper required manually creating the BIDS folder structure, etc.). The BIDS validator was indeed a researcher-oriented tool but that was useful only in the final step for validating the dataset and not for creating a FAIR dataset. We have now edited the sentence in the paper to state “To our knowledge, SODA is the first researcher-oriented tool that guides step-by-step through all the requirements for making data FAIR from organizing files, creating metadata, validating, and uploading to a data repository.”
