### 2024-02-06 fairdataihub/SODA-for-SPARC#296 @abhishektiwari
[JOSS Review] Inconsistent use of HTTP status code
Inconsistent use of HTTP status code particularly in Flask authentication and authorisation.

A few examples:

Using `400` (bad request) instead of `401` (unauthorized) [1](https://github.com/fairdataihub/SODA-for-SPARC/blob/1c9585e385083444811cfac048fbbd3478a08360/src/pyflask/users/user.py#L42)

Using `500` (Internal Server Error) instead of `400` (bad request) [1](https://github.com/fairdataihub/SODA-for-SPARC/blob/1c9585e385083444811cfac048fbbd3478a08360/src/pyflask/permissions/permissions.py#L30C15-L30C18)

Translating `400`(bad request) into `409` (conflict)  [1](https://github.com/fairdataihub/SODA-for-SPARC/blob/1c9585e385083444811cfac048fbbd3478a08360/src/pyflask/users/user.py#L41C7-L41C74)

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6140)
