### 2024-02-06 fairdataihub/SODA-for-SPARC#299 @abhishektiwari
[JOSS Review] Update the documentation portal link in README
Update link [our dedicated webpage](https://fairdataihub.org/sodaforsparc) under [documentation](https://docs.sodaforsparc.io) section of README.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6140)
