### 2024-02-06 #6140 @abhishektiwari
@bvhpatel First of great job on putting a curator friendly software workflow and comprehensive documentation for SODA. I think most of functionality works as documented and described in the paper. Just make sure the documentation, particularly images and videos are in sync with latest changes, otherwise images and videos will regress from text version.

Now coming back to software itself I see a lot of Python and JS anti-patterns in the implementation. I created a few GitHub Issues around (1) inconsistent use of HTTP status code, (2) duplicating already declared constants throughout the code base, (3) missing authentication guards for flask_restx APIs, (4) unused imports thought out the codebase, (5) outdated and vulnerable NPM packages, etc. Out of these - 3 and 5 really worries me, I would not install a software which is full of vulnerabilities and missing safeguards on access control. Again these are suggested improvements neither of them are blocking at this point.

There is no automated tests - bare minimum API tests and e2e UI tests are desirable. I saw your previous response and plans to add these at a later point and I am fine with it but my concern is that code may not be testable unless you take a major refactoring.
