### 2024-03-31 #6140 @yarikoptic
NB I somewhat lost enthusiasm/got some negative vibe from the project in that my [trivial PR](https://github.com/fairdataihub/SODA-for-SPARC/pull/230) to the project was never reviewed/merged despite my multiple pings (now - full of conflicts).

I never found quality time to do the review and will be busy/traveling for next 2 weeks.  Do you think you could proceed with two reviewers or I should strive to actually complete the review? ;)
