### 2024-04-26 47f566e
fix: inconsistent status codes #296
