+++ src/pyflask/manageDatasets/manage_datasets.py
@@ -127,10 +127,10 @@ def bf_add_account_api_key(keyname, key, secret):
     try:
         formatted_key_name = format_agent_profile_name(keyname)
         if (not formatted_key_name) or (not key) or (not secret):
-            abort(401, "Please enter valid keyname, key, and/or secret")
+            abort(400, "Please enter valid keyname, key, and/or secret")
 
         if (formatted_key_name.isspace()) or (key.isspace()) or (secret.isspace()):
-            abort(401, "Please enter valid keyname, key, and/or secret")
+            abort(400, "Please enter valid keyname, key, and/or secret")
 
         ps_path = join(userpath, ".pennsieve")
         # Load existing or create new config file
@@ -228,31 +228,28 @@ def bf_account_list():
     Action:
         Returns list of accounts stored in the system
     """
-    try:
-        accountlist = ["Select"]
-        if exists(configpath):
-            valid_account = bf_get_accounts()
-            if valid_account != "":
-                accountlist.append(valid_account)
-        return {"accounts": accountlist}
 
-    except Exception as e:
-        raise e
+    accountlist = ["Select"]
+    if exists(configpath):
+        valid_account = bf_get_accounts()
+        if valid_account != "":
+            accountlist.append(valid_account)
+    return {"accounts": accountlist}
+
 
 def bf_default_account_load():
     """
     Action:
         Returns the first valid account as the default account
     """
-    try:
-        accountlist = []
-        if exists(configpath):
-            valid_account = bf_get_accounts()
-            if valid_account != "":
-                accountlist.append(valid_account)
-        return {"defaultAccounts": accountlist}
-    except Exception as e:
-        raise e
+
+    accountlist = []
+    if exists(configpath):
+        valid_account = bf_get_accounts()
+        if valid_account != "":
+            accountlist.append(valid_account)
+    return {"defaultAccounts": accountlist}
+
 
 
 
@@ -321,10 +318,9 @@ def bf_dataset_account(accountname):
     global namespace_logger
 
     # get the datasets the user has access to
-    try:
-        datasets = get_users_dataset_list()
-    except Exception as e:
-        raise e
+
+    datasets = get_users_dataset_list()
+
 
 
     datasets_list = []
@@ -376,10 +372,8 @@ def get_username(accountname):
         abort(400, "Please select a valid Pennsieve account.")
 
     # request the user's first and last name stored on Pennsieve
-    try:
-        user_info = get_user_information(token)
-    except Exception as e:
-        abort(500, "Something went wrong while authenticating the user or connecting to Pennsieve.")
+    user_info = get_user_information(token)
+
     
     username = f"{user_info['firstName']} {user_info['lastName']}"
 
@@ -417,10 +411,9 @@ def bf_account_details(accountname):
     except Exception as e:
         abort(400, "Please select a valid Pennsieve account.")
 
-    try:
-        user_info = get_user_information(token)
-    except Exception as e:
-        abort(500, "Something went wrong while authenticating the user or connecting to Pennsieve.")
+  
+    user_info = get_user_information(token)
+
 
     user_email = user_info['email']
     organization_id = user_info['preferredOrganization']
@@ -438,14 +431,13 @@ def bf_account_details(accountname):
             organization = org["organization"]["name"]
 
 
-    try:
-        update_config_account_name(accountname)
-        
-        ## return account details and datasets where such an account has some permission
-        return {"email": user_email, "organization": organization}
+  
+    update_config_account_name(accountname)
+    
[…]
