+++ src/pyflask/datasets/datasets.py
@@ -21,9 +21,7 @@ def get_role(dataset):
         role = r.json()["role"]
         return {"role": role}
     except Exception as e:
-        if type(e).__name__ == "HTTPError":
-            abort(400, e.response.json()["message"])
-        abort(500, "An internal server error prevented the request from being fulfilled. Please try again later.")
+        abort(e.response.status_code,  e.response.json().get('message'))
 
 
 def get_dataset_by_id(dataset_name_or_id):
@@ -129,9 +127,8 @@ def reserve_dataset_doi(dataset):  # sourcery skip: extract-method
         return {"doi": doi_request.json()["doi"]}
     except Exception as e:
         print(e)
-        if type(e).__name__ == "HTTPError":
-            abort(400, e.response.json()["message"])
-        abort(500, "An internal server error prevented the request from being fulfilled. Please try again later.")
+        abort(e.response.status_code,  e.response.json().get('message'))
+
 
 def get_dataset_doi(dataset):
     """
@@ -157,9 +154,8 @@ def get_dataset_doi(dataset):
         doi_request.raise_for_status()
         return {"doi": doi_request.json()["doi"]}
     except Exception as e:
-        if type(e).__name__ == "HTTPError":
-            abort(400, e.response.json()["message"])
-        abort(500, "An internal server error prevented the request from being fulfilled. Please try again later.")
+        abort(e.response.status_code,  e.response.json().get('message'))
+        
 
 
 def get_package_type_counts(dataset_name):
