+++ src/pyflask/organizeDatasets/organize_datasets.py
@@ -1165,7 +1165,7 @@ def import_pennsieve_dataset(soda_json_structure, requested_sparc_only=True):
             curatestatus = "Done"
             raise Exception("You don't have permissions for uploading to this Pennsieve dataset")
     except Exception as e:
-        abort(401, "You do not have permissions to edit upload this Pennsieve dataset.")
+        abort(403, "You do not have permissions to edit upload this Pennsieve dataset.")
 
 
     # surface layer of dataset is pulled. then go through through the children to get information on subfolders

[…]
