### 2024-04-26 fairdataihub/SODA-for-SPARC#299 @aaronm-2112
Closing the issue as it has been addressed as per 2094938

Thanks for the feedback!
