### 2024-04-26 fairdataihub/SODA-for-SPARC#296 @aaronm-2112
This issue has been addressed as per 47f566e.

Thanks for the feedback!
