### 2024-05-01 #6140 @mattsouth
I think this question is one for the editors who may have something to say about timelines, but from my perspective I dont see that you need to withdraw and resubmit.  For me that would unnecessarily complicate things.
