### 2024-05-13 #6140 @sappelhoff
@mattsouth

could you kindly provide a status update of your review for me, now that I have taken over from `@AJQuinn` as the new editor? For example:

- Have all your comments from https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-1891021434 been addressed?
