### 2024-05-13 #6140 @sappelhoff
@bvhpatel

> We think that some of the remaining recommendations made by `@abhishektiwari`, especially around safeguard and access control, would require us more time. We do have a milestone to address these sort of things and do believe they need to be addressed but it's getting pushed back due to other priorities. `@mattsouth` `@Kevin-Mattheus-Moerman` to be respectful of everyone's time, would it make sense that we address the issues I mentioned are addressable, withdraw the paper, and resubmit when we are done address remaining issues (likely later this year)?

From my perspective (and if the other reviewers agree), a withdrawal and resubmission would unnecessarily complicate things, so I agree with @mattsouth in https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2088534640.

I suggest you address all issues that are reasonably fixable within the next weeks (let's say around 3-4 weeks), and furthermore document (and continue to create) milestones and plans to address all remaining issues within longer time frames. You could then briefly discuss the current lack of these features in the paper, together with a commitment and time table for completion.

I am making this suggestion based on @abhishektiwari's [comment](https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-1928723929):

> I will wait to hear from authors on the feedback I provided - none of which is blocking.

and based on @mattsouth's comment: https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2088534640

If @yarikoptic [agrees to review this](https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2108645256) and recommends against accepting this paper BEFORE the "milestones-that-may-take-longer" are implemented, we will have to extend this review process.
