### 2024-06-02 #6140 @sappelhoff
> we will go over the remaining ones again to double check. For those we deem will need more time/support to address, we will mention those in the paper along with a tentative plan for addressing them.

@bvhpatel do you have an update on whether you completed this?

---

@yarikoptic I haven't heard back from you, which I am interpreting as your wish to be unassigned. I will do that right away. Thanks a lot for your efforts up to this point and I hope you will consider reviewing again for JOSS in the future!

---

@mattsouth could you please reply to my comment above? --> https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2108655780
