### 2024-06-23 #6140 @bvhpatel
Hi @sappelhoff! There are two main suggestions we cannot address at this moment: adding an authentication guard for the Flask server and removing duplicate constants. As per your suggestion, I have included this paragraph in the paper: "The development of SODA is still on-going to further simplify the process of formatting datasets according to the SDS and publishing them on the SPARC data portal. Additionally, there are several software engineering improvements that need to be made, such as adding an authentication guard for the Flask server and removing duplicate constants. They are anticipated to be completed by the end of the next SODA grant cycle in August 2025.". I hope this works.

I also took the time to track Tram Ngo and added her as an author after receiving her permission (as per @mattsouth comment discussed [here](https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-1918380996))

Thanks!
