### 2024-07-08 #6140 @sappelhoff
Thanks a lot @mattsouth and @abhishektiwari for your reviews! 👏

@bvhpatel I created a "final checklist" for you and me to go through. Could you please have a look at the upper part in here: https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2214041482 ... and let me know once you are finished.

I will do several checks on my own as well and might comment in this thread if I find something that needs to be changed.
