### 2024-07-08 #6140 @sappelhoff
@bvhpatel I furthermore just noticed that there is an already existing publication on the SODA software, unless I am mistaken: https://doi.org/10.1096/fasebj.2020.34.s1.02483

Could you please let me know:

1. whether this was discussed beforehand with the previous editor or associate editor in chief ... and if so, where?
2. ... and if not: what are the reasons that warrant an additional publication of the SODA software in JOSS?
