### 2024-07-17 #6140 @bvhpatel
@sappelhoff This is a conference abstract from 4+ years ago that was published by the conference. This was really in the early days of SODA when we were just getting started. I believe I reference it in the JOSS paper and I don't think there is much overlap besides the high-level description of SODA. Do you think I should talk we the editor? If so, how?
