### 2024-07-18 c1e0763
feat: Update CHANGELOG.md
