+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to SODA will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## V15.1.0 - 2024-07-17
+## V15.1.0 - 2024-07-18
 
 ## Feature Additions:
 
