### 2024-07-20 #6140 @bvhpatel
@sappelhoff please see below. I believe we completed all the tasks. The only thing I changed in the paper is the UI screenshot (Figure 1) to showcase the latest change in the UI. Thanks!

> ## Additional Author Tasks After Review is Complete
> * [x]  ask Tram Ngo to please provide an ORCID if possible
_This would be difficult as we are not closely in touch with Tram. I checked a previous paper she published with us and she did not provide an ORCID in it._
> * [x]  check the missing references, and add them if possible, please: [[REVIEW]: SODA: Software to Support the Curation and Sharing 1 of FAIR Autonomic Nervous System Data #6140 (comment)](https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2214074224)
_All missing DOIs are GitHub repos or websites_
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
_v15.1.0_
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
_https://doi.org/10.5281/zenodo.12774826_
> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.
