### 2024-07-21 #6140 @sappelhoff
Thanks @bvhpatel all is well from my side. There is one minor comment left, namely that you don't have a year and copyright holder in your license:  https://github.com/fairdataihub/SODA-for-SPARC/blob/c1e07639f6cb16ec8723df22e3bc85c2002141e9/LICENSE#L3

This is also the case for the license in the zenodo archive. I would ask you to please update the file in your repository for future releases. Let me know once that has happened, and then I can recommend this paper for acceptance!
