### 2024-07-23 b1a75ff
Update LICENSE
https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2241568302
