### 2024-07-23 #6140 @bvhpatel
@sappelhoff Thanks! I have updated the license file in the repository for future releases.
