### 2024-07-23 #6140 @sappelhoff
@Kevin-Mattheus-Moerman please see these two conversations for your information (and in case you judge differently than I do):

1. https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2214120878
2. https://github.com/openjournals/joss-reviews/issues/6140#issuecomment-2241568302
