### 2024-07-29 #6140 @Kevin-Mattheus-Moerman
@sappelhoff thanks for pinging me on those points. I've reviewed them and think the former conference abstract style publication is okay. I will therefore proceed to process this for acceptance.
