### 2024-07-29 #6140 @Kevin-Mattheus-Moerman
On the archive:
- [ ] Please edit the ZENODO archive listed version tag to include the `v` in `v15.1.0`, this way it matches your GitHub listed release tag exactly.
- [ ] Please edit the ZENODO archive title to match the paper title.

On the paper:
- [ ] In your affiliations, please spell out `USA` as `United States of America`.
